View methods for quanteda

Redefines utils::View() as a generic, so that we can define methods for kwic to produce a nice viewer.

View(x, title)

# S3 method for default
View(x, title)

# S3 method for kwic
View(x, title)

# S3 method for dfm
View(x, title)

Arguments

x
an R object which can be coerced to a data frame with non-zero numbers of rows and columns

title
for viewer window. Defaults to name of x prefixed by Data:

See also

Examples

if (FALSE) {
## for kwic
View(kwic(data_char_ukimmig2010, "illeg*"))

## for data.frame
View(dfm(data_char_ukimmig2010))
}