remove sentences based on their token lengths or a pattern match

Removes sentences from a corpus or a character vector shorter than a specified length.

corpus_trim(x, what = c("sentences", "paragraphs", "documents"),
  min_ntoken = 1, max_ntoken = NULL, exclude_pattern = NULL)

char_trim(x, what = c("sentences", "paragraphs", "documents"),
  min_ntoken = 1, max_ntoken = NULL, exclude_pattern = NULL)

Arguments

x

corpus or character object whose sentences will be selected.

what

units of triming, "sentences" or "paragraphs", or "documents"

min_ntoken, max_ntoken

minimum and maximum lengths in word tokens (excluding punctuation)

exclude_pattern

a stringi regular expression whose match (at the sentence level) will be used to exclude sentences

Value

a corpus or character vector equal in length to the input, or a tokenized set of sentences if . If the input was a corpus, then the all docvars and metadata are preserved. For documents whose sentences have been removed entirely, a null string ("") will be returned.

Examples

txt <- c("PAGE 1. This is a single sentence.  Short sentence. Three word sentence.",
         "PAGE 2. Very short! Shorter.",
         "Very long sentence, with multiple parts, separated by commas.  PAGE 3.")
mycorp <- corpus(txt, docvars = data.frame(serial = 1:3))
texts(mycorp)

#>                                                                      text1 
#> "PAGE 1. This is a single sentence.  Short sentence. Three word sentence." 
#>                                                                      text2 
#>                                             "PAGE 2. Very short! Shorter." 
#>                                                                      text3 
#>   "Very long sentence, with multiple parts, separated by commas.  PAGE 3." 

# exclude sentences shorter than 3 tokens
texts(corpus_trim(mycorp, min_ntoken = 3))

#> Error in texts(corpus_trim(mycorp, min_ntoken = 3)): could not find function "corpus_trim"

# exclude sentences that start with "PAGE <digit(s)>"
texts(corpus_trim(mycorp, exclude_pattern = "^PAGE \\d+"))

#> Error in texts(corpus_trim(mycorp, exclude_pattern = "^PAGE \\d+")): could not find function "corpus_trim"

# trimming character objects
char_trim(txt, "sentences", min_ntoken = 3)

#> Error in eval(expr, envir, enclos): could not find function "char_trim"

char_trim(txt, "sentences", exclude_pattern = "sentence\\.")

#> Error in eval(expr, envir, enclos): could not find function "char_trim"