US presidential inaugural address texts

US presidential inaugural address texts, and metadata (for the corpus), from 1789 to present.

data_corpus_inaugural

data_char_inaugural

Format

a corpus object with docvars including year, and the last and first names of the presidents delivering the inaugural address

Details

data_corpus_inaugural is the quanteda-package corpus object of US presidents' inaugural addresses since 1789. Document variables contain the year of the address and the last name of the president.

data_char_inaugural is the character vector of US presidential inaugaration speeches

References

https://archive.org/details/Inaugural-Address-Corpus-1789-2009 and http://www.presidency.ucsb.edu/inaugurals.php.

Examples

# some operations on the inaugural corpus
summary(data_corpus_inaugural)

#> Corpus consisting of 58 documents.
#> 
#>             Text Types Tokens Sentences Year  President       FirstName
#>  1789-Washington   626   1540        23 1789 Washington          George
#>  1793-Washington    96    147         4 1793 Washington          George
#>       1797-Adams   826   2584        37 1797      Adams            John
#>   1801-Jefferson   716   1935        41 1801  Jefferson          Thomas
#>   1805-Jefferson   804   2381        45 1805  Jefferson          Thomas
#>     1809-Madison   536   1267        21 1809    Madison           James
#>     1813-Madison   542   1304        33 1813    Madison           James
#>      1817-Monroe  1040   3696       121 1817     Monroe           James
#>      1821-Monroe  1262   4898       129 1821     Monroe           James
#>       1825-Adams  1004   3154        74 1825      Adams     John Quincy
#>     1829-Jackson   517   1210        25 1829    Jackson          Andrew
#>     1833-Jackson   499   1271        29 1833    Jackson          Andrew
#>    1837-VanBuren  1315   4175        95 1837  Van Buren          Martin
#>    1841-Harrison  1893   9178       210 1841   Harrison   William Henry
#>        1845-Polk  1330   5211       153 1845       Polk      James Knox
#>      1849-Taylor   497   1185        22 1849     Taylor         Zachary
#>      1853-Pierce  1166   3657       104 1853     Pierce        Franklin
#>    1857-Buchanan   945   3106        89 1857   Buchanan           James
#>     1861-Lincoln  1075   4016       135 1861    Lincoln         Abraham
#>     1865-Lincoln   362    780        26 1865    Lincoln         Abraham
#>       1869-Grant   486   1243        40 1869      Grant      Ulysses S.
#>       1873-Grant   552   1479        43 1873      Grant      Ulysses S.
#>       1877-Hayes   829   2730        59 1877      Hayes   Rutherford B.
#>    1881-Garfield  1018   3240       111 1881   Garfield        James A.
#>   1885-Cleveland   674   1828        44 1885  Cleveland          Grover
#>    1889-Harrison  1355   4744       157 1889   Harrison        Benjamin
#>   1893-Cleveland   823   2135        58 1893  Cleveland          Grover
#>    1897-McKinley  1236   4383       130 1897   McKinley         William
#>    1901-McKinley   857   2449       100 1901   McKinley         William
#>   1905-Roosevelt   404   1089        33 1905  Roosevelt        Theodore
#>        1909-Taft  1436   5844       159 1909       Taft  William Howard
#>      1913-Wilson   661   1896        68 1913     Wilson         Woodrow
#>      1917-Wilson   549   1656        59 1917     Wilson         Woodrow
#>     1921-Harding  1172   3743       148 1921    Harding       Warren G.
#>    1925-Coolidge  1221   4442       196 1925   Coolidge          Calvin
#>      1929-Hoover  1086   3895       158 1929     Hoover         Herbert
#>   1933-Roosevelt   744   2064        85 1933  Roosevelt     Franklin D.
#>   1937-Roosevelt   729   2027        96 1937  Roosevelt     Franklin D.
#>   1941-Roosevelt   527   1552        68 1941  Roosevelt     Franklin D.
#>   1945-Roosevelt   276    651        26 1945  Roosevelt     Franklin D.
#>      1949-Truman   781   2531       116 1949     Truman        Harry S.
#>  1953-Eisenhower   903   2765       119 1953 Eisenhower       Dwight D.
#>  1957-Eisenhower   621   1933        92 1957 Eisenhower       Dwight D.
#>     1961-Kennedy   566   1568        52 1961    Kennedy         John F.
#>     1965-Johnson   569   1725        93 1965    Johnson   Lyndon Baines
#>       1969-Nixon   743   2437       103 1969      Nixon Richard Milhous
#>       1973-Nixon   545   2018        68 1973      Nixon Richard Milhous
#>      1977-Carter   528   1380        52 1977     Carter           Jimmy
#>      1981-Reagan   904   2798       128 1981     Reagan          Ronald
#>      1985-Reagan   925   2935       123 1985     Reagan          Ronald
#>        1989-Bush   795   2683       141 1989       Bush          George
#>     1993-Clinton   644   1837        81 1993    Clinton            Bill
#>     1997-Clinton   773   2451       111 1997    Clinton            Bill
#>        2001-Bush   622   1810        97 2001       Bush       George W.
#>        2005-Bush   772   2325       100 2005       Bush       George W.
#>       2009-Obama   939   2729       110 2009      Obama          Barack
#>       2013-Obama   814   2335        88 2013      Obama          Barack
#>       2017-Trump   582   1662        88 2017      Trump       Donald J.
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

head(docvars(data_corpus_inaugural), 10)

#>                 Year  President   FirstName
#> 1789-Washington 1789 Washington      George
#> 1793-Washington 1793 Washington      George
#> 1797-Adams      1797      Adams        John
#> 1801-Jefferson  1801  Jefferson      Thomas
#> 1805-Jefferson  1805  Jefferson      Thomas
#> 1809-Madison    1809    Madison       James
#> 1813-Madison    1813    Madison       James
#> 1817-Monroe     1817     Monroe       James
#> 1821-Monroe     1821     Monroe       James
#> 1825-Adams      1825      Adams John Quincy

# working with the character vector only
str(data_char_inaugural)

#>  Named chr [1:58] "Fellow-Citizens of the Senate and of the House of Representatives:\n\nAmong the vicissitudes incident to life no event could ha"| __truncated__ ...
#>  - attr(*, "names")= chr [1:58] "1789-Washington" "1793-Washington" "1797-Adams" "1801-Jefferson" ...

head(docvars(data_corpus_inaugural), 10)

#>                 Year  President   FirstName
#> 1789-Washington 1789 Washington      George
#> 1793-Washington 1793 Washington      George
#> 1797-Adams      1797      Adams        John
#> 1801-Jefferson  1801  Jefferson      Thomas
#> 1805-Jefferson  1805  Jefferson      Thomas
#> 1809-Madison    1809    Madison       James
#> 1813-Madison    1813    Madison       James
#> 1817-Monroe     1817     Monroe       James
#> 1821-Monroe     1821     Monroe       James
#> 1825-Adams      1825      Adams John Quincy

mycorpus <- corpus(data_char_inaugural)