print a dictionary object

Print/show method for dictionary objects.

Extractor for dictionary objects

Extractor for dictionary objects

Coerce a dictionary object into a list

# S4 method for dictionary2
show(object)

# S4 method for dictionary2,index,ANY,ANY
[(x, i)

# S4 method for dictionary2,index
[[(x, i)

# S3 method for dictionary2
$(x, name)

# S4 method for dictionary2
as.list(x)

Arguments

object
the dictionary to be printed

i
index for entries

name
the dictionary key

object
the dictionary to be extracted

object
the dictionary to be extracted

i
index for entries

object
the dictionary to be coerced

Slots

.Data

named list of mode character, where each element name is a dictionary "key" and each element is one or more dictionary entry "values" consisting of a pattern match

concatenator

character object specifying space between multi-word values