return the first or last part of a dfm

For a dfm object, returns the first or last n documents and first ncol features for inspection.

# S3 method for dfm
head(x, n = 6L, nfeature = 6L, ...)

# S3 method for dfm
tail(x, n = 6L, nfeature = 6L, ...)

Arguments

x

a dfm object

n

a single integer. If positive, size for the resulting object: number of first/last documents for the dfm. If negative, all but the n last/first number of documents of x.

nfeature

the number of features to return, where the resulting object will contain the first ncol features

...

additional arguments passed to other functions

Value

A dfm class object corresponding to the subset defined by n and ncol.

Examples

myDfm <- dfm(data_corpus_inaugural, ngrams = 2, verbose = FALSE)
head(myDfm)

#> Document-feature matrix of: 58 documents, 63,634 features (96.9% sparse).
#> (showing first 6 documents and first 6 features)
#>                  features
#> docs              fellow_- -_citizens citizens_of of_the the_senate senate_and
#>   1789-Washington        1          1           1     20          1          1
#>   1793-Washington        0          0           0      4          0          0
#>   1797-Adams             3          3           0     29          0          0
#>   1801-Jefferson         2          2           0     28          0          0
#>   1805-Jefferson         0          0           0     17          0          0
#>   1809-Madison           1          1           0     20          0          0

tail(myDfm)

#> Document-feature matrix of: 58 documents, 63,634 features (96.9% sparse).
#> (showing last 6 documents and last 6 features)
#>               features
#> docs           fellow_- -_citizens citizens_of of_the the_senate senate_and
#>   1997-Clinton        0          0           0     11          0          0
#>   2001-Bush           0          0           0      2          0          0
#>   2005-Bush           0          0           0     14          0          0
#>   2009-Obama          0          0           0      4          0          0
#>   2013-Obama          0          0           0      2          0          0
#>   2017-Trump          0          0           2      3          0          0

tail(myDfm, nfeature = 4)

#> Document-feature matrix of: 58 documents, 63,634 features (96.9% sparse).
#> (showing last 6 documents and last 4 features)
#>               features
#> docs           fellow_- -_citizens citizens_of of_the
#>   1997-Clinton        0          0           0     11
#>   2001-Bush           0          0           0      2
#>   2005-Bush           0          0           0     14
#>   2009-Obama          0          0           0      4
#>   2013-Obama          0          0           0      2
#>   2017-Trump          0          0           2      3