join tokens function

For a set of tokens, given a set of token sequences, join the tokens matching the sequences.

joinTokens(x, ...)

# S3 method for tokenizedTexts
joinTokens(x, ...)

# S3 method for tokens
joinTokens(x, sequences, concatenator = "_",
  valuetype = c("glob", "fixed", "regex"), verbose = FALSE,
  case_insensitive = TRUE, ...)

Arguments

x

tokens or tokenizedTexts object

...

additional arguments passed to other methods

sequences

features to concatenate, a list of characters

concatenator

character used for joining tokens

valuetype

how to interpret sequences: fixed for words as is; "regex" for regular expressions; or "glob" for "glob"-style wildcard

verbose

display progress

case_insensitive

if TRUE, ignore case when matching

Examples

toks <- tokens(data_corpus_inaugural, removePunct = TRUE)
seqs <- list(c('foreign', 'polic*'), c('United', 'States'))
kwic(toks, 'fUnited_States', window = 1) # no exisit

#> [1] NA

kwic(toks, 'foreign_policy', window = 1) # no exisit

#> [1] NA

toks <- joinTokens(toks, seqs, "_", 'glob')
kwic(toks, 'United_States', window = 1)

#>                         contextPre       keyword    contextPost
#>  [1789-Washington, 400]      the [ United_States ] a           
#>  [1789-Washington, 490]      the [ United_States ] Every       
#>       [1797-Adams, 469]      the [ United_States ] in          
#>      [1797-Adams, 1540]      the [ United_States ] and         
#>      [1797-Adams, 2227]      the [ United_States ] I           
#>   [1805-Jefferson, 405]      the [ United_States ] These       
#>     [1809-Madison, 355]      the [ United_States ] to          
#>     [1809-Madison, 474]      the [ United_States ] and         
#>     [1813-Madison, 195]      the [ United_States ] until       
#>     [1813-Madison, 534]      the [ United_States ] not         
#>     [1813-Madison, 569]      the [ United_States ] incorporated
#>    [1813-Madison, 1027]      the [ United_States ] It          
#>      [1817-Monroe, 245]      the [ United_States ] They        
#>      [1817-Monroe, 354]      the [ United_States ] have        
#>      [1817-Monroe, 657]      the [ United_States ] received    
#>      [1817-Monroe, 900]      the [ United_States ] enjoy       
#>     [1817-Monroe, 1046]      the [ United_States ] while       
#>     [1817-Monroe, 1252]      the [ United_States ] been        
#>     [1817-Monroe, 1417]      the [ United_States ] may         
#>     [1817-Monroe, 1814]      the [ United_States ] with        
#>     [1817-Monroe, 1960]      the [ United_States ] to          
#>     [1817-Monroe, 2086]      the [ United_States ] have        
#>     [1817-Monroe, 2252]      the [ United_States ] a           
#>     [1817-Monroe, 2541]      the [ United_States ] to          
#>     [1817-Monroe, 2797]      the [ United_States ] are         
#>      [1821-Monroe, 439]      the [ United_States ] had         
#>     [1821-Monroe, 1575]      the [ United_States ] to          
#>     [1821-Monroe, 1647]      the [ United_States ] regarding   
#>     [1821-Monroe, 1711]      the [ United_States ] and         
#>     [1821-Monroe, 1731]      the [ United_States ] strong      
#>     [1821-Monroe, 2072]      the [ United_States ] a           
#>     [1821-Monroe, 2177]      the [ United_States ] to          
#>     [1821-Monroe, 2223]      the [ United_States ] and         
#>     [1821-Monroe, 2274]      the [ United_States ] for         
#>     [1821-Monroe, 2319]      the [ United_States ] and         
#>     [1821-Monroe, 2345]      the [ United_States ] on          
#>     [1821-Monroe, 2399]      the [ United_States ] in          
#>     [1821-Monroe, 2490]      the [ United_States ] have        
#>     [1821-Monroe, 2554]      the [ United_States ] in          
#>     [1821-Monroe, 4200]      the [ United_States ] from        
#>     [1821-Monroe, 4248]      The [ United_States ] now         
#>      [1825-Adams, 1058]      the [ United_States ] first       
#>     [1829-Jackson, 115]      the [ United_States ] to          
#>      [1833-Jackson, 34]      the [ United_States ] for         
#>     [1833-Jackson, 381]      the [ United_States ] which       
#>     [1833-Jackson, 845]      the [ United_States ] Let         
#>   [1837-VanBuren, 2686]      the [ United_States ] including   
#>    [1841-Harrison, 678]      the [ United_States ] is          
#>   [1841-Harrison, 1403]      the [ United_States ] first       
#>   [1841-Harrison, 2209]      the [ United_States ] It          
#>   [1841-Harrison, 3134]      the [ United_States ] solely      
#>   [1841-Harrison, 3901]      the [ United_States ] If          
#>   [1841-Harrison, 5099]      the [ United_States ] Those       
#>   [1841-Harrison, 5387]      the [ United_States ] accept      
#>   [1841-Harrison, 5456]      the [ United_States ] as          
#>   [1841-Harrison, 5866]      the [ United_States ] but         
#>         [1845-Polk, 95]      the [ United_States ] even        
#>        [1845-Polk, 313]      the [ United_States ] A           
#>        [1845-Polk, 435]      the [ United_States ] is          
#>        [1845-Polk, 524]      the [ United_States ] by          
#>        [1845-Polk, 674]      the [ United_States ] has         
#>       [1845-Polk, 3341]      the [ United_States ] the         
#>       [1845-Polk, 3385]      the [ United_States ] and         
#>       [1845-Polk, 3527]      the [ United_States ] not         
#>       [1845-Polk, 3997]      the [ United_States ] or          
#>       [1845-Polk, 4132]      the [ United_States ] by          
#>       [1845-Polk, 4168]      the [ United_States ] to          
#>       [1845-Polk, 4639]      the [ United_States ] While       
#>      [1849-Taylor, 338]      the [ United_States ] with        
#>      [1853-Pierce, 262]      the [ United_States ] one         
#>     [1853-Pierce, 2370]      the [ United_States ] The         
#>     [1857-Buchanan, 25]      the [ United_States ] and         
#>     [1857-Buchanan, 42]      the [ United_States ] In          
#>    [1857-Buchanan, 349]      the [ United_States ] As          
#>    [1857-Buchanan, 442]      the [ United_States ] before      
#>    [1857-Buchanan, 540]      the [ United_States ] to          
#>    [1857-Buchanan, 601]      the [ United_States ] The         
#>   [1857-Buchanan, 2105]      the [ United_States ] shall       
#>   [1857-Buchanan, 2142]      the [ United_States ] over        
#>       [1861-Lincoln, 4]      the [ United_States ] In          
#>      [1861-Lincoln, 38]      the [ United_States ] to          
#>    [1861-Lincoln, 1041]      the [ United_States ] be          
#>    [1861-Lincoln, 1286]      the [ United_States ] are         
#>    [1861-Lincoln, 1491]      the [ United_States ] in          
#>         [1869-Grant, 4]      the [ United_States ] Your        
#>        [1869-Grant, 17]      the [ United_States ] I           
#>      [1877-Hayes, 1494]      the [ United_States ] of          
#>      [1877-Hayes, 1927]      the [ United_States ] and         
#>     [1881-Garfield, 84]      the [ United_States ] the         
#>    [1881-Garfield, 963]      the [ United_States ] Freedom     
#>   [1881-Garfield, 2051]       of [ United_States ] notes       
#>   [1881-Garfield, 2221]      the [ United_States ] afford      
#>   [1881-Garfield, 2391]      the [ United_States ] in          
#>   [1881-Garfield, 2454]      the [ United_States ] to          
#>   [1881-Garfield, 2510]      the [ United_States ] are         
#>   [1885-Cleveland, 678]      the [ United_States ] only        
#>  [1893-Cleveland, 1351]      the [ United_States ] have        
#>  [1893-Cleveland, 1776]      the [ United_States ] not         
#>     [1897-McKinley, 35]      the [ United_States ] relying     
#>    [1897-McKinley, 803]      the [ United_States ] The         
#>   [1897-McKinley, 1122]      the [ United_States ] for         
#>   [1897-McKinley, 2025]      the [ United_States ] courts      
#>   [1897-McKinley, 2088]      the [ United_States ] are         
#>   [1897-McKinley, 2597]      The [ United_States ] has         
#>   [1897-McKinley, 2733]      the [ United_States ] since       
#>   [1897-McKinley, 2981]      the [ United_States ] and         
#>   [1897-McKinley, 3150]      the [ United_States ] to          
#>   [1897-McKinley, 3576]      the [ United_States ] Again       
#>   [1897-McKinley, 3889]      the [ United_States ] and         
#>   [1897-McKinley, 3906]      the [ United_States ] This        
#>    [1901-McKinley, 613]      the [ United_States ] and         
#>   [1901-McKinley, 1283]      the [ United_States ] bore        
#>   [1901-McKinley, 1340]      the [ United_States ] and         
#>   [1901-McKinley, 1423]      the [ United_States ] under       
#>   [1901-McKinley, 1505]      the [ United_States ] The         
#>   [1901-McKinley, 1569]      the [ United_States ] in          
#>   [1901-McKinley, 1733]      the [ United_States ] throughout  
#>   [1901-McKinley, 2071]      the [ United_States ] By          
#>   [1901-McKinley, 2140]      the [ United_States ] Order       
#>       [1909-Taft, 1583]      the [ United_States ] and         
#>       [1909-Taft, 2110]      the [ United_States ] can         
#>       [1909-Taft, 2250]      the [ United_States ] in          
#>       [1909-Taft, 3255]      the [ United_States ] and         
#>       [1909-Taft, 3640]      the [ United_States ] and         
#>       [1909-Taft, 4261]      the [ United_States ] it          
#>     [1917-Wilson, 1263]      the [ United_States ] have        
#>     [1921-Harding, 771]    these [ United_States ] is          
#>     [1929-Hoover, 1754]      The [ United_States ] fully       
#>     [1929-Hoover, 2263]      The [ United_States ] seeks       
#>  [1933-Roosevelt, 1181]      the [ United_States ] a           
#>  [1933-Roosevelt, 1802]      the [ United_States ] have        
#>   [1937-Roosevelt, 431]      the [ United_States ] can         
#>  [1937-Roosevelt, 1150]        a [ United_States ] which       
#>  [1937-Roosevelt, 1735]      the [ United_States ] I           
#>    [1941-Roosevelt, 19]      the [ United_States ] In          
#>   [1941-Roosevelt, 376]      the [ United_States ] The         
#>   [1941-Roosevelt, 845]      the [ United_States ] into        
#>      [1949-Truman, 243]      the [ United_States ] as          
#>      [1949-Truman, 435]      the [ United_States ] and         
#>      [1949-Truman, 701]      the [ United_States ] alone       
#>      [1949-Truman, 771]      the [ United_States ] has         
#>     [1949-Truman, 1522]      The [ United_States ] is          
#> [1953-Eisenhower, 1642]        a [ United_States ] that        
#>      [1969-Nixon, 1745]      the [ United_States ] To          
#>     [1977-Carter, 1010]      the [ United_States ] there       
#>      [1981-Reagan, 191]    These [ United_States ] are         
#>       [1989-Bush, 2314]      the [ United_States ] of          
#>         [2005-Bush, 16]      the [ United_States ] Congress    
#>        [2005-Bush, 397]      the [ United_States ] to          
#>        [2005-Bush, 839]      The [ United_States ] will        
#>        [2005-Bush, 964]      the [ United_States ] can         
#>       [2005-Bush, 2064]      the [ United_States ] of          
#>      [2009-Obama, 2387]      the [ United_States ] of          
#>        [2013-Obama, 10]      the [ United_States ] Congress    
#>      [2013-Obama, 2093]    these [ United_States ] of          

kwic(toks, 'foreign_policy', window = 1)

#>                         contextPre        keyword  contextPost
#>    [1833-Jackson, 167]       The [ foreign_policy ] adopted   
#>  [1837-VanBuren, 3347]        of [ foreign_policy ] has       
#>      [1869-Grant, 902]        to [ foreign_policy ] I         
#> [1885-Cleveland, 1035]      that [ foreign_policy ] commended 
#>  [1897-McKinley, 2814] dignified [ foreign_policy ] which     
#>  [1897-McKinley, 3022]       our [ foreign_policy ] throughout

kwic(toks, c('foreign', 'policy'), window = 1) # no longer exisit

#> [1] NA

kwic(toks, c('united', 'states'), window = 1) # no longer exisit

#> [1] NA