compute keyness (internal functions)

Internal function used in textstat_keyness. Computes \(chi^2\) with Yates' continuity correction for 2x2 tables.

keyness_chi2_dt(x)

keyness_chi2_stats(x)

keyness_exact(x)

keyness_lr(x, correction = c("none", "Yates"))

Arguments

x
a dfm object

correction
if "Yates" implement the Yates correction for 2x2 tables, no correction if "none"

Value

a data.frame of chi2 and p-values with rows named for each feature

Details

keyness_chi2_dt uses vectorized computation from data.table objects. keyness_chi2_stats uses element-by-element application of chisq.test. keyness_exact computes Fisher's exact using element-by-element application of fisher.test, returning the odds ratio. keyness_lr computes the \(G^2\) likelihood ratio statistic using vectorized computation

References

https://en.wikipedia.org/wiki/Yates's_correction_for_continuity http://influentialpoints.com/Training/g-likelihood_ratio_test.htm

Examples

mydfm <- dfm(c(d1 = "a a a b b c c c c c c d e f g h h",
               d2 = "a a b c c d d d d e f h"))
quanteda:::keyness_chi2_dt(mydfm)

#>          chi2         p target reference
#> a  0.18510008 0.6670265      3         2
#> b  0.10251697 0.7488298      2         1
#> c  0.46729838 0.4942327      6         2
#> d -2.04024714 0.1531848      1         4
#> e -0.23758624 0.6259549      1         1
#> f -0.23758624 0.6259549      1         1
#> g  0.03173144 0.8586184      1         0
#> h  0.10251697 0.7488298      2         1

quanteda:::keyness_chi2_stats(mydfm)

#> Error in get(".SigLength", envir = env): object '.SigLength' not found

quanteda:::keyness_exact(mydfm)

#> Error in get(".SigLength", envir = env): object '.SigLength' not found

quanteda:::keyness_lr(mydfm)

#>              G          p target reference
#> a  0.004751409 0.94504497      3         2
#> b  0.091241894 0.76260424      2         1
#> c  1.274238010 0.25897314      6         2
#> d -3.779440533 0.05188601      1         4
#> e -0.064900473 0.79891178      1         1
#> f -0.064900473 0.79891178      1         1
#> g          NaN        NaN      1         0
#> h  0.091241894 0.76260424      2         1