Internal function to get, set or initialize system metadata

Sets or initializes system metadata for new objects.

meta_system(x, field = NULL)

meta_system(x, field = NULL) <- value

# S3 method for corpus
meta_system(x, field = NULL) <- value

# S3 method for tokens
meta_system(x, field = NULL) <- value

# S3 method for dfm
meta_system(x, field = NULL) <- value

# S3 method for dictionary
meta_system(x, field = NULL) <- value

meta_system_defaults()

Arguments

x
an object for which the metadata will be read or set

field
metadata field name(s); if NULL (default), return all metadata names

value
new value of the metadata field

source
character; the input object class

Value

meta_system returns a list with the object's system metadata. It is literally a wrapper to meta(x, field, type = "system")().

meta_system<- returns the object with the system metadata modified. This is an internal function and not designed for users!

meta_system_defaults returns a list of default system values, with the user setting the "source" value. This should be used to set initial system meta information.

Examples

corp <- corpus(c(d1 = "one two three", d2 = "two three four"))
# quanteda:::`meta_system<-`(corp, value = quanteda:::meta_system_defaults("example"))
quanteda:::meta_system(corp)

#> $`package-version`
#> [1] ‘2.1.0’
#> 
#> $`r-version`
#> [1] ‘4.0.2’
#> 
#> $system
#>   sysname   machine      user 
#>  "Darwin"  "x86_64" "kbenoit" 
#> 
#> $directory
#> [1] "/Users/kbenoit/Dropbox (Personal)/GitHub/quanteda/quanteda/docs/reference"
#> 
#> $created
#> [1] "2020-07-02"
#>