get or set corpus metadata

Get or set the corpus-level metadata in a corpus object.

metacorpus(x, field = NULL)

metacorpus(x, field) <- value

Arguments

x

a corpus object

field

metadata field name(s); if NULL (default), return all metadata names

value

new value of the corpus metadata field

Value

For metacorpus, a list of the metadata fields in the corpus. If a list is not what you wanted, you can wrap the results in unlist, but this will remove any metadata field that is set to NULL.

For metacorpus <-, the corpus with the updated metadata.

Examples

metacorpus(data_corpus_inaugural)

#> $source
#> [1] "/home/paul/Dropbox/code/quanteda/* on x86_64 by paul"
#> 
#> $created
#> [1] "Fri Sep 12 12:41:17 2014"
#> 
#> $notes
#> NULL
#> 
#> $citation
#> NULL
#> 

metacorpus(data_corpus_inaugural, "source")

#> $source
#> [1] "/home/paul/Dropbox/code/quanteda/* on x86_64 by paul"
#> 

metacorpus(data_corpus_inaugural, "citation") <- "Presidential Speeches Online Project (2014)."
metacorpus(data_corpus_inaugural, "citation")

#> $citation
#> [1] "Presidential Speeches Online Project (2014)."
#>