count the Scrabble letter values of text

Tally the Scrabble letter values of text given a user-supplied function, such as the sum (default) or mean of the character values.

nscrabble(x, FUN = sum)

Arguments

x

a character vector

FUN

function to be applied to the character values in the text; default is sum, but could also be mean or a user-supplied function

Value

a (named) integer vector of Scabble letter values, computed using FUN, corresponding to the input text(s)

Note

Character values are only defined for non-accented Latin a-z, A-Z letters. Lower-casing is unnecessary.

Examples

nscrabble(c("muzjiks", "excellency"))

#> [1] 29 24

nscrabble(data_char_inaugural[1:5], mean)

#> 1789-Washington 1793-Washington      1797-Adams  1801-Jefferson  1805-Jefferson 
#>        1.706789        1.721875        1.624590        1.678183        1.663654