declare a compound character to be a sequence of separate pattern matches

Declares that a whitespace-separated expression consists of multiple patterns, separated by whitespace. This is typically used as a wrapper around pattern to make it explicit that the pattern elements are to be used for matches to multi-word sequences, rather than individual, unordered matches to single words.

phrase(x)

is.phrase(x)

Arguments

x
the sequence, as a character object containing whitespace separating the patterns

Value

phrase returns a specially classed list whose white-spaced elements have been parsed into separate character elements. is.phrase returns TRUE if the object was created by phrase; FALSE otherwise.

Examples

# make phrases from characters
phrase(c("a b", "c d e", "f"))

#> [[1]]
#> [1] "a" "b"
#> 
#> [[2]]
#> [1] "c" "d" "e"
#> 
#> [[3]]
#> [1] "f"
#> 

# from a dictionary
phrase(dictionary(list(catone = c("a b"), cattwo = "c d e", catthree = "f")))

#> [[1]]
#> [1] "a" "b"
#> 
#> [[2]]
#> [1] "c" "d" "e"
#> 
#> [[3]]
#> [1] "f"
#> 

# from a collocations object
(coll <- textstat_collocations(tokens("a b c a b d e b d a b")))

#>   collocation count length   lambda         z
#> 1         a b     3      2 3.412247 1.9360826
#> 2         b d     2      2 3.218876 1.7994063
#> 3         d e     1      2 2.833213 1.5251341
#> 4         c a     1      2 2.833213 1.5251341
#> 5         b c     1      2 2.197225 1.2282859
#> 6         d a     1      2 1.609438 1.1019068
#> 7         e b     1      2 1.717651 0.9745821

phrase(coll)

#> [[1]]
#> [1] "a" "b"
#> 
#> [[2]]
#> [1] "b" "d"
#> 
#> [[3]]
#> [1] "d" "e"
#> 
#> [[4]]
#> [1] "c" "a"
#> 
#> [[5]]
#> [1] "b" "c"
#> 
#> [[6]]
#> [1] "d" "a"
#> 
#> [[7]]
#> [1] "e" "b"
#>