prediction method for Naive Bayes classifier objects

implements class predictions using trained Naive Bayes examples

# S3 method for textmodel_NB_fitted
predict(object, newdata = NULL, ...)

Arguments

object
a fitted Naive Bayes textmodel

newdata
dfm on which prediction should be made

...
not used

Value

A list of two data frames, named docs and words corresponding to word- and document-level predicted quantities

docs

data frame with document-level predictive quantities: nb.predicted, ws.predicted, bs.predicted, PcGw, wordscore.doc, bayesscore.doc, posterior.diff, posterior.logdiff. Note that the diff quantities are currently implemented only for two-class solutions.

words

data-frame with word-level predictive quantities: wordscore.word, bayesscore.word

Examples

(nbfit <- textmodel_nb(data_dfm_lbgexample, c("A", "A", "B", "C", "C", NA)))

#> Error in get(".SigLength", envir = env): object '.SigLength' not found

(nbpred <- predict(nbfit))

#> Error in predict(nbfit): object 'nbfit' not found