print a dfm object

print methods for document-feature matrices

# S4 method for dfm
print(x, show.values = NULL, show.settings = FALSE,
  show.summary = TRUE, ndoc = quanteda_options("print_dfm_max_ndoc"),
  nfeature = quanteda_options("print_dfm_max_nfeature"), ...)

# S4 method for dfm
show(object)

# S4 method for fcm
print(x, show.values = NULL, show.settings = FALSE,
  show.summary = TRUE, ndoc = quanteda_options("print_dfm_max_ndoc"),
  nfeature = quanteda_options("print_dfm_max_nfeature"), ...)

# S4 method for fcm
show(object)

# S3 method for fcm
head(x, n = 6L, nfeature = 6L, ...)

# S3 method for fcm
tail(x, n = 6L, nfeature = 6L, ...)

Arguments

x
the dfm to be printed

show.values
print the dfm values; if called explicitly this will print all values, overriding ndoc and nfeature.

show.settings
print the settings used to create the dfm. See settings.

show.summary
print a brief summary indicating the number of documents and features

ndoc
max number of documents to print; default is from the print_dfm_max_ndoc setting of quanteda_options

nfeature
max number of features to print; default is from the print_dfm_max_nfeature setting of quanteda_options

...
further arguments passed to Matrix::printSpMatrix2

object
the item to be printed

See also

quanteda_options