select features from an object

This function selects or discards features from a dfm.variety of objects, such as tokenized texts, a dfm, or a list of collocations. The most common usage for removeFeatures will be to eliminate stop words from a text or text-based object, or to select only features from a list of regular expression.

# S3 method for dfm
selectFeatures(x, features, selection = c("keep", "remove"),
  valuetype = c("glob", "regex", "fixed"), case_insensitive = TRUE,
  verbose = TRUE, ...)

selectFeatures(x, features, ...)

# S3 method for tokenizedTexts
selectFeatures(x, features, selection = c("keep",
  "remove"), valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE, padding = FALSE, indexing = FALSE,
  verbose = FALSE, ...)

# S3 method for tokens
selectFeatures(x, features, selection = c("keep", "remove"),
  valuetype = c("glob", "regex", "fixed"), case_insensitive = TRUE,
  padding = FALSE, ...)

# S3 method for collocations
selectFeatures(x, features, selection = c("keep",
  "remove"), valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE, verbose = TRUE, pos = 1:3, ...)

Arguments

x

object whose features will be selected

features

one of: a character vector of features to be selected, a dfm whose features will be used for selection, or a dictionary class object whose values (not keys) will provide the features to be selected. For dfm objects, see details in the Value section below.

selection

whether to keep or remove the features

valuetype

how to interpret keyword expressions: "glob" for "glob"-style wildcard expressions; "regex" for regular expressions; or "fixed" for exact matching. See valuetype for details.

case_insensitive

ignore the case of dictionary values if TRUE

verbose

if TRUE print message about how many features were removed

...

supplementary arguments passed to the underlying functions in stri_detect_regex. (This is how case_insensitive is passed, but you may wish to pass others.)

padding

(only for tokenizedTexts objects) if TRUE, leave an empty string where the removed tokens previously existed. This is useful if a positional match is needed between the pre- and post-selected features, for instance if a window of adjacency needs to be computed.

indexing

use dfm-based index to efficiently process large tokenizedTexts object

pos

indexes of word position if called on collocations: remove if word pos is a stopword

Value

A dfm after the feature selection has been applied.

When features is a dfm object, then the returned object will be identical in its feature set to the dfm supplied as the features argument. This means that any features in x not in features will be discarded, and that any features in found in the dfm supplied as features but not found in x will be added with all zero counts. This is useful when you have trained a model on one dfm, and need to project this onto a test set whose features must be identical.

Note

This function selects features based on their labels. To select features based on the values of a the document-feature matrix, use trim.

See also

removeFeatures, dfm_trim

Examples

## Not run: ------------------------------------
# myDfm <- dfm(c("My Christmas was ruined by your opposition tax plan.", 
#                "Does the United_States or Sweden have more progressive taxation?"),
#              tolower = FALSE, verbose = FALSE)
# mydict <- dictionary(list(countries = c("United_States", "Sweden", "France"),
#                           wordsEndingInY = c("by", "my"),
#                           notintext = "blahblah"))
# selectFeatures(myDfm, mydict)
# selectFeatures(myDfm, mydict, case_insensitive = FALSE)
# selectFeatures(myDfm, c("s$", ".y"), "keep")
# selectFeatures(myDfm, c("s$", ".y"), "keep", valuetype = "regex")
# selectFeatures(myDfm, c("s$", ".y"), "remove", valuetype = "regex")
# selectFeatures(myDfm, stopwords("english"), "keep", valuetype = "fixed")
# selectFeatures(myDfm, stopwords("english"), "remove", valuetype = "fixed")
# 
# # selecting on a dfm
# textVec1 <- c("This is text one.", "This, the second text.", "Here: the third text.")
# textVec2 <- c("Here are new words.", "New words in this text.")
# (dfm1 <- dfm(textVec1, verbose = FALSE))
# (dfm2a <- dfm(textVec2, verbose = FALSE))
# (dfm2b <- selectFeatures(dfm2a, dfm1))
# setequal(featnames(dfm1), featnames(dfm2b))
# 
# # more selection on a dfm
# selectFeatures(dfm1, dfm2a)
# selectFeatures(dfm1, dfm2a, selection = "remove")
## ---------------------------------------------
## Not run: ------------------------------------
# ## performance comparisons
# data(SOTUCorpus, package = "quantedaData")
# toks <- tokenize(SOTUCorpus, removePunct = TRUE)
# # toks <- tokenize(tokenize(SOTUCorpus, what='sentence', simplify = TRUE), removePunct = TRUE)
# # head to head, old v. new
# system.time(selectFeaturesOLD(toks, stopwords("english"), "remove", verbose = FALSE))
# system.time(selectFeatures(toks, stopwords("english"), "remove", verbose = FALSE))
# system.time(selectFeaturesOLD(toks, c("and", "of"), "remove", verbose = FALSE, valuetype = "regex"))
# system.time(selectFeatures(toks, c("and", "of"), "remove", verbose = FALSE, valuetype = "regex"))
# microbenchmark::microbenchmark(
#     old = selectFeaturesOLD(toks, stopwords("english"), "remove", verbose = FALSE),
#     new = selectFeatures(toks, stopwords("english"), "remove", verbose = FALSE),
#     times = 5, unit = "relative")
# microbenchmark::microbenchmark(
#     new = selectFeaturesOLD(toks, c("and", "of"), "remove", verbose = FALSE, valuetype = "regex"),
#     old = selectFeatures(toks, c("and", "of"), "remove", verbose = FALSE, valuetype = "regex"),
#     times = 2, unit = "relative")
#     
# types <- unique(unlist(toks))
# numbers <- types[stringi::stri_detect_regex(types, '[0-9]')]
# microbenchmark::microbenchmark(
#     new = selectFeaturesOLD(toks, numbers, "remove", verbose = FALSE, valuetype = "fixed"),
#     old = selectFeatures(toks, numbers, "remove", verbose = FALSE, valuetype = "fixed"),
#     times = 2, unit = "relative")  
#     
# # removing tokens before dfm, versus after
# microbenchmark::microbenchmark(
#     pre = dfm(selectFeaturesOLD(toks, stopwords("english"), "remove"), verbose = FALSE),
#     post = dfm(toks, remove = stopwords("english"), verbose = FALSE),
#     times = 5, unit = "relative")
## ---------------------------------------------

## with simple examples
toks <- tokenize(c("This is a sentence.", "This is a second sentence."),
                 removePunct = TRUE)
selectFeatures(toks, c("is", "a", "this"), selection = "remove",
                valuetype = "fixed", padding = TRUE, case_insensitive = TRUE)

#> Error in attr(x, "padding") && "" %in% features: invalid 'x' type in 'x && y'

# how case_insensitive works
selectFeatures(toks, c("is", "a", "this"), selection = "remove",
               valuetype = "fixed", padding = TRUE, case_insensitive = FALSE)

#> Error in attr(x, "padding") && "" %in% features: invalid 'x' type in 'x && y'

selectFeatures(toks, c("is", "a", "this"), selection = "remove",
               valuetype = "fixed", padding = TRUE, case_insensitive = TRUE)

#> Error in attr(x, "padding") && "" %in% features: invalid 'x' type in 'x && y'

selectFeatures(toks, c("is", "a", "this"), selection = "remove",
               valuetype = "glob", padding = TRUE, case_insensitive = TRUE)

#> Error in attr(x, "padding") && "" %in% features: invalid 'x' type in 'x && y'

selectFeatures(toks, c("is", "a", "this"), selection = "remove",
               valuetype = "glob", padding = TRUE, case_insensitive = FALSE)

#> Error in attr(x, "padding") && "" %in% features: invalid 'x' type in 'x && y'

# with longer texts
txts <- data_char_inaugural[1:2]
toks <- tokenize(txts)
selectFeatures(toks, stopwords("english"), "remove")

#> Error in attr(x, "padding") && "" %in% features: invalid 'x' type in 'x && y'

selectFeatures(toks, stopwords("english"), "keep")

#> tokenizedTexts from 2 documents.
#> 1789-Washington :
#>   [1] "of"         "the"        "and"        "of"         "the"       
#>   [6] "of"         "the"        "to"         "no"         "could"     
#>  [11] "have"       "me"         "with"       "than"       "that"      
#>  [16] "of"         "which"      "the"        "was"        "by"        
#>  [21] "your"       "and"        "on"         "the"        "of"        
#>  [26] "the"        "On"         "the"        "I"          "was"       
#>  [31] "by"         "my"         "I"          "but"        "with"      
#>  [36] "and"        "from"       "a"          "which"      "I"         
#>  [41] "had"        "with"       "the"        "and"        "in"        
#>  [46] "my"         "with"       "an"         "as"         "the"       
#>  [51] "of"         "my"         "a"          "which"      "was"       
#>  [56] "more"       "as"         "as"         "more"       "to"        
#>  [61] "me"         "by"         "the"        "of"         "to"        
#>  [66] "and"        "of"         "in"         "my"         "to"        
#>  [71] "the"        "on"         "it"         "by"         "On"        
#>  [76] "the"        "other"      "the"        "and"        "of"        
#>  [81] "the"        "to"         "which"      "the"        "of"        
#>  [86] "my"         "me"         "being"      "to"         "in"        
#>  [91] "the"        "and"        "most"       "of"         "her"       
#>  [96] "a"          "into"       "his"        "could"      "not"       
#> [101] "but"        "with"       "who"        "from"       "and"       
#> [106] "in"         "the"        "of"         "ought"      "to"        
#> [111] "be"         "of"         "his"        "own"        "In"        
#> [116] "this"       "of"         "all"        "I"          "is"        
#> [121] "that"       "it"         "has"        "been"       "my"        
#> [126] "to"         "my"         "from"       "a"          "of"        
#> [131] "by"         "which"      "it"         "be"         "All"       
#> [136] "I"          "is"         "that"       "if"         "in"        
#> [141] "this"       "I"          "have"       "been"       "too"       
#> [146] "by"         "a"          "of"         "or"         "by"        
#> [151] "an"         "to"         "this"       "of"         "the"       
#> [156] "of"         "my"         "and"        "have"       "too"       
#> [161] "my"         "as"         "as"         "for"        "the"       
#> [166] "and"        "before"     "me"         "my"         "be"        
#> [171] "by"         "the"        "which"      "me"         "and"       
#> [176] "its"        "be"         "by"         "my"         "with"      
#> [181] "some"       "of"         "the"        "in"         "which"     
#> [186] "they"       "Such"       "being"      "the"        "under"     
#> [191] "which"      "I"          "have"       "in"         "to"        
#> [196] "the"        "to"         "the"        "it"         "would"     
#> [201] "be"         "to"         "in"         "this"       "my"        
#> [206] "to"         "that"       "Being"      "who"        "over"      
#> [211] "the"        "who"        "in"         "the"        "of"        
#> [216] "and"        "that"       "His"        "to"         "the"       
#> [221] "and"        "of"         "the"        "of"         "the"       
#> [226] "a"          "by"         "themselves" "for"        "these"     
#> [231] "and"        "in"         "its"        "to"         "with"      
#> [236] "the"        "to"         "his"        "In"         "this"      
#> [241] "to"         "the"        "of"         "and"        "I"         
#> [246] "myself"     "that"       "it"         "your"       "not"       
#> [251] "than"       "my"         "own"        "nor"        "those"     
#> [256] "of"         "my"         "at"         "than"       "No"        
#> [261] "be"         "to"         "and"        "the"        "which"     
#> [266] "the"        "of"         "more"       "than"       "those"     
#> [271] "of"         "the"        "by"         "which"      "they"      
#> [276] "have"       "to"         "the"        "of"         "an"        
#> [281] "to"         "have"       "been"       "by"         "some"      
#> [286] "of"         "and"        "in"         "the"        "in"        
#> [291] "the"        "of"         "their"      "the"        "and"       
#> [296] "of"         "so"         "from"       "which"      "the"       
#> [301] "has"        "not"        "be"         "with"       "the"       
#> [306] "by"         "which"      "most"       "have"       "been"      
#> [311] "some"       "of"         "with"       "an"         "of"        
#> [316] "the"        "which"      "the"        "to"         "These"     
#> [321] "out"        "of"         "the"        "have"       "themselves"
#> [326] "too"        "on"         "my"         "to"         "be"        
#> [331] "You"        "with"       "me"         "I"          "in"        
#> [336] "that"       "there"      "are"        "under"      "the"       
#> [341] "of"         "which"      "the"        "of"         "a"         
#> [346] "and"        "more"       "By"         "the"        "the"       
#> [351] "it"         "is"         "the"        "of"         "the"       
#> [356] "to"         "to"         "your"       "such"       "as"        
#> [361] "he"         "and"        "The"        "under"      "which"     
#> [366] "I"          "you"        "me"         "from"       "into"      
#> [371] "that"       "further"    "than"       "to"         "to"        
#> [376] "the"        "under"      "which"      "you"        "are"       
#> [381] "and"        "which"      "in"         "your"       "the"       
#> [386] "to"         "which"      "your"       "is"         "to"        
#> [391] "be"         "It"         "be"         "more"       "with"      
#> [396] "those"      "and"        "more"       "with"       "the"       
#> [401] "which"      "me"         "to"         "in"         "of"        
#> [406] "a"          "of"         "the"        "that"       "is"        
#> [411] "to"         "the"        "the"        "and"        "the"       
#> [416] "which"      "the"        "to"         "and"        "them"      
#> [421] "In"         "these"      "I"          "the"        "that"      
#> [426] "as"         "on"         "no"         "or"         "no"        
#> [431] "nor"        "the"        "and"        "which"      "ought"     
#> [436] "to"         "over"       "this"       "of"         "and"       
#> [441] "so"         "on"         "that"       "the"        "of"        
#> [446] "our"        "be"         "in"         "the"        "and"       
#> [451] "of"         "and"        "the"        "of"         "be"        
#> [456] "by"         "all"        "the"        "which"      "the"       
#> [461] "of"         "its"        "and"        "the"        "of"        
#> [466] "the"        "I"          "on"         "this"       "with"      
#> [471] "which"      "an"         "for"        "my"         "there"     
#> [476] "is"         "no"         "more"       "than"       "that"      
#> [481] "there"      "in"         "the"        "and"        "of"        
#> [486] "an"         "between"    "and"        "between"    "and"       
#> [491] "between"    "the"        "of"         "an"         "and"       
#> [496] "and"        "the"        "of"         "and"        "we"        
#> [501] "ought"      "to"         "be"         "no"         "that"      
#> [506] "the"        "of"         "be"         "on"         "a"         
#> [511] "that"       "the"        "of"         "and"        "which"     
#> [516] "itself"     "has"        "and"        "the"        "of"        
#> [521] "the"        "of"         "and"        "the"        "of"        
#> [526] "the"        "of"         "are"        "as"         "as"        
#> [531] "on"         "the"        "to"         "the"        "of"        
#> [536] "the"        "the"        "to"         "your"       "it"        
#> [541] "with"       "your"       "to"         "how"        "an"        
#> [546] "of"         "the"        "by"         "the"        "of"        
#> [551] "the"        "is"         "at"         "the"        "by"        
#> [556] "the"        "of"         "which"      "have"       "been"      
#> [561] "against"    "the"        "or"         "by"         "the"       
#> [566] "of"         "which"      "has"        "to"         "them"      
#> [571] "of"         "on"         "this"       "in"         "which"     
#> [576] "I"          "could"      "be"         "by"         "no"        
#> [581] "from"       "I"          "again"      "to"         "my"        
#> [586] "in"         "your"       "and"        "of"         "the"       
#> [591] "for"        "I"          "myself"     "that"       "you"       
#> [596] "which"      "the"        "of"         "an"         "and"       
#> [601] "or"         "which"      "ought"      "to"         "the"       
#> [606] "of"         "a"          "for"        "the"        "of"        
#> [611] "and"        "a"          "for"        "the"        "your"      
#> [616] "on"         "the"        "how"        "the"        "be"        
#> [621] "or"         "the"        "be"         "and"        "To"        
#> [626] "the"        "I"          "have"       "to"         "which"     
#> [631] "be"         "most"       "to"         "the"        "of"        
#> [636] "It"         "myself"     "and"        "be"         "as"        
#> [641] "as"         "When"       "I"          "was"        "with"      
#> [646] "a"          "into"       "the"        "of"         "my"        
#> [651] "then"       "on"         "the"        "of"         "an"        
#> [656] "for"        "its"        "the"        "in"         "which"     
#> [661] "I"          "my"         "that"       "I"          "should"    
#> [666] "From"       "this"       "I"          "have"       "in"        
#> [671] "no"         "and"        "being"      "under"      "the"       
#> [676] "which"      "it"         "I"          "as"         "to"        
#> [681] "myself"     "any"        "in"         "the"        "which"     
#> [686] "be"         "in"         "a"          "for"        "the"       
#> [691] "and"        "that"       "the"        "for"        "the"       
#> [696] "in"         "which"      "I"          "am"         "during"    
#> [701] "my"         "in"         "it"         "be"         "to"        
#> [706] "such"       "as"         "the"        "be"         "to"        
#> [711] "Having"     "to"         "you"        "my"         "as"        
#> [716] "they"       "have"       "been"       "by"         "the"       
#> [721] "which"      "I"          "my"         "but"        "not"       
#> [726] "once"       "more"       "to"         "the"        "of"        
#> [731] "the"        "in"         "that"       "He"         "has"       
#> [736] "been"       "to"         "the"        "with"       "for"       
#> [741] "in"         "and"        "for"        "with"       "on"        
#> [746] "a"          "of"         "for"        "the"        "of"        
#> [751] "their"      "and"        "the"        "of"         "their"     
#> [756] "so"         "His"        "be"         "in"         "the"       
#> [761] "the"        "and"        "the"        "on"         "which"     
#> [766] "the"        "of"         "this"      
#> 
#> 1793-Washington :
#>  [1] "I"      "am"     "again"  "by"     "the"    "of"     "my"     "to"    
#>  [9] "the"    "of"     "its"    "When"   "the"    "for"    "it"     "I"     
#> [17] "to"     "the"    "I"      "of"     "this"   "and"    "of"     "the"   
#> [25] "which"  "has"    "been"   "in"     "me"     "by"     "the"    "of"    
#> [33] "to"     "the"    "of"     "any"    "of"     "the"    "the"    "an"    
#> [41] "of"     "This"   "I"      "am"     "about"  "to"     "and"    "in"    
#> [49] "your"   "That"   "if"     "it"     "be"     "during" "my"     "of"    
#> [57] "the"    "I"      "have"   "in"     "any"    "or"     "the"    "I"     
#> [65] "be"     "to"     "the"    "of"     "all"    "who"    "are"    "of"    
#> [73] "the"   
#> 

selectFeatures(toks, stopwords("english"), "remove", padding = TRUE)

#> Error in attr(x, "padding") && "" %in% features: invalid 'x' type in 'x && y'

selectFeatures(toks, stopwords("english"), "keep", padding = TRUE)

#> tokenizedTexts from 2 documents.
#> 1789-Washington :
#>    [1] ""           ""           ""           "of"         "the"       
#>    [6] ""           "and"        "of"         "the"        ""          
#>   [11] "of"         ""           ""           ""           "the"       
#>   [16] ""           ""           "to"         ""           "no"        
#>   [21] ""           "could"      "have"       ""           "me"        
#>   [26] "with"       ""           ""           "than"       "that"      
#>   [31] "of"         "which"      "the"        ""           "was"       
#>   [36] ""           "by"         "your"       ""           ""          
#>   [41] "and"        ""           "on"         "the"        ""          
#>   [46] ""           "of"         "the"        ""           ""          
#>   [51] ""           "On"         "the"        ""           ""          
#>   [56] ""           "I"          "was"        ""           "by"        
#>   [61] "my"         ""           ""           ""           ""          
#>   [66] "I"          ""           ""           ""           "but"       
#>   [71] "with"       ""           "and"        ""           ""          
#>   [76] "from"       "a"          ""           "which"      "I"         
#>   [81] "had"        ""           "with"       "the"        ""          
#>   [86] ""           ""           "and"        ""           "in"        
#>   [91] "my"         ""           ""           ""           "with"      
#>   [96] "an"         ""           ""           ""           "as"        
#>  [101] "the"        ""           "of"         "my"         ""          
#>  [106] ""           ""           ""           "a"          ""          
#>  [111] "which"      "was"        ""           ""           ""          
#>  [116] "more"       ""           "as"         ""           "as"        
#>  [121] "more"       ""           "to"         "me"         "by"        
#>  [126] "the"        ""           "of"         ""           "to"        
#>  [131] ""           ""           "and"        "of"         ""          
#>  [136] ""           "in"         "my"         ""           "to"        
#>  [141] "the"        ""           ""           ""           "on"        
#>  [146] "it"         "by"         ""           ""           "On"        
#>  [151] "the"        "other"      ""           ""           "the"       
#>  [156] ""           "and"        ""           "of"         "the"       
#>  [161] ""           "to"         "which"      "the"        ""          
#>  [166] "of"         "my"         ""           ""           "me"        
#>  [171] ""           "being"      ""           "to"         ""          
#>  [176] "in"         "the"        ""           "and"        "most"      
#>  [181] ""           "of"         "her"        ""           "a"         
#>  [186] ""           ""           "into"       "his"        ""          
#>  [191] ""           "could"      "not"        "but"        ""          
#>  [196] "with"       ""           ""           "who"        ""          
#>  [201] ""           ""           ""           "from"       ""          
#>  [206] "and"        ""           "in"         "the"        ""          
#>  [211] "of"         ""           ""           ""           "ought"     
#>  [216] "to"         "be"         ""           ""           "of"        
#>  [221] "his"        "own"        ""           ""           "In"        
#>  [226] "this"       ""           "of"         ""           "all"       
#>  [231] "I"          ""           ""           "is"         "that"      
#>  [236] "it"         "has"        "been"       "my"         ""          
#>  [241] ""           "to"         ""           "my"         ""          
#>  [246] "from"       "a"          ""           ""           "of"        
#>  [251] ""           ""           "by"         "which"      "it"        
#>  [256] ""           "be"         ""           ""           "All"       
#>  [261] "I"          ""           ""           "is"         "that"      
#>  [266] "if"         ""           "in"         ""           "this"      
#>  [271] ""           ""           "I"          "have"       "been"      
#>  [276] "too"        ""           ""           "by"         "a"         
#>  [281] ""           ""           "of"         ""           ""          
#>  [286] ""           "or"         "by"         "an"         ""          
#>  [291] ""           "to"         "this"       ""           ""          
#>  [296] "of"         "the"        ""           "of"         "my"        
#>  [301] ""           ""           ""           "and"        "have"      
#>  [306] ""           "too"        ""           ""           "my"        
#>  [311] ""           "as"         ""           "as"         ""          
#>  [316] "for"        "the"        ""           "and"        ""          
#>  [321] ""           "before"     "me"         ""           "my"        
#>  [326] ""           ""           "be"         ""           "by"        
#>  [331] "the"        ""           "which"      ""           "me"        
#>  [336] ""           "and"        "its"        ""           "be"        
#>  [341] ""           "by"         "my"         ""           "with"      
#>  [346] "some"       ""           "of"         "the"        ""          
#>  [351] "in"         "which"      "they"       ""           ""          
#>  [356] "Such"       "being"      "the"        ""           "under"     
#>  [361] "which"      "I"          "have"       ""           "in"        
#>  [366] ""           "to"         "the"        ""           ""          
#>  [371] ""           ""           "to"         "the"        ""          
#>  [376] ""           ""           "it"         "would"      "be"        
#>  [381] ""           ""           "to"         ""           "in"        
#>  [386] "this"       ""           ""           ""           "my"        
#>  [391] ""           ""           "to"         "that"       ""          
#>  [396] "Being"      "who"        ""           "over"       "the"       
#>  [401] ""           ""           "who"        ""           "in"        
#>  [406] "the"        ""           "of"         ""           ""          
#>  [411] "and"        ""           ""           ""           ""          
#>  [416] ""           ""           ""           ""           ""          
#>  [421] "that"       "His"        ""           ""           ""          
#>  [426] "to"         "the"        ""           "and"        ""          
#>  [431] "of"         "the"        ""           "of"         "the"       
#>  [436] ""           ""           "a"          ""           ""          
#>  [441] "by"         "themselves" "for"        "these"      ""          
#>  [446] ""           ""           "and"        ""           ""          
#>  [451] ""           ""           ""           "in"         "its"       
#>  [456] ""           "to"         ""           "with"       ""          
#>  [461] "the"        ""           ""           "to"         "his"       
#>  [466] ""           ""           "In"         ""           "this"      
#>  [471] ""           "to"         "the"        ""           ""          
#>  [476] "of"         ""           ""           "and"        ""          
#>  [481] ""           ""           "I"          ""           "myself"    
#>  [486] "that"       "it"         ""           "your"       ""          
#>  [491] "not"        ""           "than"       "my"         "own"       
#>  [496] ""           "nor"        "those"      "of"         "my"        
#>  [501] ""           ""           "at"         ""           ""          
#>  [506] "than"       ""           ""           "No"         ""          
#>  [511] ""           "be"         ""           "to"         ""          
#>  [516] "and"        ""           "the"        ""           ""          
#>  [521] "which"      ""           "the"        ""           "of"        
#>  [526] ""           "more"       "than"       "those"      "of"        
#>  [531] "the"        ""           ""           ""           ""          
#>  [536] ""           "by"         "which"      "they"       "have"      
#>  [541] ""           "to"         "the"        ""           "of"        
#>  [546] "an"         ""           ""           ""           "to"        
#>  [551] "have"       "been"       ""           "by"         "some"      
#>  [556] ""           "of"         ""           ""           ""          
#>  [561] "and"        "in"         "the"        ""           ""          
#>  [566] ""           ""           "in"         "the"        ""          
#>  [571] "of"         "their"      ""           ""           "the"       
#>  [576] ""           ""           "and"        ""           ""          
#>  [581] "of"         "so"         ""           ""           ""          
#>  [586] "from"       "which"      "the"        ""           "has"       
#>  [591] ""           ""           "not"        "be"         ""          
#>  [596] "with"       "the"        ""           "by"         "which"     
#>  [601] "most"       ""           "have"       "been"       ""          
#>  [606] ""           "some"       ""           "of"         ""          
#>  [611] ""           ""           ""           "with"       "an"        
#>  [616] ""           ""           "of"         "the"        ""          
#>  [621] ""           "which"      "the"        ""           ""          
#>  [626] "to"         ""           ""           "These"      ""          
#>  [631] ""           ""           "out"        "of"         "the"       
#>  [636] ""           ""           ""           "have"       ""          
#>  [641] "themselves" "too"        ""           "on"         "my"        
#>  [646] ""           "to"         "be"         ""           ""          
#>  [651] "You"        ""           ""           "with"       "me"        
#>  [656] ""           "I"          ""           ""           "in"        
#>  [661] ""           "that"       "there"      "are"        ""          
#>  [666] "under"      "the"        ""           "of"         "which"     
#>  [671] "the"        ""           "of"         "a"          ""          
#>  [676] "and"        ""           ""           ""           "more"      
#>  [681] ""           ""           ""           "By"         "the"       
#>  [686] ""           ""           "the"        ""           ""          
#>  [691] "it"         "is"         ""           "the"        ""          
#>  [696] "of"         "the"        ""           ""           "to"        
#>  [701] ""           "to"         "your"       ""           "such"      
#>  [706] ""           "as"         "he"         ""           ""          
#>  [711] ""           "and"        ""           ""           ""          
#>  [716] "The"        ""           "under"      "which"      "I"         
#>  [721] ""           ""           "you"        ""           ""          
#>  [726] "me"         "from"       ""           "into"       "that"      
#>  [731] ""           "further"    "than"       "to"         ""          
#>  [736] "to"         "the"        ""           ""           ""          
#>  [741] "under"      "which"      "you"        "are"        ""          
#>  [746] ""           "and"        "which"      ""           "in"        
#>  [751] ""           "your"       ""           ""           ""          
#>  [756] "the"        ""           "to"         "which"      "your"      
#>  [761] ""           "is"         "to"         "be"         ""          
#>  [766] ""           "It"         ""           "be"         "more"      
#>  [771] ""           "with"       "those"      ""           ""          
#>  [776] "and"        ""           "more"       ""           "with"      
#>  [781] "the"        ""           "which"      ""           "me"        
#>  [786] ""           "to"         ""           ""           "in"        
#>  [791] ""           "of"         "a"          ""           "of"        
#>  [796] ""           ""           ""           "the"        ""          
#>  [801] "that"       "is"         ""           "to"         "the"       
#>  [806] ""           ""           "the"        ""           ""          
#>  [811] "and"        "the"        ""           "which"      ""          
#>  [816] "the"        ""           ""           "to"         ""          
#>  [821] "and"        ""           "them"       ""           "In"        
#>  [826] "these"      ""           ""           "I"          ""          
#>  [831] "the"        ""           ""           "that"       "as"        
#>  [836] "on"         ""           ""           "no"         ""          
#>  [841] ""           "or"         ""           ""           "no"        
#>  [846] ""           ""           "nor"        ""           ""          
#>  [851] ""           ""           ""           "the"        ""          
#>  [856] "and"        ""           ""           "which"      "ought"     
#>  [861] "to"         ""           "over"       "this"       ""          
#>  [866] ""           "of"         ""           "and"        ""          
#>  [871] ""           "so"         ""           "on"         ""          
#>  [876] ""           "that"       "the"        ""           "of"        
#>  [881] "our"        ""           ""           ""           "be"        
#>  [886] ""           "in"         "the"        ""           "and"       
#>  [891] ""           ""           "of"         ""           ""          
#>  [896] ""           "and"        "the"        ""           "of"        
#>  [901] ""           ""           "be"         ""           "by"        
#>  [906] "all"        "the"        ""           "which"      ""          
#>  [911] ""           "the"        ""           "of"         "its"       
#>  [916] ""           "and"        ""           "the"        ""          
#>  [921] "of"         "the"        ""           ""           "I"         
#>  [926] ""           "on"         "this"       ""           "with"      
#>  [931] ""           ""           "which"      "an"         ""          
#>  [936] ""           "for"        "my"         ""           ""          
#>  [941] ""           ""           ""           "there"      "is"        
#>  [946] "no"         ""           "more"       ""           ""          
#>  [951] "than"       "that"       "there"      ""           "in"        
#>  [956] "the"        ""           "and"        ""           "of"        
#>  [961] ""           "an"         ""           ""           "between"   
#>  [966] ""           "and"        ""           ""           "between"   
#>  [971] ""           "and"        ""           ""           "between"   
#>  [976] "the"        ""           ""           "of"         "an"        
#>  [981] ""           "and"        ""           ""           "and"       
#>  [986] "the"        ""           ""           "of"         ""          
#>  [991] ""           "and"        ""           ""           ""          
#>  [996] "we"         "ought"      "to"         "be"         "no"        
#>  [ reached getOption("max.print") -- omitted 540 entries ]
#> 
#> 1793-Washington :
#>   [1] ""       ""       ""       "I"      "am"     "again"  ""       ""      
#>   [9] "by"     "the"    ""       "of"     "my"     ""       "to"     ""      
#>  [17] "the"    ""       "of"     "its"    ""       ""       ""       "When"  
#>  [25] "the"    ""       ""       "for"    "it"     ""       ""       ""      
#>  [33] "I"      ""       ""       "to"     ""       "the"    ""       ""      
#>  [41] "I"      ""       "of"     "this"   ""       ""       ""       "and"   
#>  [49] "of"     "the"    ""       "which"  "has"    "been"   ""       "in"    
#>  [57] "me"     "by"     "the"    ""       "of"     ""       ""       ""      
#>  [65] ""       "to"     "the"    ""       "of"     "any"    ""       ""      
#>  [73] "of"     "the"    ""       "the"    ""       ""       "an"     ""      
#>  [81] "of"     ""       ""       "This"   ""       "I"      "am"     ""      
#>  [89] "about"  "to"     ""       ""       "and"    "in"     "your"   ""      
#>  [97] ""       "That"   "if"     "it"     ""       "be"     ""       "during"
#> [105] "my"     ""       "of"     "the"    ""       "I"      "have"   "in"    
#> [113] "any"    ""       ""       ""       "or"     ""       "the"    ""      
#> [121] ""       ""       "I"      ""       ""       ""       ""       ""      
#> [129] ""       ""       "be"     ""       "to"     "the"    ""       "of"    
#> [137] "all"    "who"    "are"    ""       ""       "of"     "the"    ""      
#> [145] ""       ""       ""      
#> 

selectFeatures(tokenize(data_char_inaugural[2]), stopwords("english"), "remove", padding = TRUE)

#> Error in attr(x, "padding") && "" %in% features: invalid 'x' type in 'x && y'

toksh <- tokens(c(doc1 = "This is a SAMPLE text", doc2 = "this sample text is better"))
feats <- c("this", "sample", "is")
# keeping features
selectFeatures(toksh, feats, selection = "keep")

#> tokens from 2 documents.
#> doc1 :
#> [1] "This"   "is"     "SAMPLE"
#> 
#> doc2 :
#> [1] "this"   "sample" "is"    
#> 

selectFeatures(toksh, feats, selection = "keep", padding = TRUE)

#> tokens from 2 documents.
#> doc1 :
#> [1] "This"   "is"     ""       "SAMPLE" ""      
#> 
#> doc2 :
#> [1] "this"   "sample" ""       "is"     ""      
#> 

selectFeatures(toksh, feats, selection = "keep", case_insensitive = FALSE)

#> tokens from 2 documents.
#> doc1 :
#> [1] "is"
#> 
#> doc2 :
#> [1] "this"   "sample" "is"    
#> 

selectFeatures(toksh, feats, selection = "keep", padding = TRUE, case_insensitive = FALSE)

#> tokens from 2 documents.
#> doc1 :
#> [1] ""   "is" ""   ""   ""  
#> 
#> doc2 :
#> [1] "this"   "sample" ""       "is"     ""      
#> 

# removing features
selectFeatures(toksh, feats, selection = "remove")

#> tokens from 2 documents.
#> doc1 :
#> [1] "a"    "text"
#> 
#> doc2 :
#> [1] "text"   "better"
#> 

selectFeatures(toksh, feats, selection = "remove", padding = TRUE)

#> tokens from 2 documents.
#> doc1 :
#> [1] ""     ""     "a"    ""     "text"
#> 
#> doc2 :
#> [1] ""       ""       "text"   ""       "better"
#> 

selectFeatures(toksh, feats, selection = "remove", case_insensitive = FALSE)

#> tokens from 2 documents.
#> doc1 :
#> [1] "This"   "a"      "SAMPLE" "text"  
#> 
#> doc2 :
#> [1] "text"   "better"
#> 

selectFeatures(toksh, feats, selection = "remove", padding = TRUE, case_insensitive = FALSE)

#> tokens from 2 documents.
#> doc1 :
#> [1] "This"   ""       "a"      "SAMPLE" "text"  
#> 
#> doc2 :
#> [1] ""       ""       "text"   ""       "better"
#> 

## example for collocations
(myCollocs <- collocations(data_char_inaugural[1:3], n=20))

#>        word1    word2 word3 count       G2
#>  1:       of      the          53 62.65878
#>  2:      has     been           7 61.17854
#>  3:       my  country           8 53.59480
#>  4:      may       be           9 50.66899
#>  5: american   people           5 50.50639
#>  6:   united   states           5 49.79359
#>  7:      the   people          16 49.18676
#>  8:  foreign  nations           5 48.65287
#>  9:        i     have           8 46.76602
#> 10:     have     been           6 43.64782
#> 11:       if        a           9 41.51057
#> 12:       if       an           7 40.67846
#> 13:   fellow citizens           3 38.82186
#> 14:     well       as           4 36.76287
#> 15:      for      the          15 34.56611
#> 16:       as     well           4 34.25071
#> 17:       to       be          13 31.12807
#> 18:       it    shall           5 31.10789
#> 19:    which        i           7 30.33006
#> 20: official      act           2 30.05840

selectFeatures(myCollocs, stopwords("english"), "remove")

#> Removed 15 (75%) of 20 collocations containing one of 174 stopwords.

#>       word1    word2 word3 count       G2
#> 1: american   people           5 50.50639
#> 2:   united   states           5 49.79359
#> 3:  foreign  nations           5 48.65287
#> 4:   fellow citizens           3 38.82186
#> 5: official      act           2 30.05840