correspondence analysis of a document-feature matrix

textmodel_ca implements correspondence analysis scaling on a dfm. The method is a fast/sparse version of function ca, and returns a special class of ca object.

textmodel_ca(x, smooth = 0, nd = NA, sparse = FALSE, threads = 1,
  residual_floor = 0.1)

Arguments

x
the dfm on which the model will be fit

smooth
a smoothing parameter for word counts; defaults to zero.

nd
Number of dimensions to be included in output; if NA (the default) then the maximum possible dimensions are included.

sparse
retains the sparsity if set to TRUE; set it to TRUE if x (the dfm) is too big to be allocated after converting to dense

threads
the number of threads to be used; set to 1 to use a serial version of the function; only applicable when sparse = TRUE

residual_floor
specifies the threshold for the residual matrix for calculating the truncated svd.Larger value will reduce memory and time cost but might sacrify the accuracy; only applicable when sparse = TRUE

Details

svds in the RSpectra package is applied to enable the fast computation of the SVD.

Note

Setting threads larger than 1 (when sparse = TRUE) will trigger parallel computation, which retains sparsity of all involved matrices. You may need to increase the value of residual_floor to ignore less important information and hence to reduce the memory cost when you have a very big dfm. If your attempt to fit the model fails due to the matrix being too large, this is probably because of the memory demands of computing the \(V \times V\) residual matrix. To avoid this, consider increasing the value of residual_floor by 0.1, until the model can be fit.

References

Nenadic, O. and Greenacre, M. (2007). Correspondence analysis in R, with two- and three-dimensional graphics: The ca package. Journal of Statistical Software, 20 (3), http://www.jstatsoft.org/v20/i03/.

Examples

ieDfm <- dfm(data_corpus_irishbudget2010)
wca <- textmodel_ca(ieDfm)
summary(wca)

#> Warning: abbreviate used with non-ASCII chars

#> Warning: abbreviate used with non-ASCII chars

#> Warning: some row.names duplicated: 25,187,198,365,371,485,942,944,1109,1166,1231,1243,1610,1642,1717,1825,1877,1918,1962,2003,2024,2401,2784,2785,2804,2807,2810,2975,3021,3069,3216,3277,3364,3624,3687,3743,3762,3837,3964,4168,4205,4259,4276,4304,4395,4422,4552,4569,4647,4666,4838,4948,4955,4958,4962,4966,4968 --> row.names NOT used

#> 
#> Principal inertias (eigenvalues):
#> 
#>  dim    value      %   cum%   scree plot               
#>  1      0.185190  19.5  19.5  *****                    
#>  2      0.139466  14.7  34.2  ****                     
#>  3      0.134072  14.1  48.3  ****                     
#>  4      0.127248  13.4  61.6  ***                      
#>  5      0.125848  13.2  74.9  ***                      
#>  6      0.121379  12.8  87.6  ***                      
#>  7      0.117471  12.4 100.0  ***                      
#>         -------- -----                                 
#>  Total: 0.950673 100.0                                 
#> 
#> 
#> Rows:
#>                name   mass  qlt  inr    k=1 cor ctr    k=2 cor ctr    k=3 cor
#> 1  | 2010_BUDGET_01 |  159 1000   78 |  676 852 392 | -162  49  30 |  -14   0
#> 2  | 2010_BUDGET_02 |   82 1000   76 | -335  84  50 |  407 124  97 |  541 220
#> 3  | 2010_BUDGET_03 |  117 1000   80 | -413 162 108 | -904 776 688 |  194  36
#> 4  | 2010_BUDGET_04 |  131 1000   63 | -150 136  16 |   86  45   7 | -324 631
#> 5  | 2010_BUDGET_05 |  121 1000   74 |  633 903 262 |   48   5   2 |  142  46
#> 6  | 2010_BUDGET_06 |   78 1000   73 | -345  88  50 |   94   7   5 |   91   6
#> 7  | 2010_BUDGET_07 |   42 1000   64 | -266 440  16 |  115  83   4 |   27   5
#> 8  | 2010_BUDGET_08 |   77 1000   74 | -277  54  32 |  207  30  24 | -174  21
#> 9  | 2010_BUDGET_09 |   24 1000   75 | -478  82  29 |  509  93  44 |   96   3
#> 10 | 2010_BUDGET_10 |   24 1000   75 | -442  53  25 |  251  17  11 |  215  13
#> 11 | 2010_BUDGET_11 |   19 1000   54 |  292 412   9 |  154 115   3 |  -33   5
#> 12 | 2010_BUDGET_12 |   30 1000   66 |  -82  13   1 |  357 250  28 |  513 516
#> 13 | 2010_BUDGET_13 |   23 1000   74 |   59   1   0 |  553  90  50 |  633 118
#> 14 | 2010_BUDGET_14 |   74 1000   74 | -161  22  10 |  124  13   8 | -958 779
#>    ctr     k=4 cor ctr    k=5 cor ctr     k=6 cor ctr     k=7 cor ctr  
#> 1    0 |   -28   1   1 | -121  27  18 |    71   9   7 |  -181  61  44 |
#> 2  179 |   812 494 423 | -169  21  19 |    35   1   1 |  -269  54  50 |
#> 3   33 |    79   6   6 |  115  12  12 |   -90   8   8 |   -15   0   0 |
#> 4  102 |    77  36   6 |  -95  55   9 |  -112  76  14 |    61  23   4 |
#> 5   18 |   -23   1   1 |  110  27  12 |    12   0   0 |    89  18   8 |
#> 6    5 |  -433 138 114 | -720 383 320 |   495 182 157 |   515 196 176 |
#> 7    0 |  -180 202  11 | -180 202  11 |    63  25   1 |    84  44   3 |
#> 8   17 |  -176  22  19 |  935 618 533 |   595 250 224 |    71   4   3 |
#> 9    2 |  -700 175  91 | -171  10   5 |   264  25  14 | -1309 612 345 |
#> 10   8 | -1160 365 249 |   90   2   2 | -1349 494 353 |  -456  56  42 |
#> 11   0 |  -191 177   5 |   14   1   0 |  -232 261   8 |    78  29   1 |
#> 12  59 |  -100  20   2 |  242 115  14 |  -138  37   5 |  -159  50   6 |
#> 13  68 |    41   1   0 |  474  66  41 |  -905 241 154 |  1280 483 318 |
#> 14 508 |   350 104  72 |  -82   6   4 |  -300  76  55 |     8   0   0 |
#> 
#> Columns:
#>                name   mass  qlt  inr     k=1 cor ctr     k=2 cor ctr     k=3
#> 1    |         when |    2 1000    0 |  -379 501   1 |    73  19   0 |   -99
#> 2    |            i |    5 1000    1 |   106  45   0 |  -162 104   1 |   177
#> 3    |       prsntd |    0 1000    0 |   132  11   0 |    50   2   0 |  -802
#> 4    |          the |   66 1000    0 |   -47 453   1 |    -3   2   0 |   -27
#> 5    |     spplmntr |    0 1000    0 |   947 649   1 |  -222  36   0 |  -352
#> 6    |        budgt |    5 1000    1 |  -206 460   1 |     0   0   0 |   -30
#> 7    |           to |   30 1000    0 |    72 275   1 |    46 114   0 |    59
#> 8    |         this |   10 1000    0 |   -10   3   0 |    73 153   0 |   -54
#> 9    |         hous |    1 1000    1 |   -85   4   0 |   -25   0   0 |    55
#> 10   |         last |    1 1000    0 |   -22   3   0 |    99  68   0 |  -258
#> 11   |        april |    0 1000    0 |  -383  84   0 |   268  41   0 |  -284
#> 12   |              |   28 1000    1 |    50 131   0 |    -2   0   0 |    64
#> 13   |         said |    0 1000    0 |  -616 309   1 |  -367 110   0 |   247
#> 14   |           we |   11 1000    2 |   175 197   2 |   172 189   2 |   276
#> 15   |        could |    1 1000    1 |  -600 600   2 |   190  60   0 |   198
#> 16   |         work |    1 1000    0 |   265 318   0 |   331 496   1 |   154
#> 17   |          our |    5 1000    1 |   488 747   6 |    91  26   0 |   187
#> 18   |          way |    1 1000    0 |    32  18   0 |   101 174   0 |    42
#> 19   |       throgh |    1 1000    2 |   335  64   1 |   301  52   1 |   165
#> 20   |         perd |    0 1000    0 |   268 122   0 |   -42   3   0 |   233
#> 21   |           of |   28 1000    0 |    48 247   0 |   -28  84   0 |   -29
#> 22   |         sevr |    0 1000    0 |  1108 727   1 |  -123   9   0 |   -40
#> 23   |       econmc |    2 1000    0 |   199 426   0 |    14   2   0 |   -65
#> 24   |        dstrs |    0 1000    0 |  1570 796   0 |  -434  61   0 |   -38
#> 25   |              |   44 1000    0 |   -29 253   0 |   -19 111   0 |     3
#> 26   |         tody |    1 1000    0 |   -69   8   0 |  -611 604   2 |   358
#> 27   |          can |    2 1000    0 |    42  13   0 |   153 175   0 |    40
#> 28   |        reprt |    0 1000    0 |   445 143   0 |   162  19   0 |   704
#>      cor ctr     k=4 cor ctr     k=5 cor ctr     k=6 cor ctr     k=7 cor ctr  
#> 1     34   0 |  -286 287   1 |   -92  29   0 |  -148  77   0 |   123  53   0 |
#> 2    123   1 |  -364 521   5 |   -91  33   0 |  -190 142   1 |   -93  34   0 |
#> 3    402   0 |  -104   7   0 |  -866 469   0 |   255  41   0 |   332  69   0 |
#> 4    153   0 |    12  31   0 |   -35 249   1 |    11  25   0 |    21  86   0 |
#> 5     90   0 |   -57   2   0 |  -491 174   0 |   166  20   0 |  -199  29   0 |
#> 6     10   0 |   133 190   1 |    54  32   0 |    69  51   0 |   154 257   1 |
#> 7    183   1 |    44 104   0 |   -55 158   1 |   -27  39   0 |   -49 127   1 |
#> 8     82   0 |    83 195   1 |   -41  48   0 |  -100 287   1 |    90 232   1 |
#> 9      2   0 |  -878 415   5 |   160  14   0 |  -924 459   6 |  -446 107   2 |
#> 10   464   0 |   208 302   0 |   -71  35   0 |   -96  64   0 |    95  63   0 |
#> 11    46   0 |  -103   6   0 |  -997 567   1 |   416  99   0 |   525 157   0 |
#> 12   212   1 |  -103 549   2 |    15  12   0 |    37  72   0 |    22  25   0 |
#> 13    50   0 |  -637 331   1 |    78   5   0 |   392 126   0 |   292  70   0 |
#> 14   490   6 |    29   5   0 |   109  76   1 |   -80  41   1 |   -20   3   0 |
#> 15    65   0 |  -160  42   0 |    33   2   0 |   205  70   0 |  -311 161   1 |
#> 16   107   0 |   -37   6   0 |   -18   1   0 |   -43   9   0 |   118  63   0 |
#> 17   109   1 |    90  26   0 |   104  34   0 |  -122  47   1 |    59  11   0 |
#> 18    31   0 |    14   3   0 |  -198 674   0 |   -43  32   0 |    63  68   0 |
#> 19    16   0 |    -1   0   0 |   384  85   1 |  -628 227   4 |   983 556  10 |
#> 20    92   0 |   -82  12   0 |     2   0   0 |  -272 125   0 |   618 646   1 |
#> 21    95   0 |   -45 223   0 |   -44 210   0 |   -15  24   0 |   -33 117   0 |
#> 22     1   0 |  -159  15   0 |   385  88   0 |   470 131   0 |  -224  30   0 |
#> 23    45   0 |   208 463   1 |   -50  27   0 |    58  36   0 |    -9   1   0 |
#> 24     0   0 |   -79   2   0 |  -340  37   0 |   203  13   0 |  -528  90   0 |
#> 25     3   0 |    18 103   0 |    39 462   1 |    -5   6   0 |   -14  62   0 |
#> 26   207   1 |   330 176   1 |    17   0   0 |    38   2   0 |   -35   2   0 |
#> 27    12   0 |   140 145   0 |   201 301   1 |   -65  32   0 |   208 322   1 |
#> 28   357   1 |   652 306   1 |  -132  13   0 |    46   2   0 |  -472 160   1 |
#>  [ reached getOption("max.print") -- omitted 5112 rows ]