calculate collocation statistics

Identify and score collocations from a tokenized text.

textstat_collocations(x, method = c("lambda", "lambda1", "lr", "chi2", "pmi",
  "dice"), size = 2, min_count = 2, ...)

is.collocations(x)

Arguments

x
a tokens object whose collocations will be scored

method
association measure for detecting collocations. Let \(i\) index documents, and \(j\) index features, \(n_{ij}\) refers to observed counts, and \(m_{ij}\) the expected counts in a collocations frequency table of dimensions \((J - size + 1)^2\).

Available measures are computed as:

"chi2"

Pearson's \(\chi^2\) statistic, computed as: $$\sum_i \sum_j \frac{(n_{ij} - m_{ij})^2}{m_{ij}}$$

"pmi"

point-wise mutual information score, computed as log \(n_{11}/m_{11}\)

"dice"

the Dice coefficient, computed as \(n_{11}/n_{1.} + n_{.1}\)

"lambda1"

unigram subtuples, Blaheta and Johnson's method (called through sequences)

"lambda"

all subtuples algorithm, Blaheta and Johnson's method (called through sequences)

size
numeric argument representing the length of the collocations to be scored. The maximum size is currently 3 for all methods except "lambda" and "lambda1", which has a maximum size of 5. Use c(2,n) or 2:n to return collocations of bigram to n-gram collocations.

min_count
minimum frequency of collocations that will be scored

...
additional arguments passed to collocations2 for the first four methods, or to sequences for method = "bj_*"

Value

textstat_collocations returns a data.frame of collocations and their scores and statistsics. is.collocation returns TRUE if the object is of class collocations, FALSE otherwise.

Note

This function is under active development, and we aim to improve both its operation and efficiency in the next release of quanteda.

References

Blaheta, D., & Johnson, M. (2001). Unsupervised learning of multi-word verbs. Presented at the ACLEACL Workshop on the Computational Extraction, Analysis and Exploitation of Collocations. McInnes, B T. 2004. "Extending the Log Likelihood Measure to Improve Collocation Identification." M.Sc. Thesis, University of Minnesota.

Examples

txts <- c("quanteda is a package for quantitative text analysis",
          "quantitative text analysis is a rapidly growing field",
          "The population is rapidly growing")
toks <- tokens(txts)
textstat_collocations(toks, method = "lr")

#>         collocation count length   lambda    sigma      dice      pmi       G2
#> 1 quantitative text     2      2 5.105945 2.112015 0.8333333 1.714798 9.693488
#> 2   rapidly growing     2      2 5.105945 2.112015 0.8333333 1.714798 9.693488
#> 3     text analysis     2      2 5.105945 2.112015 0.8333333 1.714798 9.693488
#> 4              is a     2      2 3.944813 1.769515 0.7142857 1.427116 7.165912
#>        chi2        z           p
#> 1 12.925798 2.417571 0.007812243
#> 2 12.925798 2.417571 0.007812243
#> 3 12.925798 2.417571 0.007812243
#> 4  8.848039 2.229319 0.012896348

textstat_collocations(toks, method = "lr", min_count = 1)

#>          collocation count length   lambda    sigma      dice       pmi
#> 1  quantitative text     2      2 5.105945 2.112015 0.8333333 1.7147984
#> 2    rapidly growing     2      2 5.105945 2.112015 0.8333333 1.7147984
#> 3      text analysis     2      2 5.105945 2.112015 0.8333333 1.7147984
#> 4               is a     2      2 3.944813 1.769515 0.7142857 1.4271164
#> 5        package for     1      2 4.653960 2.173433 0.7500000 2.0149030
#> 6     The population     1      2 4.653960 2.173433 0.7500000 2.0149030
#> 7      growing field     1      2 4.653960 2.173433 0.7500000 2.0149030
#> 8   for quantitative     1      2 4.653960 2.173433 0.7500000 2.0149030
#> 9          a package     1      2 3.496508 1.842265 0.6000000 1.6094379
#> 10       analysis is     1      2 2.923162 1.769515 0.5000000 1.3217558
#> 11       quanteda is     1      2 2.923162 1.769515 0.5000000 1.3217558
#> 12     population is     1      2 2.923162 1.769515 0.5000000 1.3217558
#> 13         a rapidly     1      2 2.335375 1.436842 0.5000000 1.2039728
#> 14        is rapidly     1      2 1.757858 1.342497 0.4285714 0.9162907
#>          G2      chi2        z           p
#> 1  9.693488 12.925798 2.417571 0.007812243
#> 2  9.693488 12.925798 2.417571 0.007812243
#> 3  9.693488 12.925798 2.417571 0.007812243
#> 4  7.165912  8.848039 2.229319 0.012896348
#> 5  6.184479 10.432099 2.141295 0.016125131
#> 6  6.184479 10.432099 2.141295 0.016125131
#> 7  6.184479 10.432099 2.141295 0.016125131
#> 8  6.184479 10.432099 2.141295 0.016125131
#> 9  4.332928  6.274510 1.897940 0.028852007
#> 10 3.260868  4.201389 1.651957 0.049271708
#> 11 3.260868  4.201389 1.651957 0.049271708
#> 12 3.260868  4.201389 1.651957 0.049271708
#> 13 2.602663  3.391003 1.625352 0.052043772
#> 14 1.659725  1.985294 1.309394 0.095200451

textstat_collocations(toks, method = "lr", size = 2:3, min_count = 1)

#>                   collocation count length     lambda    sigma      dice
#> 1           quantitative text     2      2  5.1059455 2.112015 0.8333333
#> 2             rapidly growing     2      2  5.1059455 2.112015 0.8333333
#> 3               text analysis     2      2  5.1059455 2.112015 0.8333333
#> 4                        is a     2      2  3.9448128 1.769515 0.7142857
#> 5                 package for     1      2  4.6539604 2.173433 0.7500000
#> 6              The population     1      2  4.6539604 2.173433 0.7500000
#> 7               growing field     1      2  4.6539604 2.173433 0.7500000
#> 8            for quantitative     1      2  4.6539604 2.173433 0.7500000
#> 9                   a package     1      2  3.4965076 1.842265 0.6000000
#> 10                analysis is     1      2  2.9231616 1.769515 0.5000000
#> 11                quanteda is     1      2  2.9231616 1.769515 0.5000000
#> 12              population is     1      2  2.9231616 1.769515 0.5000000
#> 13                  a rapidly     1      2  2.3353749 1.436842 0.5000000
#> 14                 is rapidly     1      2  1.7578579 1.342497 0.4285714
#> 15      population is rapidly     1      3  0.6711683 3.133734 0.3750000
#> 16               is a rapidly     1      3 -0.9382696 2.958652 0.3461538
#> 17          The population is     1      3 -1.0986123 3.377485 0.4500000
#> 18              a package for     1      3 -1.0986123 3.377485 0.4500000
#> 19           text analysis is     1      3 -1.0986123 3.377485 0.4500000
#> 20 quantitative text analysis     2      3 -1.6863990 3.531866 0.6250000
#> 21         is rapidly growing     1      3 -1.5260563 3.133734 0.3461538
#> 22   package for quantitative     1      3 -2.2686835 3.568702 0.5000000
#> 23      rapidly growing field     1      3 -2.2686835 3.568702 0.5000000
#> 24      for quantitative text     1      3 -2.2686835 3.568702 0.5000000
#> 25               is a package     1      3 -2.1202635 3.174902 0.3750000
#> 26              analysis is a     1      3 -2.1202635 3.174902 0.3750000
#> 27              quanteda is a     1      3 -2.1202635 3.174902 0.3750000
#> 28          a rapidly growing     1      3 -2.1202635 3.174902 0.3750000
#>          pmi       G2      chi2          z           p
#> 1  1.7147984 9.693488 12.925798  2.4175710 0.007812243
#> 2  1.7147984 9.693488 12.925798  2.4175710 0.007812243
#> 3  1.7147984 9.693488 12.925798  2.4175710 0.007812243
#> 4  1.4271164 7.165912  8.848039  2.2293189 0.012896348
#> 5  2.0149030 6.184479 10.432099  2.1412949 0.016125131
#> 6  2.0149030 6.184479 10.432099  2.1412949 0.016125131
#> 7  2.0149030 6.184479 10.432099  2.1412949 0.016125131
#> 8  2.0149030 6.184479 10.432099  2.1412949 0.016125131
#> 9  1.6094379 4.332928  6.274510  1.8979398 0.028852007
#> 10 1.3217558 3.260868  4.201389  1.6519565 0.049271708
#> 11 1.3217558 3.260868  4.201389  1.6519565 0.049271708
#> 12 1.3217558 3.260868  4.201389  1.6519565 0.049271708
#> 13 1.2039728 2.602663  3.391003  1.6253522 0.052043772
#> 14 0.9162907 1.659725  1.985294  1.3093941 0.095200451
#> 15 1.4469190 3.842145  5.130357  0.2141752 0.415205201
#> 16 1.0414539 2.151098  2.623724 -0.3171274 0.624426525
#> 17 1.4469190 3.842145  5.130357 -0.3252752 0.627513607
#> 18 1.8523841 5.681476  8.731389 -0.3252752 0.627513607
#> 19 1.4469190 3.842145  5.130357 -0.3252752 0.627513607
#> 20 1.5522795 8.826477 10.815334 -0.4774811 0.683490204
#> 21 1.4469190 3.842145  5.130357 -0.4869769 0.686862667
#> 22 1.8523841 5.681476  8.731389 -0.6357168 0.737519477
#> 23 1.8523841 5.681476  8.731389 -0.6357168 0.737519477
#> 24 1.8523841 5.681476  8.731389 -0.6357168 0.737519477
#> 25 1.4469190 3.842145  5.130357 -0.6678202 0.747875828
#> 26 1.4469190 3.842145  5.130357 -0.6678202 0.747875828
#> 27 1.4469190 3.842145  5.130357 -0.6678202 0.747875828
#> 28 1.4469190 3.842145  5.130357 -0.6678202 0.747875828

(cols <- textstat_collocations(toks, method = "lr", size = 2:3, min_count = 2))

#>                  collocation count length    lambda    sigma      dice      pmi
#> 1          quantitative text     2      2  5.105945 2.112015 0.8333333 1.714798
#> 2            rapidly growing     2      2  5.105945 2.112015 0.8333333 1.714798
#> 3              text analysis     2      2  5.105945 2.112015 0.8333333 1.714798
#> 4                       is a     2      2  3.944813 1.769515 0.7142857 1.427116
#> 5 quantitative text analysis     2      3 -1.686399 3.531866 0.6250000 1.552279
#>         G2      chi2          z           p
#> 1 9.693488 12.925798  2.4175710 0.007812243
#> 2 9.693488 12.925798  2.4175710 0.007812243
#> 3 9.693488 12.925798  2.4175710 0.007812243
#> 4 7.165912  8.848039  2.2293189 0.012896348
#> 5 8.826477 10.815334 -0.4774811 0.683490204

# extracting multi-part proper nouns (capitalized terms)
toks2 <- tokens(corpus_segment(data_corpus_inaugural, what = "sentence"))
toks2 <- tokens_select(toks2, stopwords("english"), "remove", padding = TRUE)
toks2 <- tokens_select(toks2, "^([A-Z][a-z\\-]{2,})", valuetype="regex",
                     case_insensitive = FALSE, padding = TRUE)
seqs <- textstat_collocations(toks2, method = "lambda")
head(seqs, 10)

#>            collocation count length   lambda     sigma      dice      pmi
#> 1        United States   157      2 6.218998 0.4289252 0.9319527 1.194925
#> 2   Federal Government    32      2 3.635310 0.3816227 0.5652174 1.829632
#> 3         Almighty God    15      2 6.677447 0.8297427 0.8157895 3.249998
#> 4        Chief Justice    13      2 5.428468 0.7415552 0.6279070 2.929526
#> 5        South America     6      2 6.927623 1.1055370 0.7647059 3.990398
#> 6     Chief Magistrate    10      2 5.489063 0.9078897 0.5384615 2.965893
#> 7        Supreme Court     4      2 6.563291 1.1360380 0.6923077 4.161669
#> 8    Divine Providence     3      2 5.801151 1.0427885 0.5833333 4.064506
#> 9  National Government    11      2 2.741278 0.5094379 0.2584270 1.684557
#> 10           Old World    10      2 8.162516 1.5435878 0.8400000 3.709684
#>           G2      chi2         z            p
#> 1  522.92449 490.85679 14.499029 0.000000e+00
#> 2  105.77793 174.77489  9.525927 0.000000e+00
#> 3  108.08883 393.94665  8.047612 4.440892e-16
#> 4   77.91105 243.61956  7.320383 1.236788e-13
#> 5   53.39048 348.43563  6.266297 1.848673e-10
#> 6   60.69400 195.94825  6.045958 7.426236e-10
#> 7   37.12681 286.04826  5.777352 3.794277e-09
#> 8   26.69744 200.89086  5.563114 1.325013e-08
#> 9   28.64978  48.09379  5.380986 3.703957e-08
#> 10  85.92909 425.66979  5.288016 6.182522e-08

# compounding tokens is more efficient when applied to the same tokens object 
toks_comp <- tokens_compound(toks2, seqs)