Calculate keyness statistics

Calculate "keyness", a score for features that occur differentially across different categories. Here, the categories are defined by reference to a "target" document index in the dfm, with the reference group consisting of all other documents.

textstat_keyness(
  x,
  target = 1L,
  measure = c("chi2", "exact", "lr", "pmi"),
  sort = TRUE,
  correction = c("default", "yates", "williams", "none")
)

Arguments

x
a dfm containing the features to be examined for keyness

target
the document index (numeric, character or logical) identifying the document forming the "target" for computing keyness; all other documents' feature frequencies will be combined for use as a reference

measure
(signed) association measure to be used for computing keyness. Currently available: "chi2"; "exact" (Fisher's exact test); "lr" for the likelihood ratio; "pmi" for pointwise mutual information.

sort
logical; if TRUE sort features scored in descending order of the measure, otherwise leave in original feature order

correction
if "default", Yates correction is applied to "chi2"; William's correction is applied to "lr"; and no correction is applied for the "exact" and "pmi" measures. Specifying a value other than the default can be used to override the defaults, for instance to apply the Williams correction to the chi2 measure. Specifying a correction for the "exact" and "pmi" measures has no effect and produces a warning.

Value

a data.frame of computed statistics and associated p-values, where the features scored name each row, and the number of occurrences for both the target and reference groups. For measure = "chi2" this is the chi-squared value, signed positively if the observed value in the target exceeds its expected value; for measure = "exact" this is the estimate of the odds ratio; for measure = "lr" this is the likelihood ratio \(G2\) statistic; for "pmi" this is the pointwise mutual information statistics.

textstat_keyness returns a data.frame of features and their keyness scores and frequency counts.

References

Bondi, M. & Scott, M. (eds) (2010). Keyness in Texts. Amsterdam, Philadelphia: John Benjamins.

Stubbs, M. (2010). Three Concepts of Keywords. In Keyness in Texts, Bondi, M. & Scott, M. (eds): 1--42. Amsterdam, Philadelphia: John Benjamins.

Scott, M. & Tribble, C. (2006). Textual Patterns: Keyword and Corpus Analysis in Language Education. Amsterdam: Benjamins: 55.

Dunning, T. (1993). Accurate Methods for the Statistics of Surprise and Coincidence. Computational Linguistics, 19(1): 61--74.

Examples

# compare pre- v. post-war terms using grouping
period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
dfmat1 <- dfm(data_corpus_inaugural, groups = period)
head(dfmat1) # make sure 'post-war' is in the first row

#> Document-feature matrix of: 2 documents, 9,399 features (35.0% sparse).
#>           features
#> docs       fellow-citizens   of  the senate  and house representatives   :
#>   post-war               0 1437 1988      2 1455     3               0 105
#>   pre-war               39 5665 8093     13 3851     8              19  29
#>           features
#> docs       among vicissitudes
#>   post-war    22            0
#>   pre-war     86            5
#> [ reached max_nfeat ... 9,389 more features ]

head(tstat1 <- textstat_keyness(dfmat1), 10)

#>      feature     chi2 p n_target n_reference
#> 1         we 707.0479 0      960         779
#> 2          . 230.7676 0     1804        3141
#> 3         -- 197.5479 0      199         123
#> 4        our 190.8214 0      874        1306
#> 5         us 187.3596 0      262         216
#> 6          : 178.9679 0      105          29
#> 7    america 177.5366 0      130          54
#> 8      world 176.2401 0      188         123
#> 9  americans 148.6186 0       66           7
#> 10       new 142.2551 0      150          97

tail(tstat1, 10)

#>           feature       chi2            p n_target n_reference
#> 9390           it  -51.27459 8.031353e-13      257        1129
#> 9391         upon  -51.53143 7.045475e-13       39         332
#> 9392       public  -55.70987 8.404388e-14       11         213
#> 9393       states  -58.75698 1.787459e-14       28         305
#> 9394 constitution  -60.89244 5.995204e-15        6         200
#> 9395           be  -71.23360 0.000000e+00      257        1224
#> 9396       should  -82.69550 0.000000e+00       15         309
#> 9397        which -158.73301 0.000000e+00       95         910
#> 9398           of -175.43330 0.000000e+00     1437        5665
#> 9399          the -294.16169 0.000000e+00     1988        8093

# compare pre- v. post-war terms using logical vector
dfmat2 <- dfm(data_corpus_inaugural)
head(textstat_keyness(dfmat2, docvars(data_corpus_inaugural, "Year") >= 1945), 10)

#>      feature     chi2 p n_target n_reference
#> 1         we 707.0479 0      960         779
#> 2          . 230.7676 0     1804        3141
#> 3         -- 197.5479 0      199         123
#> 4        our 190.8214 0      874        1306
#> 5         us 187.3596 0      262         216
#> 6          : 178.9679 0      105          29
#> 7    america 177.5366 0      130          54
#> 8      world 176.2401 0      188         123
#> 9  americans 148.6186 0       66           7
#> 10       new 142.2551 0      150          97

# compare Trump 2017 to other post-war preseidents
dfmat3 <- dfm(corpus_subset(data_corpus_inaugural, period == "post-war"))
head(textstat_keyness(dfmat3, target = "2017-Trump"), 10)

#>      feature     chi2            p n_target n_reference
#> 1  protected 76.16493 0.000000e+00        5           1
#> 2          - 69.35184 1.110223e-16       11          20
#> 3       will 50.84346 1.000311e-12       40         299
#> 4      while 47.90052 4.484080e-12        6           7
#> 5      obama 47.56110 5.331624e-12        3           0
#> 6      we've 47.56110 5.331624e-12        3           0
#> 7    america 31.12584 2.418301e-08       18         112
#> 8      again 28.59131 8.938220e-08        9          32
#> 9   everyone 27.48612 1.582257e-07        4           5
#> 10      your 26.43285 2.728609e-07       11          50

# using the likelihood ratio method
head(textstat_keyness(dfm_smooth(dfmat3), measure = "lr", target = "2017-Trump"), 10)

#>      feature        G2            p n_target n_reference
#> 1       will 24.503820 7.416265e-07       41         317
#> 2          - 19.342361 1.092554e-05       12          38
#> 3    america 13.990375 1.837489e-04       19         130
#> 4       your 10.402855 1.258206e-03       12          68
#> 5      again  9.972178 1.589234e-03       10          50
#> 6      while  9.484004 2.072712e-03        7          25
#> 7   american  8.847068 2.935601e-03       12          76
#> 8  protected  8.802374 3.008388e-03        6          19
#> 9       back  6.834410 8.941822e-03        7          34
#> 10       you  6.681709 9.740719e-03       14         121