Calculate readability

Calculate the readability of text(s) using one of a variety of computed indexes.

textstat_readability(x, measure = c("all", "ARI", "ARI.simple",
  "Bormuth", "Bormuth.GP", "Coleman", "Coleman.C2", "Coleman.Liau",
  "Coleman.Liau.grade", "Coleman.Liau.short", "Dale.Chall",
  "Dale.Chall.old", "Dale.Chall.PSK", "Danielson.Bryan",
  "Danielson.Bryan.2", "Dickes.Steiwer", "DRP", "ELF",
  "Farr.Jenkins.Paterson", "Flesch", "Flesch.PSK", "Flesch.Kincaid", "FOG",
  "FOG.PSK", "FOG.NRI", "FORCAST", "FORCAST.RGL", "Fucks", "Linsear.Write",
  "LIW", "nWS", "nWS.2", "nWS.3", "nWS.4", "RIX", "Scrabble", "SMOG",
  "SMOG.C", "SMOG.simple",      "SMOG.de", "Spache", "Spache.old",
  "Strain", "Traenkle.Bailer", "Traenkle.Bailer.2", "Wheeler.Smith",
  "meanSentenceLength", "meanWordSyllables"), remove_hyphens = TRUE,
  min_sentence_length = 1, max_sentence_length = 10000, ...)

Arguments

x
a character or corpus object containing the texts

measure
character vector defining the readability measure to calculate. Matches are case-insensitive.

remove_hyphens
if TRUE, treat constituent words in hyphenated as separate terms, for purposes of computing word lengths, e.g. "decision-making" as two terms of lengths 8 and 6 characters respectively, rather than as a single word of 15 characters

min_sentence_length, max_sentence_length
set the minimum and maximum sentence lengths (in tokens, excluding punctuation) to include in the computation of readability. This makes it easy to exclude "sentences" that may not really be sentences, such as section titles, table elements, and other cruft that might be in the texts following conversion.

For finer-grained control, consider filtering sentences prior first, including through pattern-matching, using corpus_trim.

...
not used

Value

textstat_readability returns a data.frame of documents and their readability scores.

Examples

txt <- c("Readability zero one.  Ten, Eleven.", "The cat in a dilapidated tophat.")
textstat_readability(txt, "Flesch.Kincaid")

#>   document Flesch.Kincaid
#> 1    text1      13.705000
#> 2    text2       8.383333

textstat_readability(txt, c("FOG", "FOG.PSK", "FOG.NRI"))

#>   document       FOG  FOG.PSK FOG.NRI
#> 1    text1 17.000000 4.608659 -1.3875
#> 2    text2  9.066667 3.254382 -1.2600

inaugReadability <- textstat_readability(data_corpus_inaugural, "all")
cor(inaugReadability[,-1])

#>                               ARI   ARI.simple    Bormuth  Bormuth.GP
#> ARI                    1.00000000  0.999985813 -0.9918403  0.90528275
#> ARI.simple             0.99998581  1.000000000 -0.9925049  0.90658337
#> Bormuth               -0.99184030 -0.992504914  1.0000000 -0.92940312
#> Bormuth.GP             0.90528275  0.906583371 -0.9294031  1.00000000
#> Coleman               -0.57842561 -0.574425719  0.4781632 -0.34486865
#> Coleman.C2            -0.76781514 -0.764705749  0.6875936 -0.52057935
#> Coleman.Liau          -0.80785899 -0.804736665  0.7269951 -0.56325168
#> Coleman.Liau.grade     0.80785899  0.804736665 -0.7269951  0.56325168
#> Coleman.Liau.short     0.80788731  0.804765212 -0.7270287  0.56327984
#> Dale.Chall            -0.98133172 -0.980599105  0.9562768 -0.84252184
#> Dale.Chall.old         0.92539614  0.923745184 -0.8789003  0.74156962
#> Dale.Chall.PSK         0.96402587  0.962928175 -0.9304090  0.80682534
#> Danielson.Bryan        0.99708146  0.996677212 -0.9795650  0.88789861
#> Danielson.Bryan.2     -0.31030344 -0.305238155  0.1867526 -0.04281093
#> Dickes.Steiwer        -0.99844204 -0.998703929  0.9968931 -0.91616350
#> DRP                    0.99184030  0.992504914 -1.0000000  0.92940312
#> ELF                    0.99594813  0.995835853 -0.9853675  0.90129404
#> Farr.Jenkins.Paterson -0.99154201 -0.992219215  0.9999862 -0.92965691
#> Flesch                -0.96831214 -0.967086924  0.9315157 -0.82575026
#> Flesch.PSK             0.98543988  0.984630595 -0.9584409  0.85941333
#> Flesch.Kincaid         0.99828644  0.998096618 -0.9859779  0.89805311
#>                          Coleman Coleman.C2 Coleman.Liau Coleman.Liau.grade
#> ARI                   -0.5784256 -0.7678151   -0.8078590          0.8078590
#> ARI.simple            -0.5744257 -0.7647057   -0.8047367          0.8047367
#> Bormuth                0.4781632  0.6875936    0.7269951         -0.7269951
#> Bormuth.GP            -0.3448687 -0.5205794   -0.5632517          0.5632517
#> Coleman                1.0000000  0.9578621    0.9027073         -0.9027073
#> Coleman.C2             0.9578621  1.0000000    0.9676044         -0.9676044
#> Coleman.Liau           0.9027073  0.9676044    1.0000000         -1.0000000
#> Coleman.Liau.grade    -0.9027073 -0.9676044   -1.0000000          1.0000000
#> Coleman.Liau.short    -0.9026909 -0.9676038   -1.0000000          1.0000000
#> Dale.Chall             0.6688835  0.8384839    0.8758856         -0.8758856
#> Dale.Chall.old        -0.7659290 -0.8999305   -0.9341311          0.9341311
#> Dale.Chall.PSK        -0.7103061 -0.8667060   -0.9030272          0.9030272
#> Danielson.Bryan       -0.6320177 -0.8068048   -0.8471402          0.8471402
#> Danielson.Bryan.2      0.8922077  0.7931428    0.8066665         -0.8066665
#> Dickes.Steiwer         0.5390065  0.7368282    0.7764052         -0.7764052
#> DRP                   -0.4781632 -0.6875936   -0.7269951          0.7269951
#> ELF                   -0.6127006 -0.7903882   -0.8134246          0.8134246
#> Farr.Jenkins.Paterson  0.4768093  0.6862258    0.7252233         -0.7252233
#> Flesch                 0.7524555  0.8921701    0.9145879         -0.9145879
#> Flesch.PSK            -0.6991742 -0.8564052   -0.8829007          0.8829007
#> Flesch.Kincaid        -0.6124186 -0.7933597   -0.8255002          0.8255002
#>                       Coleman.Liau.short  Dale.Chall Dale.Chall.old
#> ARI                            0.8078873 -0.98133172     0.92539614
#> ARI.simple                     0.8047652 -0.98059910     0.92374518
#> Bormuth                       -0.7270287  0.95627679    -0.87890030
#> Bormuth.GP                     0.5632798 -0.84252184     0.74156962
#> Coleman                       -0.9026909  0.66888351    -0.76592897
#> Coleman.C2                    -0.9676038  0.83848386    -0.89993050
#> Coleman.Liau                  -1.0000000  0.87588559    -0.93413115
#> Coleman.Liau.grade             1.0000000 -0.87588559     0.93413115
#> Coleman.Liau.short             1.0000000 -0.87590905     0.93414653
#> Dale.Chall                    -0.8759090  1.00000000    -0.97997540
#> Dale.Chall.old                 0.9341465 -0.97997540     1.00000000
#> Dale.Chall.PSK                 0.9030476 -0.99692238     0.99256930
#> Danielson.Bryan                0.8471650 -0.98783325     0.94438014
#> Danielson.Bryan.2             -0.8066351  0.43323704    -0.58015544
#> Dickes.Steiwer                -0.7764357  0.97294240    -0.90823264
#> DRP                            0.7270287 -0.95627679     0.87890030
#> ELF                            0.8134514 -0.98005641     0.92765949
#> Farr.Jenkins.Paterson         -0.7252569  0.95552255    -0.87766706
#> Flesch                        -0.9146054  0.98219047    -0.96805512
#> Flesch.PSK                     0.8829221 -0.98818840     0.95944764
#> Flesch.Kincaid                 0.8255270 -0.98469390     0.93491117
#>                       Dale.Chall.PSK Danielson.Bryan Danielson.Bryan.2
#> ARI                       0.96402587      0.99708146       -0.31030344
#> ARI.simple                0.96292817      0.99667721       -0.30523815
#> Bormuth                  -0.93040902     -0.97956505        0.18675264
#> Bormuth.GP                0.80682534      0.88789861       -0.04281093
#> Coleman                  -0.71030609     -0.63201768        0.89220768
#> Coleman.C2               -0.86670597     -0.80680483        0.79314280
#> Coleman.Liau             -0.90302717     -0.84714024        0.80666655
#> Coleman.Liau.grade        0.90302717      0.84714024       -0.80666655
#> Coleman.Liau.short        0.90304756      0.84716497       -0.80663507
#> Dale.Chall               -0.99692238     -0.98783325        0.43323704
#> Dale.Chall.old            0.99256930      0.94438014       -0.58015544
#> Dale.Chall.PSK            1.00000000      0.97547312       -0.49316250
#> Danielson.Bryan           0.97547312      1.00000000       -0.37865977
#> Danielson.Bryan.2        -0.49316250     -0.37865977        1.00000000
#> Dickes.Steiwer           -0.95214175     -0.99178586        0.26101392
#> DRP                       0.93040902      0.97956505       -0.18675264
#> ELF                       0.96413764      0.99535015       -0.32598561
#> Farr.Jenkins.Paterson    -0.92946257     -0.97914589        0.18438177
#> Flesch                   -0.98134591     -0.98231683        0.51737927
#> Flesch.PSK                0.98162237      0.99363576       -0.44828841
#> Flesch.Kincaid            0.96982665      0.99786963       -0.34148774
#>                       Dickes.Steiwer        DRP         ELF
#> ARI                      -0.99844204  0.9918403  0.99594813
#> ARI.simple               -0.99870393  0.9925049  0.99583585
#> Bormuth                   0.99689308 -1.0000000 -0.98536748
#> Bormuth.GP               -0.91616350  0.9294031  0.90129404
#> Coleman                   0.53900648 -0.4781632 -0.61270055
#> Coleman.C2                0.73682822 -0.6875936 -0.79038825
#> Coleman.Liau              0.77640516 -0.7269951 -0.81342458
#> Coleman.Liau.grade       -0.77640516  0.7269951  0.81342458
#> Coleman.Liau.short       -0.77643567  0.7270287  0.81345135
#> Dale.Chall                0.97294240 -0.9562768 -0.98005641
#> Dale.Chall.old           -0.90823264  0.8789003  0.92765949
#> Dale.Chall.PSK           -0.95214175  0.9304090  0.96413764
#> Danielson.Bryan          -0.99178586  0.9795650  0.99535015
#> Danielson.Bryan.2         0.26101392 -0.1867526 -0.32598561
#> Dickes.Steiwer            1.00000000 -0.9968931 -0.99345440
#> DRP                      -0.99689308  1.0000000  0.98536748
#> ELF                      -0.99345440  0.9853675  1.00000000
#> Farr.Jenkins.Paterson     0.99672456 -0.9999862 -0.98524897
#> Flesch                    0.95524130 -0.9315157 -0.97375531
#> Flesch.PSK               -0.97634394  0.9584409  0.98827872
#> Flesch.Kincaid           -0.99512741  0.9859779  0.99727425
#>                       Farr.Jenkins.Paterson     Flesch  Flesch.PSK
#> ARI                             -0.99154201 -0.9683121  0.98543988
#> ARI.simple                      -0.99221922 -0.9670869  0.98463060
#> Bormuth                          0.99998621  0.9315157 -0.95844094
#> Bormuth.GP                      -0.92965691 -0.8257503  0.85941333
#> Coleman                          0.47680930  0.7524555 -0.69917416
#> Coleman.C2                       0.68622584  0.8921701 -0.85640522
#> Coleman.Liau                     0.72522331  0.9145879 -0.88290067
#> Coleman.Liau.grade              -0.72522331 -0.9145879  0.88290067
#> Coleman.Liau.short              -0.72525694 -0.9146054  0.88292206
#> Dale.Chall                       0.95552255  0.9821905 -0.98818840
#> Dale.Chall.old                  -0.87766706 -0.9680551  0.95944764
#> Dale.Chall.PSK                  -0.92946257 -0.9813459  0.98162237
#> Danielson.Bryan                 -0.97914589 -0.9823168  0.99363576
#> Danielson.Bryan.2                0.18438177  0.5173793 -0.44828841
#> Dickes.Steiwer                   0.99672456  0.9552413 -0.97634394
#> DRP                             -0.99998621 -0.9315157  0.95844094
#> ELF                             -0.98524897 -0.9737553  0.98827872
#> Farr.Jenkins.Paterson            1.00000000  0.9307176 -0.95781641
#> Flesch                           0.93071764  1.0000000 -0.99656281
#> Flesch.PSK                      -0.95781641 -0.9965628  1.00000000
#> Flesch.Kincaid                  -0.98561541 -0.9791403  0.99260681
#>                       Flesch.Kincaid         FOG    FOG.PSK    FOG.NRI
#> ARI                       0.99828644  0.99534334  0.9980495  0.4905310
#> ARI.simple                0.99809662  0.99501238  0.9983218  0.4892834
#> Bormuth                  -0.98597795 -0.97967399 -0.9967705 -0.4570760
#> Bormuth.GP                0.89805311  0.88578683  0.9165208  0.3389056
#> Coleman                  -0.61241856 -0.62791369 -0.5401789 -0.4679026
#> Coleman.C2               -0.79335974 -0.80512019 -0.7376615 -0.5350724
#> Coleman.Liau             -0.82550024 -0.83906193 -0.7749720 -0.5426918
#> Coleman.Liau.grade        0.82550024  0.83906193  0.7749720  0.5426918
#> Coleman.Liau.short        0.82552702  0.83908765  0.7750026  0.5426992
#> Dale.Chall               -0.98469390 -0.98559808 -0.9725009 -0.5221445
#> Dale.Chall.old            0.93491117  0.94068676  0.9076041  0.5406606
#> Dale.Chall.PSK            0.96982665  0.97265309  0.9516245  0.5319440
#> Danielson.Bryan           0.99786963  0.99695239  0.9912643  0.5028375
#> Danielson.Bryan.2        -0.34148774 -0.36653699 -0.2590925 -0.3671130
#> Dickes.Steiwer           -0.99512741 -0.99112313 -0.9994543 -0.4649778
#> DRP                       0.98597795  0.97967399  0.9967705  0.4570760
#> ELF                       0.99727425  0.99443432  0.9938624  0.4952706
#> Farr.Jenkins.Paterson    -0.98561541 -0.97924413 -0.9966030 -0.4562044
#> Flesch                   -0.97914032 -0.98344165 -0.9567470 -0.5313693
#> Flesch.PSK                0.99260681  0.99454929  0.9775024  0.5208520
#> Flesch.Kincaid            1.00000000  0.99846226  0.9957626  0.4989686
#>                          FORCAST FORCAST.RGL        Fucks Linsear.Write
#> ARI                    0.5784256   0.5784256  0.997730647    0.98324412
#> ARI.simple             0.5744257   0.5744257  0.998049346    0.98272675
#> Bormuth               -0.4781632  -0.4781632 -0.997494320   -0.96309746
#> Bormuth.GP             0.3448687   0.3448687  0.922666503    0.88512274
#> Coleman               -1.0000000  -1.0000000 -0.530595415   -0.65408730
#> Coleman.C2            -0.9578621  -0.9578621 -0.727965258   -0.81398206
#> Coleman.Liau          -0.9027073  -0.9027073 -0.768184772   -0.84460735
#> Coleman.Liau.grade     0.9027073   0.9027073  0.768184772    0.84460735
#> Coleman.Liau.short     0.9026909   0.9026909  0.768215251    0.84462952
#> Dale.Chall            -0.6688835  -0.6688835 -0.970055732   -0.97694426
#> Dale.Chall.old         0.7659290   0.7659290  0.903083345    0.93782490
#> Dale.Chall.PSK         0.7103061   0.7103061  0.948350389    0.96623803
#> Danielson.Bryan        0.6320177   0.6320177  0.990775706    0.98940268
#> Danielson.Bryan.2     -0.8922077  -0.8922077 -0.249745678   -0.39525523
#> Dickes.Steiwer        -0.5390065  -0.5390065 -0.999407002   -0.97750750
#> DRP                    0.4781632   0.4781632  0.997494320    0.96309746
#> ELF                    0.6127006   0.6127006  0.993623225    0.98844709
#> Farr.Jenkins.Paterson -0.4768093  -0.4768093 -0.997402917   -0.96273644
#> Flesch                -0.7524555  -0.7524555 -0.951964715   -0.98424870
#> Flesch.PSK             0.6991742   0.6991742  0.973922677    0.99144816
#> Flesch.Kincaid         0.6124186   0.6124186  0.993989409    0.98967358
#>                               LIW        nWS       nWS.2       nWS.3
#> ARI                    0.99534363  0.9716012  0.97550799  0.98501356
#> ARI.simple             0.99493616  0.9704382  0.97443531  0.98427247
#> Bormuth               -0.97775695 -0.9361760 -0.94221224 -0.95962677
#> Bormuth.GP             0.88322111  0.8251255  0.83249872  0.85720370
#> Coleman               -0.64151385 -0.7435480 -0.72971888 -0.68262459
#> Coleman.C2            -0.81448329 -0.8869623 -0.87759182 -0.84444185
#> Coleman.Liau          -0.84663781 -0.9108590 -0.90420615 -0.87578843
#> Coleman.Liau.grade     0.84663781  0.9108590  0.90420615  0.87578843
#> Coleman.Liau.short     0.84666264  0.9108773  0.90422529  0.87581059
#> Dale.Chall            -0.98837716 -0.9840797 -0.98570948 -0.98616948
#> Dale.Chall.old         0.94646298  0.9678892  0.96642656  0.95532169
#> Dale.Chall.PSK         0.97662553  0.9824351  0.98285518  0.97876419
#> Danielson.Bryan        0.99817496  0.9848656  0.98751970  0.99235519
#> Danielson.Bryan.2     -0.37953455 -0.5077365 -0.49287612 -0.43696629
#> Dickes.Steiwer        -0.99028645 -0.9594714 -0.96426773 -0.97697873
#> DRP                    0.97775695  0.9361760  0.94221224  0.95962677
#> ELF                    0.99591724  0.9767864  0.97966902  0.98635632
#> Farr.Jenkins.Paterson -0.97739468 -0.9354875 -0.94154405 -0.95902099
#> Flesch                -0.98395156 -0.9980699 -0.99747205 -0.99333954
#> Flesch.PSK             0.99452041  0.9961226  0.99702731  0.99774824
#> Flesch.Kincaid         0.99764432  0.9808885  0.98398342  0.99180645
#>                             nWS.4         RIX      Scrabble        SMOG
#> ARI                    0.99487148  0.99222014 -0.0112182516  0.97393941
#> ARI.simple             0.99451326  0.99191614 -0.0093650938  0.97295999
#> Bormuth               -0.97855446 -0.97697958 -0.0334415038 -0.94293454
#> Bormuth.GP             0.88408125  0.89405554 -0.0148915693  0.82457236
#> Coleman               -0.63167564 -0.62921060  0.3522407942 -0.71333148
#> Coleman.C2            -0.80789970 -0.79924879  0.2472067525 -0.86811865
#> Coleman.Liau          -0.84167664 -0.82960560  0.1918831342 -0.89524825
#> Coleman.Liau.grade     0.84167664  0.82960560 -0.1918831342  0.89524825
#> Coleman.Liau.short     0.84170213  0.82963012 -0.1918617971  0.89526894
#> Dale.Chall            -0.98584692 -0.98186614  0.0389953291 -0.98295534
#> Dale.Chall.old         0.94185846  0.93626682 -0.0867076637  0.96144728
#> Dale.Chall.PSK         0.97326647  0.96863234 -0.0579675641  0.97921174
#> Danielson.Bryan        0.99686096  0.99500484 -0.0367962483  0.98460337
#> Danielson.Bryan.2     -0.37128519 -0.35912065  0.3211432536 -0.47640589
#> Dickes.Steiwer        -0.99039744 -0.98825075 -0.0143478880 -0.96367176
#> DRP                    0.97855446  0.97697958  0.0334415038  0.94293454
#> ELF                    0.99411156  0.99677384 -0.0411360767  0.97747350
#> Farr.Jenkins.Paterson -0.97811285 -0.97683788 -0.0335901823 -0.94224804
#> Flesch                -0.98430842 -0.97797435  0.1056725636 -0.99453689
#> Flesch.PSK             0.99497427  0.98969704 -0.0753648675  0.99489006
#> Flesch.Kincaid         0.99823466  0.99457144 -0.0300226725  0.98304183
#>                            SMOG.C SMOG.simple     SMOG.de       Spache
#> ARI                    0.97457741  0.97393941  0.97393941  0.996573524
#> ARI.simple             0.97361454  0.97295999  0.97295999  0.996476926
#> Bormuth               -0.94395484 -0.94293454 -0.94293454 -0.986528797
#> Bormuth.GP             0.82681059  0.82457236  0.82457236  0.893133790
#> Coleman               -0.71131513 -0.71333148 -0.71333148 -0.592936253
#> Coleman.C2            -0.86636103 -0.86811865 -0.86811865 -0.780802409
#> Coleman.Liau          -0.89368168 -0.89524825 -0.89524825 -0.816010788
#> Coleman.Liau.grade     0.89368168  0.89524825  0.89524825  0.816010788
#> Coleman.Liau.short     0.89370243  0.89526894  0.89526894  0.816038974
#> Dale.Chall            -0.98305197 -0.98295534 -0.98295534 -0.987459093
#> Dale.Chall.old         0.96090393  0.96144728  0.96144728  0.938951188
#> Dale.Chall.PSK         0.97905687  0.97921174  0.97921174  0.973107042
#> Danielson.Bryan        0.98509384  0.98460337  0.98460337  0.994618356
#> Danielson.Bryan.2     -0.47358728 -0.47640589 -0.47640589 -0.324070196
#> Dickes.Steiwer        -0.96447120 -0.96367176 -0.96367176 -0.994355885
#> DRP                    0.94395484  0.94293454  0.94293454  0.986528797
#> ELF                    0.97820253  0.97747350  0.97747350  0.993741729
#> Farr.Jenkins.Paterson -0.94328383 -0.94224804 -0.94224804 -0.986160349
#> Flesch                -0.99442699 -0.99453689 -0.99453689 -0.968134928
#> Flesch.PSK             0.99503836  0.99489006  0.99489006  0.984074923
#> Flesch.Kincaid         0.98356657  0.98304183  0.98304183  0.995198218
#>                        Spache.old        Strain Traenkle.Bailer
#> ARI                    0.99724569  0.9976064239    -0.999223850
#> ARI.simple             0.99721960  0.9978635733    -0.999411499
#> Bormuth               -0.98887702 -0.9959353364     0.995903103
#> Bormuth.GP             0.89746461  0.9228647606    -0.913053802
#> Coleman               -0.58297612 -0.5461292792     0.549146837
#> Coleman.C2            -0.77303867 -0.7394040240     0.745171965
#> Coleman.Liau          -0.80865782 -0.7744192348     0.785328440
#> Coleman.Liau.grade     0.80865782  0.7744192348    -0.785328440
#> Coleman.Liau.short     0.80868653  0.7744490784    -0.785358433
#> Dale.Chall            -0.98568443 -0.9719760523     0.975968274
#> Dale.Chall.old         0.93445485  0.9073025337    -0.913852878
#> Dale.Chall.PSK         0.97025230  0.9511850254    -0.956203593
#> Danielson.Bryan        0.99433483  0.9916468644    -0.993491902
#> Danielson.Bryan.2     -0.31169025 -0.2608961180     0.274099704
#> Dickes.Steiwer        -0.99570107 -0.9987560895     0.999594489
#> DRP                    0.98887702  0.9959353364    -0.995903103
#> ELF                    0.99408659  0.9957646921    -0.994372531
#> Farr.Jenkins.Paterson -0.98854268 -0.9958154056     0.995687365
#> Flesch                -0.96580819 -0.9579641510     0.959086871
#> Flesch.PSK             0.98278698  0.9782782859    -0.979131224
#> Flesch.Kincaid         0.99544826  0.9958824210    -0.996329486
#>                       Traenkle.Bailer.2 Wheeler.Smith meanSentenceLength
#> ARI                          -0.7549178    0.99594813         0.99099423
#> ARI.simple                   -0.7514487    0.99583585         0.99169343
#> Bormuth                       0.6660785   -0.98536748        -0.99996928
#> Bormuth.GP                   -0.5182389    0.90129404         0.93014855
#> Coleman                       0.9206771   -0.61270055        -0.47287100
#> Coleman.C2                    0.9574552   -0.79038825        -0.68300831
#> Coleman.Liau                  0.9904793   -0.81342458        -0.72238083
#> Coleman.Liau.grade           -0.9904793    0.81342458         0.72238083
#> Coleman.Liau.short           -0.9904734    0.81345135         0.72241463
#> Dale.Chall                    0.8323812   -0.98005641        -0.95442718
#> Dale.Chall.old               -0.9046281    0.92765949         0.87588943
#> Dale.Chall.PSK               -0.8648262    0.96413764         0.92809333
#> Danielson.Bryan              -0.7999586    0.99535015         0.97829531
#> Danielson.Bryan.2             0.8509255   -0.32598561        -0.18028533
#> Dickes.Steiwer                0.7195290   -0.99345440        -0.99638999
#> DRP                          -0.6660785    0.98536748         0.99996928
#> ELF                          -0.7636349    1.00000000         0.98451149
#> Farr.Jenkins.Paterson         0.6642605   -0.98524897        -0.99998999
#> Flesch                        0.8800874   -0.97375531        -0.92913685
#> Flesch.PSK                   -0.8419885    0.98827872         0.95657238
#> Flesch.Kincaid               -0.7757025    0.99727425         0.98488025
#>                       meanWordSyllables
#> ARI                           0.6762172
#> ARI.simple                    0.6724762
#> Bormuth                      -0.5816275
#> Bormuth.GP                    0.4510043
#> Coleman                      -0.9645525
#> Coleman.C2                   -0.9629188
#> Coleman.Liau                 -0.9543023
#> Coleman.Liau.grade            0.9543023
#> Coleman.Liau.short            0.9542911
#> Dale.Chall                   -0.7609139
#> Dale.Chall.old                0.8456970
#> Dale.Chall.PSK                0.7979510
#> Danielson.Bryan               0.7259320
#> Danielson.Bryan.2            -0.8772134
#> Dickes.Steiwer               -0.6393556
#> DRP                           0.5816275
#> ELF                           0.6978253
#> Farr.Jenkins.Paterson        -0.5798330
#> Flesch                       -0.8376402
#> Flesch.PSK                    0.7895116
#> Flesch.Kincaid                0.7091829
#>  [ reached getOption("max.print") -- omitted 26 rows ]

textstat_readability(data_corpus_inaugural, measure = "Flesch.Kincaid")

#>           document Flesch.Kincaid
#> 1  1789-Washington      28.432155
#> 2  1793-Washington      17.151759
#> 3       1797-Adams      28.692087
#> 4   1801-Jefferson      19.295924
#> 5   1805-Jefferson      21.917088
#> 6     1809-Madison      25.557611
#> 7     1813-Madison      18.034452
#> 8      1817-Monroe      14.443950
#> 9      1821-Monroe      16.773058
#> 10      1825-Adams      19.748273
#> 11    1829-Jackson      21.893148
#> 12    1833-Jackson      19.958757
#> 13   1837-VanBuren      20.263524
#> 14   1841-Harrison      19.355855
#> 15       1845-Polk      16.406306
#> 16     1849-Taylor      23.748805
#> 17     1853-Pierce      16.438241
#> 18   1857-Buchanan      16.110868
#> 19    1861-Lincoln      13.661936
#> 20    1865-Lincoln      11.774929
#> 21      1869-Grant      14.008426
#> 22      1873-Grant      15.082444
#> 23      1877-Hayes      20.297851
#> 24   1881-Garfield      14.042973
#> 25  1885-Cleveland      19.052937
#> 26   1889-Harrison      14.944583
#> 27  1893-Cleveland      18.248838
#> 28   1897-McKinley      15.854298
#> 29   1901-McKinley      12.728640
#> 30  1905-Roosevelt      13.959254
#> 31       1909-Taft      17.099757
#> 32     1913-Wilson      11.707659
#> 33     1917-Wilson      11.780913
#> 34    1921-Harding      13.214469
#> 35   1925-Coolidge      11.626699
#> 36     1929-Hoover      13.259377
#> 37  1933-Roosevelt      11.640683
#> 38  1937-Roosevelt      10.619558
#> 39  1941-Roosevelt       9.687985
#> 40  1945-Roosevelt       9.175680
#> 41     1949-Truman      11.328244
#> 42 1953-Eisenhower      10.328852
#> 43 1957-Eisenhower       8.094908
#> 44    1961-Kennedy      11.822997
#> 45    1965-Johnson       7.550602
#> 46      1969-Nixon       9.235663
#> 47      1973-Nixon      12.288965
#> 48     1977-Carter      11.670742
#> 49     1981-Reagan       9.798608
#> 50     1985-Reagan      10.420294
#> 51       1989-Bush       7.147029
#> 52    1993-Clinton      10.374204
#> 53    1997-Clinton       9.828863
#> 54       2001-Bush       8.918201
#> 55       2005-Bush      11.036277
#> 56      2009-Obama      10.229426
#> 57      2013-Obama      11.734767
#> 58      2017-Trump       9.163084

inaugReadability <- textstat_readability(data_corpus_inaugural, "all")
cor(inaugReadability[,-1])

#>                               ARI   ARI.simple    Bormuth  Bormuth.GP
#> ARI                    1.00000000  0.999985813 -0.9918403  0.90528275
#> ARI.simple             0.99998581  1.000000000 -0.9925049  0.90658337
#> Bormuth               -0.99184030 -0.992504914  1.0000000 -0.92940312
#> Bormuth.GP             0.90528275  0.906583371 -0.9294031  1.00000000
#> Coleman               -0.57842561 -0.574425719  0.4781632 -0.34486865
#> Coleman.C2            -0.76781514 -0.764705749  0.6875936 -0.52057935
#> Coleman.Liau          -0.80785899 -0.804736665  0.7269951 -0.56325168
#> Coleman.Liau.grade     0.80785899  0.804736665 -0.7269951  0.56325168
#> Coleman.Liau.short     0.80788731  0.804765212 -0.7270287  0.56327984
#> Dale.Chall            -0.98133172 -0.980599105  0.9562768 -0.84252184
#> Dale.Chall.old         0.92539614  0.923745184 -0.8789003  0.74156962
#> Dale.Chall.PSK         0.96402587  0.962928175 -0.9304090  0.80682534
#> Danielson.Bryan        0.99708146  0.996677212 -0.9795650  0.88789861
#> Danielson.Bryan.2     -0.31030344 -0.305238155  0.1867526 -0.04281093
#> Dickes.Steiwer        -0.99844204 -0.998703929  0.9968931 -0.91616350
#> DRP                    0.99184030  0.992504914 -1.0000000  0.92940312
#> ELF                    0.99594813  0.995835853 -0.9853675  0.90129404
#> Farr.Jenkins.Paterson -0.99154201 -0.992219215  0.9999862 -0.92965691
#> Flesch                -0.96831214 -0.967086924  0.9315157 -0.82575026
#> Flesch.PSK             0.98543988  0.984630595 -0.9584409  0.85941333
#> Flesch.Kincaid         0.99828644  0.998096618 -0.9859779  0.89805311
#>                          Coleman Coleman.C2 Coleman.Liau Coleman.Liau.grade
#> ARI                   -0.5784256 -0.7678151   -0.8078590          0.8078590
#> ARI.simple            -0.5744257 -0.7647057   -0.8047367          0.8047367
#> Bormuth                0.4781632  0.6875936    0.7269951         -0.7269951
#> Bormuth.GP            -0.3448687 -0.5205794   -0.5632517          0.5632517
#> Coleman                1.0000000  0.9578621    0.9027073         -0.9027073
#> Coleman.C2             0.9578621  1.0000000    0.9676044         -0.9676044
#> Coleman.Liau           0.9027073  0.9676044    1.0000000         -1.0000000
#> Coleman.Liau.grade    -0.9027073 -0.9676044   -1.0000000          1.0000000
#> Coleman.Liau.short    -0.9026909 -0.9676038   -1.0000000          1.0000000
#> Dale.Chall             0.6688835  0.8384839    0.8758856         -0.8758856
#> Dale.Chall.old        -0.7659290 -0.8999305   -0.9341311          0.9341311
#> Dale.Chall.PSK        -0.7103061 -0.8667060   -0.9030272          0.9030272
#> Danielson.Bryan       -0.6320177 -0.8068048   -0.8471402          0.8471402
#> Danielson.Bryan.2      0.8922077  0.7931428    0.8066665         -0.8066665
#> Dickes.Steiwer         0.5390065  0.7368282    0.7764052         -0.7764052
#> DRP                   -0.4781632 -0.6875936   -0.7269951          0.7269951
#> ELF                   -0.6127006 -0.7903882   -0.8134246          0.8134246
#> Farr.Jenkins.Paterson  0.4768093  0.6862258    0.7252233         -0.7252233
#> Flesch                 0.7524555  0.8921701    0.9145879         -0.9145879
#> Flesch.PSK            -0.6991742 -0.8564052   -0.8829007          0.8829007
#> Flesch.Kincaid        -0.6124186 -0.7933597   -0.8255002          0.8255002
#>                       Coleman.Liau.short  Dale.Chall Dale.Chall.old
#> ARI                            0.8078873 -0.98133172     0.92539614
#> ARI.simple                     0.8047652 -0.98059910     0.92374518
#> Bormuth                       -0.7270287  0.95627679    -0.87890030
#> Bormuth.GP                     0.5632798 -0.84252184     0.74156962
#> Coleman                       -0.9026909  0.66888351    -0.76592897
#> Coleman.C2                    -0.9676038  0.83848386    -0.89993050
#> Coleman.Liau                  -1.0000000  0.87588559    -0.93413115
#> Coleman.Liau.grade             1.0000000 -0.87588559     0.93413115
#> Coleman.Liau.short             1.0000000 -0.87590905     0.93414653
#> Dale.Chall                    -0.8759090  1.00000000    -0.97997540
#> Dale.Chall.old                 0.9341465 -0.97997540     1.00000000
#> Dale.Chall.PSK                 0.9030476 -0.99692238     0.99256930
#> Danielson.Bryan                0.8471650 -0.98783325     0.94438014
#> Danielson.Bryan.2             -0.8066351  0.43323704    -0.58015544
#> Dickes.Steiwer                -0.7764357  0.97294240    -0.90823264
#> DRP                            0.7270287 -0.95627679     0.87890030
#> ELF                            0.8134514 -0.98005641     0.92765949
#> Farr.Jenkins.Paterson         -0.7252569  0.95552255    -0.87766706
#> Flesch                        -0.9146054  0.98219047    -0.96805512
#> Flesch.PSK                     0.8829221 -0.98818840     0.95944764
#> Flesch.Kincaid                 0.8255270 -0.98469390     0.93491117
#>                       Dale.Chall.PSK Danielson.Bryan Danielson.Bryan.2
#> ARI                       0.96402587      0.99708146       -0.31030344
#> ARI.simple                0.96292817      0.99667721       -0.30523815
#> Bormuth                  -0.93040902     -0.97956505        0.18675264
#> Bormuth.GP                0.80682534      0.88789861       -0.04281093
#> Coleman                  -0.71030609     -0.63201768        0.89220768
#> Coleman.C2               -0.86670597     -0.80680483        0.79314280
#> Coleman.Liau             -0.90302717     -0.84714024        0.80666655
#> Coleman.Liau.grade        0.90302717      0.84714024       -0.80666655
#> Coleman.Liau.short        0.90304756      0.84716497       -0.80663507
#> Dale.Chall               -0.99692238     -0.98783325        0.43323704
#> Dale.Chall.old            0.99256930      0.94438014       -0.58015544
#> Dale.Chall.PSK            1.00000000      0.97547312       -0.49316250
#> Danielson.Bryan           0.97547312      1.00000000       -0.37865977
#> Danielson.Bryan.2        -0.49316250     -0.37865977        1.00000000
#> Dickes.Steiwer           -0.95214175     -0.99178586        0.26101392
#> DRP                       0.93040902      0.97956505       -0.18675264
#> ELF                       0.96413764      0.99535015       -0.32598561
#> Farr.Jenkins.Paterson    -0.92946257     -0.97914589        0.18438177
#> Flesch                   -0.98134591     -0.98231683        0.51737927
#> Flesch.PSK                0.98162237      0.99363576       -0.44828841
#> Flesch.Kincaid            0.96982665      0.99786963       -0.34148774
#>                       Dickes.Steiwer        DRP         ELF
#> ARI                      -0.99844204  0.9918403  0.99594813
#> ARI.simple               -0.99870393  0.9925049  0.99583585
#> Bormuth                   0.99689308 -1.0000000 -0.98536748
#> Bormuth.GP               -0.91616350  0.9294031  0.90129404
#> Coleman                   0.53900648 -0.4781632 -0.61270055
#> Coleman.C2                0.73682822 -0.6875936 -0.79038825
#> Coleman.Liau              0.77640516 -0.7269951 -0.81342458
#> Coleman.Liau.grade       -0.77640516  0.7269951  0.81342458
#> Coleman.Liau.short       -0.77643567  0.7270287  0.81345135
#> Dale.Chall                0.97294240 -0.9562768 -0.98005641
#> Dale.Chall.old           -0.90823264  0.8789003  0.92765949
#> Dale.Chall.PSK           -0.95214175  0.9304090  0.96413764
#> Danielson.Bryan          -0.99178586  0.9795650  0.99535015
#> Danielson.Bryan.2         0.26101392 -0.1867526 -0.32598561
#> Dickes.Steiwer            1.00000000 -0.9968931 -0.99345440
#> DRP                      -0.99689308  1.0000000  0.98536748
#> ELF                      -0.99345440  0.9853675  1.00000000
#> Farr.Jenkins.Paterson     0.99672456 -0.9999862 -0.98524897
#> Flesch                    0.95524130 -0.9315157 -0.97375531
#> Flesch.PSK               -0.97634394  0.9584409  0.98827872
#> Flesch.Kincaid           -0.99512741  0.9859779  0.99727425
#>                       Farr.Jenkins.Paterson     Flesch  Flesch.PSK
#> ARI                             -0.99154201 -0.9683121  0.98543988
#> ARI.simple                      -0.99221922 -0.9670869  0.98463060
#> Bormuth                          0.99998621  0.9315157 -0.95844094
#> Bormuth.GP                      -0.92965691 -0.8257503  0.85941333
#> Coleman                          0.47680930  0.7524555 -0.69917416
#> Coleman.C2                       0.68622584  0.8921701 -0.85640522
#> Coleman.Liau                     0.72522331  0.9145879 -0.88290067
#> Coleman.Liau.grade              -0.72522331 -0.9145879  0.88290067
#> Coleman.Liau.short              -0.72525694 -0.9146054  0.88292206
#> Dale.Chall                       0.95552255  0.9821905 -0.98818840
#> Dale.Chall.old                  -0.87766706 -0.9680551  0.95944764
#> Dale.Chall.PSK                  -0.92946257 -0.9813459  0.98162237
#> Danielson.Bryan                 -0.97914589 -0.9823168  0.99363576
#> Danielson.Bryan.2                0.18438177  0.5173793 -0.44828841
#> Dickes.Steiwer                   0.99672456  0.9552413 -0.97634394
#> DRP                             -0.99998621 -0.9315157  0.95844094
#> ELF                             -0.98524897 -0.9737553  0.98827872
#> Farr.Jenkins.Paterson            1.00000000  0.9307176 -0.95781641
#> Flesch                           0.93071764  1.0000000 -0.99656281
#> Flesch.PSK                      -0.95781641 -0.9965628  1.00000000
#> Flesch.Kincaid                  -0.98561541 -0.9791403  0.99260681
#>                       Flesch.Kincaid         FOG    FOG.PSK    FOG.NRI
#> ARI                       0.99828644  0.99534334  0.9980495  0.4905310
#> ARI.simple                0.99809662  0.99501238  0.9983218  0.4892834
#> Bormuth                  -0.98597795 -0.97967399 -0.9967705 -0.4570760
#> Bormuth.GP                0.89805311  0.88578683  0.9165208  0.3389056
#> Coleman                  -0.61241856 -0.62791369 -0.5401789 -0.4679026
#> Coleman.C2               -0.79335974 -0.80512019 -0.7376615 -0.5350724
#> Coleman.Liau             -0.82550024 -0.83906193 -0.7749720 -0.5426918
#> Coleman.Liau.grade        0.82550024  0.83906193  0.7749720  0.5426918
#> Coleman.Liau.short        0.82552702  0.83908765  0.7750026  0.5426992
#> Dale.Chall               -0.98469390 -0.98559808 -0.9725009 -0.5221445
#> Dale.Chall.old            0.93491117  0.94068676  0.9076041  0.5406606
#> Dale.Chall.PSK            0.96982665  0.97265309  0.9516245  0.5319440
#> Danielson.Bryan           0.99786963  0.99695239  0.9912643  0.5028375
#> Danielson.Bryan.2        -0.34148774 -0.36653699 -0.2590925 -0.3671130
#> Dickes.Steiwer           -0.99512741 -0.99112313 -0.9994543 -0.4649778
#> DRP                       0.98597795  0.97967399  0.9967705  0.4570760
#> ELF                       0.99727425  0.99443432  0.9938624  0.4952706
#> Farr.Jenkins.Paterson    -0.98561541 -0.97924413 -0.9966030 -0.4562044
#> Flesch                   -0.97914032 -0.98344165 -0.9567470 -0.5313693
#> Flesch.PSK                0.99260681  0.99454929  0.9775024  0.5208520
#> Flesch.Kincaid            1.00000000  0.99846226  0.9957626  0.4989686
#>                          FORCAST FORCAST.RGL        Fucks Linsear.Write
#> ARI                    0.5784256   0.5784256  0.997730647    0.98324412
#> ARI.simple             0.5744257   0.5744257  0.998049346    0.98272675
#> Bormuth               -0.4781632  -0.4781632 -0.997494320   -0.96309746
#> Bormuth.GP             0.3448687   0.3448687  0.922666503    0.88512274
#> Coleman               -1.0000000  -1.0000000 -0.530595415   -0.65408730
#> Coleman.C2            -0.9578621  -0.9578621 -0.727965258   -0.81398206
#> Coleman.Liau          -0.9027073  -0.9027073 -0.768184772   -0.84460735
#> Coleman.Liau.grade     0.9027073   0.9027073  0.768184772    0.84460735
#> Coleman.Liau.short     0.9026909   0.9026909  0.768215251    0.84462952
#> Dale.Chall            -0.6688835  -0.6688835 -0.970055732   -0.97694426
#> Dale.Chall.old         0.7659290   0.7659290  0.903083345    0.93782490
#> Dale.Chall.PSK         0.7103061   0.7103061  0.948350389    0.96623803
#> Danielson.Bryan        0.6320177   0.6320177  0.990775706    0.98940268
#> Danielson.Bryan.2     -0.8922077  -0.8922077 -0.249745678   -0.39525523
#> Dickes.Steiwer        -0.5390065  -0.5390065 -0.999407002   -0.97750750
#> DRP                    0.4781632   0.4781632  0.997494320    0.96309746
#> ELF                    0.6127006   0.6127006  0.993623225    0.98844709
#> Farr.Jenkins.Paterson -0.4768093  -0.4768093 -0.997402917   -0.96273644
#> Flesch                -0.7524555  -0.7524555 -0.951964715   -0.98424870
#> Flesch.PSK             0.6991742   0.6991742  0.973922677    0.99144816
#> Flesch.Kincaid         0.6124186   0.6124186  0.993989409    0.98967358
#>                               LIW        nWS       nWS.2       nWS.3
#> ARI                    0.99534363  0.9716012  0.97550799  0.98501356
#> ARI.simple             0.99493616  0.9704382  0.97443531  0.98427247
#> Bormuth               -0.97775695 -0.9361760 -0.94221224 -0.95962677
#> Bormuth.GP             0.88322111  0.8251255  0.83249872  0.85720370
#> Coleman               -0.64151385 -0.7435480 -0.72971888 -0.68262459
#> Coleman.C2            -0.81448329 -0.8869623 -0.87759182 -0.84444185
#> Coleman.Liau          -0.84663781 -0.9108590 -0.90420615 -0.87578843
#> Coleman.Liau.grade     0.84663781  0.9108590  0.90420615  0.87578843
#> Coleman.Liau.short     0.84666264  0.9108773  0.90422529  0.87581059
#> Dale.Chall            -0.98837716 -0.9840797 -0.98570948 -0.98616948
#> Dale.Chall.old         0.94646298  0.9678892  0.96642656  0.95532169
#> Dale.Chall.PSK         0.97662553  0.9824351  0.98285518  0.97876419
#> Danielson.Bryan        0.99817496  0.9848656  0.98751970  0.99235519
#> Danielson.Bryan.2     -0.37953455 -0.5077365 -0.49287612 -0.43696629
#> Dickes.Steiwer        -0.99028645 -0.9594714 -0.96426773 -0.97697873
#> DRP                    0.97775695  0.9361760  0.94221224  0.95962677
#> ELF                    0.99591724  0.9767864  0.97966902  0.98635632
#> Farr.Jenkins.Paterson -0.97739468 -0.9354875 -0.94154405 -0.95902099
#> Flesch                -0.98395156 -0.9980699 -0.99747205 -0.99333954
#> Flesch.PSK             0.99452041  0.9961226  0.99702731  0.99774824
#> Flesch.Kincaid         0.99764432  0.9808885  0.98398342  0.99180645
#>                             nWS.4         RIX      Scrabble        SMOG
#> ARI                    0.99487148  0.99222014 -0.0112182516  0.97393941
#> ARI.simple             0.99451326  0.99191614 -0.0093650938  0.97295999
#> Bormuth               -0.97855446 -0.97697958 -0.0334415038 -0.94293454
#> Bormuth.GP             0.88408125  0.89405554 -0.0148915693  0.82457236
#> Coleman               -0.63167564 -0.62921060  0.3522407942 -0.71333148
#> Coleman.C2            -0.80789970 -0.79924879  0.2472067525 -0.86811865
#> Coleman.Liau          -0.84167664 -0.82960560  0.1918831342 -0.89524825
#> Coleman.Liau.grade     0.84167664  0.82960560 -0.1918831342  0.89524825
#> Coleman.Liau.short     0.84170213  0.82963012 -0.1918617971  0.89526894
#> Dale.Chall            -0.98584692 -0.98186614  0.0389953291 -0.98295534
#> Dale.Chall.old         0.94185846  0.93626682 -0.0867076637  0.96144728
#> Dale.Chall.PSK         0.97326647  0.96863234 -0.0579675641  0.97921174
#> Danielson.Bryan        0.99686096  0.99500484 -0.0367962483  0.98460337
#> Danielson.Bryan.2     -0.37128519 -0.35912065  0.3211432536 -0.47640589
#> Dickes.Steiwer        -0.99039744 -0.98825075 -0.0143478880 -0.96367176
#> DRP                    0.97855446  0.97697958  0.0334415038  0.94293454
#> ELF                    0.99411156  0.99677384 -0.0411360767  0.97747350
#> Farr.Jenkins.Paterson -0.97811285 -0.97683788 -0.0335901823 -0.94224804
#> Flesch                -0.98430842 -0.97797435  0.1056725636 -0.99453689
#> Flesch.PSK             0.99497427  0.98969704 -0.0753648675  0.99489006
#> Flesch.Kincaid         0.99823466  0.99457144 -0.0300226725  0.98304183
#>                            SMOG.C SMOG.simple     SMOG.de       Spache
#> ARI                    0.97457741  0.97393941  0.97393941  0.996573524
#> ARI.simple             0.97361454  0.97295999  0.97295999  0.996476926
#> Bormuth               -0.94395484 -0.94293454 -0.94293454 -0.986528797
#> Bormuth.GP             0.82681059  0.82457236  0.82457236  0.893133790
#> Coleman               -0.71131513 -0.71333148 -0.71333148 -0.592936253
#> Coleman.C2            -0.86636103 -0.86811865 -0.86811865 -0.780802409
#> Coleman.Liau          -0.89368168 -0.89524825 -0.89524825 -0.816010788
#> Coleman.Liau.grade     0.89368168  0.89524825  0.89524825  0.816010788
#> Coleman.Liau.short     0.89370243  0.89526894  0.89526894  0.816038974
#> Dale.Chall            -0.98305197 -0.98295534 -0.98295534 -0.987459093
#> Dale.Chall.old         0.96090393  0.96144728  0.96144728  0.938951188
#> Dale.Chall.PSK         0.97905687  0.97921174  0.97921174  0.973107042
#> Danielson.Bryan        0.98509384  0.98460337  0.98460337  0.994618356
#> Danielson.Bryan.2     -0.47358728 -0.47640589 -0.47640589 -0.324070196
#> Dickes.Steiwer        -0.96447120 -0.96367176 -0.96367176 -0.994355885
#> DRP                    0.94395484  0.94293454  0.94293454  0.986528797
#> ELF                    0.97820253  0.97747350  0.97747350  0.993741729
#> Farr.Jenkins.Paterson -0.94328383 -0.94224804 -0.94224804 -0.986160349
#> Flesch                -0.99442699 -0.99453689 -0.99453689 -0.968134928
#> Flesch.PSK             0.99503836  0.99489006  0.99489006  0.984074923
#> Flesch.Kincaid         0.98356657  0.98304183  0.98304183  0.995198218
#>                        Spache.old        Strain Traenkle.Bailer
#> ARI                    0.99724569  0.9976064239    -0.999223850
#> ARI.simple             0.99721960  0.9978635733    -0.999411499
#> Bormuth               -0.98887702 -0.9959353364     0.995903103
#> Bormuth.GP             0.89746461  0.9228647606    -0.913053802
#> Coleman               -0.58297612 -0.5461292792     0.549146837
#> Coleman.C2            -0.77303867 -0.7394040240     0.745171965
#> Coleman.Liau          -0.80865782 -0.7744192348     0.785328440
#> Coleman.Liau.grade     0.80865782  0.7744192348    -0.785328440
#> Coleman.Liau.short     0.80868653  0.7744490784    -0.785358433
#> Dale.Chall            -0.98568443 -0.9719760523     0.975968274
#> Dale.Chall.old         0.93445485  0.9073025337    -0.913852878
#> Dale.Chall.PSK         0.97025230  0.9511850254    -0.956203593
#> Danielson.Bryan        0.99433483  0.9916468644    -0.993491902
#> Danielson.Bryan.2     -0.31169025 -0.2608961180     0.274099704
#> Dickes.Steiwer        -0.99570107 -0.9987560895     0.999594489
#> DRP                    0.98887702  0.9959353364    -0.995903103
#> ELF                    0.99408659  0.9957646921    -0.994372531
#> Farr.Jenkins.Paterson -0.98854268 -0.9958154056     0.995687365
#> Flesch                -0.96580819 -0.9579641510     0.959086871
#> Flesch.PSK             0.98278698  0.9782782859    -0.979131224
#> Flesch.Kincaid         0.99544826  0.9958824210    -0.996329486
#>                       Traenkle.Bailer.2 Wheeler.Smith meanSentenceLength
#> ARI                          -0.7549178    0.99594813         0.99099423
#> ARI.simple                   -0.7514487    0.99583585         0.99169343
#> Bormuth                       0.6660785   -0.98536748        -0.99996928
#> Bormuth.GP                   -0.5182389    0.90129404         0.93014855
#> Coleman                       0.9206771   -0.61270055        -0.47287100
#> Coleman.C2                    0.9574552   -0.79038825        -0.68300831
#> Coleman.Liau                  0.9904793   -0.81342458        -0.72238083
#> Coleman.Liau.grade           -0.9904793    0.81342458         0.72238083
#> Coleman.Liau.short           -0.9904734    0.81345135         0.72241463
#> Dale.Chall                    0.8323812   -0.98005641        -0.95442718
#> Dale.Chall.old               -0.9046281    0.92765949         0.87588943
#> Dale.Chall.PSK               -0.8648262    0.96413764         0.92809333
#> Danielson.Bryan              -0.7999586    0.99535015         0.97829531
#> Danielson.Bryan.2             0.8509255   -0.32598561        -0.18028533
#> Dickes.Steiwer                0.7195290   -0.99345440        -0.99638999
#> DRP                          -0.6660785    0.98536748         0.99996928
#> ELF                          -0.7636349    1.00000000         0.98451149
#> Farr.Jenkins.Paterson         0.6642605   -0.98524897        -0.99998999
#> Flesch                        0.8800874   -0.97375531        -0.92913685
#> Flesch.PSK                   -0.8419885    0.98827872         0.95657238
#> Flesch.Kincaid               -0.7757025    0.99727425         0.98488025
#>                       meanWordSyllables
#> ARI                           0.6762172
#> ARI.simple                    0.6724762
#> Bormuth                      -0.5816275
#> Bormuth.GP                    0.4510043
#> Coleman                      -0.9645525
#> Coleman.C2                   -0.9629188
#> Coleman.Liau                 -0.9543023
#> Coleman.Liau.grade            0.9543023
#> Coleman.Liau.short            0.9542911
#> Dale.Chall                   -0.7609139
#> Dale.Chall.old                0.8456970
#> Dale.Chall.PSK                0.7979510
#> Danielson.Bryan               0.7259320
#> Danielson.Bryan.2            -0.8772134
#> Dickes.Steiwer               -0.6393556
#> DRP                           0.5816275
#> ELF                           0.6978253
#> Farr.Jenkins.Paterson        -0.5798330
#> Flesch                       -0.8376402
#> Flesch.PSK                    0.7895116
#> Flesch.Kincaid                0.7091829
#>  [ reached getOption("max.print") -- omitted 26 rows ]