pattern matching using valuetype

Pattern matching in quanteda using the valuetype argument.

Arguments

valuetype

how to interpret keyword expressions: "glob" for "glob"-style wildcard expressions; "regex" for regular expressions; or "fixed" for exact matching. See valuetype for details.

Details

Pattern matching in in quanteda uses "glob"-style pattern matching as the default, because this is simpler than regular expression matching while addressing most users' needs. It is also has the advantage of being identical to fixed pattern matching when the wildcard characters (`*` and `?`) are not used. Finally, most dictionary formats use glob matching.

"glob"

"glob"-style wildcard expressions, the quanteda default. The implementation used in quanteda uses `*` to match any number of any characters including none, and `?` to match any single character. See also glob2rx and References below.

"regex"

Regular expression matching.

"fixed"

Fixed (literal) pattern matching.

Note

If "fixed" is used with case_insensitive = TRUE, the text will typically be lowercased internally prior to matching. Also, glob matches are converted to regular expressions (using glob2rx) when they contain wild card characters, and to fixed pattern matches when they do not.

See also

glob2rx, https://en.wikipedia.org/wiki/Glob_(programming), regex