coerce a dfm to a matrix or data.frame

Methods for coercing a dfm object to a matrix or data.frame object.

# S3 method for dfm
as.matrix(x, ...)

# S3 method for dfm
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

Arguments

x

dfm to be coerced

...

not used

row.names

if FALSE, do not set the row names of the data.frame to the docnames of the dfm (default); or a vector of values to which the row names will be set.

optional

not applicable to this method

Examples

# coercion to matrix
mydfm <- dfm(data_char_inaugural)
str(as.matrix(mydfm))

#>  num [1:57, 1:9174] 3 1 3 7 8 1 1 6 10 3 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ docs    : chr [1:57] "1789-Washington" "1793-Washington" "1797-Adams" "1801-Jefferson" ...
#>   ..$ features: atomic [1:9174] fellow - citizens of ...
#>   .. ..- attr(*, ".match.hash")=Class 'match.hash' <externalptr> 

# coercion to a data.frame
inaugDfm <- dfm(data_char_inaugural[1:5])
as.data.frame(inaugDfm[, 1:10])

#>                 fellow  - citizens  of the senate and house representatives :
#> 1789-Washington      3  3        5  71 116      1  48     2               2 1
#> 1793-Washington      1  0        1  11  13      0   2     0               0 1
#> 1797-Adams           3  5        5 140 163      1 130     0               2 0
#> 1801-Jefferson       7 12        7 104 130      0  81     0               0 1
#> 1805-Jefferson       8  2       10 101 143      0  93     0               0 0

as.data.frame(inaugDfm[, 1:10], row.names = FALSE)

#>   fellow  - citizens  of the senate and house representatives :
#> 1      3  3        5  71 116      1  48     2               2 1
#> 2      1  0        1  11  13      0   2     0               0 1
#> 3      3  5        5 140 163      1 130     0               2 0
#> 4      7 12        7 104 130      0  81     0               0 1
#> 5      8  2       10 101 143      0  93     0               0 0