coercion, checking, and combining functions for tokens objects

Coercion functions to and from tokens objects, checks for whether an object is a tokens object, and functions to combine tokens objects.

as.tokens(x)

# S3 method for list
as.tokens(x)

# S3 method for collocations
as.tokens(x)

# S3 method for tokens
as.list(x, ...)

# S3 method for tokens
unlist(x, recursive = FALSE, use.names = TRUE)

# S3 method for tokens
as.character(x, use.names = FALSE, ...)

is.tokens(x)

# S3 method for tokens
+(t1, t2)

# S3 method for tokens
c(...)

Arguments

x
object to be coerced or checked

...
for c.tokens only, tokens objects to be concatenated

recursive
a required argument for unlist but inapplicable to tokens objects

use.names
logical; preserve names if TRUE. For as.character and unlist only.

t1
tokens one to be added

t2
tokens two to be added

Value

as.tokens returns a quanteda tokens object as.list returns a simple list of characters from a tokens object unlist returns a simple vector of characters from a tokens object as.character returns a character vector from a tokens object is.tokens returns TRUE if the object is of class tokens, FALSE otherwise. c(...) and + return a tokens object whose documents have been added as a single sequence of documents.

Examples

# combining tokens
toks1 <- tokens(c("a b c d e", "f g h"))
toks2 <- tokens(c("1 2 3"))
toks1 + toks2

#> tokens from 3 documents.
#> Component 1 :
#> [1] "a" "b" "c" "d" "e"
#> 
#> Component 2 :
#> [1] "f" "g" "h"
#> 
#> Component 3 :
#> [1] "1" "2" "3"
#> 

c(toks1, toks2)

#> tokens from 3 documents.
#> Component 1 :
#> [1] "a" "b" "c" "d" "e"
#> 
#> Component 2 :
#> [1] "f" "g" "h"
#> 
#> Component 3 :
#> [1] "1" "2" "3"
#>