coercion and checking functions for tokens objects

Coerce a list of character elements to a quanteda tokens object, or check whether an object is a tokens object.

# S3 method for collocations
as.tokens(x)

as.tokens(x)

# S3 method for list
as.tokens(x)

# S3 method for tokens
as.list(x, ...)

# S3 method for tokens
as.character(x, use.names = FALSE, ...)

is.tokens(x)

# S3 method for tokens
+(t1, t2)

# S3 method for tokens
c(..., recursive = FALSE)

Arguments

x

list of character elements

...

unused

use.names

logical; preserve names if TRUE. For as.character only.

t1

tokens one to be added

t2

tokens two to be added

recursive

logical used by `c()` method, always set to `FALSE`

Value

as.tokens returns a quanteda tokens object

as.list returns a simple list of characters from a tokens object

as.character returns a character vector from a tokens object

is.tokens returns TRUE if the object is of class tokens, FALSE otherwise.

Examples

toks1 <- tokens(data_corpus_inaugural[1:5])
toks2 <- tokens(data_corpus_inaugural[21:25])
toks3 <- toks1 + toks2

toks1 <- tokens(data_corpus_inaugural[1:5])
toks2 <- tokens(data_corpus_inaugural[21:25])
toks3 <- tokens(data_corpus_inaugural[41:45])
summary(c(toks1, toks2, toks3))

#>                 Length Class  Mode     
#> 1789-Washington 1540   -none- character
#> 1793-Washington  147   -none- character
#> 1797-Adams      2584   -none- character
#> 1801-Jefferson  1935   -none- character
#> 1805-Jefferson  2381   -none- character
#> 1869-Grant      1243   -none- character
#> 1873-Grant      1479   -none- character
#> 1877-Hayes      2730   -none- character
#> 1881-Garfield   3240   -none- character
#> 1885-Cleveland  1828   -none- character
#> 1949-Truman     2531   -none- character
#> 1953-Eisenhower 2765   -none- character
#> 1957-Eisenhower 1933   -none- character
#> 1961-Kennedy    1568   -none- character
#> 1965-Johnson    1725   -none- character