convert quanteda dictionary objects to the YAML format

Converts a quanteda dictionary object constructed by the dictionary function into the YAML format. The YAML files can be edited in text editors and imported into quanteda again.

as.yaml(x)

Arguments

x
a dictionary object

Value

as.yaml a dictionary in the YAML format, as a character object

Examples

# NOT RUN {
dict <- dictionary(list(one = c("a b", "c*"), two = c("x", "y", "z??")))
cat(yaml <- as.yaml(dict))
cat(yaml, file = (yamlfile <- paste0(tempfile(), ".yml")))
dictionary(file = yamlfile)
# }