bootstrap a dfm

Create an array of resampled dfms.

bootstrap_dfm(x, n = 10, ..., verbose = getOption("verbose"))

Arguments

x

a character or corpus object

n

number of resamples

...

additional arguments passed to dfm

verbose

if TRUE print status messages

Value

A named list of dfm objects, where the first, dfm_0, is the dfm from the original texts, and subsequent elements are the sentence-resampled dfms.

Details

This code loops the creation of a dfm from a corpus, keeping a constant set of features based on the original dfm. Resampling of the corpus is done at the sentence level, within document.

Examples

txt <- c(textone = "This is a sentence.  Another sentence.  Yet another.",
         texttwo = "Premiere phrase.  Deuxieme phrase.")
bootstrap_dfm(txt, n = 3)

#> $dfm_0
#> Document-feature matrix of: 2 documents, 10 features (45% sparse).
#> 2 x 10 sparse Matrix of class "dfmSparse"
#>          features
#> docs      this is a sentence . another yet premiere phrase deuxieme
#>   textone    1  1 1        2 3       2   1        0      0        0
#>   texttwo    0  0 0        0 2       0   0        1      2        1
#> 
#> $dfm_1
#> Document-feature matrix of: 2 documents, 9 features (44.4% sparse).
#> 2 x 9 sparse Matrix of class "dfmSparse"
#>          features
#> docs      yet another . this is a sentence deuxieme phrase
#>   textone   2       2 3    1  1 1        1        0      0
#>   texttwo   0       0 2    0  0 0        0        2      2
#> 
#> $dfm_2
#> Document-feature matrix of: 2 documents, 5 features (40% sparse).
#> 2 x 5 sparse Matrix of class "dfmSparse"
#>          features
#> docs      yet another . premiere phrase
#>   textone   3       3 3        0      0
#>   texttwo   0       0 2        2      2
#> 
#> $dfm_3
#> Document-feature matrix of: 2 documents, 7 features (42.9% sparse).
#> 2 x 7 sparse Matrix of class "dfmSparse"
#>          features
#> docs      yet another . sentence deuxieme phrase premiere
#>   textone   1       3 3        2        0      0        0
#>   texttwo   0       0 2        0        1      2        1
#> 
#> attr(,"class")
#> [1] "dfm_bootstrap" "list"