Detect collocations from text

Detects collocations from texts or a corpus, returning a data.frame of collocations and their scores, sorted in descending order of the association measure. Words separated by punctuation delimiters are not counted by default (spanPunct = FALSE) as adjacent and hence are not eligible to be collocations.

collocations(x, ...)

# S3 method for corpus
collocations(x, method = c("lr", "chi2", "pmi", "dice",
  "all"), size = 2, n = NULL, toLower = TRUE,
  punctuation = c("dontspan", "ignore", "include"), ...)

# S3 method for character
collocations(x, method = c("lr", "chi2", "pmi", "dice",
  "all"), size = 2, n = NULL, toLower = TRUE,
  punctuation = c("dontspan", "ignore", "include"), ...)

# S3 method for tokens
collocations(x, ...)

# S3 method for tokenizedTexts
collocations(x, method = c("lr", "chi2", "pmi",
  "dice", "all"), size = 2, n = NULL, toLower = FALSE,
  punctuation = c("dontspan", "ignore", "include"), ...)

Arguments

x

a text, a character vector of texts, or a corpus

...

additional parameters passed to tokenize

method

association measure for detecting collocations. Let \(i\) index documents, and \(j\) index features, \(n_{ij}\) refers to observed counts, and \(m_{ij}\) the expected counts in a collocations frequency table of dimensions \((J - size + 1)^2\). Available measures are computed as:

"lr"

The likelihood ratio statistic \(G^2\), computed as: $$2 * \sum_i \sum_j ( n_{ij} * log \frac{n_{ij}}{m_{ij}} )$$

"chi2"

Pearson's \(\chi^2\) statistic, computed as: $$\sum_i \sum_j \frac{(n_{ij} - m_{ij})^2}{m_{ij}}$$

"pmi"

point-wise mutual information score, computed as log \(n_{11}/m_{11}\)

"dice"

the Dice coefficient, computed as \(n_{11}/n_{1.} + n_{.1}\)

"all"

returns all of the above

size

length of the collocation. Only bigram (n=2) and trigram (n=3) collocations are currently implemented. Can be c(2,3) (or 2:3) to return both bi- and tri-gram collocations.

n

the number of collocations to return, sorted in descending order of the requested statistic, or \(G^2\) if none is specified.

toLower

convert collocations to lower case if TRUE (default)

punctuation

how to handle tokens separated by punctuation characters. Options are:

dontspan

do not form collocations from tokens separated by punctuation characters (default)

ignore

ignore punctuation characters when forming collocations, meaning that collocations will include those separated by punctuation characters in the text. The punctuation characters themselves are not returned.

include

do not treat punctuation characters specially, meaning that collocations will include punctuation characters as tokens

Value

A data.table of collocations, their frequencies, and the computed association measure(s).

References

McInnes, B T. 2004. "Extending the Log Likelihood Measure to Improve Collocation Identification." M.Sc. Thesis, University of Minnesota.

See also

ngrams

Examples

txt <- c("This is software testing: looking for (word) pairs!  
         This [is] a software testing again. For.",
         "Here: this is more Software Testing, looking again for word pairs.")
collocations(txt, punctuation = "dontspan") # default

#>        word1    word2 word3 count        G2
#>  1: software  testing           3 15.442482
#>  2:     this       is           2 12.056645
#>  3:      for     word           1  7.481333
#>  4:     word    pairs           1  7.481333
#>  5:    again      for           1  4.708744
#>  6:       is     more           1  4.708744
#>  7:  testing    again           1  4.708744
#>  8:        a software           1  3.662248
#>  9:     more software           1  3.662248
#> 10:  looking    again           1  2.079135
#> 11:  looking      for           1  2.079135
#> 12:       is software           1  1.186636

collocations(txt, punctuation = "dontspan", removePunct = TRUE)  # includes "testing looking"

#>        word1    word2 word3 count        G2
#>  1: software  testing           3 18.836933
#>  2:     this       is           3 18.836933
#>  3:     word    pairs           2 14.258811
#>  4:    again      for           2 10.439726
#>  5:      for     word           2 10.439726
#>  6:  testing  looking           2 10.439726
#>  7:     here     this           1  5.781582
#>  8:        a software           1  4.735086
#>  9:       is        a           1  4.735086
#> 10:       is     more           1  4.735086
#> 11:     more software           1  4.735086
#> 12:  looking    again           1  3.089015
#> 13:    pairs     this           1  3.089015
#> 14:  looking      for           1  2.125876
#> 15:  testing    again           1  2.125876
#> 16:       is software           1  1.249721

collocations(txt, punctuation = "ignore", removePunct = TRUE)    # same as previous 

#>        word1    word2 word3 count        G2
#>  1: software  testing           3 18.836933
#>  2:     this       is           3 18.836933
#>  3:     word    pairs           2 14.258811
#>  4:    again      for           2 10.439726
#>  5:      for     word           2 10.439726
#>  6:  testing  looking           2 10.439726
#>  7:     here     this           1  5.781582
#>  8:        a software           1  4.735086
#>  9:       is        a           1  4.735086
#> 10:       is     more           1  4.735086
#> 11:     more software           1  4.735086
#> 12:  looking    again           1  3.089015
#> 13:    pairs     this           1  3.089015
#> 14:  looking      for           1  2.125876
#> 15:  testing    again           1  2.125876
#> 16:       is software           1  1.249721

collocations(txt, punctuation = "include", removePunct = FALSE)  # keep punctuation as tokens

#>        word1    word2 word3 count        G2
#>  1: software  testing           3 20.990510
#>  2:        ]        a           1  9.248623
#>  3:     this       is           2  8.089577
#>  4:        !     this           1  6.476034
#>  5:        (     word           1  6.476034
#>  6:        )    pairs           1  6.476034
#>  7:        ,  looking           1  6.476034
#>  8:     here        :           1  6.476034
#>  9:    pairs        !           1  6.476034
#> 10:     word        )           1  6.476034
#> 11:        [       is           1  5.429538
#> 12:        a software           1  5.429538
#> 13:      for        (           1  5.429538
#> 14:       is        ]           1  5.429538
#> 15:       is     more           1  5.429538
#> 16:     more software           1  5.429538
#> 17:  testing        ,           1  5.429538
#> 18:     this        [           1  5.429538
#> 19:        :  looking           1  3.759008
#> 20:        :     this           1  3.759008
#> 21:  looking    again           1  3.759008
#> 22:     word    pairs           1  3.759008
#> 23:    again        .           1  2.769662
#> 24:    again      for           1  2.769662
#> 25:      for     word           1  2.769662
#> 26:  looking      for           1  2.769662
#> 27:    pairs        .           1  2.769662
#> 28:  testing        :           1  2.769662
#> 29:  testing    again           1  2.769662
#> 30:        .      for           1  1.839149
#> 31:      for        .           1  1.839149
#> 32:       is software           1  1.839149
#>        word1    word2 word3 count        G2

collocations(txt, size = 2:3)

#>        word1    word2    word3 count        G2
#>  1: software  testing              3 15.442482
#>  2:     this       is              2 12.056645
#>  3:      for     word              1  7.481333
#>  4:     word    pairs              1  7.481333
#>  5:    again      for              1  4.708744
#>  6:       is     more              1  4.708744
#>  7:  testing    again              1  4.708744
#>  8:        a software              1  3.662248
#>  9:     more software              1  3.662248
#> 10:  looking    again              1  2.079135
#> 11:  looking      for              1  2.079135
#> 12:       is software              1  1.186636
#> 13:        a software  testing     1 14.899861
#> 14:     more software  testing     1 14.899861
#> 15:     this       is     more     1 13.737119
#> 16:  looking    again      for     1 13.003319
#> 17:    again      for     word     1 13.003319
#> 18: software  testing    again     1 13.003319
#> 19:      for     word    pairs     1 13.003319
#> 20:       is software  testing     1 12.664621
#> 21:     this       is software     1 11.215186
#> 22:       is     more software     1  7.708797
#>        word1    word2    word3 count        G2

removeFeatures(collocations(txt, size = 2:3), stopwords("english"))

#> Removed 20 (90.9%) of 22 collocations containing one of 174 stopwords.

#>       word1   word2 word3 count        G2
#> 1: software testing           3 15.442482
#> 2:     word   pairs           1  7.481333

collocations("@textasdata We really, really love the #quanteda package - thanks!!")

#>          word1     word2 word3 count       G2
#> 1:   #quanteda   package           1 5.406735
#> 2: @textasdata        we           1 5.406735
#> 3:        love       the           1 5.406735
#> 4:      really      love           1 5.406735
#> 5:         the #quanteda           1 5.406735
#> 6:          we    really           1 5.406735

collocations("@textasdata We really, really love the #quanteda package - thanks!!",
              removeTwitter = TRUE)

#>         word1   word2 word3 count       G2
#> 1:       love     the           1 5.004024
#> 2:   quanteda package           1 5.004024
#> 3:     really    love           1 5.004024
#> 4: textasdata      we           1 5.004024
#> 5:         we  really           1 5.004024

collocations(data_char_inaugural[49:57], n = 10)

#>     word1  word2 word3 count       G2
#>  1:    we   will          73 284.6103
#>  2:   let     us          34 281.6678
#>  3:    we   must          50 235.2163
#>  4:    we   have          56 217.7637
#>  5: those    who          27 217.2483
#>  6:    of    our         101 202.6810
#>  7:   the  world          53 197.2617
#>  8:    it     is          39 180.2728
#>  9:    my fellow          17 173.7008
#> 10:   god  bless          13 168.7802

collocations(data_char_inaugural[49:57], method = "all", n = 10)

#>     word1  word2 word3 count       G2        X2      pmi       dice
#>  1:    we   will          73 284.6103  882.7964 2.605588 0.21823617
#>  2:   let     us          34 281.6678 2789.3414 4.419122 0.33170732
#>  3:    we   must          50 235.2163  875.0199 2.936472 0.17421603
#>  4:    we   have          56 217.7637  683.3833 2.615762 0.17948718
#>  5: those    who          27 217.2483 2538.3791 4.556226 0.38848921
#>  6:    of    our         101 202.6810  384.8322 1.679914 0.17504333
#>  7:   the  world          53 197.2617  496.2047 2.365627 0.09566787
#>  8:    it     is          39 180.2728  742.8811 3.023608 0.20365535
#>  9:    my fellow          17 173.7008 3927.2881 5.446699 0.45945946
#> 10:   god  bless          13 168.7802 8400.8567 6.472020 0.70270270

collocations(data_char_inaugural[49:57], method = "chi2", size = 3, n = 10)

#>         word1           word2      word3 count        X2
#>  1:      pork            chop       hill     1 206755642
#>  2:    amidst       gathering     clouds     1 206755642
#>  3:    wholly         herself     unless     1 206755642
#>  4:    feared       illnesses       seem     1 206755642
#>  5: broadcast instantaneously tobillions     1 206755642
#>  6:   hundred            rice    paddies     1 206755642
#>  7:   nourish         starved     bodies     1 206755642
#>  8:     wield    technology's    wonders     1 206755642
#>  9:      news       travelled     slowly     1 206755642
#> 10:   present        troubles   parallel     1 206755642

collocations(corpus_subset(data_corpus_inaugural, Year>1980), method = "pmi", size = 3, n = 10)

#>         word1           word2      word3 count      pmi
#>  1:      pork            chop       hill     1 19.14705
#>  2:    amidst       gathering     clouds     1 19.14705
#>  3:    wholly         herself     unless     1 19.14705
#>  4:    feared       illnesses       seem     1 19.14705
#>  5: broadcast instantaneously tobillions     1 19.14705
#>  6:   hundred            rice    paddies     1 19.14705
#>  7:   nourish         starved     bodies     1 19.14705
#>  8:     wield    technology's    wonders     1 19.14705
#>  9:      news       travelled     slowly     1 19.14705
#> 10:   present        troubles   parallel     1 19.14705