convert a dfm to a non-quanteda format

Convert a quanteda dfm-class object to a format useable by other text analysis packages. The general function convert provides easy conversion from a dfm to the document-term representations used in all other text analysis packages for which conversions are defined. See also convert-wrappers for convenience functions for specific package converters.

convert(x, to = c("lda", "tm", "stm", "austin", "topicmodels", "matrix",
  "data.frame"), docvars = NULL, ...)

Arguments

x

dfm to be converted

to

target conversion format, consisting of the name of the package into whose document-term matrix representation the dfm will be converted:

"lda"

a list with components "documents" and "vocab" as needed by lda.collapsed.gibbs.sampler from the lda package

"tm"

a DocumentTermMatrix from the tm package

"stm"

the format for the stm package

"austin"

the wfm format from the austin package

"topicmodels"

the "dtm" format as used by the topicmodels package

docvars

optional data.frame of document variables used as the meta information in conversion to the STM package format. This aids in selecting the document variables only corresponding to the documents with non-zero counts.

...

unused

Value

A converted object determined by the value of to (see above). See conversion target package documentation for more detailed descriptions of the return formats.

Note

There also exist a variety of converter shortcut commands, designed to mimic the idioms of the packages into whose format they convert. See convert-wrappers for details.

Examples

mycorpus <- corpus_subset(data_corpus_inaugural, Year > 1970)
quantdfm <- dfm(mycorpus, verbose = FALSE)

# austin's wfm format
identical(dim(quantdfm), dim(convert(quantdfm, to = "austin")))

#> [1] TRUE

# tm's DocumentTermMatrix format
tmdfm <- convert(quantdfm, to = "tm")
str(tmdfm)

#> List of 6
#>  $ i       : int [1:7848] 1 3 5 9 11 1 2 3 4 5 ...
#>  $ j       : int [1:7848] 1 1 1 1 1 2 2 2 2 2 ...
#>  $ v       : num [1:7848] 3 3 6 1 1 69 52 130 124 142 ...
#>  $ nrow    : int 11
#>  $ ncol    : int 3325
#>  $ dimnames:List of 2
#>   ..$ Docs : chr [1:11] "1973-Nixon" "1977-Carter" "1981-Reagan" "1985-Reagan" ...
#>   ..$ Terms: chr [1:3325] "mr" "." "vice" "president" ...
#>  - attr(*, "class")= chr [1:2] "DocumentTermMatrix" "simple_triplet_matrix"
#>  - attr(*, "weighting")= chr [1:2] "term frequency" "tf"

# stm package format
stmdfm <- convert(quantdfm, to = "stm")
str(stmdfm)

#> List of 3
#>  $ documents:List of 11
#>   ..$ 1973-Nixon  : int [1:2, 1:516] 1 37 2 96 3 1 4 5 6 3 ...
#>   ..$ 1977-Carter : int [1:2, 1:502] 1 20 2 65 3 7 4 4 7 52 ...
#>   ..$ 1981-Reagan : int [1:2, 1:852] 1 23 2 174 3 7 4 3 6 5 ...
#>   ..$ 1985-Reagan : int [1:2, 1:876] 1 31 2 177 3 13 4 7 6 3 ...
#>   ..$ 1989-Bush   : int [1:2, 1:756] 1 16 2 166 3 14 4 16 6 5 ...
#>   ..$ 1993-Clinton: int [1:2, 1:606] 1 2 2 139 3 6 4 5 7 81 ...
#>   ..$ 1997-Clinton: int [1:2, 1:726] 1 27 2 131 3 13 4 7 6 3 ...
#>   ..$ 2001-Bush   : int [1:2, 1:593] 1 5 2 110 3 4 4 7 6 1 ...
#>   ..$ 2005-Bush   : int [1:2, 1:734] 1 4 2 120 3 3 4 8 6 2 ...
#>   ..$ 2009-Obama  : int [1:2, 1:901] 1 53 2 130 3 22 4 4 5 1 ...
#>   ..$ 2013-Obama  : int [1:2, 1:786] 1 22 2 99 3 14 4 5 7 89 ...
#>  $ vocab    : chr [1:3325] "-" "," ";" ":" ...
#>  $ meta     : NULL

# illustrate what happens with zero-length documents
quantdfm2 <- dfm(c(punctOnly = "!!!", mycorpus[-1]), verbose = FALSE)
rowSums(quantdfm2)

#>    punctOnly  1977-Carter  1981-Reagan  1985-Reagan    1989-Bush 1993-Clinton 
#>            3         1380         2798         2935         2683         1837 
#> 1997-Clinton    2001-Bush    2005-Bush   2009-Obama   2013-Obama 
#>         2451         1810         2325         2729         2335 

stmdfm2 <- convert(quantdfm2, to = "stm", docvars = docvars(mycorpus))
str(stmdfm2)

#> List of 3
#>  $ documents:List of 11
#>   ..$ punctOnly   : int [1:2, 1] 5 3
#>   ..$ 1977-Carter : int [1:2, 1:502] 1 20 2 65 3 7 4 4 7 52 ...
#>   ..$ 1981-Reagan : int [1:2, 1:852] 1 23 2 174 3 7 4 3 6 5 ...
#>   ..$ 1985-Reagan : int [1:2, 1:876] 1 31 2 177 3 13 4 7 6 3 ...
#>   ..$ 1989-Bush   : int [1:2, 1:756] 1 16 2 166 3 14 4 16 6 5 ...
#>   ..$ 1993-Clinton: int [1:2, 1:606] 1 2 2 139 3 6 4 5 7 81 ...
#>   ..$ 1997-Clinton: int [1:2, 1:726] 1 27 2 131 3 13 4 7 6 3 ...
#>   ..$ 2001-Bush   : int [1:2, 1:593] 1 5 2 110 3 4 4 7 6 1 ...
#>   ..$ 2005-Bush   : int [1:2, 1:734] 1 4 2 120 3 3 4 8 6 2 ...
#>   ..$ 2009-Obama  : int [1:2, 1:901] 1 53 2 130 3 22 4 4 5 1 ...
#>   ..$ 2013-Obama  : int [1:2, 1:786] 1 22 2 99 3 14 4 5 7 89 ...
#>  $ vocab    : chr [1:3239] "-" "," ";" ":" ...
#>  $ meta     :'data.frame':	11 obs. of  3 variables:
#>   ..$ Year     : int [1:11] 1973 1977 1981 1985 1989 1993 1997 2001 2005 2009 ...
#>   ..$ President: chr [1:11] "Nixon" "Carter" "Reagan" "Reagan" ...
#>   ..$ FirstName: chr [1:11] "Richard Milhous" "Jimmy" "Ronald" "Ronald" ...

# topicmodels package format
str(convert(quantdfm, to = "topicmodels"))

#> List of 6
#>  $ i       : int [1:7848] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ j       : int [1:7848] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ v       : int [1:7848] 3 69 1 1 96 1 1 1 1 1 ...
#>  $ nrow    : int 11
#>  $ ncol    : int 3325
#>  $ dimnames:List of 2
#>   ..$ Docs : chr [1:11] "1973-Nixon" "1977-Carter" "1981-Reagan" "1985-Reagan" ...
#>   ..$ Terms: chr [1:3325] "mr" "." "vice" "president" ...
#>  - attr(*, "class")= chr [1:2] "DocumentTermMatrix" "simple_triplet_matrix"
#>  - attr(*, "weighting")= chr [1:2] "term frequency" "tf"

# lda package format
ldadfm <- convert(quantdfm, to = "lda")
str(ldadfm)

#> List of 2
#>  $ documents:List of 11
#>   ..$ 1973-Nixon  : int [1:2, 1:516] 0 3 1 69 2 1 3 1 4 96 ...
#>   ..$ 1977-Carter : int [1:2, 1:502] 1 52 3 3 4 65 7 1 12 48 ...
#>   ..$ 1981-Reagan : int [1:2, 1:852] 0 3 1 130 2 2 3 5 4 174 ...
#>   ..$ 1985-Reagan : int [1:2, 1:876] 1 124 2 1 3 3 4 177 5 1 ...
#>   ..$ 1989-Bush   : int [1:2, 1:756] 0 6 1 142 2 1 3 6 4 166 ...
#>   ..$ 1993-Clinton: int [1:2, 1:606] 1 81 3 2 4 139 12 66 13 7 ...
#>   ..$ 1997-Clinton: int [1:2, 1:726] 1 108 3 1 4 131 7 1 12 94 ...
#>   ..$ 2001-Bush   : int [1:2, 1:593] 1 96 2 1 3 3 4 110 7 3 ...
#>   ..$ 2005-Bush   : int [1:2, 1:734] 0 1 1 98 2 1 3 4 4 120 ...
#>   ..$ 2009-Obama  : int [1:2, 1:901] 1 118 3 1 4 130 12 111 13 2 ...
#>   ..$ 2013-Obama  : int [1:2, 1:786] 0 1 1 89 2 1 3 2 4 99 ...
#>  $ vocab    : chr [1:3325] "mr" "." "vice" "president" ...