recast the document units of a corpus

For a corpus, reshape (or recast) the documents to a different level of aggregation. Units of aggregation can be defined as documents, paragraphs, or sentences. Because the corpus object records its current "units" status, it is possible to move from recast units back to original units, for example from documents, to sentences, and then back to documents (possibly after modifying the sentences).

corpus_reshape(x, to = c("sentences", "paragraphs", "documents"), ...)

Arguments

x

corpus whose document units will be reshaped

to

new document units in which the corpus will be recast

...

not used

Value

A corpus object with the documents defined as the new units, including document-level meta-data identifying the original documents.

Examples

# simple example
mycorpus <- corpus(c(textone = "This is a sentence.  Another sentence.  Yet another.",
                     textwo = "Premiere phrase.  Deuxieme phrase."),
                   docvars = data.frame(country=c("UK", "USA"), year=c(1990, 2000)),
                   metacorpus = list(notes = "Example showing how corpus_reshape() works."))
summary(mycorpus)

#> Corpus consisting of 2 documents.
#> 
#>     Text Types Tokens Sentences country year
#>  textone     8     11         3      UK 1990
#>   textwo     4      6         2     USA 2000
#> 
#> Source:  /Users/kbenoit/Dropbox (Personal)/GitHub/quanteda/docs/reference/* on x86_64 by kbenoit
#> Created: Mon Apr  3 19:19:31 2017
#> Notes:   Example showing how corpus_reshape() works.
#> 

summary(corpus_reshape(mycorpus, to = "sentences"), showmeta=TRUE)

#> Corpus consisting of 5 documents.
#> 
#>      Text Types Tokens Sentences country year _document _serialno
#>  textone1     5      5         1      UK 1990   textone         1
#>  textone2     3      3         1      UK 1990   textone         2
#>  textone3     3      3         1      UK 1990   textone         3
#>   textwo1     3      3         1     USA 2000    textwo         1
#>   textwo2     3      3         1     USA 2000    textwo         2
#> 
#> Source:  /Users/kbenoit/Dropbox (Personal)/GitHub/quanteda/docs/reference/* on x86_64 by kbenoit
#> Created: Mon Apr  3 19:19:31 2017
#> Notes:   Example showing how corpus_reshape() works.
#> 

# example with inaugural corpus speeches
(mycorpus2 <- corpus_subset(data_corpus_inaugural, Year>2004))

#> Corpus consisting of 4 documents and 3 docvars.

paragCorpus <- corpus_reshape(mycorpus2, to="paragraphs")
paragCorpus

#> Corpus consisting of 138 documents and 3 docvars.

summary(paragCorpus, 100, showmeta=TRUE)

#> Corpus consisting of 138 documents, showing 100 documents.
#> 
#>          Text Types Tokens Sentences Year President FirstName  _document
#>     2005-Bush   772   2325       100 2005      Bush George W.  2005-Bush
#>   2009-Obama1     4      4         1 2009     Obama    Barack 2009-Obama
#>   2009-Obama2    42     53         2 2009     Obama    Barack 2009-Obama
#>   2009-Obama3    64     88         4 2009     Obama    Barack 2009-Obama
#>   2009-Obama4    12     15         2 2009     Obama    Barack 2009-Obama
#>   2009-Obama5    78    110         5 2009     Obama    Barack 2009-Obama
#>   2009-Obama6    39     47         2 2009     Obama    Barack 2009-Obama
#>   2009-Obama7    36     47         4 2009     Obama    Barack 2009-Obama
#>   2009-Obama8    19     22         1 2009     Obama    Barack 2009-Obama
#>   2009-Obama9    31     35         1 2009     Obama    Barack 2009-Obama
#>  2009-Obama10    58     84         2 2009     Obama    Barack 2009-Obama
#>  2009-Obama11    73    110         5 2009     Obama    Barack 2009-Obama
#>  2009-Obama12    21     21         1 2009     Obama    Barack 2009-Obama
#>  2009-Obama13    20     24         1 2009     Obama    Barack 2009-Obama
#>  2009-Obama14    17     20         1 2009     Obama    Barack 2009-Obama
#>  2009-Obama15    43     51         2 2009     Obama    Barack 2009-Obama
#>  2009-Obama16    73    108         7 2009     Obama    Barack 2009-Obama
#>  2009-Obama17    85    144         8 2009     Obama    Barack 2009-Obama
#>  2009-Obama18    52     62         3 2009     Obama    Barack 2009-Obama
#>  2009-Obama19    97    150         5 2009     Obama    Barack 2009-Obama
#>  2009-Obama20    76    117         3 2009     Obama    Barack 2009-Obama
#>  2009-Obama21    90    137         4 2009     Obama    Barack 2009-Obama
#>  2009-Obama22    60     83         3 2009     Obama    Barack 2009-Obama
#>  2009-Obama23    93    144         5 2009     Obama    Barack 2009-Obama
#>  2009-Obama24    83    128         3 2009     Obama    Barack 2009-Obama
#>  2009-Obama25    65    103         3 2009     Obama    Barack 2009-Obama
#>  2009-Obama26    63     84         3 2009     Obama    Barack 2009-Obama
#>  2009-Obama27    82    117         4 2009     Obama    Barack 2009-Obama
#>  2009-Obama28    69     96         3 2009     Obama    Barack 2009-Obama
#>  2009-Obama29    95    158         7 2009     Obama    Barack 2009-Obama
#>  2009-Obama30     9     10         1 2009     Obama    Barack 2009-Obama
#>  2009-Obama31    20     22         1 2009     Obama    Barack 2009-Obama
#>  2009-Obama32    53     65         1 2009     Obama    Barack 2009-Obama
#>  2009-Obama33    67     95         6 2009     Obama    Barack 2009-Obama
#>  2009-Obama34    34     53         1 2009     Obama    Barack 2009-Obama
#>  2009-Obama35    75    106         4 2009     Obama    Barack 2009-Obama
#>  2009-Obama36    11     16         3 2009     Obama    Barack 2009-Obama
#>   2013-Obama1    20     23         1 2013     Obama    Barack 2013-Obama
#>   2013-Obama2    56     82         4 2013     Obama    Barack 2013-Obama
#>   2013-Obama3    35     43         1 2013     Obama    Barack 2013-Obama
#>   2013-Obama4    78    117         4 2013     Obama    Barack 2013-Obama
#>   2013-Obama5    10     10         1 2013     Obama    Barack 2013-Obama
#>   2013-Obama6    36     46         2 2013     Obama    Barack 2013-Obama
#>   2013-Obama7    22     26         1 2013     Obama    Barack 2013-Obama
#>   2013-Obama8    21     21         1 2013     Obama    Barack 2013-Obama
#>   2013-Obama9    23     24         1 2013     Obama    Barack 2013-Obama
#>  2013-Obama10    44     54         2 2013     Obama    Barack 2013-Obama
#>  2013-Obama11    89    130         4 2013     Obama    Barack 2013-Obama
#>  2013-Obama12    68     93         5 2013     Obama    Barack 2013-Obama
#>  2013-Obama13    90    131         4 2013     Obama    Barack 2013-Obama
#>  2013-Obama14    68     97         5 2013     Obama    Barack 2013-Obama
#>  2013-Obama15    66     97         4 2013     Obama    Barack 2013-Obama
#>  2013-Obama16    72    109         4 2013     Obama    Barack 2013-Obama
#>  2013-Obama17    54     76         3 2013     Obama    Barack 2013-Obama
#>  2013-Obama18    70    104         6 2013     Obama    Barack 2013-Obama
#>  2013-Obama19    81    122         5 2013     Obama    Barack 2013-Obama
#>  2013-Obama20    44     56         2 2013     Obama    Barack 2013-Obama
#>  2013-Obama21    92    140         4 2013     Obama    Barack 2013-Obama
#>  2013-Obama22    66     98         1 2013     Obama    Barack 2013-Obama
#>  2013-Obama23   110    189         6 2013     Obama    Barack 2013-Obama
#>  2013-Obama24    62    101         4 2013     Obama    Barack 2013-Obama
#>  2013-Obama25    65     95         4 2013     Obama    Barack 2013-Obama
#>  2013-Obama26    75    109         4 2013     Obama    Barack 2013-Obama
#>  2013-Obama27    45     72         3 2013     Obama    Barack 2013-Obama
#>  2013-Obama28    39     52         2 2013     Obama    Barack 2013-Obama
#>  2013-Obama29    15     18         2 2013     Obama    Barack 2013-Obama
#>   2017-Trump1    20     28         1 2017     Trump Donald J. 2017-Trump
#>   2017-Trump2    26     29         1 2017     Trump Donald J. 2017-Trump
#>   2017-Trump3    17     20         1 2017     Trump Donald J. 2017-Trump
#>   2017-Trump4    13     18         3 2017     Trump Donald J. 2017-Trump
#>   2017-Trump5    40     48         3 2017     Trump Donald J. 2017-Trump
#>   2017-Trump6    35     49         2 2017     Trump Donald J. 2017-Trump
#>   2017-Trump7    23     25         1 2017     Trump Donald J. 2017-Trump
#>   2017-Trump8    13     13         1 2017     Trump Donald J. 2017-Trump
#>   2017-Trump9    12     13         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump10    13     13         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump11    30     38         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump12    21     24         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump13    13     14         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump14     6     10         2 2017     Trump Donald J. 2017-Trump
#>  2017-Trump15    12     13         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump16    18     21         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump17    18     21         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump18    13     14         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump19     7      7         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump20    21     25         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump21    19     20         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump22    16     20         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump23    12     14         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump24    62     84         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump25     9     11         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump26    23     39         3 2017     Trump Donald J. 2017-Trump
#>  2017-Trump27    14     16         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump28    47     63         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump29    20     22         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump30    25     30         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump31    20     20         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump32    14     16         2 2017     Trump Donald J. 2017-Trump
#>  2017-Trump33    23     28         1 2017     Trump Donald J. 2017-Trump
#>  2017-Trump34    13     13         1 2017     Trump Donald J. 2017-Trump
#>  _serialno
#>          1
#>          1
#>          2
#>          3
#>          4
#>          5
#>          6
#>          7
#>          8
#>          9
#>         10
#>         11
#>         12
#>         13
#>         14
#>         15
#>         16
#>         17
#>         18
#>         19
#>         20
#>         21
#>         22
#>         23
#>         24
#>         25
#>         26
#>         27
#>         28
#>         29
#>         30
#>         31
#>         32
#>         33
#>         34
#>         35
#>         36
#>          1
#>          2
#>          3
#>          4
#>          5
#>          6
#>          7
#>          8
#>          9
#>         10
#>         11
#>         12
#>         13
#>         14
#>         15
#>         16
#>         17
#>         18
#>         19
#>         20
#>         21
#>         22
#>         23
#>         24
#>         25
#>         26
#>         27
#>         28
#>         29
#>          1
#>          2
#>          3
#>          4
#>          5
#>          6
#>          7
#>          8
#>          9
#>         10
#>         11
#>         12
#>         13
#>         14
#>         15
#>         16
#>         17
#>         18
#>         19
#>         20
#>         21
#>         22
#>         23
#>         24
#>         25
#>         26
#>         27
#>         28
#>         29
#>         30
#>         31
#>         32
#>         33
#>         34
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

## Note that Bush 2005 is recorded as a single paragraph because that text used a single
## \n to mark the end of a paragraph.