randomly sample documents from a corpus

Takes a random sample or documents or features of the specified size from a corpus or document-feature matrix, with or without replacement. Works just as sample works for the documents and their associated document-level variables.

corpus_sample(x, size = ndoc(x), replace = FALSE, prob = NULL, ...)

Arguments

x

a corpus object whose documents will be sampled

size

a positive number, the number of documents to select

replace

Should sampling be with replacement?

prob

A vector of probability weights for obtaining the elements of the vector being sampled.

...

unused

Value

A corpus object with number of documents equal to size, drawn from the corpus x. The returned corpus object will contain all of the meta-data of the original corpus, and the same document variables for the documents selected.

See also

sample

Examples

# sampling from a corpus
summary(corpus_sample(data_corpus_inaugural, 5))

#> Corpus consisting of 5 documents.
#> 
#>             Text Types Tokens Sentences Year  President   FirstName
#>     1865-Lincoln   362    780        26 1865    Lincoln     Abraham
#>       1825-Adams  1004   3154        74 1825      Adams John Quincy
#>  1953-Eisenhower   903   2765       119 1953 Eisenhower   Dwight D.
#>   1945-Roosevelt   276    651        26 1945  Roosevelt Franklin D.
#>  1793-Washington    96    147         4 1793 Washington      George
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

summary(corpus_sample(data_corpus_inaugural, 10, replace=TRUE))

#> Corpus consisting of 10 documents.
#> 
#>               Text Types Tokens Sentences Year  President      FirstName
#>         1873-Grant   552   1479        43 1873      Grant     Ulysses S.
#>      1901-McKinley   857   2449       100 1901   McKinley        William
#>    1957-Eisenhower   621   1933        92 1957 Eisenhower      Dwight D.
#>  1957-Eisenhower.1   621   1933        92 1957 Eisenhower      Dwight D.
#>     1905-Roosevelt   404   1089        33 1905  Roosevelt       Theodore
#>       1865-Lincoln   362    780        26 1865    Lincoln        Abraham
#>         1797-Adams   826   2584        37 1797      Adams           John
#>      1881-Garfield  1018   3240       111 1881   Garfield       James A.
#>         1825-Adams  1004   3154        74 1825      Adams    John Quincy
#>          1909-Taft  1436   5844       159 1909       Taft William Howard
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#>