randomly sample documents from a corpus

Takes a random sample or documents or features of the specified size from a corpus or document-feature matrix, with or without replacement. Works just as sample works for the documents and their associated document-level variables.

corpus_sample(x, size = ndoc(x), replace = FALSE, prob = NULL, ...)

Arguments

x

a corpus object whose documents will be sampled

size

a positive number, the number of documents to select

replace

Should sampling be with replacement?

prob

A vector of probability weights for obtaining the elements of the vector being sampled.

...

unused

Value

A corpus object with number of documents equal to size, drawn from the corpus x. The returned corpus object will contain all of the meta-data of the original corpus, and the same document variables for the documents selected.

See also

sample

Examples

# sampling from a corpus
summary(corpus_sample(data_corpus_inaugural, 5))

#> Corpus consisting of 5 documents.
#> 
#>            Text Types Tokens Sentences Year President     FirstName
#>      1877-Hayes   829   2730        59 1877     Hayes Rutherford B.
#>  1893-Cleveland   823   2135        58 1893 Cleveland        Grover
#>   1897-McKinley  1236   4383       130 1897  McKinley       William
#>      1797-Adams   826   2584        37 1797     Adams          John
#>  1905-Roosevelt   404   1089        33 1905 Roosevelt      Theodore
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

summary(corpus_sample(data_corpus_inaugural, 10, replace=TRUE))

#> Corpus consisting of 10 documents.
#> 
#>             Text Types Tokens Sentences Year  President       FirstName
#>    1889-Harrison  1355   4744       157 1889   Harrison        Benjamin
#>    1857-Buchanan   945   3106        89 1857   Buchanan           James
#>   1933-Roosevelt   744   2064        85 1933  Roosevelt     Franklin D.
#>       1877-Hayes   829   2730        59 1877      Hayes   Rutherford B.
#>       1869-Grant   486   1243        40 1869      Grant      Ulysses S.
#>     1921-Harding  1172   3743       148 1921    Harding       Warren G.
#>       1797-Adams   826   2584        37 1797      Adams            John
#>  1957-Eisenhower   621   1933        92 1957 Eisenhower       Dwight D.
#>       1973-Nixon   545   2018        68 1973      Nixon Richard Milhous
#>    1841-Harrison  1893   9178       210 1841   Harrison   William Henry
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#>