randomly sample documents from a corpus

Takes a random sample or documents or features of the specified size from a corpus or document-feature matrix, with or without replacement. Works just as sample works for the documents and their associated document-level variables.

corpus_sample(x, size = ndoc(x), replace = FALSE, prob = NULL, ...)

Arguments

x

a corpus object whose documents will be sampled

size

a positive number, the number of documents to select

replace

Should sampling be with replacement?

prob

A vector of probability weights for obtaining the elements of the vector being sampled.

...

unused

Value

A corpus object with number of documents equal to size, drawn from the corpus x. The returned corpus object will contain all of the meta-data of the original corpus, and the same document variables for the documents selected.

See also

sample

Examples

# sampling from a corpus
summary(corpus_sample(data_corpus_inaugural, 5))

#> Corpus consisting of 5 documents.
#> 
#>             Text Types Tokens Sentences Year  President     FirstName
#>  1957-Eisenhower   621   1933        92 1957 Eisenhower     Dwight D.
#>    1925-Coolidge  1221   4442       196 1925   Coolidge        Calvin
#>      1985-Reagan   925   2935       123 1985     Reagan        Ronald
#>     1813-Madison   542   1304        33 1813    Madison         James
#>       1877-Hayes   829   2730        59 1877      Hayes Rutherford B.
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

summary(corpus_sample(data_corpus_inaugural, 10, replace=TRUE))

#> Corpus consisting of 10 documents.
#> 
#>             Text Types Tokens Sentences Year  President       FirstName
#>     1993-Clinton   644   1837        81 1993    Clinton            Bill
#>      1985-Reagan   925   2935       123 1985     Reagan          Ronald
#>  1793-Washington    96    147         4 1793 Washington          George
#>        1909-Taft  1436   5844       159 1909       Taft  William Howard
#>       1973-Nixon   545   2018        68 1973      Nixon Richard Milhous
#>    1881-Garfield  1018   3240       111 1881   Garfield        James A.
#>     1921-Harding  1172   3743       148 1921    Harding       Warren G.
#>  1881-Garfield.1  1018   3240       111 1881   Garfield        James A.
#>   1805-Jefferson   804   2381        45 1805  Jefferson          Thomas
#>   1921-Harding.1  1172   3743       148 1921    Harding       Warren G.
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#>