randomly sample documents from a corpus

Takes a random sample or documents or features of the specified size from a corpus or document-feature matrix, with or without replacement. Works just as sample works for the documents and their associated document-level variables.

corpus_sample(x, size = ndoc(x), replace = FALSE, prob = NULL, ...)

Arguments

x

a corpus object whose documents will be sampled

size

a positive number, the number of documents to select

replace

Should sampling be with replacement?

prob

A vector of probability weights for obtaining the elements of the vector being sampled.

...

unused

Value

A corpus object with number of documents equal to size, drawn from the corpus x. The returned corpus object will contain all of the meta-data of the original corpus, and the same document variables for the documents selected.

See also

sample

Examples

# sampling from a corpus
summary(corpus_sample(data_corpus_inaugural, 5))

#> Corpus consisting of 5 documents.
#> 
#>             Text Types Tokens Sentences Year  President  FirstName
#>  1789-Washington   626   1540        23 1789 Washington     George
#>     1921-Harding  1172   3743       148 1921    Harding  Warren G.
#>        1845-Polk  1330   5211       153 1845       Polk James Knox
#>      1913-Wilson   661   1896        68 1913     Wilson    Woodrow
#>     1865-Lincoln   362    780        26 1865    Lincoln    Abraham
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

summary(corpus_sample(data_corpus_inaugural, 10, replace=TRUE))

#> Corpus consisting of 10 documents.
#> 
#>            Text Types Tokens Sentences Year President FirstName
#>    1833-Jackson   499   1271        29 1833   Jackson    Andrew
#>       2001-Bush   622   1810        97 2001      Bush George W.
#>    1961-Kennedy   566   1568        52 1961   Kennedy   John F.
#>     1929-Hoover  1086   3895       158 1929    Hoover   Herbert
#>  1905-Roosevelt   404   1089        33 1905 Roosevelt  Theodore
#>       2005-Bush   772   2325       100 2005      Bush George W.
#>       1989-Bush   795   2683       141 1989      Bush    George
#>     1817-Monroe  1040   3696       121 1817    Monroe     James
#>   1881-Garfield  1018   3240       111 1881  Garfield  James A.
#>     1949-Truman   781   2531       116 1949    Truman  Harry S.
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#>