randomly sample documents from a corpus

Takes a random sample or documents or features of the specified size from a corpus or document-feature matrix, with or without replacement. Works just as sample works for the documents and their associated document-level variables.

corpus_sample(x, size = ndoc(x), replace = FALSE, prob = NULL, ...)

Arguments

x

a corpus object whose documents will be sampled

size

a positive number, the number of documents to select

replace

Should sampling be with replacement?

prob

A vector of probability weights for obtaining the elements of the vector being sampled.

...

unused

Value

A corpus object with number of documents equal to size, drawn from the corpus x. The returned corpus object will contain all of the meta-data of the original corpus, and the same document variables for the documents selected.

See also

sample

Examples

# sampling from a corpus
summary(corpus_sample(data_corpus_inaugural, 5))

#> Corpus consisting of 5 documents.
#> 
#>            Text Types Tokens Sentences Year President   FirstName
#>      1797-Adams   826   2584        37 1797     Adams        John
#>       2005-Bush   772   2325       100 2005      Bush   George W.
#>  1893-Cleveland   823   2135        58 1893 Cleveland      Grover
#>  1941-Roosevelt   527   1552        68 1941 Roosevelt Franklin D.
#>   1881-Garfield  1018   3240       111 1881  Garfield    James A.
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

summary(corpus_sample(data_corpus_inaugural, 10, replace=TRUE))

#> Corpus consisting of 10 documents.
#> 
#>             Text Types Tokens Sentences Year President   FirstName
#>        2005-Bush   772   2325       100 2005      Bush   George W.
#>    1837-VanBuren  1315   4175        95 1837 Van Buren      Martin
#>     1813-Madison   542   1304        33 1813   Madison       James
#>     1861-Lincoln  1075   4016       135 1861   Lincoln     Abraham
#>     1809-Madison   536   1267        21 1809   Madison       James
#>   1937-Roosevelt   729   2027        96 1937 Roosevelt Franklin D.
#>    1897-McKinley  1236   4383       130 1897  McKinley     William
#>   1809-Madison.1   536   1267        21 1809   Madison       James
#>       1873-Grant   552   1479        43 1873     Grant  Ulysses S.
#>  1837-VanBuren.1  1315   4175        95 1837 Van Buren      Martin
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#>