randomly sample documents from a corpus

Takes a random sample or documents or features of the specified size from a corpus or document-feature matrix, with or without replacement. Works just as sample works for the documents and their associated document-level variables.

corpus_sample(x, size = ndoc(x), replace = FALSE, prob = NULL,
  by = NULL, ...)

Arguments

x

a corpus object whose documents will be sampled

size

a positive number, the number of documents to select

replace

Should sampling be with replacement?

prob

A vector of probability weights for obtaining the elements of the vector being sampled.

by

a grouping variable for sampling. Useful for resampling sub-document units such as sentences, for instance by specifying by = "document"

...

unused

Value

A corpus object with number of documents equal to size, drawn from the corpus x. The returned corpus object will contain all of the meta-data of the original corpus, and the same document variables for the documents selected.

Examples

# sampling from a corpus
summary(corpus_sample(data_corpus_inaugural, 5))

#> Corpus consisting of 5 documents.
#> 
#>             Text Types Tokens Sentences Year  President FirstName
#>     1921-Harding  1172   3743       148 1921    Harding Warren G.
#>      1821-Monroe  1262   4898       129 1821     Monroe     James
#>  1793-Washington    96    147         4 1793 Washington    George
#>  1953-Eisenhower   903   2765       119 1953 Eisenhower Dwight D.
#>        2001-Bush   622   1810        97 2001       Bush George W.
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

summary(corpus_sample(data_corpus_inaugural, 10, replace=TRUE))

#> Corpus consisting of 10 documents.
#> 
#>              Text Types Tokens Sentences Year President       FirstName
#>     1889-Harrison  1355   4744       157 1889  Harrison        Benjamin
#>        1969-Nixon   743   2437       103 1969     Nixon Richard Milhous
#>      1829-Jackson   517   1210        25 1829   Jackson          Andrew
#>     1901-McKinley   857   2449       100 1901  McKinley         William
#>    1945-Roosevelt   276    651        26 1945 Roosevelt     Franklin D.
#>    1829-Jackson.1   517   1210        25 1829   Jackson          Andrew
#>     1837-VanBuren  1315   4175        95 1837 Van Buren          Martin
#>  1945-Roosevelt.1   276    651        26 1945 Roosevelt     Franklin D.
#>     1881-Garfield  1018   3240       111 1881  Garfield        James A.
#>        2009-Obama   939   2729       110 2009     Obama          Barack
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

# sampling sentences within document
doccorpus <- corpus(c(one = "Sentence one.  Sentence two.  Third sentence.",
                      two = "First sentence, doc2.  Second sentence, doc2."))
sentcorpus <- corpus_reshape(doccorpus, to = "sentences")
texts(sentcorpus)

#>                     one1                     one2                     one3 
#>          "Sentence one."          "Sentence two."        "Third sentence." 
#>                     two1                     two2 
#>  "First sentence, doc2." "Second sentence, doc2." 

texts(corpus_sample(sentcorpus, replace = TRUE, by = "document"))

#>                     one2                     one3                     one1 
#>          "Sentence two."        "Third sentence."          "Sentence one." 
#>                     two2                     two1 
#> "Second sentence, doc2."  "First sentence, doc2."