segment texts into component elements

Segment corpus text(s) or a character vector into tokens, sentences, paragraphs, or other sections. segment works on a character vector or corpus object, and allows the delimiters to be user-defined. This is useful for breaking the texts of a corpus into smaller documents based on sentences, or based on a user defined "tag" pattern. See details.

corpus_segment(x, what = c("tokens", "sentences", "paragraphs", "tags",
  "other"), delimiter = ifelse(what == "tokens", " ", ifelse(what ==
  "sentences", "[.!?:;]", ifelse(what == "paragraphs", "\\n{2}", ifelse(what
  == "tags", "##\\w+\\b", NULL)))), valuetype = c("regex", "fixed",
  "glob"), perl = FALSE, keepdocvars = TRUE, ...)

char_segment(x, what = c("tokens", "sentences", "paragraphs", "tags",
  "other"), delimiter = ifelse(what == "tokens", " ", ifelse(what ==
  "sentences", "[.!?:;]", ifelse(what == "paragraphs", "\\n{2}", ifelse(what
  == "tags", "##\\w+\\b", NULL)))), valuetype = c("regex", "fixed",
  "glob"), perl = FALSE, ...)

Arguments

x

corpus object whose texts will be segmented

what

unit of segmentation. Current options are "tokens" (default), "sentences", "paragraphs", "tags", and "other". Segmenting on other allows segmentation of a text on any user-defined value, and must be accompanied by the delimiter argument. Segmenting on tags performs the same function but preserves the tags as a document variable in the segmented corpus.

delimiter

delimiter defined as a regex for segmentation. Each type has its own default, except other, which requires a value to be specified.

valuetype

how to interpret the delimiter: fixed for exact matching; "regex" for regular expressions; or "glob" for "glob"-style wildcard patterns

perl

logical. Should Perl-compatible regular expressions be used?

keepdocvars

(for corpus objects) if TRUE, repeat the docvar values for each segmented text; if FALSE, drop the docvars in the segmented corpus. Dropping the docvars might be useful in order to conserve space or if these are not desired for the segmented corpus.

...

provides additional arguments passed to tokens, if what = "tokens" is used

Value

A corpus of segmented texts.

Details

Tokens are delimited by separators. For sentences, the delimiter can be defined by the user. The default for sentences includes ., !, ?, plus ; and :.

For paragraphs, the default is two carriage returns, although this could be changed to a single carriage return by changing the value of delimiter to "\\n{1}" which is the R version of the regex for one newline character. (You might need this if the document was created in a word processor, for instance, and the lines were wrapped in the window rather than being hard-wrapped with a newline character.)

Note

Does not currently record document segments if segmenting a multi-text corpus into smaller units. For this, use changeunits instead.

Examples

testCorpus <- corpus(c("##INTRO This is the introduction. 
                       ##DOC1 This is the first document.  
                       Second sentence in Doc 1.  
                       ##DOC3 Third document starts here.  
                       End of third document.",
                      "##INTRO Document ##NUMBER Two starts before ##NUMBER Three."))
# add a docvar
testCorpus[["serialno"]] <- paste0("textSerial", 1:ndoc(testCorpus))
testCorpusSeg <- corpus_segment(testCorpus, "tags")
summary(testCorpusSeg)

#> Corpus consisting of 6 documents.
#> 
#>     Text Types Tokens Sentences      tag    serialno
#>  text1.1     5      5         1  ##INTRO textSerial1
#>  text1.2    11     12         2   ##DOC1 textSerial1
#>  text1.3     8     10         2   ##DOC3 textSerial1
#>  text2.1     1      1         1  ##INTRO textSerial2
#>  text2.2     3      3         1 ##NUMBER textSerial2
#>  text2.3     2      2         1 ##NUMBER textSerial2
#> 
#> Source:  /Users/kbenoit/Dropbox (Personal)/GitHub/quanteda/docs/reference/* on x86_64 by kbenoit
#> Created: Mon Dec  5 22:37:23 2016
#> Notes:   corpus_segment(corpus_segment)corpus_segment(testCorpus)corpus_segment(tags)
#> 

texts(testCorpusSeg)

#>                                                                                                       text1.1 
#>                                                        " This is the introduction. \n                       " 
#>                                                                                                       text1.2 
#> " This is the first document.  \n                       Second sentence in Doc 1.  \n                       " 
#>                                                                                                       text1.3 
#>                               " Third document starts here.  \n                       End of third document." 
#>                                                                                                       text2.1 
#>                                                                                                  " Document " 
#>                                                                                                       text2.2 
#>                                                                                         " Two starts before " 
#>                                                                                                       text2.3 
#>                                                                                                     " Three." 

# segment a corpus into sentences
segmentedCorpus <- corpus_segment(corpus(data_char_ukimmig2010), "sentences")
summary(segmentedCorpus)

#> Corpus consisting of 207 documents, showing 100 documents.
#> 
#>            Text Types Tokens Sentences
#>           BNP.1   145    259         1
#>           BNP.2    18     18         1
#>           BNP.3    26     29         1
#>           BNP.4    31     40         1
#>           BNP.5   103    191         1
#>           BNP.6    25     29         1
#>           BNP.7    58     90         1
#>           BNP.8    40     48         1
#>           BNP.9    72    106         1
#>          BNP.10    40     69         1
#>          BNP.11    40     49         1
#>          BNP.12    44     56         1
#>          BNP.13    28     33         1
#>          BNP.14    29     31         1
#>          BNP.15    32     36         1
#>          BNP.16     9      9         1
#>          BNP.17    10     10         1
#>          BNP.18    42     67         1
#>          BNP.19    50     63         1
#>          BNP.20    41     52         1
#>          BNP.21   103    169         1
#>          BNP.22    37     42         1
#>          BNP.23    25     29         1
#>          BNP.24    45     54         1
#>          BNP.25    25     27         1
#>          BNP.26    18     19         1
#>          BNP.27    36     40         1
#>          BNP.28    29     32         1
#>          BNP.29    28     30         1
#>          BNP.30    21     22         1
#>          BNP.31    28     30         1
#>          BNP.32    29     31         1
#>          BNP.33    18     21         1
#>          BNP.34    14     14         1
#>          BNP.35    29     33         1
#>          BNP.36    30     35         1
#>          BNP.37    14     14         1
#>          BNP.38    34     40         1
#>          BNP.39     2      2         1
#>          BNP.40    16     19         1
#>          BNP.41    23     24         1
#>          BNP.42    17     18         1
#>          BNP.43    18     21         1
#>          BNP.44    28     33         1
#>          BNP.45    23     25         1
#>          BNP.46     2      2         1
#>          BNP.47    24     27         1
#>          BNP.48    20     21         1
#>          BNP.49    19     23         1
#>          BNP.50    28     34         1
#>          BNP.51    30     35         1
#>          BNP.52    25     28         1
#>          BNP.53    31     34         1
#>          BNP.54    24     26         1
#>          BNP.55     2      2         1
#>          BNP.56    39     45         1
#>          BNP.57    25     30         1
#>          BNP.58    23     24         1
#>          BNP.59     2      2         1
#>          BNP.60    26     36         1
#>          BNP.61    12     15         1
#>          BNP.62    31     35         1
#>          BNP.63     2      2         1
#>          BNP.64    20     22         1
#>          BNP.65    18     21         1
#>          BNP.66     2      2         1
#>          BNP.67    34     36         1
#>          BNP.68     2      2         1
#>          BNP.69    16     18         1
#>          BNP.70    18     21         1
#>          BNP.71     2      2         1
#>          BNP.72    83    147         1
#>          BNP.73    38     49         1
#>          BNP.74    48     64         1
#>          BNP.75    32     34         1
#>          BNP.76    25     26         1
#>          BNP.77    37     43         1
#>          BNP.78    23     26         1
#>          BNP.79    32     41         1
#>          BNP.80    34     41         1
#>          BNP.81    21     22         1
#>          BNP.82    22     23         1
#>          BNP.83    20     21         1
#>          BNP.84    18     22         1
#>          BNP.85    31     39         1
#>          BNP.86    28     30         1
#>          BNP.87    14     15         1
#>          BNP.88    30     33         1
#>     Coalition.1     2      2         1
#>     Coalition.2    26     29         1
#>     Coalition.3    72    121         1
#>     Coalition.4    74    116         1
#>  Conservative.1     9      9         1
#>  Conservative.2    25     30         1
#>  Conservative.3    12     12         1
#>  Conservative.4    24     26         1
#>  Conservative.5    22     28         1
#>  Conservative.6    58     78         1
#>  Conservative.7    34     37         1
#>  Conservative.8    16     16         1
#> 
#> Source:  /Users/kbenoit/Dropbox (Personal)/GitHub/quanteda/docs/reference/* on x86_64 by kbenoit
#> Created: Mon Dec  5 22:37:23 2016
#> Notes:   corpus_segment(corpus_segment)corpus_segment(corpus(data_char_ukimmig2010))corpus_segment(sentences)
#> 

# same as tokenize()
identical(tokenize(data_char_ukimmig2010), char_segment(data_char_ukimmig2010))

#> [1] TRUE

# segment into paragraphs
char_segment(data_char_ukimmig2010[3:4], "paragraphs")

#> $Conservative
#>  [1] "Attract the brightest and best to our country."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#>  [2] "Immigration has enriched our nation over the years and we want to attract the brightest and the best people who can make a real difference to our economic growth. But immigration today is too high and needs to be reduced. We do not need to attract people to do jobs that could be carried out by British citizens, given the right training and support. So we will take steps to take net migration back to the levels of the 1990s - tens of thousands a year, not hundreds of thousands."                                                   
#>  [3] "To help achieve this goal, we will introduce a number of measures, such as:"                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [4] "- Setting an annual limit on the number of non-EU economic migrants admitted into the UK to live and work; "                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [5] "- Limiting access only to those who will bring the most value to the British economy; and, "                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [6] "- Applying transitional controls as a matter of course in the future for all new EU member States. "                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#>  [7] "In addition, we will promote integration into British society, as we believe that everyone coming to this country must be ready to embrace our core values and become a part of their local community. So there will be an English language test for anyone coming here to get married. "                                                                                                                                                                                                                                                            
#>  [8] "We want to encourage students to come to our universities and colleges, but our student visa system has become the biggest weakness in our border controls. A Conservative government will strengthen the system of granting student visas so that it is less open to abuse. We want to make it easier for reputable universities and colleges to accept applications, while putting extra scrutiny on new institutions looking to accept foreign students or existing institutions not registered with Companies house. In addition, we will:"      
#>  [9] "- Insist foreign students at new or unregistered institutions pay a bond in order to study in this country, to be repaid after the student has left the country at the end of their studies; "                                                                                                                                                                                                                                                                                                                                                       
#> [10] "- Ensure foreign students can prove that they have the financial means to support themselves in the UK; and, "                                                                                                                                                                                                                                                                                                                                                                                                                                       
#> [11] "- Require that students must usually leave the country and reapply if they want to switch to another course or apply for a work permit. "                                                                                                                                                                                                                                                                                                                                                                                                            
#> [12] "Extremists, serious criminals and others find our borders far too easy to penetrate. That is why we will create a dedicated border Police force, as part of a refocused Serious Organised Crime agency, to enhance national security, improve immigration controls, and crack down on the trafficking of people, weapons and drugs. We will work with police forces to strengthen arrangements to deal with serious crime and other cross-boundary policing challenges, and extend collaboration between forces to deliver better value for money.\n"
#> 
#> $Greens
#>  [1] "Immigration."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [2] "Migration is a fact of life.  People have always moved from one country to another, and as a practical matter the ability to control borders without oppressive measures is more limited than most politicians like to pretend. Much of our language, culture and way of life have been enriched by successive new arrivals over two thousand years. It is not just a matter of immigration: over 5 million British Citizens benefit from other countries' liberal immigration policies by living abroad."                                                                                                                                                                                                                                                                                                     
#>  [3] "The causes of a person moving to the UK are complex. For the person concerned, there may be escape from persecution and improved economic prospects, but also separation from home, friends and family. For the country of origin, there may be the loss of skilled workers, especially health professionals, but also the receipt of remittances from the immigrant, and many migrants return with improved skills. For the community that receives the immigrant there may be the benefits of getting done jobs that no one in that community wants to or can do, more taxes being paid and the creation of a more cosmopolitan atmosphere. But there may also be costs in terms of unwelcome competition for jobs, pressure on housing and other resources and longer-term pressures on overall population."
#>  [4] "In deciding policy on immigration it is important that all these points are considered and balanced against each other. We must accept too our legal and moral obligations to give sanctuary to those fleeing persecution, and the principle of free movement throughout the European Union. Against this background our policy is as follows: "                                                                                                                                                                                                                                                                                                                                                                                                                                                               
#>  [5] "- Where we are limiting numbers, our priority must be to meet our obligations to refugees and those seeking sanctuary, including the increasing numbers of people displaced by environmental change, above the needs of our economy."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
#>  [6] "- Our immigration policies must be fair and non-discriminatory, respect the integrity of families and be applied promptly and effectively."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#>  [7] "- Our international policies should every- where seek to reduce the economic, political and environmental factors that force people to migrate. Emigration should be a positive choice, not the outcome of desperation. In particular, free movement within the EU is a fact. We should press for EU policies that make all parts of the EU an attractive place to live."                                                                                                                                                                                                                                                                                                                                                                                                                                      
#>  [8] "- We reject the use of immigration as a political issue to mask problems such as a lack of high-quality social housing. The proper solution is to provide enough social housing, as we propose elsewhere in this manifesto."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#>  [9] "- We should not tolerate the long-term presence of large numbers of people whose immigration status is not defined. Such immigrants are vulnerable to exploitation by unscrupulous employers and others, under- mining national terms and conditions of employment. We would open up ways for existing illegal migrants who have been here for three years to become legal. In particular, a legal status must be provided for people who have not succeeded in their claim for humanitarian protection but who cannot be returned to their country of origin due to the political situation there."                                                                                                                                                                                                           
#> [10] "- We would review the asylum procedures to ensure that destitution plays no role in the asylum process by allowing those seeking sanctuary to work."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> [11] "- We would review the Nationality, Immigration and Asylum Act 2002, particularly with regard to issues of access to legal advice, childcare and levels of subsistence allowance."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
#> [12] "- Those who have been trafficked should not be subject to summary deportation. They should receive a temporary right to stay and have the same right to apply to remain as others seeking to migrate."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [13] "- Those seeking sanctuary should not be detained, and in particular the administrative detention of children is unacceptable and should cease immediately.\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
#> 

# segment a text into sentences
segmentedChar <- char_segment(data_char_ukimmig2010, "sentences")
segmentedChar[2]

#> $Coalition
#> [1] "IMMIGRATION."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#> [2] "The Government believes that immigration has enriched our culture and strengthened our economy, but that it must be controlled so that people have confidence in the system."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#> [3] "We also recognise that to ensure cohesion and protect our public services, we need to introduce a cap on immigration and reduce the number of non-EU immigrants.  - We will introduce an annual limit on the number of non-EU economic migrants admitted into the UK to live and work.  - We will consider jointly the mechanism for implementing the limit.  - We will end the detention of children for immigration purposes.  - We will create a dedicated Border Police Force, as part of a refocused Serious Organised Crime Agency, to enhance national security, improve immigration controls and crack down on the trafficking of people, weapons and drugs."
#> [4] "We will work with police forces to strengthen arrangements to deal with serious crime and other cross-boundary policing challenges, and extend collaboration between forces to deliver better value for money.  - We support E-borders and will reintroduce exit checks.  - We will apply transitional controls as a matter of course in the future for all new EU Member States.  - We will introduce new measures to minimise abuse of the immigration system, for example via student routes, and will tackle human trafficking as a priority.  - We will explore new ways to improve the current asylum system to speed up the processing of applications."      
#>