extract a subset of a corpus

Returns subsets of a corpus that meet certain conditions, including direct logical operations on docvars (document-level variables). corpus_subset functions identically to subset.data.frame, using non-standard evaluation to evaluate conditions based on the docvars in the corpus.

corpus_subset(x, subset, select, ...)

Arguments

x

corpus object to be subsetted

subset

logical expression indicating elements or rows to keep: missing values are taken as false

select

expression, indicating the attributes to select from the corpus

...

not used

Value

corpus object, with a subset of documents (and docvars) selected according to arguments

See also

subset.data.frame

Examples

summary(corpus_subset(data_corpus_inaugural, Year > 1980))

#> Corpus consisting of 10 documents.
#> 
#>          Text Types Tokens Sentences Year President FirstName
#>   1981-Reagan   904   2798       128 1981    Reagan    Ronald
#>   1985-Reagan   925   2935       123 1985    Reagan    Ronald
#>     1989-Bush   795   2683       141 1989      Bush    George
#>  1993-Clinton   644   1837        81 1993   Clinton      Bill
#>  1997-Clinton   773   2451       111 1997   Clinton      Bill
#>     2001-Bush   622   1810        97 2001      Bush George W.
#>     2005-Bush   772   2325       100 2005      Bush George W.
#>    2009-Obama   939   2729       110 2009     Obama    Barack
#>    2013-Obama   814   2335        88 2013     Obama    Barack
#>    2017-Trump   582   1662        88 2017     Trump Donald J.
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

summary(corpus_subset(data_corpus_inaugural, Year > 1930 & President == "Roosevelt",
                      select = Year))

#> Corpus consisting of 4 documents.
#> 
#>            Text Types Tokens Sentences Year
#>  1933-Roosevelt   744   2064        85 1933
#>  1937-Roosevelt   729   2027        96 1937
#>  1941-Roosevelt   527   1552        68 1941
#>  1945-Roosevelt   276    651        26 1945
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#>