Irish budget speeches from 2010

Speeches and document-level variables from the debate over the Irish budget of 2010.

data_corpus_irishbudget2010

Format

The corpus object for the 2010 budget speeches, with document-level variables for year, debate, serial number, first and last name of the speaker, and the speaker's party.

Source

Dáil Éireann Debate, Budget Statement 2010. 9 December 2009. vol. 697, no. 3.

Details

At the time of the debate, Fianna Fáil (FF) and the Greens formed the government coalition, while Fine Gael (FG), Labour (LAB), and Sinn Féin (SF) were in opposition.

References

Lowe, W. & Benoit, K.R. (2013). Validating Estimates of Latent Traits From Textual Data Using Human Judgment as a Benchmark. Political Analysis, 21(3), 298--313.

Examples

summary(data_corpus_irishbudget2010)

#> Corpus consisting of 14 documents:
#> 
#>                       Text Types Tokens Sentences year debate number      foren
#>        Lenihan, Brian (FF)  1953   8641       374 2010 BUDGET     01      Brian
#>       Bruton, Richard (FG)  1040   4446       217 2010 BUDGET     02    Richard
#>         Burton, Joan (LAB)  1624   6393       307 2010 BUDGET     03       Joan
#>        Morgan, Arthur (SF)  1595   7107       343 2010 BUDGET     04     Arthur
#>          Cowen, Brian (FF)  1629   6599       250 2010 BUDGET     05      Brian
#>           Kenny, Enda (FG)  1148   4232       153 2010 BUDGET     06       Enda
#>      ODonnell, Kieran (FG)   678   2297       133 2010 BUDGET     07     Kieran
#>       Gilmore, Eamon (LAB)  1181   4177       201 2010 BUDGET     08      Eamon
#>     Higgins, Michael (LAB)   488   1286        44 2010 BUDGET     09    Michael
#>        Quinn, Ruairi (LAB)   439   1284        59 2010 BUDGET     10     Ruairi
#>      Gormley, John (Green)   401   1030        49 2010 BUDGET     11       John
#>        Ryan, Eamon (Green)   510   1643        90 2010 BUDGET     12      Eamon
#>      Cuffe, Ciaran (Green)   442   1240        45 2010 BUDGET     13     Ciaran
#>  OCaolain, Caoimhghin (SF)  1188   4044       176 2010 BUDGET     14 Caoimhghin
#>      name party
#>   Lenihan    FF
#>    Bruton    FG
#>    Burton   LAB
#>    Morgan    SF
#>     Cowen    FF
#>     Kenny    FG
#>  ODonnell    FG
#>   Gilmore   LAB
#>   Higgins   LAB
#>     Quinn   LAB
#>   Gormley Green
#>      Ryan Green
#>     Cuffe Green
#>  OCaolain    SF
#> 
#> Source: /Users/kbenoit/Dropbox (Personal)/GitHub/quanteda/* on x86_64 by kbenoit
#> Created: Wed Jun 28 22:04:18 2017
#> Notes: