recombine a dfm or fcm by combining identical dimension elements

"Compresses" or groups a dfm or fcm whose dimension names are the same, for either documents or features. This may happen, for instance, if features are made equivalent through application of a thesaurus. It may also occur after lower-casing or stemming the features of a dfm, but this should only be done in very rare cases (approaching never: it's better to do this before constructing the dfm.) It could also be needed after a cbind.dfm or rbind.dfm operation. dfm_group allows combining dfm documents by a grouping variable, which can also be one of the docvars attached to the dfm. This is identical in functionality to using the "groups" argument in dfm.

dfm_compress(x, margin = c("both", "documents", "features"))

dfm_group(x, groups = NULL)

fcm_compress(x)

Arguments

x
input object, a dfm or fcm

margin
character indicating on which margin to compress a dfm, either "documents", "features", or "both" (default). For fcm objects, "documents" has no effect.

groups
either: a character vector containing the names of document variables to be used for grouping; or a factor or object that can be coerced into a factor equal in length or rows to the number of documents

...
additional arguments passed from generic to specific methods

Value

dfm_compress returns a dfm whose dimensions have been recombined by summing the cells across identical dimension names (docnames or featnames). The docvars will be preserved for combining by features but not when documents are combined. dfm_group returns a dfm whose documents are equal to the unique group combinations, and whose cell values are the sums of the previous values summed by group. This currently erases any docvars in the dfm.

Note

fcm_compress works only when the fcm was created with a document context.

Examples

# dfm_compress examples
mat <- rbind(dfm(c("b A A", "C C a b B"), tolower = FALSE),
             dfm("A C C C C C", tolower = FALSE))

#> 

colnames(mat) <- char_tolower(featnames(mat))
mat

#> Document-feature matrix of: 3 documents, 5 features (46.7% sparse).
#> 3 x 5 sparse Matrix of class "dfmSparse"
#>        features
#> docs    a c b a b
#>   text1 2 0 1 0 0
#>   text2 0 2 1 1 1
#>   text1 1 5 0 0 0

dfm_compress(mat, margin = "documents")

#> Document-feature matrix of: 2 documents, 5 features (30% sparse).
#> 2 x 5 sparse Matrix of class "dfmSparse"
#>        features
#> docs    a c b a b
#>   text1 3 5 1 0 0
#>   text2 0 2 1 1 1

dfm_compress(mat, margin = "features")

#> Document-feature matrix of: 3 documents, 3 features (22.2% sparse).
#> 3 x 3 sparse Matrix of class "dfmSparse"
#>        features
#> docs    a c b
#>   text1 2 0 1
#>   text2 1 2 2
#>   text1 1 5 0

dfm_compress(mat)

#> Document-feature matrix of: 2 documents, 3 features (0% sparse).
#> 2 x 3 sparse Matrix of class "dfmSparse"
#>        features
#> docs    a c b
#>   text1 3 5 1
#>   text2 1 2 2

# no effect if no compression needed
compactdfm <- dfm(data_corpus_inaugural[1:5])
dim(compactdfm)

#> [1]    5 1947

dim(dfm_compress(compactdfm))

#> [1]    5 1947

# dfm_group examples
mycorpus <- corpus(c("a a b", "a b c c", "a c d d", "a c c d"),
                   docvars = data.frame(grp = c("grp1", "grp1", "grp2", "grp2")))
mydfm <- dfm(mycorpus)
dfm_group(mydfm, groups = "grp")

#> Document-feature matrix of: 2 documents, 4 features (25% sparse).
#> 2 x 4 sparse Matrix of class "dfmSparse"
#>       features
#> docs   a b c d
#>   grp1 3 2 2 0
#>   grp2 2 0 3 3

dfm_group(mydfm, groups = c(1, 1, 2, 2))

#> Document-feature matrix of: 2 documents, 4 features (25% sparse).
#> 2 x 4 sparse Matrix of class "dfmSparse"
#>     features
#> docs a b c d
#>    1 3 2 2 0
#>    2 2 0 3 3

# equivalent
dfm(mydfm, groups = "grp")

#> Document-feature matrix of: 2 documents, 4 features (25% sparse).
#> 2 x 4 sparse Matrix of class "dfmSparse"
#>       features
#> docs   a b c d
#>   grp1 3 2 2 0
#>   grp2 2 0 3 3

dfm(mydfm, groups = c(1, 1, 2, 2))

#> Document-feature matrix of: 2 documents, 4 features (25% sparse).
#> 2 x 4 sparse Matrix of class "dfmSparse"
#>     features
#> docs a b c d
#>    1 3 2 2 0
#>    2 2 0 3 3

# compress an fcm
myfcm <- fcm(tokens("A D a C E a d F e B A C E D"),
             context = "window", window = 3)
## this will produce an error:
# fcm_compress(myfcm)

txt <- c("The fox JUMPED over the dog.",
         "The dog jumped over the fox.")
toks <- tokens(txt, remove_punct = TRUE)
myfcm <- fcm(toks, context = "document")

#> Error in get(".SigLength", envir = env): object '.SigLength' not found

colnames(myfcm) <- rownames(myfcm) <- tolower(colnames(myfcm))
colnames(myfcm)[5] <- rownames(myfcm)[5] <- "fox"
myfcm

#> Feature co-occurrence matrix of: 9 by 9 features.
#> 9 x 9 sparse Matrix of class "fcm"
#>         features
#> features a d a c fox d f e b
#>      a   . 2 1 2   1 . 1 1 1
#>      d   . . 1 2   2 . . . .
#>      a   . . 1 2   2 1 1 1 .
#>      c   . . . .   2 1 . 1 1
#>      fox . . . .   . 1 1 . 1
#>      d   . . . .   . . 1 1 1
#>      f   . . . .   . . . 1 1
#>      e   . . . .   . . . . 1
#>      b   . . . .   . . . . .

fcm_compress(myfcm)

#> Error in fcm_compress(myfcm): compress_fcm invalid if fcm was created with a window context