randomly sample documents or features from a dfm

Sample randomly from a dfm object, from documents or features.

dfm_sample(x, size = ndoc(x), replace = FALSE, prob = NULL,
  what = c("documents", "features"))

Arguments

x

the dfm object whose documents or features will be sampled

size

a positive number, the number of documents or features to select

replace

logical; should sampling be with replacement?

prob

a vector of probability weights for obtaining the elements of the vector being sampled.

what

dimension (of a dfm) to sample: can be documents or features

Value

A dfm object with number of documents or features equal to size, drawn from the dfm x.

See also

sample

Examples

myDfm <- dfm(data_corpus_inaugural[1:10])
dfm_sample(myDfm)[, 1:10]

#> Document-feature matrix of: 10 documents, 10 features (31% sparse).
#> 10 x 10 sparse Matrix of class "dfmSparse"
#>                  features
#> docs              fellow  - citizens  of the senate and house representatives :
#>   1821-Monroe         10  6       15 197 360      0 141     0               0 0
#>   1805-Jefferson       8  2       10 101 143      0  93     0               0 0
#>   1801-Jefferson       7 12        7 104 130      0  81     0               0 1
#>   1789-Washington      3  3        5  71 116      1  48     2               2 1
#>   1825-Adams           3 11        2 245 304      0 116     0               1 0
#>   1817-Monroe          6 14        9 164 275      0 122     0               1 0
#>   1813-Madison         1  1        4  65 100      0  44     0               0 0
#>   1793-Washington      1  0        1  11  13      0   2     0               0 1
#>   1809-Madison         1  6        1  69 104      0  43     0               0 0
#>   1797-Adams           3  5        5 140 163      1 130     0               2 0

dfm_sample(myDfm, replace = TRUE)[, 1:10]

#> Document-feature matrix of: 10 documents, 10 features (32% sparse).
#> 10 x 10 sparse Matrix of class "dfmSparse"
#>                  features
#> docs              fellow  - citizens  of the senate and house representatives :
#>   1793-Washington      1  0        1  11  13      0   2     0               0 1
#>   1825-Adams           3 11        2 245 304      0 116     0               1 0
#>   1797-Adams           3  5        5 140 163      1 130     0               2 0
#>   1797-Adams           3  5        5 140 163      1 130     0               2 0
#>   1825-Adams           3 11        2 245 304      0 116     0               1 0
#>   1793-Washington      1  0        1  11  13      0   2     0               0 1
#>   1821-Monroe         10  6       15 197 360      0 141     0               0 0
#>   1793-Washington      1  0        1  11  13      0   2     0               0 1
#>   1805-Jefferson       8  2       10 101 143      0  93     0               0 0
#>   1797-Adams           3  5        5 140 163      1 130     0               2 0

dfm_sample(myDfm, what = "features")[1:10, ]

#> Document-feature matrix of: 10 documents, 10 features (85% sparse).
#> 10 x 10 sparse Matrix of class "dfmSparse"
#>                  features
#> docs              honest inadequacy pretense practiced forthwith dominion
#>   1789-Washington      1          0        0         0         0        0
#>   1793-Washington      0          0        0         0         0        0
#>   1797-Adams           1          0        1         0         0        0
#>   1801-Jefferson       4          0        0         1         0        0
#>   1805-Jefferson       0          0        0         0         0        0
#>   1809-Madison         0          1        0         0         0        0
#>   1813-Madison         0          0        0         0         0        0
#>   1817-Monroe          0          0        0         0         0        0
#>   1821-Monroe          0          0        0         0         1        0
#>   1825-Adams           1          0        0         0         0        1
#>                  features
#> docs              disapproved particular attachment equivalent
#>   1789-Washington           0          2          0          0
#>   1793-Washington           0          0          0          0
#>   1797-Adams                0          1          2          0
#>   1801-Jefferson            0          0          1          0
#>   1805-Jefferson            1          0          0          0
#>   1809-Madison              0          0          0          0
#>   1813-Madison              0          0          0          0
#>   1817-Monroe               0          0          0          0
#>   1821-Monroe               0          0          0          1
#>   1825-Adams                0          0          0          0