#' @noRd #' @export #' @examples #' # for dfmDense #' mat <- rbind(dfm(c("b A A", "C C a b B"), tolower = FALSE, verbose = FALSE), #' dfm("A C C C C C", tolower = FALSE, verbose = FALSE)) #' matd <- dfm_smooth(mat) #' colnames(matd) <- char_tolower(featnames(mat)) #' matd #' dfm_compress(matd, margin = "documents") #' dfm_compress(matd, margin = "features") #' dfm_compress(matd) dfm_compress.dfmDense <- function(x, ...) dfm_compress(new("dfm", Matrix::Matrix(as.matrix(x), sparse = TRUE), settings = x@settings, weightTf = x@weightTf, weightDf = x@weightDf, smooth = x@smooth, ngrams = x@ngrams, skip = x@skip, concatenator = x@concatenator, docvars = x@docvars), ...) sort a dfm by frequency of one or more margins

Sorts a dfm by descending frequency of total features, total features in documents, or both.

dfm_sort(x, decreasing = TRUE, margin = c("features", "documents", "both"))

Arguments

x
Document-feature matrix created by dfm

decreasing
logical; if TRUE, the sort will be in descending order, otherwise sort in increasing order

margin
which margin to sort on features to sort by frequency of features, documents to sort by total feature counts in documents, and both to sort by both

Value

A sorted dfm matrix object

Examples

dtm <- dfm(data_corpus_inaugural)
head(dtm)

#> Document-feature matrix of: 6 documents, 6 features (30.6% sparse).
#> 6 x 6 sparse Matrix of class "dfm"
#>                  features
#> docs              fellow-citizens  of the senate and house
#>   1789-Washington               1  71 116      1  48     2
#>   1793-Washington               0  11  13      0   2     0
#>   1797-Adams                    3 140 163      1 130     0
#>   1801-Jefferson                2 104 130      0  81     0
#>   1805-Jefferson                0 101 143      0  93     0
#>   1809-Madison                  1  69 104      0  43     0

head(dfm_sort(dtm))

#> Error in get(".SigLength", envir = env): object '.SigLength' not found

head(dfm_sort(dtm, decreasing = FALSE, "both"))

#> Error in get(".SigLength", envir = env): object '.SigLength' not found