sort a dfm by frequency of one or more margins

Sorts a dfm by descending frequency of total features, total features in documents, or both.

dfm_sort(x, decreasing = TRUE, margin = c("features", "documents", "both"))

Arguments

x
Document-feature matrix created by dfm

decreasing
logical; if TRUE, the sort will be in descending order, otherwise sort in increasing order

margin
which margin to sort on features to sort by frequency of features, documents to sort by total feature counts in documents, and both to sort by both

Value

A sorted dfm matrix object

Examples

dtm <- dfm(data_corpus_inaugural)
head(dtm)

#> Document-feature matrix of: 6 documents, 6 features (30.6% sparse).
#> 6 x 6 sparse Matrix of class "dfm"
#>                  features
#> docs              fellow-citizens  of the senate and house
#>   1789-Washington               1  71 116      1  48     2
#>   1793-Washington               0  11  13      0   2     0
#>   1797-Adams                    3 140 163      1 130     0
#>   1801-Jefferson                2 104 130      0  81     0
#>   1805-Jefferson                0 101 143      0  93     0
#>   1809-Madison                  1  69 104      0  43     0

head(dfm_sort(dtm))

#> Document-feature matrix of: 6 documents, 6 features (0% sparse).
#> 6 x 6 sparse Matrix of class "dfm"
#>                  features
#> docs              the  of   , and  . to
#>   1789-Washington 116  71  70  48 23 48
#>   1793-Washington  13  11   5   2  4  5
#>   1797-Adams      163 140 201 130 33 72
#>   1801-Jefferson  130 104 128  81 37 61
#>   1805-Jefferson  143 101 142  93 41 83
#>   1809-Madison    104  69  47  43 21 61

head(dfm_sort(dtm, decreasing = FALSE, "both"))

#> Document-feature matrix of: 6 documents, 6 features (100% sparse).
#> 6 x 6 sparse Matrix of class "dfm"
#>                  features
#> docs              notification 14th fondest predilection flattering asylum
#>   1793-Washington            0    0       0            0          0      0
#>   1945-Roosevelt             0    0       0            0          0      0
#>   1865-Lincoln               0    0       0            0          0      0
#>   1905-Roosevelt             0    0       0            0          0      0
#>   1849-Taylor                0    0       0            0          0      0
#>   1829-Jackson               0    0       0            0          0      0