sort a dfm by frequency of one or more margins

Sorts a dfm by descending frequency of total features, total features in documents, or both.

dfm_sort(x, decreasing = TRUE, margin = c("features", "documents", "both"))

Arguments

x

Document-feature matrix created by dfm

decreasing

TRUE (default) if sort will be in descending order, otherwise sort in increasing order

margin

which margin to sort on features to sort by frequency of features, documents to sort by total feature counts in documents, and both to sort by both

Value

A sorted dfm matrix object

Examples

dtm <- dfm(data_corpus_inaugural)
dtm[1:10, 1:5]

#> Document-feature matrix of: 10 documents, 5 features (2% sparse).
#> 10 x 5 sparse Matrix of class "dfmSparse"
#>                  features
#> docs              fellow  - citizens  of the
#>   1789-Washington      3  3        5  71 116
#>   1793-Washington      1  0        1  11  13
#>   1797-Adams           3  5        5 140 163
#>   1801-Jefferson       7 12        7 104 130
#>   1805-Jefferson       8  2       10 101 143
#>   1809-Madison         1  6        1  69 104
#>   1813-Madison         1  1        4  65 100
#>   1817-Monroe          6 14        9 164 275
#>   1821-Monroe         10  6       15 197 360
#>   1825-Adams           3 11        2 245 304

dfm_sort(dtm)[1:10, 1:5]

#> Document-feature matrix of: 10 documents, 5 features (0% sparse).
#> 10 x 5 sparse Matrix of class "dfmSparse"
#>                  features
#> docs              the  of   , and   .
#>   1789-Washington 116  71  70  48  23
#>   1793-Washington  13  11   5   2   4
#>   1797-Adams      163 140 201 130  33
#>   1801-Jefferson  130 104 128  81  37
#>   1805-Jefferson  143 101 142  93  41
#>   1809-Madison    104  69  47  43  21
#>   1813-Madison    100  65  53  44  31
#>   1817-Monroe     275 164 169 122 110
#>   1821-Monroe     360 197 271 141 130
#>   1825-Adams      304 245 116 116  72

dfm_sort(dtm, decreasing = FALSE, "both")[1:10, 1:5]

#> Document-feature matrix of: 10 documents, 5 features (100% sparse).
#> 10 x 5 sparse Matrix of class "dfmSparse"
#>                  features
#> docs              notification 14th fondest predilection flattering
#>   1793-Washington            0    0       0            0          0
#>   1945-Roosevelt             0    0       0            0          0
#>   1865-Lincoln               0    0       0            0          0
#>   1905-Roosevelt             0    0       0            0          0
#>   1849-Taylor                0    0       0            0          0
#>   1829-Jackson               0    0       0            0          0
#>   1869-Grant                 0    0       0            0          0
#>   1809-Madison               0    0       0            0          0
#>   1833-Jackson               0    0       0            0          0
#>   1813-Madison               0    0       0            0          0