convert the case of the features of a dfm and combine

dfm_tolower and dfm_toupper convert the features of the dfm to lower and upper case, respectively, and then recombine the counts.

dfm_tolower(x)

dfm_toupper(x)

fcm_tolower(x)

fcm_toupper(x)

Arguments

x

the dfm or fcm object

Details

fcm_tolower and fcm_toupper convert both dimensions of the fcm to lower and upper case, respectively, and then recombine the counts. This works only on fcm objects created with context = "document".

Examples

# for a document-feature matrix
mydfm <- dfm(c("b A A", "C C a b B"),
             toLower = FALSE, verbose = FALSE)

#> Warning: Argument toLower not used.

mydfm

#> Document-feature matrix of: 2 documents, 3 features (16.7% sparse).
#> 2 x 3 sparse Matrix of class "dfmSparse"
#>        features
#> docs    b a c
#>   text1 1 2 0
#>   text2 2 1 2

dfm_tolower(mydfm)

#> Document-feature matrix of: 2 documents, 3 features (16.7% sparse).
#> 2 x 3 sparse Matrix of class "dfmSparse"
#>        features
#> docs    b a c
#>   text1 1 2 0
#>   text2 2 1 2

dfm_toupper(mydfm)

#> Document-feature matrix of: 2 documents, 3 features (16.7% sparse).
#> 2 x 3 sparse Matrix of class "dfmSparse"
#>        features
#> docs    B A C
#>   text1 1 2 0
#>   text2 2 1 2

# for a feature co-occurrence matrix
myfcm <- fcm(tokens(c("b A A d", "C C a b B e")),
             context = "document")
myfcm

#> Feature co-occurrence matrix of: 7 by 7 features.
#> 7 x 7 sparse Matrix of class "fcm"
#>         features
#> features b A d C a B e
#>        b 0 2 1 2 1 1 1
#>        A . 1 2 . . . .
#>        d . . 0 . . . .
#>        C . . . 1 2 2 2
#>        a . . . . 0 1 1
#>        B . . . . . 0 1
#>        e . . . . . . 0

fcm_tolower(myfcm)
fcm_toupper(myfcm)