trim a dfm using frequency threshold-based feature selection

Returns a document by feature matrix reduced in size based on document and term frequency, usually in terms of a minimum frequencies, but may also be in terms of maximum frequencies. Setting a combination of minimum and maximum frequencies will select features based on a range.

dfm_trim(x, min_count = 1, min_docfreq = 1, max_count = NULL,
  max_docfreq = NULL, sparsity = NULL, verbose = TRUE)

Arguments

x

a dfm object

min_count

minimum count or fraction of features across all documents, below which features will be removed

min_docfreq

minimum number or fraction of documents in which a feature appears, below which features will be removed

max_count

maximum count or fraction of features across all documents, above which features will be removed. (Default is no upper limit.)

max_docfreq

maximum number or fraction of documents in which a feature appears, above which features will be removed. (Default is no upper limit.)

sparsity

equivalent to 1 - min_docfreq, included for comparison with tm

verbose

print messages

Value

A dfm reduced in features (with the same number of documents)

Note

Trimming a dfm object is an operation based on the values in the document-feature matrix. To select subsets of a dfm based on attributes of the features themselves -- such as selecting features matching a regular expression, or removing features matching a stopword list, use dfm_select.

See also

dfm_select, dfm_sample

Examples

(myDfm <- dfm(data_corpus_inaugural[1:5]))

#> Document-feature matrix of: 5 documents, 1,947 features (69.5% sparse).

# keep only words occuring >=10 times and in >=2 docs
dfm_trim(myDfm, min_count = 10, min_docfreq = 2)

#> Removing features occurring: 

#>   - fewer than 10 times: 1,840

#>   - in fewer than 2 documents: 1,379

#>   Total features removed: 1,840 (94.5%).

#> Document-feature matrix of: 5 documents, 107 features (16.1% sparse).

# keep only words occuring >=10 times and in at least 0.4 of the documents
dfm_trim(myDfm, min_count = 10, min_docfreq = 0.4)

#> Removing features occurring: 

#>   - fewer than 10 times: 1,840

#>   - in fewer than 0.4 * 5 = 2 documents: 1,379

#>   Total features removed: 1,840 (94.5%).

#> Document-feature matrix of: 5 documents, 107 features (16.1% sparse).

# keep only words occuring <=10 times and in <=2 docs
dfm_trim(myDfm, max_count = 10, max_docfreq = 2)

#> Removing features occurring: 

#>   - more than 10 times: 93

#>   - in more than 2 documents: 271

#>   Total features removed: 272 (14.0%).

#> Document-feature matrix of: 5 documents, 1,675 features (76.5% sparse).

# keep only words occuring <=10 times and in at most 3/4 of the documents
dfm_trim(myDfm, max_count = 10, max_docfreq = 0.75)

#> Removing features occurring: 

#>   - more than 10 times: 93

#>   - in more than 0.75 * 5 = 3.75 documents: 139

#>   Total features removed: 149 (7.65%).

#> Document-feature matrix of: 5 documents, 1,798 features (74% sparse).

# keep only words occuring at least 0.01 times and in >=2 documents
dfm_trim(myDfm, min_count = .01, min_docfreq = 2)

#> Removing features occurring: 

#>   - fewer than 0.01 * 1,947 = 19.47 times: 1,889

#>   - in fewer than 2 documents: 1,379

#>   Total features removed: 1,889 (97.0%).

#> Document-feature matrix of: 5 documents, 58 features (8.28% sparse).

# keep only words occuring 5 times in 1000, and in 2 of 5 of documents
dfm_trim(myDfm, min_docfreq = 0.4, min_count = 0.005)

#> Removing features occurring: 

#>   - fewer than 0.005 * 1,947 = 9.735 times: 1,840

#>   - in fewer than 0.4 * 5 = 2 documents: 1,379

#>   Total features removed: 1,840 (94.5%).

#> Document-feature matrix of: 5 documents, 107 features (16.1% sparse).

## Not run: ------------------------------------
# # compare to removeSpareTerms from the tm package 
# if (require(tm)) {
#     (tmdtm <- convert(myDfm, "tm"))
#     removeSparseTerms(tmdtm, 0.7)
#     dfm_trim(td, min_docfreq = 0.3)
#     dfm_trim(td, sparsity = 0.7)
# }
## ---------------------------------------------