weight the feature frequencies in a dfm

Returns a document by feature matrix with the feature frequencies weighted according to one of several common methods. Some shortcut functions that offer finer-grained control are:

tf compute term frequency weights

tfidf compute term frequency-inverse document frequency weights

docfreq compute document frequencies of features

dfm_weight(x, type = c("frequency", "relFreq", "relMaxFreq", "logFreq",
  "tfidf"), weights = NULL)

dfm_smooth(x, smoothing = 1)

Arguments

x
document-feature matrix created by dfm

type
a label of the weight type:

"frequency"

integer feature count (default when a dfm is created)

"relFreq"

the proportion of the feature counts of total feature counts (aka relative frequency)

"relMaxFreq"

the proportion of the feature counts of the highest feature count in a document

"logFreq"

take the logarithm of 1 + the feature count, for base 10

"tfidf"

Term-frequency * inverse document frequency. For a full explanation, see, for example, http://nlp.stanford.edu/IR-book/html/htmledition/term-frequency-and-weighting-1.html. This implementation will not return negative values. For finer-grained control, call tfidf directly.

weights
if type is unused, then weights can be a named numeric vector of weights to be applied to the dfm, where the names of the vector correspond to feature labels of the dfm, and the weights will be applied as multipliers to the existing feature counts for the corresponding named fatures. Any features not named will be assigned a weight of 1.0 (meaning they will be unchanged).

smoothing
constant added to the dfm cells for smoothing, default is 1

Value

The dfm with weighted values.

Details

This converts a matrix from sparse to dense format, so may exceed memory requirements depending on the size of your input matrix.

Note

For finer grained control, consider calling the convenience functions directly.

References

Manning, Christopher D., Prabhakar Raghavan, and Hinrich Schutze. Introduction to Information Retrieval. Vol. 1. Cambridge: Cambridge University Press, 2008.

See also

tf, tfidf, docfreq

Examples

dtm <- dfm(data_corpus_inaugural)

x <- apply(dtm, 1, function(tf) tf/max(tf))
topfeatures(dtm)

#> Error in get(".SigLength", envir = env): object '.SigLength' not found

normDtm <- dfm_weight(dtm, "relFreq")

#> Error in get(".SigLength", envir = env): object '.SigLength' not found

topfeatures(normDtm)

#> Error in topfeatures(normDtm): object 'normDtm' not found

maxTfDtm <- dfm_weight(dtm, type = "relMaxFreq")
topfeatures(maxTfDtm)

#> Error in get(".SigLength", envir = env): object '.SigLength' not found

logTfDtm <- dfm_weight(dtm, type = "logFreq")
topfeatures(logTfDtm)

#> Error in get(".SigLength", envir = env): object '.SigLength' not found

tfidfDtm <- dfm_weight(dtm, type = "tfidf")

#> Error in getMethod("t", "dgCMatrix"): no generic function found for 't'

topfeatures(tfidfDtm)

#> Error in topfeatures(tfidfDtm): object 'tfidfDtm' not found

# combine these methods for more complex dfm_weightings, e.g. as in Section 6.4
# of Introduction to Information Retrieval
head(logTfDtm <- dfm_weight(dtm, type = "logFreq"))

#> Document-feature matrix of: 58 documents, 9,232 features (91.6% sparse).
#> (showing first 6 documents and first 6 features)
#>                  features
#> docs                fellow        - citizens       of      the senate
#>   1789-Washington 1.477121 1.477121 1.698970 2.851258 3.064458      1
#>   1793-Washington 1.000000 0.000000 1.000000 2.041393 2.113943      0
#>   1797-Adams      1.477121 1.698970 1.698970 3.146128 3.212188      1
#>   1801-Jefferson  1.845098 2.079181 1.845098 3.017033 3.113943      0
#>   1805-Jefferson  1.903090 1.301030 2.000000 3.004321 3.155336      0
#>   1809-Madison    1.000000 1.778151 1.000000 2.838849 3.017033      0

head(tfidf(logTfDtm, normalize = FALSE))

#> Error in getMethod("t", "dgCMatrix"): no generic function found for 't'

#' # apply numeric weights
str <- c("apple is better than banana", "banana banana apple much better")
(mydfm <- dfm(str, remove = stopwords("english")))

#> Document-feature matrix of: 2 documents, 4 features (12.5% sparse).
#> 2 x 4 sparse Matrix of class "dfmSparse"
#>        features
#> docs    apple better banana much
#>   text1     1      1      1    0
#>   text2     1      1      2    1

dfm_weight(mydfm, weights = c(apple = 5, banana = 3, much = 0.5))

#> Document-feature matrix of: 2 documents, 4 features (12.5% sparse).
#> 2 x 4 sparse Matrix of class "dfmSparse"
#>        features
#> docs    apple better banana much
#>   text1     5      1      3  0  
#>   text2     5      1      6  0.5

# smooth the dfm
dfm_smooth(mydfm, 0.5)

#> Document-feature matrix of: 2 documents, 4 features (0% sparse).
#> 2 x 4 Matrix of class "dfmDense"
#>        features
#> docs    apple better banana much
#>   text1   1.5    1.5    1.5  0.5
#>   text2   1.5    1.5    2.5  1.5