compute keyness (internal functions)

Internal function used in textstat_keyness. Computes \(chi^2\) with Yates' continuity correction for 2x2 tables.

keyness_chi2_dt(x)

keyness_chi2_stats(x)

keyness_exact(x)

keyness_lr(x, correction = c("none", "Yates"))

Arguments

x

a dfm object

correction

if "Yates" implement the Yates correction for 2x2 tables, no correction if "none"

Value

named numeric vector containing the chi2 value for each feature

Details

keyness_chi2_dt uses vectorized computation from data.table objects.

keyness_chi2_stats uses element-by-element application of chisq.test.

keyness_exact computes Fisher's exact using element-by-element application of fisher.test.

keyness_lr computes the \(G^2\) likelihood ratio statistic using vectorized computation

References

https://en.wikipedia.org/wiki/Yates's_correction_for_continuity

http://influentialpoints.com/Training/g-likelihood_ratio_test.htm

Examples

mydfm <- dfm(c(d1 = "a a a b b c c c c c c d e f g h h",
               d2 = "a a b c c d d d d e f h"))
quanteda:::keyness_chi2_dt(mydfm)

#>           a           b           c           d           e           f 
#>  0.18510008  0.10251697  0.46729838 -2.04024714 -0.23758624 -0.23758624 
#>           g           h 
#>  0.03173144  0.10251697 

quanteda:::keyness_chi2_stats(mydfm)

#>             a             b             c             d             e 
#>  1.626059e-31  6.775245e-32  4.672984e-01 -2.040247e+00 -2.242191e-32 
#>             f             g             h 
#> -2.242191e-32  2.207170e-31  6.775245e-32 

quanteda:::keyness_exact(mydfm)

#>         a         b         c         d         e         f         g         h 
#> 1.0689130 1.4480473 2.6375984 0.1347360 0.6966446 0.6966446       Inf 1.4480473 

quanteda:::keyness_lr(mydfm)

#>            a            b            c            d            e            f 
#>  0.004751409  0.091241894  1.274238010 -3.779440533 -0.064900473 -0.064900473 
#>            g            h 
#>          NaN  0.091241894