locate keywords-in-context

For a text or a collection of texts (in a quanteda corpus object), return a list of a keyword supplied by the user in its immediate context, identifying the source text and the word index number within the source text. (Not the line number, since the text may or may not be segmented using end-of-line delimiters.)

kwic(x, keywords, window = 5, valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE, ...)

is.kwic(x)

# S3 method for kwic
as.tokens(x)

Arguments

x

a character, corpus, or tokens object

keywords

a keyword pattern or phrase consisting of multiple keyword patterns, possibly including punctuation. If a phrase, keywords will be tokenized using the ... options.

window

the number of context words to be displayed around the keyword.

valuetype

how to interpret keyword expressions: "glob" for "glob"-style wildcard expressions; "regex" for regular expressions; or "fixed" for exact matching. See valuetype for details.

case_insensitive

match without respect to case if TRUE

...

additional arguments passed to tokens, for applicable object types

Value

A kwic object classed data.frame, with the document name (docname), the token index position (position), the context before (contextPre), the keyword in its original format (keyword, preserving case and attached punctuation), and the context after (contextPost).

Examples

head(kwic(data_corpus_inaugural, "secure*", window = 3, valuetype = "glob"))

#>                                                                       
#>       [1797-Adams, 479]  welfare, and | secure  | the blessings of    
#>      [1797-Adams, 1517]  nations, and | secured | immortal glory with 
#>  [1805-Jefferson, 2368]   , and shall | secure  | to you the          
#>     [1817-Monroe, 1765] cherished. To | secure  | us against these    
#>     [1817-Monroe, 1825] defense as to | secure  | our cities and      
#>     [1817-Monroe, 3026]      I can to | secure  | economy and fidelity

head(kwic(data_corpus_inaugural, "secur", window = 3, valuetype = "regex"))

#>                                                                               
#>  [1789-Washington, 1499] government for the | security | of their union       
#>        [1797-Adams, 479]       welfare, and |  secure  | the blessings of     
#>       [1797-Adams, 1517]       nations, and | secured  | immortal glory with  
#>   [1805-Jefferson, 2368]        , and shall |  secure  | to you the           
#>      [1813-Madison, 321]       seas and the | security | of an important      
#>      [1817-Monroe, 1620]      may form some | security | against these dangers

head(kwic(data_corpus_inaugural, "security", window = 3, valuetype = "fixed"))

#>                                                         
#>  [1789-Washington, 1499] government for the | security |
#>      [1813-Madison, 321]       seas and the | security |
#>      [1817-Monroe, 1620]      may form some | security |
#>      [1817-Monroe, 3444]           and as a | security |
#>       [1825-Adams, 1373]      that the best | security |
#>       [1825-Adams, 1445]   that the firmest | security |
#>                         
#>  of their union         
#>  of an important        
#>  against these dangers  
#>  against foreign dangers
#>  for the beneficence    
#>  of peace is            

toks <- tokens(data_corpus_inaugural)
kwic(data_corpus_inaugural, "war against")

#>                                                                         
#>   [1857-Buchanan, 2933:2934] advantage of the fortune of | war against |
#>   [1901-McKinley, 2284:2285]         . We are not waging | war against |
#>   [1901-McKinley, 2299:2300]  portion of them are making | war against |
#>   [1901-McKinley, 2413:2414]    used when those who make | war against |
#>  [1933-Roosevelt, 1851:1852]   Executive power to wage a | war against |
#>                                   
#>  a sister republic, we            
#>  the inhabitants of the Philippine
#>  the United States. By            
#>  us shall make it no              
#>  the emergency, as great          

kwic(data_corpus_inaugural, "war against", valuetype = "regex")

#>                                                                
#>  [1801-Jefferson, 1277:1278] domestic concerns and the surest |
#>         [1845-Polk, 692:693] domestic concerns and the surest |
#>   [1857-Buchanan, 2933:2934]      advantage of the fortune of |
#>   [1901-McKinley, 2284:2285]              . We are not waging |
#>   [1901-McKinley, 2299:2300]       portion of them are making |
#>   [1901-McKinley, 2413:2414]         used when those who make |
#>  [1933-Roosevelt, 1851:1852]        Executive power to wage a |
#>       [1977-Carter, 927:928]            and we will fight our |
#>                                                                
#>  bulwarks against | antirepublican tendencies; the preservation
#>  bulwark against  | antirepublican tendencies," and            
#>    war against    | a sister republic, we                      
#>    war against    | the inhabitants of the Philippine          
#>    war against    | the United States. By                      
#>    war against    | us shall make it no                        
#>    war against    | the emergency, as great                    
#>    wars against   | poverty, ignorance, and                    

mykwic <- kwic(data_corpus_inaugural, "provident*")
is.kwic(mykwic)

#> [1] TRUE

is.kwic("Not a kwic")

#> [1] FALSE