get or set document-level meta-data

Get or set the document-level meta-data.

metadoc(x, field = NULL)

metadoc(x, field = NULL) <- value

Arguments

x
a corpus object

field
character, the name of the metadata field(s) to be queried or set

value
the new value of the new meta-data field

Value

For texts, a character vector of the texts in the corpus. For texts <-, the corpus with the updated texts.

Note

Document-level meta-data names are preceded by an underscore character, such as _language, but when named in in the field argument, do not need the underscore character.

Examples

mycorp <- corpus_subset(data_corpus_inaugural, Year > 1990)
summary(mycorp, showmeta = TRUE)

#> Corpus consisting of 7 documents.
#> 
#>          Text Types Tokens Sentences Year President FirstName
#>  1993-Clinton   644   1837        81 1993   Clinton      Bill
#>  1997-Clinton   773   2451       111 1997   Clinton      Bill
#>     2001-Bush   622   1810        97 2001      Bush George W.
#>     2005-Bush   772   2325       100 2005      Bush George W.
#>    2009-Obama   939   2729       110 2009     Obama    Barack
#>    2013-Obama   814   2335        88 2013     Obama    Barack
#>    2017-Trump   582   1662        88 2017     Trump Donald J.
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

metadoc(mycorp, "encoding") <- "UTF-8"
metadoc(mycorp)

#>              _encoding
#> 1993-Clinton     UTF-8
#> 1997-Clinton     UTF-8
#> 2001-Bush        UTF-8
#> 2005-Bush        UTF-8
#> 2009-Obama       UTF-8
#> 2013-Obama       UTF-8
#> 2017-Trump       UTF-8

metadoc(mycorp, "language") <- "english"
summary(mycorp, showmeta = TRUE)

#> Corpus consisting of 7 documents.
#> 
#>          Text Types Tokens Sentences Year President FirstName _encoding
#>  1993-Clinton   644   1837        81 1993   Clinton      Bill     UTF-8
#>  1997-Clinton   773   2451       111 1997   Clinton      Bill     UTF-8
#>     2001-Bush   622   1810        97 2001      Bush George W.     UTF-8
#>     2005-Bush   772   2325       100 2005      Bush George W.     UTF-8
#>    2009-Obama   939   2729       110 2009     Obama    Barack     UTF-8
#>    2013-Obama   814   2335        88 2013     Obama    Barack     UTF-8
#>    2017-Trump   582   1662        88 2017     Trump Donald J.     UTF-8
#>  _language
#>    english
#>    english
#>    english
#>    english
#>    english
#>    english
#>    english
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#>