get or set package options for quanteda

Get or set global options affecting functions across quatneda.

quanteda_options(..., reset = FALSE)

Arguments

...

options to be set, as key-value pair, same as options

reset

logical; if TRUE, reset all quanteda options to their default values

Value

When called using a key = value pair (where key can be a label or quoted character name)), the option is set and TRUE is returned invisibly.

When called with no arguments, a named list of the package options is returned.

When called with reset = TRUE as an argument, all arguments are options are reset to their default values, and TRUE is returned invisibly.

Details

Currently available options are:

verbose

logical; if TRUE then use this as the default for all functions with a verbose argument

threads

integer; specifies the number of threads to use in use this as the setting in all functions that uee parallelization

Examples

quanteda_options()

#> $threads
#> [1] 7
#> 
#> $verbose
#> [1] FALSE
#> 

quanteda_options(verbose = FALSE)
quanteda_options("verbose" = FALSE)
quanteda_options("threads")

#> [1] 7

## Not run: ------------------------------------
# quanteda_options(reset = TRUE) 
## ---------------------------------------------