Find variable-length collocations

This function automatically identifies contiguous collocations consisting of variable-length term sequences whose frequency is unlikely to have occurred by chance. The algorithm is based on Blaheta and Johnson's (2001) "Unsupervised Learning of Multi-Word Verbs".

sequences(x, min_count = 2, min_size = 2, max_size = 5,
  method = c("unigram", "all_subtuples"), nested = TRUE)

is.sequences(x)

Arguments

x
a tokens object

min_count
minimum frequency of sequences for which parameters are estimated

min_size
minimum length of sequences which are collected

max_size
maximum length of sequences which are collected

method
default is "unigram" and option is "all_subtuples"

nested
if TRUE, collect all the subsequences of a longer sequence as separate entities. e.g. in a sequence of capitalized words "United States Congress", "States Congress" is considered as a subsequence. But "United States" is not a subsequence because it is followed by "Congress".

Value

sequences returns TRUE if the object is of class sequences, FALSE otherwise.

References

Blaheta, D., & Johnson, M. (2001). Unsupervised learning of multi-word verbs. Presented at the ACLEACL Workshop on the Computational Extraction, Analysis and Exploitation of Collocations.

Examples

toks <- tokens(corpus_segment(data_corpus_inaugural, what = "sentence"), remove_punct=TRUE)
toks <- tokens_select(toks, stopwords("english"), "remove", padding = TRUE)
# extracting multi-part proper nouns (capitalized terms)
toks <- tokens_select(toks, "^([A-Z][a-z\\-]{2,})", valuetype="regex",
                     case_insensitive = FALSE, padding = TRUE)

seqs <- sequences(toks)
head(seqs, 10)

#>             collocation count length   lambda     sigma         z            p
#> 266       United States   155      2 6.329748 0.4286645 14.766205 0.000000e+00
#> 103  Federal Government    32      2 3.614841 0.3734731  9.678987 0.000000e+00
#> 169        Almighty God    15      2 6.697117 0.8323334  8.046195 4.440892e-16
#> 191       Chief Justice     7      2 5.438884 0.7454415  7.296191 1.479927e-13
#> 284       South America     6      2 5.741934 0.8975573  6.397290 7.907930e-11
#> 200    Chief Magistrate    10      2 5.476210 0.9133139  5.995978 1.011323e-09
#> 239       Supreme Court     4      2 6.396374 1.1635881  5.497112 1.930311e-08
#> 115 National Government    11      2 2.745355 0.5173342  5.306734 5.580353e-08
#> 269           Old World    10      2 8.147660 1.5467865  5.267476 6.915622e-08
#> 74    Divine Providence     3      2 5.549076 1.0961553  5.062308 2.071052e-07

# to return only trigrams
seqs <- sequences(toks, min_size = 3, max_size = 3)
head(seqs, 10)

#>                   collocation count length    lambda    sigma        z
#> 8      United States Congress     2      3 10.306583 2.319391 4.443659
#> 45        Vice President Bush     2      3  8.105361 2.200615 3.683225
#> 80    Chief Justice President     2      3  7.504587 2.188510 3.429085
#> 71   President Vice President     2      3  5.444652 1.736877 3.134738
#> 11   President Bush President     2      3  5.134497 1.722863 2.980213
#> 32 President Carter President     2      3  4.898109 1.714697 2.856545
#>               p
#> 8  4.422095e-06
#> 45 1.151510e-04
#> 80 3.028102e-04
#> 71 8.600395e-04
#> 11 1.440240e-03
#> 32 2.141397e-03