find variable-length collocations with filtering

This function automatically identifies contiguous collocations consisting of variable-length term sequences whose frequency is unlikey to have occurred by chance. The algorithm is based on Blaheta and Johnson's "Unsupervised Learning of Multi-Word Verbs".

sequences(x, features, valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE, min_count = 2, max_length = 5, nested = TRUE,
  ordered = FALSE)

Arguments

x

a tokens object

features

a regular expression for filtering the features to be located in sequences

valuetype

how to interpret keyword expressions: "glob" for "glob"-style wildcard expressions; "regex" for regular expressions; or "fixed" for exact matching. See valuetype for details.

case_insensitive

ignore case when matching, if TRUE

min_count

minimum frequency of sequences for which parameters are estimated

max_length

maxium length of sequences which are collected

nested

if true, collect all the subsequences of a longer sequence as separate entities. e.g. in a sequence of capitalized words "United States Congress", "States Congress" is considered as a subsequence. But "United States" is not a subsequence because it is followed by "Congress".

ordered

if true, use the Blaheta-Johnson method that distinguishs between the order of words, and tends to promote rare sequences.

References

Blaheta, D., & Johnson, M. (2001). http://web.science.mq.edu.au/~mjohnson/papers/2001/dpb-colloc01.pdf. Presented at the ACLEACL Workshop on the Computational Extraction, Analysis and Exploitation of Collocations.

Examples

toks <- tokens(corpus_segment(data_corpus_inaugural, what = "sentence"))
toks <- tokens_select(toks, stopwords("english"), "remove", padding = TRUE)

# extracting multi-part proper nouns (capitalized terms)
seqs <- sequences(toks, "^([A-Z][a-z\\-]{2,})", valuetype="regex", case_insensitive = FALSE)
head(seqs, 10)

#>               sequence   lambda      sigma count length         z p
#> 1        United States 20.56459 0.04324223   155      2 475.56728 0
#> 2   Federal Government 16.07807 0.09312539    32      2 172.64966 0
#> 3   General Government 14.50954 0.12914263    24      2 112.35286 0
#> 4  National Government 12.88826 0.15847837    11      2  81.32504 0
#> 5         Almighty God 13.10835 0.16301521    15      2  80.41181 0
#> 6        Chief Justice 12.17009 0.18109481    11      2  67.20288 0
#> 7       President Bush 11.87632 0.18741675     6      2  63.36850 0
#> 8     Chief Magistrate 11.37940 0.22063436    10      2  51.57583 0
#> 9    President Clinton 10.67352 0.22913765     4      2  46.58124 0
#> 10      United Nations 10.72087 0.24598156     8      2  43.58402 0

# types can be any words
seqs2 <- sequences(toks, "^([a-z]+)$", valuetype="regex", case_insensitive = FALSE,
                   min_count = 2, ordered = TRUE)
head(seqs2, 10)

#>                sequence   lambda    sigma count length        z p
#> 1       fellow citizens 55.46514 1.015421    55      2 54.62278 0
#> 2             years ago 55.27680 1.035780    24      2 53.36732 0
#> 3            four years 52.49221 1.045142    17      2 50.22496 0
#> 4           one another 51.30887 1.074872    14      2 47.73487 0
#> 5            go forward 50.82484 1.090404     6      2 46.61100 0
#> 6         every citizen 47.42716 1.033517    13      2 45.88908 0
#> 7  executive department 52.43918 1.147450     6      2 45.70060 0
#> 8            good faith 48.93755 1.078625     7      2 45.37030 0
#> 9     religious liberty 50.67227 1.123021     6      2 45.12140 0
#> 10         public money 47.22302 1.049822     8      2 44.98191 0