find variable-length collocations with filtering

This function automatically identifies contiguous collocations consisting of variable-length term sequences whose frequency is unlikey to have occurred by chance. The algorithm is based on Blaheta and Johnson's "Unsupervised Learning of Multi-Word Verbs".

sequences(x, features, valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE, count_min = 2, nested = TRUE)

Arguments

x

a tokens object

features

a regular expression for filtering the features to be located in sequences

valuetype

how to interpret keyword expressions: "glob" for "glob"-style wildcard expressions; "regex" for regular expressions; or "fixed" for exact matching. See valuetype for details.

case_insensitive

ignore case when matching, if TRUE

count_min

minimum frequency of sequences

nested

collect nested sequences

References

Blaheta, D., & Johnson, M. (2001). http://web.science.mq.edu.au/~mjohnson/papers/2001/dpb-colloc01.pdf. Presented at the ACLEACL Workshop on the Computational Extraction, Analysis and Exploitation of Collocations.

Examples

toks <- tokens(corpus_segment(data_corpus_inaugural, what = "sentence"))
toks <- tokens_select(toks, stopwords("english"), "remove", padding = TRUE)

# extracting multi-part proper nouns (capitalized terms)
seqs <- sequences(toks, "^([A-Z][a-z\\-]{2,})", valuetype="regex", case_insensitive = FALSE)
head(seqs, 10)

#>               sequence   lambda      sigma count length          z            p
#> 1        United States 4.912248 0.03397118   152      2 144.600475 0.000000e+00
#> 2   Federal Government 3.455271 0.07315512    32      2  47.232120 0.000000e+00
#> 3   General Government 2.460985 0.11068747    24      2  22.233638 0.000000e+00
#> 4  National Government 2.397005 0.12440066    11      2  19.268430 0.000000e+00
#> 5       President Bush 2.348624 0.14723627     5      2  15.951398 0.000000e+00
#> 6         Almighty God 1.998062 0.13963675    15      2  14.308997 0.000000e+00
#> 7        Chief Justice 1.647330 0.16313666    11      2  10.097851 0.000000e+00
#> 8    President Clinton 1.850778 0.18943641     3      2   9.769916 0.000000e+00
#> 9        Federal Union 1.490737 0.19251392     5      2   7.743529 4.884981e-15
#> 10    Chief Magistrate 1.270018 0.19257293    10      2   6.594995 2.126355e-11

# types can be any words
seqs2 <- sequences(toks, "^([a-z]+)$", valuetype="regex", case_insensitive = FALSE, count_min = 10)
head(seqs2, 10)

#>           sequence   lambda      sigma count length        z p
#> 1          upon us 6.806521 0.01873042    21      2 363.3939 0
#> 2        good will 6.530191 0.02150380    14      2 303.6762 0
#> 3    will continue 6.428101 0.02287065    10      2 281.0633 0
#> 4  fellow citizens 6.300411 0.02396993    55      2 262.8464 0
#> 5    every citizen 6.217178 0.02527487    13      2 245.9825 0
#> 6    national life 5.992374 0.02836887    13      2 211.2306 0
#> 7      public debt 5.660241 0.03333407    10      2 169.8035 0
#> 8  foreign nations 5.648590 0.03341090    15      2 169.0643 0
#> 9      one another 5.571175 0.03458396    14      2 161.0913 0
#> 10      first time 5.417941 0.03753724    11      2 144.3351 0