compute similarities between documents and/or features

Compute similarities between documents and/or features from a dfm. Uses the similarity measures defined in simil. See pr_DB for available distance measures, or how to create your own.

similarity(x, selection = NULL, n = NULL, margin = c("documents",
  "features"), method = "correlation", sorted = TRUE, normalize = FALSE)

# S4 method for dfm
similarity(x, selection = NULL, n = NULL,
  margin = c("documents", "features"), method = "correlation",
  sorted = TRUE, normalize = FALSE)

# S3 method for similMatrix
as.matrix(x, ...)

Arguments

x

a dfm object

selection

character or character vector of document names or feature labels from the dfm

n

the top n most similar items will be returned, sorted in descending order. If n is NULL, return all items.

margin

identifies the margin of the dfm on which similarity will be computed: documents for documents or features for word/term features.

method

a valid method for computing similarity from pr_DB

sorted

sort results in descending order if TRUE

normalize

a deprecated argument retained (temporarily) for legacy reasons. If you want to compute similarity on a "normalized" dfm objects (e.g. x), wrap it in weight(x, "relFreq").

...

unused

Value

a named list of the selection labels, with a sorted named vector of similarity measures.

Note

The method for computing feature similarities can be quite slow when there are large numbers of feature types. Future implementations will hopefully speed this up.

Examples

# create a dfm from inaugural addresses from Reagan onwards
presDfm <- dfm(corpus_subset(data_corpus_inaugural, Year > 1980), stem = TRUE,
               remove = stopwords("english"))

#> 

# compute some document similarities
(tmp <- similarity(presDfm, margin = "documents"))

#> similarity Matrix:
#> $`1981-Reagan`
#>  1985-Reagan    1989-Bush    2001-Bush 1993-Clinton   2013-Obama   2009-Obama 
#>    0.9685409    0.9599217    0.9498797    0.9391137    0.9389438    0.9387730 
#> 1997-Clinton    2005-Bush 
#>    0.9386791    0.9346990 
#> 
#> $`1985-Reagan`
#>  1981-Reagan    1989-Bush 1997-Clinton 1993-Clinton   2009-Obama   2013-Obama 
#>    0.9685409    0.9537449    0.9468983    0.9452609    0.9422163    0.9409353 
#>    2001-Bush    2005-Bush 
#>    0.9371453    0.9297108 
#> 
#> $`1989-Bush`
#>  1981-Reagan    2001-Bush  1985-Reagan   2009-Obama 1997-Clinton   2013-Obama 
#>    0.9599217    0.9543776    0.9537449    0.9407657    0.9381904    0.9370933 
#> 1993-Clinton    2005-Bush 
#>    0.9358174    0.9349331 
#> 
#> $`1993-Clinton`
#>  1985-Reagan  1981-Reagan    1989-Bush    2001-Bush 1997-Clinton    2005-Bush 
#>    0.9452609    0.9391137    0.9358174    0.9258829    0.9241745    0.9142198 
#>   2013-Obama   2009-Obama 
#>    0.9118394    0.9000505 
#> 
#> $`1997-Clinton`
#>  1985-Reagan   2013-Obama  1981-Reagan   2009-Obama    1989-Bush    2001-Bush 
#>    0.9468983    0.9387808    0.9386791    0.9382904    0.9381904    0.9321614 
#> 1993-Clinton    2005-Bush 
#>    0.9241745    0.9077993 
#> 
#> $`2001-Bush`
#>    1989-Bush  1981-Reagan    2005-Bush  1985-Reagan   2013-Obama 1997-Clinton 
#>    0.9543776    0.9498797    0.9437295    0.9371453    0.9363852    0.9321614 
#> 1993-Clinton   2009-Obama 
#>    0.9258829    0.9240388 
#> 
#> $`2005-Bush`
#>    2001-Bush    1989-Bush  1981-Reagan  1985-Reagan   2013-Obama 1993-Clinton 
#>    0.9437295    0.9349331    0.9346990    0.9297108    0.9195065    0.9142198 
#> 1997-Clinton   2009-Obama 
#>    0.9077993    0.8976144 
#> 
#> $`2009-Obama`
#>   2013-Obama  1985-Reagan    1989-Bush  1981-Reagan 1997-Clinton    2001-Bush 
#>    0.9458185    0.9422163    0.9407657    0.9387730    0.9382904    0.9240388 
#> 1993-Clinton    2005-Bush 
#>    0.9000505    0.8976144 
#> 
#> $`2013-Obama`
#>   2009-Obama  1985-Reagan  1981-Reagan 1997-Clinton    1989-Bush    2001-Bush 
#>    0.9458185    0.9409353    0.9389438    0.9387808    0.9370933    0.9363852 
#>    2005-Bush 1993-Clinton 
#>    0.9195065    0.9118394 
#> 

# output as a matrix
as.matrix(tmp)

#>              1981-Reagan 1985-Reagan 1989-Bush 1993-Clinton 1997-Clinton
#> 1981-Reagan    1.0000000   0.9685409 0.9599217    0.9391137    0.9386791
#> 1985-Reagan    0.9685409   1.0000000 0.9537449    0.9452609    0.9468983
#> 1989-Bush      0.9599217   0.9537449 1.0000000    0.9358174    0.9381904
#> 2001-Bush      0.9498797   0.9371453 0.9543776    0.9258829    0.9321614
#> 1993-Clinton   0.9391137   0.9452609 0.9358174    1.0000000    0.9241745
#> 2013-Obama     0.9389438   0.9409353 0.9370933    0.9118394    0.9387808
#> 2009-Obama     0.9387730   0.9422163 0.9407657    0.9000505    0.9382904
#> 1997-Clinton   0.9386791   0.9468983 0.9381904    0.9241745    1.0000000
#> 2005-Bush      0.9346990   0.9297108 0.9349331    0.9142198    0.9077993
#>              2001-Bush 2005-Bush 2009-Obama 2013-Obama
#> 1981-Reagan  0.9498797 0.9346990  0.9387730  0.9389438
#> 1985-Reagan  0.9371453 0.9297108  0.9422163  0.9409353
#> 1989-Bush    0.9543776 0.9349331  0.9407657  0.9370933
#> 2001-Bush    1.0000000 0.9437295  0.9240388  0.9363852
#> 1993-Clinton 0.9258829 0.9142198  0.9000505  0.9118394
#> 2013-Obama   0.9363852 0.9195065  0.9458185  1.0000000
#> 2009-Obama   0.9240388 0.8976144  1.0000000  0.9458185
#> 1997-Clinton 0.9321614 0.9077993  0.9382904  0.9387808
#> 2005-Bush    0.9437295 1.0000000  0.8976144  0.9195065

# for specific comparisons
similarity(presDfm, "1985-Reagan", n = 5, margin = "documents")

#> similarity Matrix:
#> $`1985-Reagan`
#>  1981-Reagan    1989-Bush 1997-Clinton 1993-Clinton   2009-Obama 
#>    0.9685409    0.9537449    0.9468983    0.9452609    0.9422163 
#> 

similarity(presDfm, c("2009-Obama" , "2013-Obama"), n = 5, margin = "documents")

#> similarity Matrix:
#> $`2009-Obama`
#>   2013-Obama  1985-Reagan    1989-Bush  1981-Reagan 1997-Clinton 
#>    0.9458185    0.9422163    0.9407657    0.9387730    0.9382904 
#> 
#> $`2013-Obama`
#>   2009-Obama  1985-Reagan  1981-Reagan 1997-Clinton    1989-Bush 
#>    0.9458185    0.9409353    0.9389438    0.9387808    0.9370933 
#> 

similarity(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents")

#> similarity Matrix:
#> $`2009-Obama`
#>   2013-Obama  1985-Reagan    1989-Bush  1981-Reagan 1997-Clinton    2001-Bush 
#>    0.9458185    0.9422163    0.9407657    0.9387730    0.9382904    0.9240388 
#> 1993-Clinton    2005-Bush 
#>    0.9000505    0.8976144 
#> 
#> $`2013-Obama`
#>   2009-Obama  1985-Reagan  1981-Reagan 1997-Clinton    1989-Bush    2001-Bush 
#>    0.9458185    0.9409353    0.9389438    0.9387808    0.9370933    0.9363852 
#>    2005-Bush 1993-Clinton 
#>    0.9195065    0.9118394 
#> 

similarity(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents", method = "cosine")

#> similarity Matrix:
#> $`2009-Obama`
#>   2013-Obama  1985-Reagan    1989-Bush  1981-Reagan 1997-Clinton    2001-Bush 
#>    0.9473229    0.9436339    0.9416945    0.9399460    0.9399306    0.9255036 
#> 1993-Clinton    2005-Bush 
#>    0.9015783    0.9004034 
#> 
#> $`2013-Obama`
#>   2009-Obama  1985-Reagan 1997-Clinton  1981-Reagan    1989-Bush    2001-Bush 
#>    0.9473229    0.9422001    0.9403480    0.9397146    0.9376157    0.9371368 
#>    2005-Bush 1993-Clinton 
#>    0.9217499    0.9125611 
#> 

similarity(presDfm, "2005-Bush", margin = "documents", method = "eJaccard", sorted = FALSE)

#> similarity Matrix:
#> $`2005-Bush`
#>  1981-Reagan  1985-Reagan    1989-Bush 1993-Clinton 1997-Clinton    2001-Bush 
#>    0.8094123    0.7992198    0.8102824    0.8441215    0.8252637    0.8887129 
#>   2009-Obama   2013-Obama 
#>    0.8051778    0.8416028 
#> 

# compute some term similarities
similarity(presDfm, c("fair", "health", "terror"), method="cosine", margin = "features", 20)

#> similarity Matrix:
#> $fair
#>   economi     begin      mani jefferson    author     howev     faith      call 
#> 0.9080252 0.9075951 0.9038769 0.8981462 0.8944272 0.8944272 0.8866586 0.8736360 
#>       god   struggl     order     never    courag      best     creat    believ 
#> 0.8687784 0.8660254 0.8606630 0.8516884 0.8391464 0.8366600 0.8347300 0.8322504 
#>      much      well     pledg   compass 
#> 0.8315668 0.8315668 0.8292884 0.8280787 
#> 
#> $terror
#>    factori  adversari commonplac     miracl     racial     bounti     martin 
#>  0.9525793  0.9525793  0.9428090  0.9428090  0.9428090  0.9428090  0.9428090 
#>   guarante       solv    potenti      solut      whose     cultur   maintain 
#>  0.8944272  0.8944272  0.8944272  0.8888889  0.8845380  0.8819171  0.8728716 
#>       upon      dream       told      polit   industri       open 
#>  0.8699502  0.8677218  0.8660254  0.8540168  0.8432740  0.8432740 
#> 
#> $health
#>  knowledg     shape   generat     wrong     defin    common     child      fear 
#> 0.9428090 0.9045340 0.8971180 0.8944272 0.8892973 0.8888889 0.8888889 0.8888889 
#>    demand    planet     power      task     everi       eye      forc      even 
#> 0.8845380 0.8819171 0.8796317 0.8728716 0.8682693 0.8660254 0.8641586 0.8606630 
#>      long      born    danger     choos 
#> 0.8571594 0.8432740 0.8432740 0.8432740 
#>