summarize a corpus or a vector of texts

Displays information about a corpus or vector of texts. For a corpus, this includes attributes and metadata such as date of number of texts, creation and source. For texts, prints to the console a desription of the texts, including number of types, tokens, and sentences.

# S3 method for character
summary(object, n = 100, verbose = TRUE,
  toLower = FALSE, ...)

# S3 method for corpus
summary(object, n = 100, verbose = TRUE,
  showmeta = FALSE, toLower = FALSE, ...)

Arguments

object

corpus or texts to be summarized

n

maximum number of texts to describe, default=100

verbose

set to FALSE to turn off printed output, for instance if you simply want to assign the output to a data.frame

toLower

convert texts to lower case before counting types

...

additional arguments passed through to tokenize

showmeta

for a corpus, set to TRUE to include document-level meta-data

Examples

# summarize texts
summary(c("Testing this text.  Second sentence.", "And this one."))

#>    Text Types Tokens Sentences
#> 1 text1     6      7         2
#> 2 text2     4      4         1

summary(data_char_ukimmig2010)

#>           Text Types Tokens Sentences
#> 1          BNP  1126   3330        88
#> 2    Coalition   144    268         4
#> 3 Conservative   252    503        15
#> 4       Greens   325    687        21
#> 5       Labour   296    703        29
#> 6       LibDem   257    499        14
#> 7           PC    80    118         5
#> 8          SNP    90    136         4
#> 9         UKIP   346    739        27

myTextSummaryDF <- summary(data_char_ukimmig2010, verbose = FALSE)
head(myTextSummaryDF)

#>           Text Types Tokens Sentences
#> 1          BNP  1126   3330        88
#> 2    Coalition   144    268         4
#> 3 Conservative   252    503        15
#> 4       Greens   325    687        21
#> 5       Labour   296    703        29
#> 6       LibDem   257    499        14

# summarize corpus information
summary(data_corpus_inaugural)

#> Corpus consisting of 58 documents.
#> 
#>             Text Types Tokens Sentences Year  President       FirstName
#>  1789-Washington   626   1540        23 1789 Washington          George
#>  1793-Washington    96    147         4 1793 Washington          George
#>       1797-Adams   826   2584        37 1797      Adams            John
#>   1801-Jefferson   716   1935        41 1801  Jefferson          Thomas
#>   1805-Jefferson   804   2381        45 1805  Jefferson          Thomas
#>     1809-Madison   536   1267        21 1809    Madison           James
#>     1813-Madison   542   1304        33 1813    Madison           James
#>      1817-Monroe  1040   3696       121 1817     Monroe           James
#>      1821-Monroe  1262   4898       129 1821     Monroe           James
#>       1825-Adams  1004   3154        74 1825      Adams     John Quincy
#>     1829-Jackson   517   1210        25 1829    Jackson          Andrew
#>     1833-Jackson   499   1271        29 1833    Jackson          Andrew
#>    1837-VanBuren  1315   4175        95 1837  Van Buren          Martin
#>    1841-Harrison  1893   9178       210 1841   Harrison   William Henry
#>        1845-Polk  1330   5211       153 1845       Polk      James Knox
#>      1849-Taylor   497   1185        22 1849     Taylor         Zachary
#>      1853-Pierce  1166   3657       104 1853     Pierce        Franklin
#>    1857-Buchanan   945   3106        89 1857   Buchanan           James
#>     1861-Lincoln  1075   4016       135 1861    Lincoln         Abraham
#>     1865-Lincoln   362    780        26 1865    Lincoln         Abraham
#>       1869-Grant   486   1243        40 1869      Grant      Ulysses S.
#>       1873-Grant   552   1479        43 1873      Grant      Ulysses S.
#>       1877-Hayes   829   2730        59 1877      Hayes   Rutherford B.
#>    1881-Garfield  1018   3240       111 1881   Garfield        James A.
#>   1885-Cleveland   674   1828        44 1885  Cleveland          Grover
#>    1889-Harrison  1355   4744       157 1889   Harrison        Benjamin
#>   1893-Cleveland   823   2135        58 1893  Cleveland          Grover
#>    1897-McKinley  1236   4383       130 1897   McKinley         William
#>    1901-McKinley   857   2449       100 1901   McKinley         William
#>   1905-Roosevelt   404   1089        33 1905  Roosevelt        Theodore
#>        1909-Taft  1436   5844       159 1909       Taft  William Howard
#>      1913-Wilson   661   1896        68 1913     Wilson         Woodrow
#>      1917-Wilson   549   1656        59 1917     Wilson         Woodrow
#>     1921-Harding  1172   3743       148 1921    Harding       Warren G.
#>    1925-Coolidge  1221   4442       196 1925   Coolidge          Calvin
#>      1929-Hoover  1086   3895       158 1929     Hoover         Herbert
#>   1933-Roosevelt   744   2064        85 1933  Roosevelt     Franklin D.
#>   1937-Roosevelt   729   2027        96 1937  Roosevelt     Franklin D.
#>   1941-Roosevelt   527   1552        68 1941  Roosevelt     Franklin D.
#>   1945-Roosevelt   276    651        26 1945  Roosevelt     Franklin D.
#>      1949-Truman   781   2531       116 1949     Truman        Harry S.
#>  1953-Eisenhower   903   2765       119 1953 Eisenhower       Dwight D.
#>  1957-Eisenhower   621   1933        92 1957 Eisenhower       Dwight D.
#>     1961-Kennedy   566   1568        52 1961    Kennedy         John F.
#>     1965-Johnson   569   1725        93 1965    Johnson   Lyndon Baines
#>       1969-Nixon   743   2437       103 1969      Nixon Richard Milhous
#>       1973-Nixon   545   2018        68 1973      Nixon Richard Milhous
#>      1977-Carter   528   1380        52 1977     Carter           Jimmy
#>      1981-Reagan   904   2798       128 1981     Reagan          Ronald
#>      1985-Reagan   925   2935       123 1985     Reagan          Ronald
#>        1989-Bush   795   2683       141 1989       Bush          George
#>     1993-Clinton   644   1837        81 1993    Clinton            Bill
#>     1997-Clinton   773   2451       111 1997    Clinton            Bill
#>        2001-Bush   622   1810        97 2001       Bush       George W.
#>        2005-Bush   772   2325       100 2005       Bush       George W.
#>       2009-Obama   939   2729       110 2009      Obama          Barack
#>       2013-Obama   814   2335        88 2013      Obama          Barack
#>       2017-Trump   582   1662        88 2017      Trump       Donald J.
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

summary(data_corpus_inaugural, n=10)

#> Corpus consisting of 58 documents, showing 10 documents.
#> 
#>             Text Types Tokens Sentences Year  President   FirstName
#>  1789-Washington   626   1540        23 1789 Washington      George
#>  1793-Washington    96    147         4 1793 Washington      George
#>       1797-Adams   826   2584        37 1797      Adams        John
#>   1801-Jefferson   716   1935        41 1801  Jefferson      Thomas
#>   1805-Jefferson   804   2381        45 1805  Jefferson      Thomas
#>     1809-Madison   536   1267        21 1809    Madison       James
#>     1813-Madison   542   1304        33 1813    Madison       James
#>      1817-Monroe  1040   3696       121 1817     Monroe       James
#>      1821-Monroe  1262   4898       129 1821     Monroe       James
#>       1825-Adams  1004   3154        74 1825      Adams John Quincy
#> 
#> Source:  /home/paul/Dropbox/code/quanteda/* on x86_64 by paul
#> Created: Fri Sep 12 12:41:17 2014
#> Notes:   
#> 

mycorpus <- corpus(data_char_ukimmig2010,
                   docvars = data.frame(party=names(data_char_ukimmig2010)))
summary(mycorpus, showmeta=TRUE)  # show the meta-data

#> Corpus consisting of 9 documents.
#> 
#>          Text Types Tokens Sentences        party
#>           BNP  1126   3330        88          BNP
#>     Coalition   144    268         4    Coalition
#>  Conservative   252    503        15 Conservative
#>        Greens   325    687        21       Greens
#>        Labour   296    703        29       Labour
#>        LibDem   257    499        14       LibDem
#>            PC    80    118         5           PC
#>           SNP    90    136         4          SNP
#>          UKIP   346    739        27         UKIP
#> 
#> Source:  /Users/kbenoit/Dropbox (Personal)/GitHub/quanteda/docs/reference/* on x86_64 by kbenoit
#> Created: Fri Mar 10 19:12:06 2017
#> Notes:   
#> 

mysummary <- summary(mycorpus, verbose=FALSE)  # (quietly) assign the results
mysummary$Types / mysummary$Tokens             # crude type-token ratio

#> [1] 0.3381381 0.5373134 0.5009940 0.4730713 0.4210526 0.5150301 0.6779661
#> [8] 0.6617647 0.4682003