Naive Bayes classifier for texts

Currently working for vectors of texts -- not specially defined for a dfm.

textmodel_NB(x, y, smooth = 1, prior = c("uniform", "docfreq", "termfreq"),
  distribution = c("multinomial", "Bernoulli"), ...)

Arguments

x
the dfm on which the model will be fit. Does not need to contain only the training documents.

y
vector of training labels associated with each document identified in train. (These will be converted to factors if not already factors.)

smooth
smoothing parameter for feature counts by class

prior
prior distribution on texts, see details

distribution
count model for text features, can be multinomial or Bernoulli

...
more arguments passed through

Value

A list of return values, consisting of:

call

original function call

PwGc

probability of the word given the class (empirical likelihood)

Pc

class prior probability

PcGw

posterior class probability given the word

Pw

baseline probability of the word

data

list consisting of x training class, and y test class

distribution

the distribution argument

prior

argument passed as a prior

smooth

smoothing parameter

Details

This naive Bayes model works on word counts, with smoothing.

Predict Methods

A predict method is also available for a fitted Naive Bayes object, see predict.textmodel_NB_fitted.

Examples

## Example from 13.1 of _An Introduction to Information Retrieval_
trainingset <- as.dfm(matrix(c(1, 2, 0, 0, 0, 0,
                        0, 2, 0, 0, 1, 0,
                        0, 1, 0, 1, 0, 0,
                        0, 1, 1, 0, 0, 1,
                        0, 3, 1, 0, 0, 1),
                      ncol=6, nrow=5, byrow=TRUE,
                      dimnames = list(docs = paste("d", 1:5, sep = ""),
                                      features = c("Beijing", "Chinese",  "Japan", "Macao",
                                                   "Shanghai", "Tokyo"))))
trainingclass <- factor(c("Y", "Y", "Y", "N", NA), ordered = TRUE)
## replicate IIR p261 prediction for test set (document 5)
(nb.p261 <- textmodel_NB(trainingset, trainingclass))

#> Fitted Naive Bayes model:
#> Call:
#> 	textmodel_NB(x = trainingset, y = trainingclass)
#> 
#> 
#> Training classes and priors:
#>   Y   N 
#> 0.5 0.5 
#> 
#> 		  Likelihoods:		Class Posteriors:
#> 6 x 4 Matrix of class "dgeMatrix"
#>                   Y         N         Y         N
#> Beijing  0.14285714 0.1111111 0.5625000 0.4375000
#> Chinese  0.42857143 0.2222222 0.6585366 0.3414634
#> Japan    0.07142857 0.2222222 0.2432432 0.7567568
#> Macao    0.14285714 0.1111111 0.5625000 0.4375000
#> Shanghai 0.14285714 0.1111111 0.5625000 0.4375000
#> Tokyo    0.07142857 0.2222222 0.2432432 0.7567568
#> 

predict(nb.p261, newdata = trainingset[5, ])

#> Predicted textmodel of type: Naive Bayes
#> 
#>        lp(Y)     lp(N)     Pr(Y)  Pr(N) Predicted
#> d5 -8.513155 -8.213534    0.4257 0.5743         N
#> 

# contrast with other priors
predict(textmodel_NB(trainingset, trainingclass, prior = "docfreq"))

#> Predicted textmodel of type: Naive Bayes
#> 
#>        lp(Y)     lp(N)     Pr(Y)  Pr(N) Predicted
#> d1 -3.928188 -6.591674    0.9348 0.0652         Y
#> d2 -3.928188 -6.591674    0.9348 0.0652         Y
#> d3 -3.080890 -5.087596    0.8815 0.1185         Y
#> d4 -6.413095 -5.898527    0.3741 0.6259         N
#> d5 -8.107690 -8.906681    0.6898 0.3102         Y
#> 

predict(textmodel_NB(trainingset, trainingclass, prior = "termfreq"))

#> Predicted textmodel of type: Naive Bayes
#> 
#>        lp(Y)     lp(N)     Pr(Y)  Pr(N) Predicted
#> d1 -3.958960 -6.504662    0.9273 0.0727         Y
#> d2 -3.958960 -6.504662    0.9273 0.0727         Y
#> d3 -3.111662 -5.000585    0.8686 0.1314         Y
#> d4 -6.443866 -5.811515    0.3470 0.6530         N
#> d5 -8.138462 -8.819670    0.6640 0.3360         Y
#>