class affinity maximum likelihood text scaling model

textmodel_affinity implements the maximum likelihood supervised text scaling method described in Perry and Benoit (2017).

textmodel_affinity(x, y, exclude = NULL, smooth = 0.5, ref_smooth = 0.5,
  verbose = FALSE)

Arguments

x
the dfm or bootstrap_dfm object on which the model will be fit. Does not need to contain only the training documents, since the index of these will be matched automatically.

y
vector of training classes/scores associated with each document identified in data

exclude
a set of words to exclude from the model

smooth
a smoothing parameter for class affinities; defaults to 0.5 (Jeffreys prior). A plausible alternative would be 1.0 (Laplace prior).

ref_smooth
a smoothing parameter for token distributions; defaults to 0.5

verbose
logical; if TRUE print diagnostic information during fitting.

References

Perry, Patrick O. and Kenneth Benoit. (2017) "Scaling Text with the Class Affinity Model". arXiv:1710.08963 [stat.ML].

Examples

(fitted <- textmodel_affinity(data_dfm_lbgexample, y = c("L", NA, NA, NA, "R", NA)))

#> Call:	
#> textmodel_affinity.dfm(x = data_dfm_lbgexample, y = c("L", NA, 
#>     NA, NA, "R", NA))
#> 
#>     Training documents per class: L: 1, R: 1
#>          Total training features: 37

predict(fitted)

#> Predicted textmodel of type: affinity
#> 
#> Estimated coefficients:
#> 
#>          L    s.e.       R    s.e.    chi2
#> R1 0.99950 0.00071 0.00050 0.00071     9.3
#> R2 0.99941 0.00083 0.00059 0.00083  9415.8
#> R3 0.50000 0.02734 0.50000 0.02734 24864.5
#> R4 0.00059 0.00083 0.99941 0.00083  9415.8
#> R5 0.00050 0.00071 0.99950 0.00071     9.3
#> V1 0.99867 0.00187 0.00133 0.00187 19458.5
#> 
#> Some diagnostics here about how many words were not found in training vocabulary.

predict(fitted, newdata = data_dfm_lbgexample[6, ])

#> Predicted textmodel of type: affinity
#> 
#> Estimated coefficients:
#> 
#>    L   s.e.      R   s.e.  chi2
#> V1 1 0.0019 0.0013 0.0019 19459
#> 
#> Some diagnostics here about how many words were not found in training vocabulary.

# compute bootstrapped SEs
bsdfm <- bootstrap_dfm(data_corpus_dailnoconf1991, n = 10, remove_punct = TRUE)
textmodel_affinity(bsdfm, y = c("Govt", "Opp", "Opp", rep(NA, 55)))

#> Bootstrapping textmodel_affinity for 58 documents:

#>    ...computing model for dfm from original texts

#>    ...computing bootstrapped models and saving coefficients
#>      

#>  1

#>  2

#>  3

#>  4

#>  5

#>  6

#>  7

#>  8

#>  9

#>  10

#>  11

#>    ...replacing original SEs with bootstrapped SEs

#>    ...finished.

#> Predicted textmodel of type: affinity
#> (showing 30 of 58 documents)
#> 
#> Estimated coefficients:
#> 
#>                              Govt    s.e.     Opp    s.e.  chi2
#> Haughey_FF_Taois.txt      0.99955 3.2e-05 0.00045 3.2e-05   874
#> Spring_Lab_Leader.txt     0.00044 2.8e-05 0.99956 2.8e-05  1847
#> deRossa_DL_Leader.txt     0.00047 4.8e-05 0.99953 4.8e-05  1879
#> vAhearn_FG_.txt           0.18175 3.9e-02 0.81825 3.9e-02  3346
#> vAhernD_FF_.txt           0.20235 4.8e-02 0.79765 4.8e-02  4279
#> vBarry_FG_.txt            0.44786 8.9e-02 0.55214 8.9e-02  3513
#> vBlaney_FF_Indep.txt      0.28803 4.9e-02 0.71197 4.9e-02  4374
#> vBoylan_FG_.txt           0.18318 5.6e-02 0.81682 5.6e-02  3338
#> vBrennan_FF_Min.txt       0.39177 4.9e-02 0.60823 4.9e-02  4862
#> vBrowne_FG_.txt           0.26126 8.4e-02 0.73874 8.4e-02  3605
#> vBrutonJ_FG_Leader.txt    0.27156 3.2e-02 0.72844 3.2e-02  4722
#> vBrutonR_FG_.txt          0.22384 6.2e-02 0.77616 6.2e-02  2976
#> vBurke_FF_Min.txt         0.37467 3.0e-02 0.62533 3.0e-02  7914
#> vCollins_FF_Min.txt       0.71582 1.7e-01 0.28418 1.7e-01  5264
#> vConnaughton_FG_.txt      0.17719 6.4e-02 0.82281 6.4e-02  3080
#> vCowen_FF_.txt            0.53006 1.4e-01 0.46994 1.4e-01  3327
#> vCreed_FG_.txt            0.30043 2.2e-02 0.69957 2.2e-02  3306
#> vCullimore_FF_.txt        0.58683 1.4e-01 0.41317 1.4e-01  2253
#> vCurrie_FG_.txt           0.13869 5.6e-02 0.86131 5.6e-02  2705
#> vDaly_FF_Min.txt          0.53370 1.2e-01 0.46630 1.2e-01  4979
#> vDavern_FF_.txt           0.15070 6.2e-02 0.84930 6.2e-02  3300
#> vDeasy_FG_.txt            0.13819 6.2e-02 0.86181 6.2e-02  3849
#> vDeenihan_FG_.txt         0.18397 7.5e-02 0.81603 7.5e-02  2397
#> vDurkan_FG_.txt           0.06387 1.1e-01 0.93613 1.1e-01  2745
#> vFerris_Lab_.txt          0.16892 6.7e-02 0.83108 6.7e-02  4033
#> vFinucane_FG_.txt         0.15331 6.5e-02 0.84669 6.5e-02  3193
#> vFitzGerald_FG_.txt       0.44962 8.6e-02 0.55038 8.6e-02  3662
#> vFlaherty_FG_.txt         0.20939 5.6e-02 0.79061 5.6e-02  3525
#> vFlynn_FF_Min.txt         0.43254 6.8e-02 0.56746 6.8e-02  5463
#> vGarland_Green_Leader.txt 0.27677 6.1e-02 0.72323 6.1e-02  3599
#> vGilmore_DL_.txt          0.25704 4.1e-02 0.74296 4.1e-02  4036
#> vHigginsJ_FG_.txt         0.15266 5.5e-02 0.84734 5.5e-02 16156
#> vHigginsMD_Lab_.txt       0.17709 6.0e-02 0.82291 6.0e-02  3672
#> vHillery_FF_.txt          0.53479 9.0e-02 0.46521 9.0e-02  6744
#> vHowlin_Lab_.txt          0.23440 5.2e-02 0.76560 5.2e-02  2776
#> vKenny_FG_.txt            0.29380 7.6e-02 0.70620 7.6e-02  2473
#> vLenihan_FF_.txt          0.34527 5.4e-02 0.65473 5.4e-02  5127
#> vLeyden_FF_JMin.txt       0.52413 1.0e-01 0.47587 1.0e-01  3828
#> vMcDaid_FF_.txt           0.17517 7.7e-02 0.82483 7.7e-02  4042
#> vNolan_FF_.txt            0.79026 2.1e-01 0.20974 2.1e-01  2379
#> vNoonan_FG_.txt           0.22302 6.3e-02 0.77698 6.3e-02  3801
#> vODonoghue_FF_.txt        0.27694 4.8e-02 0.72306 4.8e-02  2755
#> vOHanlon_FF_Min.txt       0.62640 1.4e-01 0.37360 1.4e-01  8316
#> vOKennedy_FF_Min.txt      0.50431 8.4e-02 0.49569 8.4e-02  5860
#> vOMalley_PD_Leader.txt    0.44588 6.4e-02 0.55412 6.4e-02  4100
#> vORourke_FF_Min.txt       0.41646 5.8e-02 0.58354 5.8e-02  8401
#> vOShea_Lab_.txt           0.12647 7.6e-02 0.87353 7.6e-02  6645
#> vOwen_FG_.txt             0.11156 6.8e-02 0.88844 6.8e-02  6885
#> vQuinn_Lab_.txt           0.29982 2.7e-02 0.70018 2.7e-02  3783
#> vRabbitte_DL_.txt         0.07838 1.3e-01 0.92162 1.3e-01  4262
#> vReynoldsG_FG_.txt        0.17913 2.4e-02 0.82087 2.4e-02  2339
#> vReynolds_FF_Min.txt      0.68655 1.5e-01 0.31345 1.5e-01  4036
#> vRoche_FF_.txt            0.20856 4.8e-02 0.79144 4.8e-02  4976
#> vStagg_Lab_.txt           0.43570 8.4e-02 0.56430 8.4e-02  5393
#> vTaylorQuinn_FG_.txt      0.20615 4.5e-02 0.79385 4.5e-02  3522
#> vWilson_FF_Tan.txt        0.72483 1.8e-01 0.27517 1.8e-01  3533
#> vWoods_FF_Min.txt         0.59174 1.4e-01 0.40826 1.4e-01  4226
#> vYates_FG_.txt            0.38047 3.3e-02 0.61953 3.3e-02  3860
#> 
#> Some diagnostics here about how many words were not found in training vocabulary.