Wordfish text model

Estimate Slapin and Proksch's (2008) "wordfish" Poisson scaling model of one-dimensional document positions using conditional maximum likelihood.

textmodel_wordfish(x, dir = c(1, 2), priors = c(Inf, Inf, 3, 1),
  tol = c(1e-06, 1e-08), dispersion = c("poisson", "quasipoisson"),
  dispersion_level = c("feature", "overall"), dispersion_floor = 0,
  sparse = TRUE, abs_err = FALSE, svd_sparse = TRUE,
  residual_floor = 0.5)

Arguments

x
the dfm on which the model will be fit

dir
set global identification by specifying the indexes for a pair of documents such that \(\hat{\theta}_{dir[1]} < \hat{\theta}_{dir[2]}\).

priors
prior precisions for the estimated parameters \(\alpha_i\), \(\psi_j\), \(\beta_j\), and \(\theta_i\), where \(i\) indexes documents and \(j\) indexes features

tol
tolerances for convergence. The first value is a convergence threshold for the log-posterior of the model, the second value is the tolerance in the difference in parameter values from the iterative conditional maximum likelihood (from conditionally estimating document-level, then feature-level parameters).

dispersion
sets whether a quasi-Poisson quasi-likelihood should be used based on a single dispersion parameter ("poisson"), or quasi-Poisson ("quasipoisson")

dispersion_level
sets the unit level for the dispersion parameter, options are "feature" for term-level variances, or "overall" for a single dispersion parameter

dispersion_floor
constraint for the minimal underdispersion multiplier in the quasi-Poisson model. Used to minimize the distorting effect of terms with rare term or document frequencies that appear to be severely underdispersed. Default is 0, but this only applies if dispersion = "quasipoisson".

sparse
specifies whether the "dfm" is coerced to dense

abs_err
specifies how the convergence is considered

svd_sparse
uses svd to initialize the starting values of theta, only applies when sparse = TRUE

residual_floor
specifies the threshold for residual matrix when calculating the svds, only applies when sparse = TRUE

Value

An object of class textmodel_fitted_wordfish. This is a list containing:

dir

global identification of the dimension

theta

estimated document positions

alpha

estimated document fixed effects

beta

estimated feature marginal effects

psi

estimated word fixed effects

docs

document labels

features

feature labels

sigma

regularization parameter for betas in Poisson form

ll

log likelihood at convergence

se.theta

standard errors for theta-hats

x

dfm to which the model was fit

Details

The returns match those of Will Lowe's R implementation of wordfish (see the austin package), except that here we have renamed words to be features. (This return list may change.) We have also followed the practice begun with Slapin and Proksch's early implementation of the model that used a regularization parameter of se\((\sigma) = 3\), through the third element in priors.

Note

In the rare situation where a warning message of "The algorithm did not converge." shows up, removing some documents may work.

References

Jonathan Slapin and Sven-Oliver Proksch. 2008. "A Scaling Model for Estimating Time-Series Party Positions from Texts." American Journal of Political Science 52(3):705-772.

Lowe, Will and Kenneth Benoit. 2013. "Validating Estimates of Latent Traits from Textual Data Using Human Judgment as a Benchmark." Political Analysis 21(3), 298-313. http://doi.org/10.1093/pan/mpt002

See also

predict.textmodel_wordfish

Examples

(wf <- textmodel_wordfish(data_dfm_lbgexample, dir = c(1,5)))

#> 
#> Call:
#> textmodel_wordfish.dfm(x = data_dfm_lbgexample, dir = c(1, 5))
#> 
#> Dispersion: poisson; direction: 1 < 5; 6 documents; 37 features.

summary(wf, n = 10)

#> 
#> Call:
#> textmodel_wordfish.dfm(x = data_dfm_lbgexample, dir = c(1, 5))
#> 
#> Estimated Document Positions:
#>       theta      se
#> R1 -1.33148 0.01573
#> R2 -0.60623 0.01216
#> R3  0.06347 0.01129
#> R4  0.70272 0.01266
#> R5  1.50107 0.01776
#> V1 -0.32955 0.01195
#> 
#> Estimated Feature Scores:
#>           A      B      C       D       E      F       G       H      I      J
#> beta -7.406 -8.077 -9.733 -10.713 -11.581 -11.36 -11.557 -10.453 -9.732 -8.670
#> psi  -6.287 -6.539 -7.231  -7.677  -8.089  -7.23  -7.102  -5.394 -4.350 -3.012

coef(wf)

#> $documents
#>          theta     alpha
#> R1 -1.33147776 -3.630926
#> R2 -0.60623046  1.520180
#> R3  0.06347106  3.030810
#> R4  0.70272242  1.228497
#> R5  1.50106629 -4.691470
#> V1 -0.32955155  2.542908
#> 
#> $features
#>           beta          psi
#> A   -7.4056616 -6.287342603
#> B   -8.0765727 -6.539026323
#> C   -9.7333519 -7.231193879
#> D  -10.7126585 -7.676597943
#> E  -11.5809685 -8.088635210
#> F  -11.3588028 -7.229822718
#> G  -11.5568119 -7.102127834
#> H  -10.4531092 -5.393549038
#> I   -9.7315797 -4.349544198
#> J   -8.6704659 -3.012054132
#> K   -7.6174882 -1.844025903
#> L   -6.6301576 -0.886141079
#> M   -5.6167414 -0.036220698
#> N   -4.6846792  0.613322531
#> O   -3.7786358  1.123282661
#> P   -2.7242771  1.547467746
#> Q   -1.8088196  1.784735597
#> R   -0.7116621  1.927776133
#> S    0.2595271  1.907993628
#> T    1.2560437  1.765688168
#> U    2.3937160  1.467079446
#> V    3.3682392  1.088026854
#> W    4.4778063  0.516825126
#> X    5.2779710  0.001071918
#> Y    6.1129118 -0.660849864
#> Z    6.9840413 -1.491957182
#> ZA   7.8350989 -2.420953925
#> ZB   8.8107305 -3.650413209
#> ZC   9.7296495 -4.955515332
#> ZD  10.5005691 -6.193372980
#> ZE  11.3285987 -7.676879219
#> ZF  11.1530950 -7.804360220
#> ZG  11.4371452 -8.791719487
#> ZH  10.6069111 -8.287671264
#> ZI   9.6634407 -7.726540581
#> ZJ   8.0544224 -6.814514464
#> ZK   7.4028588 -6.470757738
#> 

predict(wf)

#>          R1          R2          R3          R4          R5          V1 
#> -1.33147776 -0.60623046  0.06347106  0.70272242  1.50106629 -0.32955155 

predict(wf, se.fit = TRUE)

#> $fit
#>          R1          R2          R3          R4          R5          V1 
#> -1.33147776 -0.60623046  0.06347106  0.70272242  1.50106629 -0.32955155 
#> 
#> $se.fit
#> [1] 0.01573421 0.01215946 0.01129486 0.01266336 0.01775535 0.01194835
#> 

predict(wf, interval = "confidence")

#> $fit
#>            fit         lwr         upr
#> R1 -1.33147776 -1.36231625 -1.30063926
#> R2 -0.60623046 -0.63006258 -0.58239835
#> R3  0.06347106  0.04133354  0.08560858
#> R4  0.70272242  0.67790269  0.72754215
#> R5  1.50106629  1.46626644  1.53586615
#> V1 -0.32955155 -0.35296990 -0.30613321
#> 

# NOT RUN {
ie2010dwf <- dfm(data_corpus_irishbudget2010, verbose = FALSE)
(wf1 <- textmodel_wordfish(ie2010dfm, dir = c(6,5)))
(wf2a <- textmodel_wordfish(ie2010dfm, dir = c(6,5),
                             dispersion = "quasipoisson", dispersion_floor = 0))
(wf2b <- textmodel_wordfish(ie2010dfm, dir = c(6,5),
                             dispersion = "quasipoisson", dispersion_floor = .5))
plot(wf2a$phi, wf2b$phi, xlab = "Min underdispersion = 0", ylab = "Min underdispersion = .5",
     xlim = c(0, 1.0), ylim = c(0, 1.0))
plot(wf2a$phi, wf2b$phi, xlab = "Min underdispersion = 0", ylab = "Min underdispersion = .5",
     xlim = c(0, 1.0), ylim = c(0, 1.0), type = "n")
underdispersedTerms <- sample(which(wf2a$phi < 1.0), 5)
which(featnames(ie2010dfm) %in% names(topfeatures(ie2010dfm, 20)))
text(wf2a$phi, wf2b$phi, wf2a$features,
     cex = .8, xlim = c(0, 1.0), ylim = c(0, 1.0), col = "grey90")
text(wf2a$phi['underdispersedTerms'], wf2b$phi['underdispersedTerms'],
     wf2a$features['underdispersedTerms'],
     cex = .8, xlim = c(0, 1.0), ylim = c(0, 1.0), col = "black")
if (require(austin)) {
    wf_austin <- austin::wordfish(quanteda::as.wfm(ie2010dfm), dir = c(6,5))
    cor(wf1$theta, wf_austin$theta)
}
# }