wordfish text model

Estimate Slapin and Proksch's (2008) "wordfish" Poisson scaling model of one-dimensional document positions using conditional maximum likelihood.

textmodel_wordfish(x, dir = c(1, 2), priors = c(Inf, Inf, 3, 1),
  tol = c(1e-06, 1e-08), dispersion = c("poisson", "quasipoisson"),
  dispersion_level = c("feature", "overall"), dispersion_floor = 0,
  sparse = TRUE, threads = quanteda_options("threads"), abs_err = FALSE,
  svd_sparse = TRUE, residual_floor = 0.5)

Arguments

x

the dfm on which the model will be fit

dir

set global identification by specifying the indexes for a pair of documents such that \(\hat{\theta}_{dir[1]} < \hat{\theta}_{dir[2]}\).

priors

prior precisions for the estimated parameters \(\alpha_i\), \(\psi_j\), \(\beta_j\), and \(\theta_i\), where \(i\) indexes documents and \(j\) indexes features

tol

tolerances for convergence. The first value is a convergence threshold for the log-posterior of the model, the second value is the tolerance in the difference in parameter values from the iterative conditional maximum likelihood (from conditionally estimating document-level, then feature-level parameters).

dispersion

sets whether a quasi-poisson quasi-likelihood should be used based on a single dispersion parameter ("poisson"), or quasi-Poisson ("quasipoisson")

dispersion_level

sets the unit level for the dispersion parameter, options are "feature" for term-level variances, or "overall" for a single dispersion parameter

dispersion_floor

constraint for the minimal underdispersion multiplier in the quasi-Poisson model. Used to minimize the distorting effect of terms with rare term or document frequencies that appear to be severely underdispersed. Default is 0, but this only applies if dispersion = "quasipoisson".

sparse

specifies whether the "dfm" is coerced to dense

threads

specifies the number of threads to use; set to 1 to override the package settings and use a serial version of the function

abs_err

specifies how the convergence is considered

svd_sparse

uses svd to initialize the starting values of theta, only applies when sparse = TRUE

residual_floor

specifies the threshold for residual matrix when calculating the svds, only applies when sparse = TRUE

Value

An object of class textmodel_fitted_wordfish. This is a list containing:

Details

The returns match those of Will Lowe's R implementation of wordfish (see the austin package), except that here we have renamed words to be features. (This return list may change.) We have also followed the practice begun with Slapin and Proksch's early implementation of the model that used a regularization parameter of se\((\sigma) = 3\), through the third element in priors.

References

Jonathan Slapin and Sven-Oliver Proksch. 2008. "A Scaling Model for Estimating Time-Series Party Positions from Texts." American Journal of Political Science 52(3):705-772.

Lowe, Will and Kenneth Benoit. 2013. "Validating Estimates of Latent Traits from Textual Data Using Human Judgment as a Benchmark." Political Analysis 21(3), 298-313. http://doi.org/10.1093/pan/mpt002

Examples

textmodel_wordfish(data_dfm_LBGexample, dir = c(1,5))

#> Fitted wordfish model:
#> Call:
#> 	textmodel_wordfish(x = data_dfm_LBGexample, dir = c(1, 5))
#> 
#> Estimated document positions:
#> 
#>   Documents       theta         SE       lower       upper
#> 1        R1 -1.29155112 0.01450459 -1.31998012 -1.26312212
#> 2        R2 -0.61629592 0.01061247 -0.63709636 -0.59549548
#> 3        R3  0.03431077 0.01085587  0.01303327  0.05558828
#> 4        R4  0.70995750 0.01212683  0.68618892  0.73372608
#> 5        R5  1.52250730 0.01741995  1.48836419  1.55665040
#> 6        V1 -0.35892854 0.01071309 -0.37992620 -0.33793087
#> 
#> Estimated feature scores: showing first 30 beta-hats for features
#> 
#>           A           B           C           D           E           F 
#>  -8.4392704  -9.2914593 -11.2444304 -12.3486004 -13.3149107 -13.2180169 
#>           G           H           I           J           K           L 
#> -13.4665198 -12.3392234 -11.5662044 -10.4070576  -9.2412225  -8.1354780 
#>           M           N           O           P           Q           R 
#>  -6.9710012  -5.8689151  -4.7877759  -3.5677039  -2.5471639  -1.3418545 
#>           S           T           U           V           W           X 
#>  -0.2855031   0.7731068   1.9392648   2.9340387   4.1076564   5.0083465 
#>           Y           Z          ZA          ZB          ZC          ZD 
#>   5.9450536   6.8813261   7.7624218   8.7566699   9.6840037  10.4605090 

## Not run: ------------------------------------
# ie2010dfm <- dfm(data_corpus_irishbudget2010, verbose = FALSE)
# (wfm1 <- textmodel_wordfish(ie2010dfm, dir = c(6,5)))
# (wfm2a <- textmodel_wordfish(ie2010dfm, dir = c(6,5), 
#                              dispersion = "quasipoisson", dispersion_floor = 0))
# (wfm2b <- textmodel_wordfish(ie2010dfm, dir = c(6,5), 
#                              dispersion = "quasipoisson", dispersion_floor = .5))
# plot(wfm2a@phi, wfm2b@phi, xlab = "Min underdispersion = 0", ylab = "Min underdispersion = .5",
#      xlim = c(0, 1.0), ylim = c(0, 1.0))
# plot(wfm2a@phi, wfm2b@phi, xlab = "Min underdispersion = 0", ylab = "Min underdispersion = .5",
#      xlim = c(0, 1.0), ylim = c(0, 1.0), type = "n")
# underdispersedTerms <- sample(which(wfm2a@phi < 1.0), 5)
# which(featnames(ie2010dfm) %in% names(topfeatures(ie2010dfm, 20)))
# text(wfm2a@phi, wfm2b@phi, wfm2a@features, 
#      cex = .8, xlim = c(0, 1.0), ylim = c(0, 1.0), col = "grey90")
# text(wfm2a@phi[underdispersedTerms], wfm2b@phi[underdispersedTerms], 
#      wfm2a@features[underdispersedTerms], 
#      cex = .8, xlim = c(0, 1.0), ylim = c(0, 1.0), col = "black")
# if (require(austin)) {
#     wfmodelAustin <- austin::wordfish(quanteda::as.wfm(ie2010dfm), dir = c(6,5))
#     cor(wfm1@theta, wfmodelAustin$theta)
# }
## ---------------------------------------------