calculate collocation statistics

Identify and score collocations from a corpus, character, or tokens object, with targeted selection.

#' @rdname textstat_collocations #' @noRd #' @export textstat_collocations.character <- function(x, method = c("lr", "chi2", "pmi", "dice", "bj"), ...) method <- match.arg(method) textstat_collocations(tokens(x), method = method, ...)

textstat_collocations(x, method = c("lr", "chi2", "pmi", "dice", "bj_uni",
  "bj_all"), min_size = 2, max_size = 3, min_count = 2, ...)

is.collocations(x)

Arguments

x
a character, corpus, or tokens object to be mined for collocations

method
association measure for detecting collocations. Let \(i\) index documents, and \(j\) index features, \(n_{ij}\) refers to observed counts, and \(m_{ij}\) the expected counts in a collocations frequency table of dimensions \((J - size + 1)^2\). Available measures are computed as:

"lr"

The likelihood ratio statistic \(G^2\), computed as: $$2 * \sum_i \sum_j ( n_{ij} * log \frac{n_{ij}}{m_{ij}} )$$

"chi2"

Pearson's \(\chi^2\) statistic, computed as: $$\sum_i \sum_j \frac{(n_{ij} - m_{ij})^2}{m_{ij}}$$

"pmi"

point-wise mutual information score, computed as log \(n_{11}/m_{11}\)

"dice"

the Dice coefficient, computed as \(n_{11}/n_{1.} + n_{.1}\)

"bj_uni"

unigram subtuples, Blaheta and Johnson's method (called through sequences)

"bj_all"

all subtuples algorithm, Blaheta and Johnson's method (called through sequences)

min_size
minimum length of sequences which are collected

max_size
numeric argument representing the maximum length of the collocations to be scored. The maximum size is currently 3 for all methods except "bj_uni" and "bj_all", which has a maximum size of 5.

min_count
minimum frequency of collocations that will be scored

...
additional arguments passed to collocations2 for the first four methods, or to sequences for method = "bj"

Value

is.collocation returns TRUE if the object is of class collocations, FALSE otherwise.

Details

#' @rdname textstat_collocations #' @noRd #' @export textstat_collocations.tokenizedTexts <- function(x, method = c("lr", "chi2", "pmi", "dice", "bj"), ...) method <- match.arg(method) textstat_collocations(as.tokens(x), method = method, ...)

check if an object is collocations object

Note

This function is under active development, and we aim to improve both its operation and efficiency in the next release of quanteda.

References

Blaheta, D., & Johnson, M. (2001). Unsupervised learning of multi-word verbs. Presented at the ACLEACL Workshop on the Computational Extraction, Analysis and Exploitation of Collocations. McInnes, B T. 2004. "Extending the Log Likelihood Measure to Improve Collocation Identification." M.Sc. Thesis, University of Minnesota.

Examples

txts <- c("quanteda is a package for quantitative text analysis",
          "quantitative text analysis is a rapidly growing field",
          "The population is rapidly growing")
toks <- tokens(txts)
textstat_collocations(toks, method = "lr")

#>                  collocation length count       G2
#> 1 quantitative text analysis      3     2 23.56047
#> 2          quantitative text      2     2 12.55796
#> 3              text analysis      2     2 12.55796
#> 4            rapidly growing      2     2 12.55796
#> 5                       is a      2     2  8.73887

textstat_collocations(toks, method = "lr", min_count = 1)

#>                   collocation length count        G2
#> 1  quantitative text analysis      3     2 23.560468
#> 2    package for quantitative      3     1 14.695802
#> 3       for quantitative text      3     1 14.695802
#> 4       rapidly growing field      3     1 14.695802
#> 5           a rapidly growing      3     1 13.736956
#> 6          is rapidly growing      3     1 12.857319
#> 7           quantitative text      2     2 12.557955
#> 8               text analysis      2     2 12.557955
#> 9             rapidly growing      2     2 12.557955
#> 10              quanteda is a      3     1 12.536461
#> 11              analysis is a      3     1 12.536461
#> 12               is a package      3     1 12.536461
#> 13              a package for      3     1 11.923213
#> 14           text analysis is      3     1 11.923213
#> 15          The population is      3     1 11.923213
#> 16               is a rapidly      3     1 10.084731
#> 17                       is a      2     2  8.738870
#> 18      population is rapidly      3     1  8.424986
#> 19                package for      2     1  7.724130
#> 20           for quantitative      2     1  7.724130
#> 21              growing field      2     1  7.724130
#> 22             The population      2     1  7.724130
#> 23                  a package      2     1  4.951541
#> 24                quanteda is      2     1  3.905045
#> 25                analysis is      2     1  3.905045
#> 26              population is      2     1  3.905045
#> 27                  a rapidly      2     1  2.304034
#> 28                 is rapidly      2     1  1.390970

textstat_collocations(toks, method = "lr", max_size = 3, min_count = 1)

#>                   collocation length count        G2
#> 1  quantitative text analysis      3     2 23.560468
#> 2    package for quantitative      3     1 14.695802
#> 3       for quantitative text      3     1 14.695802
#> 4       rapidly growing field      3     1 14.695802
#> 5           a rapidly growing      3     1 13.736956
#> 6          is rapidly growing      3     1 12.857319
#> 7           quantitative text      2     2 12.557955
#> 8               text analysis      2     2 12.557955
#> 9             rapidly growing      2     2 12.557955
#> 10              quanteda is a      3     1 12.536461
#> 11              analysis is a      3     1 12.536461
#> 12               is a package      3     1 12.536461
#> 13              a package for      3     1 11.923213
#> 14           text analysis is      3     1 11.923213
#> 15          The population is      3     1 11.923213
#> 16               is a rapidly      3     1 10.084731
#> 17                       is a      2     2  8.738870
#> 18      population is rapidly      3     1  8.424986
#> 19                package for      2     1  7.724130
#> 20           for quantitative      2     1  7.724130
#> 21              growing field      2     1  7.724130
#> 22             The population      2     1  7.724130
#> 23                  a package      2     1  4.951541
#> 24                quanteda is      2     1  3.905045
#> 25                analysis is      2     1  3.905045
#> 26              population is      2     1  3.905045
#> 27                  a rapidly      2     1  2.304034
#> 28                 is rapidly      2     1  1.390970

(cols <- textstat_collocations(toks, method = "lr", max_size = 3, min_count = 2))

#>                  collocation length count       G2
#> 1 quantitative text analysis      3     2 23.56047
#> 2          quantitative text      2     2 12.55796
#> 3              text analysis      2     2 12.55796
#> 4            rapidly growing      2     2 12.55796
#> 5                       is a      2     2  8.73887

# extracting multi-part proper nouns (capitalized terms)
toks2 <- tokens(corpus_segment(data_corpus_inaugural, what = "sentence"))
toks2 <- tokens_select(toks2, stopwords("english"), "remove", padding = TRUE)
toks2 <- tokens_select(toks2, "^([A-Z][a-z\\-]{2,})", valuetype="regex",
                     case_insensitive = FALSE, padding = TRUE)
seqs <- textstat_collocations(toks2, method = "bj_uni")
head(seqs, 10)

#>            collocation count length    lambda     sigma         z            p
#> 1        United States   155      2  6.212629 0.4289724 14.482583 0.000000e+00
#> 2   Federal Government    32      2  3.604058 0.3829004  9.412521 0.000000e+00
#> 3         Almighty God    15      2  6.610756 0.8324196  7.941615 9.992007e-16
#> 4        Chief Justice    11      2  5.351507 0.7455401  7.178027 3.536060e-13
#> 5        South America     6      2  6.760569 1.1181163  6.046392 7.406286e-10
#> 6     Chief Magistrate    10      2  5.388979 0.9133941  5.899950 1.818057e-09
#> 7  Vice President Bush     2      3 15.219565 2.5833183  5.891479 1.913775e-09
#> 8        Supreme Court     4      2  6.311977 1.1636470  5.424305 2.909019e-08
#> 9            Old World    10      2  8.062433 1.5468317  5.212224 9.329519e-08
#> 10      Vice President     5      2 10.427772 2.0175932  5.168422 1.180396e-07

# compounding tokens is more efficient when applied to the same tokens object 
toks_comp <- tokens_compound(toks2, seqs)