calculate collocation statistics

Identify and score collocations from a corpus, character, or tokens object, with targeted selection.

#' @rdname textstat_collocations #' @noRd #' @export textstat_collocations.character <- function(x, method = c("lr", "chi2", "pmi", "dice", "bj"), ...) method <- match.arg(method) textstat_collocations(tokens(x), method = method, ...)

textstat_collocations(x, method = c("lr", "chi2", "pmi", "dice", "bj"),
  max_size = 3, min_count = 2, ...)

is.collocations(x)

Arguments

x

a character, corpus, or tokens object to be mined for collocations

method

association measure for detecting collocations. Let \(i\) index documents, and \(j\) index features, \(n_{ij}\) refers to observed counts, and \(m_{ij}\) the expected counts in a collocations frequency table of dimensions \((J - size + 1)^2\). Available measures are computed as:

"lr"

The likelihood ratio statistic \(G^2\), computed as: $$2 * \sum_i \sum_j ( n_{ij} * log \frac{n_{ij}}{m_{ij}} )$$

"chi2"

Pearson's \(\chi^2\) statistic, computed as: $$\sum_i \sum_j \frac{(n_{ij} - m_{ij})^2}{m_{ij}}$$

"pmi"

point-wise mutual information score, computed as log \(n_{11}/m_{11}\)

"dice"

the Dice coefficient, computed as \(n_{11}/n_{1.} + n_{.1}\)

"bj"

Blaheta and Johnson's method (called through sequences)

max_size

numeric argument representing the maximum length of the collocations to be scored. The maximum size is currently 3 for all methods except "bj", which has a maximum size of 5.

min_count

minimum frequency of collocations that will be scored

...

additional arguments passed to collocations2 for the first four methods, or to sequences for method = "bj"

Value

is.collocation returns TRUE if the object is of class collocations, FALSE otherwise.

Details

#' @rdname textstat_collocations #' @noRd #' @export textstat_collocations.tokenizedTexts <- function(x, method = c("lr", "chi2", "pmi", "dice", "bj"), ...) method <- match.arg(method) textstat_collocations(as.tokens(x), method = method, ...)

check if an object is collocations object

Note

This function is under active development, and we aim to improve both its operation and efficiency in the next release of quanteda.

References

Blaheta, D., & Johnson, M. (2001). http://web.science.mq.edu.au/~mjohnson/papers/2001/dpb-colloc01.pdf. Presented at the ACLEACL Workshop on the Computational Extraction, Analysis and Exploitation of Collocations.

McInnes, B T. 2004. "Extending the Log Likelihood Measure to Improve Collocation Identification." M.Sc. Thesis, University of Minnesota.

Examples

txts <- c("quanteda is a package for quantitative text analysis",
          "quantitative text analysis is a rapidly growing field",
          "The population is rapidly growing")
toks <- tokens(txts)
textstat_collocations(toks, method = "lr")

#>                  collocation count       G2
#> 1          quantitative text     2 12.55796
#> 2              text analysis     2 12.55796
#> 3            rapidly growing     2 12.55796
#> 4                       is a     2  8.73887
#> 5 quantitative text analysis     2 23.56047

textstat_collocations(toks, method = "lr", min_count = 1)

#>                   collocation count        G2
#> 1           quantitative text     2 12.557955
#> 2               text analysis     2 12.557955
#> 3             rapidly growing     2 12.557955
#> 4                        is a     2  8.738870
#> 5                 package for     1  7.724130
#> 6            for quantitative     1  7.724130
#> 7               growing field     1  7.724130
#> 8              The population     1  7.724130
#> 9                   a package     1  4.951541
#> 10                quanteda is     1  3.905045
#> 11                analysis is     1  3.905045
#> 12              population is     1  3.905045
#> 13                  a rapidly     1  2.304034
#> 14                 is rapidly     1  1.390970
#> 15 quantitative text analysis     2 23.560468
#> 16   package for quantitative     1 14.695802
#> 17      for quantitative text     1 14.695802
#> 18      rapidly growing field     1 14.695802
#> 19          a rapidly growing     1 13.736956
#> 20         is rapidly growing     1 12.857319
#> 21              quanteda is a     1 12.536461
#> 22              analysis is a     1 12.536461
#> 23               is a package     1 12.536461
#> 24              a package for     1 11.923213
#> 25           text analysis is     1 11.923213
#> 26          The population is     1 11.923213
#> 27               is a rapidly     1 10.084731
#> 28      population is rapidly     1  8.424986

textstat_collocations(toks, method = "lr", max_size = 3, min_count = 1)

#>                   collocation count        G2
#> 1           quantitative text     2 12.557955
#> 2               text analysis     2 12.557955
#> 3             rapidly growing     2 12.557955
#> 4                        is a     2  8.738870
#> 5                 package for     1  7.724130
#> 6            for quantitative     1  7.724130
#> 7               growing field     1  7.724130
#> 8              The population     1  7.724130
#> 9                   a package     1  4.951541
#> 10                quanteda is     1  3.905045
#> 11                analysis is     1  3.905045
#> 12              population is     1  3.905045
#> 13                  a rapidly     1  2.304034
#> 14                 is rapidly     1  1.390970
#> 15 quantitative text analysis     2 23.560468
#> 16   package for quantitative     1 14.695802
#> 17      for quantitative text     1 14.695802
#> 18      rapidly growing field     1 14.695802
#> 19          a rapidly growing     1 13.736956
#> 20         is rapidly growing     1 12.857319
#> 21              quanteda is a     1 12.536461
#> 22              analysis is a     1 12.536461
#> 23               is a package     1 12.536461
#> 24              a package for     1 11.923213
#> 25           text analysis is     1 11.923213
#> 26          The population is     1 11.923213
#> 27               is a rapidly     1 10.084731
#> 28      population is rapidly     1  8.424986

(cols <- textstat_collocations(toks, method = "lr", max_size = 3, min_count = 2))

#>                  collocation count       G2
#> 1          quantitative text     2 12.55796
#> 2              text analysis     2 12.55796
#> 3            rapidly growing     2 12.55796
#> 4                       is a     2  8.73887
#> 5 quantitative text analysis     2 23.56047

as.tokens(cols)

#> tokens from 5 documents.
#> Component 1 :
#> [1] "quantitative" "text"        
#> 
#> Component 2 :
#> [1] "text"     "analysis"
#> 
#> Component 3 :
#> [1] "rapidly" "growing"
#> 
#> Component 4 :
#> [1] "is" "a" 
#> 
#> Component 5 :
#> [1] "quantitative" "text"         "analysis"    
#> 

# extracting multi-part proper nouns (capitalized terms)
toks2 <- tokens(corpus_segment(data_corpus_inaugural, what = "sentence"))
toks2 <- tokens_select(toks2, stopwords("english"), "remove", padding = TRUE)
seqs <- textstat_collocations(toks2, method = "bj",
                              features = "^([A-Z][a-z\\-]{2,})",
                              valuetype = "regex", case_insensitive = FALSE)
head(seqs, 10)

#>            collocation   lambda      sigma count         z p
#> 1        United States 20.56459 0.04324223   155 475.56728 0
#> 2   Federal Government 16.07807 0.09312539    32 172.64966 0
#> 3   General Government 14.50954 0.12914263    24 112.35286 0
#> 4  National Government 12.88826 0.15847837    11  81.32504 0
#> 5         Almighty God 13.10835 0.16301521    15  80.41181 0
#> 6        Chief Justice 12.17009 0.18109481    11  67.20288 0
#> 7       President Bush 11.87632 0.18741675     6  63.36850 0
#> 8     Chief Magistrate 11.37940 0.22063436    10  51.57583 0
#> 9    President Clinton 10.67352 0.22913765     4  46.58124 0
#> 10      United Nations 10.72087 0.24598156     8  43.58402 0

# compounding tokens is more efficient when applied to the same tokens object 
toks_comp <- tokens_compound(toks2, seqs)