calculate keyness statistics

Calculate "keyness", a score for features that occur differentially across different categories. Here, the categories are defined by reference to a "target" document index in the dfm, with the reference group consisting of all other documents.

textstat_keyness(x, target = 1L, measure = c("chi2", "exact", "lr", "pmi"),
  sort = TRUE, correction = c("default", "yates", "williams", "none"))

Arguments

x
a dfm containing the features to be examined for keyness

target
the document index (numeric, character or logical) identifying the document forming the "target" for computing keyness; all other documents' feature frequencies will be combined for use as a reference

measure
(signed) association measure to be used for computing keyness. Currenly available: "chi2"; "exact" (Fisher's exact test); "lr" for the likelihood ratio; "pmi" for pointwise mutual information.

sort
logical; if TRUE sort features scored in descending order of the measure, otherwise leave in original feature order

correction
if "default", Yates correction is applied to "chi2"; William's correction is applied to "lr"; and no correction is applied for the "exact" and "pmi" measures. Specifying a value other than the default can be used to override the defaults, for instance to apply the Williams correction to the chi2 measure. Specying a correction for the "exact" and "pmi" measures has no effect and produces a warning.

Value

a data.frame of computed statistics and associated p-values, where the features scored name each row, and the number of occurrences for both the target and reference groups. For measure = "chi2" this is the chi-squared value, signed positively if the observed value in the target exceeds its expected value; for measure = "exact" this is the estimate of the odds ratio; for measure = "lr" this is the likelihood ratio \(G2\) statistic; for "pmi" this is the pointwise mutual information statistics.

References

Bondi, Marina, and Mike Scott, eds. 2010. Keyness in Texts. Amsterdam, Philadelphia: John Benjamins, 2010. Stubbs, Michael. 2010. "Three Concepts of Keywords". In Keyness in Texts, Marina Bondi and Mike Scott, eds. pp21–42. Amsterdam, Philadelphia: John Benjamins. Scott, M. & Tribble, C. 2006. Textual Patterns: keyword and corpus analysis in language education. Amsterdam: Benjamins, p. 55. Dunning, Ted. 1993. "Accurate Methods for the Statistics of Surprise and Coincidence", Computational Linguistics, Vol 19, No. 1, pp. 61-74.

Examples

# compare pre- v. post-war terms using grouping
period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
mydfm <- dfm(data_corpus_inaugural, groups = period)
head(mydfm) # make sure 'post-war' is in the first row

#> Document-feature matrix of: 2 documents, 9,357 features (34.9% sparse).
#> (showing first 2 documents and first 6 features)
#>           features
#> docs       chief justice    , mr    . vice
#>   post-war    12      40 2194 25 1804   15
#>   pre-war     24      97 4832  6 3141    2

head(result <- textstat_keyness(mydfm), 10)

#>               chi2 p n_target n_reference
#> we        701.9371 0      960         779
#> -         395.1029 0      450         312
#> .         226.6128 0     1804        3141
#> our       187.8002 0      874        1307
#> us        185.9959 0      262         216
#> :         178.1175 0      105          29
#> america   176.5995 0      130          54
#> world     175.1212 0      188         123
#> americans 150.6399 0       67           7
#> new       141.3576 0      150          97

tail(result, 10)

#>                    chi2            p n_target n_reference
#> upon          -51.91683 5.789813e-13       39         332
#> it            -52.70987 3.866907e-13      257        1132
#> public        -55.99591 7.260859e-14       11         213
#> states        -59.13492 1.476597e-14       28         305
#> constitution  -61.17021 5.218048e-15        6         200
#> be            -72.23208 0.000000e+00      257        1224
#> should        -83.11275 0.000000e+00       15         309
#> which        -160.15966 0.000000e+00       95         911
#> of           -179.22512 0.000000e+00     1437        5666
#> the          -299.92925 0.000000e+00     1988        8094

# compare pre- v. post-war terms using logical vector
mydfm2 <- dfm(data_corpus_inaugural)
textstat_keyness(mydfm2, docvars(data_corpus_inaugural, "Year") >= 1945)

#>                                chi2            p n_target n_reference
#> we                     7.019371e+02 0.000000e+00      960         779
#> -                      3.951029e+02 0.000000e+00      450         312
#> .                      2.266128e+02 0.000000e+00     1804        3141
#> our                    1.878002e+02 0.000000e+00      874        1307
#> us                     1.859959e+02 0.000000e+00      262         216
#> :                      1.781175e+02 0.000000e+00      105          29
#> america                1.765995e+02 0.000000e+00      130          54
#> world                  1.751212e+02 0.000000e+00      188         123
#> americans              1.506399e+02 0.000000e+00       67           7
#> new                    1.413576e+02 0.000000e+00      150          97
#> freedom                1.373359e+02 0.000000e+00      121          64
#> today                  1.288651e+02 0.000000e+00       76          21
#> let                    1.115078e+02 0.000000e+00      100          54
#> together               1.049739e+02 0.000000e+00       64          19
#> you                    1.020582e+02 0.000000e+00      116          80
#> america's              9.406430e+01 0.000000e+00       35           0
#> help                   9.213230e+01 0.000000e+00       46           8
#> work                   9.077268e+01 0.000000e+00       78          40
#> ,                      6.282676e+01 2.220446e-15     2194        4832
#> know                   6.042877e+01 7.660539e-15       63          40
#> earth                  6.032286e+01 8.104628e-15       44          18
#> history                5.769436e+01 3.064216e-14       60          38
#> god                    5.416903e+01 1.839640e-13       55          34
#> live                   5.336664e+01 2.767786e-13       39          16
#> time                   5.273163e+01 3.823608e-13      106         110
#> do                     5.259918e+01 4.090062e-13      112         120
#> are                    5.086586e+01 9.889867e-13      311         503
#> journey                4.929934e+01 2.197353e-12       20           1
#> dignity                4.863845e+01 3.077760e-12       24           4
#> what                   4.847095e+01 3.352207e-12       88          86
#> century                4.724528e+01 6.263656e-12       39          19
#> will                   4.720020e+01 6.409429e-12      339         572
#> promise                4.641132e+01 9.585999e-12       34          14
#> mr                     4.493591e+01 2.035905e-11       25           6
#> build                  4.426885e+01 2.862344e-11       21           3
#> thank                  4.394952e+01 3.369549e-11       18           1
#> bless                  4.000791e+01 2.529360e-10       18           2
#> strength               3.960798e+01 3.104119e-10       45          31
#> for                    3.954402e+01 3.207475e-10      421         776
#> challenges             3.938708e+01 3.475887e-10       16           0
#> cannot                 3.914181e+01 3.941127e-10       36          20
#> lives                  3.798479e+01 7.129835e-10       32          16
#> across                 3.750232e+01 9.130433e-10       22           6
#> dream                  3.737364e+01 9.753134e-10       17           2
#> generation             3.721345e+01 1.058811e-09       28          12
#> moment                 3.710940e+01 1.116838e-09       27          11
#> must                   3.708057e+01 1.133474e-09      151         215
#> again                  3.687284e+01 1.260900e-09       42          29
#> young                  3.582870e+01 2.154502e-09       19           4
#> words                  3.485706e+01 3.548204e-09       26          11
#> weapons                3.402350e+01 5.445038e-09       14           0
#> resolve                3.389803e+01 5.807758e-09       17           3
#> faith                  3.349866e+01 7.131329e-09       48          40
#> heart                  3.323535e+01 8.165248e-09       28          14
#> day                    3.314579e+01 8.550135e-09       45          36
#> nation                 3.278191e+01 1.031001e-08      123         170
#> friends                3.262894e+01 1.115414e-08       25          11
#> children               3.239258e+01 1.259667e-08       31          18
#> because                3.217192e+01 1.411158e-08       59          58
#> dreams                 3.169042e+01 1.808117e-08       16           2
#> your                   3.144598e+01 2.050668e-08       61          62
#> role                   3.134307e+01 2.162292e-08       13           0
#> peoples                3.086078e+01 2.772182e-08       26          13
#> forward                3.042712e+01 3.466481e-08       24          11
#> unity                  3.010072e+01 4.101794e-08       21           8
#> meaning                2.910633e+01 6.851291e-08       15           2
#> vice                   2.910633e+01 6.851291e-08       15           2
#> communism              2.866383e+01 8.609686e-08       12           0
#> begin                  2.866383e+01 8.609686e-08       12           0
#> yes                    2.866383e+01 8.609686e-08       12           0
#> this                   2.860926e+01 8.855778e-08      295         540
#> join                   2.829412e+01 1.042122e-07       16           4
#> president              2.817614e+01 1.107614e-07       46          42
#> world's                2.783617e+01 1.320347e-07       18           6
#> man                    2.765126e+01 1.452771e-07       45          41
#> story                  2.737176e+01 1.678644e-07       13           1
#> strong                 2.685147e+01 2.197069e-07       32          23
#> families               2.653518e+01 2.587820e-07       14           2
#> back                   2.611192e+01 3.221896e-07       22          11
#> believe                2.607958e+01 3.276300e-07       40          35
#> jobs                   2.598610e+01 3.438840e-07       11           0
#> seek                   2.527910e+01 4.960543e-07       34          27
#> celebrate              2.474698e+01 6.537130e-07       12           1
#> when                   2.470508e+01 6.680797e-07       90         123
#> women                  2.400406e+01 9.613302e-07       21          11
#> senator                2.331033e+01 1.378579e-06       10           0
#> nuclear                2.331033e+01 1.378579e-06       10           0
#> remember               2.320192e+01 1.458514e-06       17           7
#> ask                    2.275541e+01 1.839859e-06       26          18
#> free                   2.227014e+01 2.368613e-06       78         105
#> change                 2.191660e+01 2.847600e-06       36          33
#> peace                  2.186958e+01 2.918237e-06      102         152
#> old                    2.157249e+01 3.407032e-06       39          38
#> courage                2.145599e+01 3.620428e-06       23          15
#> centuries              2.144282e+01 3.645380e-06       12           2
#> make                   2.125829e+01 4.013703e-06       64          81
#> born                   2.105791e+01 4.456087e-06       13           3
#> pledge                 2.104440e+01 4.487625e-06       22          14
#> need                   2.104283e+01 4.491306e-06       38          37
#> values                 2.096359e+01 4.680959e-06       16           7
#> challenge              2.087629e+01 4.899226e-06       14           4
#> hard                   2.084164e+01 4.988641e-06       14           5
#> third                  2.063714e+01 5.550876e-06        9           0
#> come                   2.013304e+01 7.223784e-06       38          38
#> hope                   1.957075e+01 9.694199e-06       53          64
#> who                    1.942541e+01 1.046062e-05      138         232
#> american               1.909419e+01 1.244229e-05       69          94
#> here                   1.894053e+01 1.348571e-05       39          41
#> threat                 1.892927e+01 1.356552e-05       11           2
#> tyranny                1.892927e+01 1.356552e-05       11           2
#> go                     1.890993e+01 1.370373e-05       26          21
#> nation's               1.876611e+01 1.477694e-05       15           7
#> poverty                1.876611e+01 1.477694e-05       15           7
#> answer                 1.862883e+01 1.588007e-05       12           3
#> renew                  1.852375e+01 1.678003e-05       13           4
#> voices                 1.852375e+01 1.678003e-05       13           4
#> each                   1.802778e+01 2.177048e-05       55          70
#> areas                  1.796750e+01 2.247084e-05        8           0
#> heal                   1.796750e+01 2.247084e-05        8           0
#> define                 1.796750e+01 2.247084e-05        8           0
#> compassion             1.796750e+01 2.247084e-05        8           0
#> speaker                1.796750e+01 2.247084e-05        8           0
#> quiet                  1.796750e+01 2.247084e-05        8           0
#> don't                  1.796750e+01 2.247084e-05        8           0
#> bring                  1.791263e+01 2.312817e-05       24          19
#> democracy              1.776574e+01 2.498415e-05       30          28
#> but                    1.761623e+01 2.702720e-05      225         429
#> child                  1.694788e+01 3.842026e-05        9           1
#> founding               1.694788e+01 3.842026e-05        9           1
#> age                    1.688853e+01 3.964045e-05       15           8
#> land                   1.682733e+01 4.093943e-05       35          37
#> creed                  1.644471e+01 5.008970e-05       10           2
#> bold                   1.644471e+01 5.008970e-05       10           2
#> simple                 1.639618e+01 5.138856e-05       13           6
#> achieve                1.639618e+01 5.138856e-05       13           6
#> schools                1.631250e+01 5.370846e-05       13           5
#> strive                 1.631250e+01 5.370846e-05       13           5
#> friend                 1.623668e+01 5.590136e-05       11           3
#> leaders                1.623668e+01 5.590136e-05       11           3
#> learned                1.623668e+01 5.590136e-05       11           3
#> man's                  1.623668e+01 5.590136e-05       11           3
#> program                1.623668e+01 5.590136e-05       11           3
#> planet                 1.530292e+01 9.157466e-05        7           0
#> sick                   1.530292e+01 9.157466e-05        7           0
#> productivity           1.530292e+01 9.157466e-05        7           0
#> decency                1.530292e+01 9.157466e-05        7           0
#> decades                1.530292e+01 9.157466e-05        7           0
#> goals                  1.530292e+01 9.157466e-05        7           0
#> \\                     1.530292e+01 9.157466e-05        7           0
#> a                      1.493776e+01 1.111165e-04      690        1556
#> factories              1.438717e+01 1.488132e-04        8           1
#> historic               1.438717e+01 1.488132e-04        8           1
#> can                    1.416016e+01 1.678878e-04      164         307
#> way                    1.411781e+01 1.717101e-04       37          44
#> ceremony               1.399704e+01 1.830985e-04        9           2
#> decent                 1.399704e+01 1.830985e-04        9           2
#> everyone               1.399704e+01 1.830985e-04        9           2
#> commitment             1.399704e+01 1.830985e-04        9           2
#> goal                   1.399704e+01 1.830985e-04        9           2
#> prayer                 1.395270e+01 1.874678e-04       11           4
#> move                   1.395270e+01 1.874678e-04       11           4
#> small                  1.373758e+01 2.102062e-04       18          14
#> lead                   1.324991e+01 2.725930e-04       17          13
#> generations            1.324991e+01 2.725930e-04       17          13
#> is                     1.320179e+01 2.796824e-04      458        1004
#> fellow                 1.316010e+01 2.859732e-04       45          60
#> life                   1.312163e+01 2.919062e-04       56          81
#> human                  1.294591e+01 3.206218e-04       41          53
#> instead                1.285460e+01 3.366500e-04       13           8
#> misery                 1.264591e+01 3.763887e-04        6           0
#> hunger                 1.264591e+01 3.763887e-04        6           0
#> quest                  1.264591e+01 3.763887e-04        6           0
#> night                  1.264591e+01 3.763887e-04        6           0
#> deepest                1.264591e+01 3.763887e-04        6           0
#> celebration            1.264591e+01 3.763887e-04        6           0
#> civility               1.264591e+01 3.763887e-04        6           0
#> choices                1.264591e+01 3.763887e-04        6           0
#> ensure                 1.264591e+01 3.763887e-04        6           0
#> adversaries            1.264591e+01 3.763887e-04        6           0
#> timeless               1.264591e+01 3.763887e-04        6           0
#> bush                   1.264591e+01 3.763887e-04        6           0
#> workers                1.264591e+01 3.763887e-04        6           0
#> streets                1.264591e+01 3.763887e-04        6           0
#> budget                 1.264591e+01 3.763887e-04        6           0
#> enemies                1.231091e+01 4.503192e-04       12           6
#> sacrifice              1.229407e+01 4.543997e-04       15          11
#> end                    1.218334e+01 4.821830e-04       23          23
#> better                 1.193495e+01 5.509066e-04       34          42
#> security               1.191505e+01 5.568221e-04       30          35
#> renewal                1.185811e+01 5.741080e-04        7           1
#> basic                  1.185811e+01 5.741080e-04        7           1
#> learn                  1.175332e+01 6.073467e-04       10           4
#> ideas                  1.160003e+01 6.595082e-04        9           3
#> try                    1.160003e+01 6.595082e-04        9           3
#> truly                  1.160003e+01 6.595082e-04        9           3
#> welcome                1.160003e+01 6.595082e-04        9           3
#> play                   1.160003e+01 6.595082e-04        9           3
#> shape                  1.159731e+01 6.604713e-04        8           2
#> heroes                 1.159731e+01 6.604713e-04        8           2
#> working                1.159731e+01 6.604713e-04        8           2
#> shared                 1.159731e+01 6.604713e-04        8           2
#> anew                   1.159731e+01 6.604713e-04        8           2
#> reach                  1.154531e+01 6.792028e-04       16          13
#> think                  1.154531e+01 6.792028e-04       16          13
#> act                    1.130155e+01 7.744240e-04       26          29
#> nations                1.125353e+01 7.947173e-04       75         124
#> once                   1.112362e+01 8.523535e-04       21          21
#> stand                  1.108058e+01 8.723658e-04       27          31
#> chance                 1.093825e+01 9.419912e-04       12           8
#> era                    1.093825e+01 9.419912e-04       12           8
#> responsibility         1.071729e+01 1.061388e-03       29          35
#> ourselves              1.061066e+01 1.124378e-03       35          46
#> that                   1.057059e+01 1.149012e-03      546        1243
#> "                      1.056188e+01 1.154439e-03       90         158
#> proud                  1.032220e+01 1.314399e-03       11           6
#> policies               1.032220e+01 1.314399e-03       11           6
#> programs               1.000097e+01 1.564576e-03        5           0
#> cold                   1.000097e+01 1.564576e-03        5           0
#> brothers               1.000097e+01 1.564576e-03        5           0
#> self                   1.000097e+01 1.564576e-03        5           0
#> hungry                 1.000097e+01 1.564576e-03        5           0
#> knows                  1.000097e+01 1.564576e-03        5           0
#> ended                  1.000097e+01 1.564576e-03        5           0
#> hatreds                1.000097e+01 1.564576e-03        5           0
#> dawn                   1.000097e+01 1.564576e-03        5           0
#> today's                1.000097e+01 1.564576e-03        5           0
#> breeze                 1.000097e+01 1.564576e-03        5           0
#> drugs                  1.000097e+01 1.564576e-03        5           0
#> edge                   1.000097e+01 1.564576e-03        5           0
#> god's                  9.951189e+00 1.607455e-03       10           5
#> building               9.951189e+00 1.607455e-03       10           5
#> few                    9.935423e+00 1.621282e-03       17          16
#> light                  9.910947e+00 1.642985e-03       15          13
#> say                    9.731862e+00 1.811009e-03       20          21
#> fate                   9.629033e+00 1.915254e-03        9           4
#> choice                 9.575758e+00 1.971634e-03       12           9
#> fight                  9.382365e+00 2.190827e-03        8           3
#> allies                 9.382365e+00 2.190827e-03        8           3
#> lift                   9.382365e+00 2.190827e-03        8           3
#> turning                9.382365e+00 2.190827e-03        8           3
#> watching               9.374260e+00 2.200535e-03        6           1
#> bible                  9.374260e+00 2.200535e-03        6           1
#> big                    9.374260e+00 2.200535e-03        6           1
#> job                    9.374260e+00 2.200535e-03        6           1
#> tens                   9.374260e+00 2.200535e-03        6           1
#> feed                   9.374260e+00 2.200535e-03        6           1
#> long                   9.302682e+00 2.288186e-03       40          58
#> hear                   9.261501e+00 2.340211e-03        7           2
#> gift                   9.261501e+00 2.340211e-03        7           2
#> loyalty                9.261501e+00 2.340211e-03        7           2
#>  [ reached getOption("max.print") -- omitted 9107 rows ]

# compare Trump 2017 to other post-war preseidents
pwdfm <- dfm(corpus_subset(data_corpus_inaugural, period == "post-war"))
head(textstat_keyness(pwdfm, target = "2017-Trump"), 10)

#>                  chi2            p n_target n_reference
#> protected    76.64873 0.000000e+00        5           1
#> will         51.45307 7.333023e-13       40         299
#> while        48.23301 3.784639e-12        6           7
#> obama        47.85978 4.578227e-12        3           0
#> we've        47.85978 4.578227e-12        3           0
#> america      31.45816 2.037841e-08       18         112
#> again        27.81347 1.335930e-07        9          33
#> everyone     27.68039 1.431061e-07        4           5
#> your         26.68107 2.399608e-07       11          50
#> transferring 25.54705 4.317259e-07        2           0

# using the likelihood ratio method
head(textstat_keyness(dfm_smooth(pwdfm), measure = "lr", target = "2017-Trump"), 10)

#>                  G2            p n_target n_reference
#> will      24.604906 7.037235e-07       41         317
#> america   14.040653 1.789008e-04       19         130
#> your      10.435398 1.236229e-03       12          68
#> again      9.758737 1.784725e-03       10          51
#> while      9.505157 2.048952e-03        7          25
#> american   8.877934 2.886380e-03       12          76
#> protected  8.820707 2.978313e-03        6          19
#> back       6.853678 8.845897e-03        7          34
#> you        6.713453 9.568826e-03       14         121
#> country    5.821785 1.582887e-02       10          72