calculate keyness statistics

calculate keyness statistics

textstat_keyness(x, target = 1L, measure = c("chi2", "exact", "lr", "pmi"),
  sort = TRUE, correction = c("default", "yates", "williams", "none"))

Arguments

x
a dfm containing the features to be examined for keyness

target
the document index (numeric, character or logical) identifying the document forming the "target" for computing keyness; all other documents' feature frequencies will be combined for use as a reference

measure
(signed) association measure to be used for computing keyness. Currenly available: "chi2"; "exact" (Fisher's exact test); "lr" for the likelihood ratio; "pmi" for pointwise mutual information.

sort
logical; if TRUE sort features scored in descending order of the measure, otherwise leave in original feature order

correction
if "default", Yates correction is applied to "chi2"; William's correction is applied to "lr"; and no correction is applied for the "exact" and "pmi" measures. Specifying a value other than the default can be used to override the defaults, for instance to apply the Williams correction to the chi2 measure. Specying a correction for the "exact" and "pmi" measures has no effect and produces a warning.

Value

a data.frame of computed statistics and associated p-values, where the features scored name each row, and the number of occurrences for both the target and reference groups. For measure = "chi2" this is the chi-squared value, signed positively if the observed value in the target exceeds its expected value; for measure = "exact" this is the estimate of the odds ratio; for measure = "lr" this is the likelihood ratio \(G2\) statistic; for "pmi" this is the pointwise mutual information statistics.

References

Bondi, Marina, and Mike Scott, eds. 2010. Keyness in Texts. Amsterdam, Philadelphia: John Benjamins, 2010. Stubbs, Michael. 2010. "Three Concepts of Keywords". In Keyness in Texts, Marina Bondi and Mike Scott, eds. pp21–42. Amsterdam, Philadelphia: John Benjamins. Scott, M. & Tribble, C. 2006. Textual Patterns: keyword and corpus analysis in language education. Amsterdam: Benjamins, p. 55. Dunning, Ted. 1993. "Accurate Methods for the Statistics of Surprise and Coincidence", Computational Linguistics, Vol 19, No. 1, pp. 61-74.

Examples

# compare pre- v. post-war terms using grouping
period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
mydfm <- dfm(data_corpus_inaugural, groups = period)
head(mydfm) # make sure 'post-war' is in the first row

#> Document-feature matrix of: 2 documents, 9,232 features (34.6% sparse).
#> (showing first 2 documents and first 6 features)
#>           features
#> docs       chief justice    , mr    . vice
#>   post-war    12      40 2194 25 1804   15
#>   pre-war     24      97 4832  6 3141    3

head(result <- textstat_keyness(mydfm), 10)

#>               chi2 p n_target n_reference
#> we        703.3183 0      960         779
#> -         253.3973 0      510         532
#> .         227.7287 0     1804        3141
#> our       188.5010 0      874        1307
#> us        186.3718 0      262         216
#> :         178.3530 0      105          29
#> america   176.8588 0      130          54
#> world     171.0441 0      188         126
#> americans 150.8152 0       67           7
#> freedom   139.6468 0      122          64

tail(result, 10)

#>                    chi2            p n_target n_reference
#> upon          -51.80884 6.117329e-13       39         332
#> it            -52.47374 4.359846e-13      257        1132
#> public        -55.91584 7.560619e-14       11         213
#> states        -59.02901 1.554312e-14       28         305
#> constitution  -61.09244 5.440093e-15        6         200
#> be            -72.17238 0.000000e+00      257        1225
#> should        -82.99568 0.000000e+00       15         309
#> which        -159.84877 0.000000e+00       95         911
#> of           -178.17359 0.000000e+00     1437        5666
#> the          -298.27243 0.000000e+00     1988        8094

# compare pre- v. post-war terms using logical vector
mydfm2 <- dfm(data_corpus_inaugural)
textstat_keyness(mydfm2, docvars(data_corpus_inaugural, "Year") >= 1945)

#>                            chi2            p n_target n_reference
#> we                 7.033183e+02 0.000000e+00      960         779
#> -                  2.533973e+02 0.000000e+00      510         532
#> .                  2.277287e+02 0.000000e+00     1804        3141
#> our                1.885010e+02 0.000000e+00      874        1307
#> us                 1.863718e+02 0.000000e+00      262         216
#> :                  1.783530e+02 0.000000e+00      105          29
#> america            1.768588e+02 0.000000e+00      130          54
#> world              1.710441e+02 0.000000e+00      188         126
#> americans          1.508152e+02 0.000000e+00       67           7
#> freedom            1.396468e+02 0.000000e+00      122          64
#> new                1.343087e+02 0.000000e+00      150         102
#> today              1.290356e+02 0.000000e+00       76          21
#> let                1.116881e+02 0.000000e+00      100          54
#> together           1.051150e+02 0.000000e+00       64          19
#> you                1.022438e+02 0.000000e+00      116          80
#> america's          9.416566e+01 0.000000e+00       35           0
#> work               9.091650e+01 0.000000e+00       78          40
#> help               8.590415e+01 0.000000e+00       46          10
#> ,                  6.350021e+01 1.554312e-15     2194        4832
#> know               6.053412e+01 7.216450e-15       63          40
#> earth              6.041119e+01 7.660539e-15       44          18
#> history            5.779483e+01 2.908784e-14       60          38
#> god                5.660697e+01 5.317968e-14       57          35
#> live               5.344486e+01 2.658984e-13       39          16
#> do                 5.192717e+01 5.759837e-13      112         121
#> time               5.118605e+01 8.401058e-13      106         112
#> are                5.108571e+01 8.841816e-13      311         503
#> century            4.947576e+01 2.008393e-12       40          19
#> journey            4.935450e+01 2.136402e-12       20           1
#> dignity            4.869781e+01 2.985945e-12       24           4
#> what               4.858280e+01 3.166356e-12       88          86
#> will               4.742202e+01 5.723644e-12      339         572
#> promise            4.647943e+01 9.258483e-12       34          14
#> mr                 4.499384e+01 1.976552e-11       25           6
#> build              4.432193e+01 2.785772e-11       21           3
#> thank              4.399890e+01 3.285616e-11       18           1
#> bless              4.005476e+01 2.469419e-10       18           2
#> for                3.977202e+01 2.854045e-10      421         776
#> strength           3.968008e+01 2.991609e-10       45          31
#> challenges         3.943058e+01 3.399322e-10       16           0
#> cannot             3.920598e+01 3.813707e-10       36          20
#> lives              3.804444e+01 6.915142e-10       32          16
#> across             3.755188e+01 8.901374e-10       22           6
#> dream              3.741762e+01 9.535663e-10       17           2
#> generation         3.726878e+01 1.029192e-09       28          12
#> must               3.721043e+01 1.060448e-09      151         215
#> moment             3.716370e+01 1.086171e-09       27          11
#> again              3.694006e+01 1.218174e-09       42          29
#> young              3.587389e+01 2.105106e-09       19           4
#> words              3.490867e+01 3.455383e-09       26          11
#> weapons            3.406121e+01 5.340536e-09       14           0
#> resolve            3.393971e+01 5.684663e-09       17           3
#> faith              3.356720e+01 6.884392e-09       48          40
#> heart              3.328755e+01 7.948983e-09       28          14
#> friends            3.267788e+01 1.087678e-08       25          11
#> children           3.244674e+01 1.225046e-08       31          18
#> nation             3.237397e+01 1.271785e-08      123         171
#> because            3.224657e+01 1.357968e-08       59          58
#> dreams             3.172895e+01 1.772601e-08       16           2
#> your               3.152108e+01 1.972873e-08       61          62
#> role               3.137788e+01 2.123865e-08       13           0
#> peoples            3.090925e+01 2.703800e-08       26          13
#> unity              3.014388e+01 4.011501e-08       21           8
#> meaning            2.914201e+01 6.726262e-08       15           2
#> this               2.877155e+01 8.143857e-08      295         540
#> communism          2.869575e+01 8.468931e-08       12           0
#> begin              2.869575e+01 8.468931e-08       12           0
#> yes                2.869575e+01 8.468931e-08       12           0
#> forward            2.853104e+01 9.220816e-08       24          12
#> join               2.833088e+01 1.022519e-07       16           4
#> world's            2.787467e+01 1.294328e-07       18           6
#> story              2.740403e+01 1.650865e-07       13           1
#> president          2.724162e+01 1.795518e-07       46          43
#> strong             2.690155e+01 2.140874e-07       32          23
#> families           2.656803e+01 2.544198e-07       14           2
#> day                2.627587e+01 2.959632e-07       46          44
#> back               2.615293e+01 3.154177e-07       22          11
#> believe            2.613483e+01 3.183883e-07       40          35
#> vice               2.603583e+01 3.351402e-07       15           3
#> jobs               2.601513e+01 3.387520e-07       11           0
#> man                2.574569e+01 3.894954e-07       45          43
#> seek               2.532921e+01 4.833346e-07       34          27
#> when               2.478678e+01 6.403522e-07       90         123
#> celebrate          2.477638e+01 6.438161e-07       12           1
#> women              2.404246e+01 9.423443e-07       21          11
#> centuries          2.400959e+01 9.585730e-07       13           2
#> free               2.337626e+01 1.332121e-06       79         105
#> senator            2.333646e+01 1.359972e-06       10           0
#> nuclear            2.333646e+01 1.359972e-06       10           0
#> remember           2.323599e+01 1.432905e-06       17           7
#> peace              2.283776e+01 1.762679e-06      103         152
#> ask                2.279696e+01 1.800503e-06       26          18
#> change             2.196468e+01 2.777140e-06       36          33
#> courage            2.149395e+01 3.549472e-06       23          15
#> make               2.132199e+01 3.882521e-06       64          81
#> need               2.109129e+01 4.379149e-06       38          37
#> pledge             2.108117e+01 4.402340e-06       22          14
#> values             2.099498e+01 4.604885e-06       16           7
#> challenge          2.090517e+01 4.825908e-06       14           4
#> hard               2.088831e+01 4.868573e-06       15           6
#> old                2.071181e+01 5.338574e-06       39          39
#> come               2.018047e+01 7.046882e-06       38          38
#> hope               1.962627e+01 9.416547e-06       53          64
#> who                1.951627e+01 9.974638e-06      138         232
#> american           1.915709e+01 1.203897e-05       69          94
#> here               1.898718e+01 1.315994e-05       39          41
#> threat             1.895365e+01 1.339328e-05       11           2
#> tyranny            1.895365e+01 1.339328e-05       11           2
#> go                 1.894784e+01 1.343414e-05       26          21
#> nation's           1.879485e+01 1.455595e-05       15           7
#> poverty            1.879485e+01 1.455595e-05       15           7
#> answer             1.865405e+01 1.567142e-05       12           3
#> renew              1.854990e+01 1.655138e-05       13           4
#> voices             1.854990e+01 1.655138e-05       13           4
#> each               1.808218e+01 2.115726e-05       55          70
#> areas              1.798786e+01 2.223178e-05        8           0
#> heal               1.798786e+01 2.223178e-05        8           0
#> define             1.798786e+01 2.223178e-05        8           0
#> compassion         1.798786e+01 2.223178e-05        8           0
#> speaker            1.798786e+01 2.223178e-05        8           0
#> quiet              1.798786e+01 2.223178e-05        8           0
#> don't              1.798786e+01 2.223178e-05        8           0
#> bring              1.794807e+01 2.270149e-05       24          19
#> democracy          1.780528e+01 2.447035e-05       30          28
#> but                1.772791e+01 2.548600e-05      225         429
#> founding           1.696872e+01 3.800076e-05        9           1
#> third              1.696872e+01 3.800076e-05        9           1
#> age                1.691575e+01 3.907594e-05       15           8
#> land               1.686900e+01 4.005041e-05       35          37
#> creed              1.646629e+01 4.952272e-05       10           2
#> bold               1.646629e+01 4.952272e-05       10           2
#> simple             1.642119e+01 5.071498e-05       13           6
#> achieve            1.642119e+01 5.071498e-05       13           6
#> schools            1.633698e+01 5.301924e-05       13           5
#> born               1.633698e+01 5.301924e-05       13           5
#> strive             1.633698e+01 5.301924e-05       13           5
#> friend             1.625914e+01 5.524243e-05       11           3
#> leaders            1.625914e+01 5.524243e-05       11           3
#> learned            1.625914e+01 5.524243e-05       11           3
#> man's              1.625914e+01 5.524243e-05       11           3
#> program            1.625914e+01 5.524243e-05       11           3
#> planet             1.532040e+01 9.073152e-05        7           0
#> sick               1.532040e+01 9.073152e-05        7           0
#> productivity       1.532040e+01 9.073152e-05        7           0
#> decency            1.532040e+01 9.073152e-05        7           0
#> decades            1.532040e+01 9.073152e-05        7           0
#> goals              1.532040e+01 9.073152e-05        7           0
#> \\                 1.532040e+01 9.073152e-05        7           0
#> a                  1.512173e+01 1.007965e-04      690        1556
#> factories          1.440517e+01 1.473968e-04        8           1
#> historic           1.440517e+01 1.473968e-04        8           1
#> can                1.424551e+01 1.604435e-04      164         307
#> ceremony           1.401584e+01 1.812767e-04        9           2
#> decent             1.401584e+01 1.812767e-04        9           2
#> everyone           1.401584e+01 1.812767e-04        9           2
#> child              1.401584e+01 1.812767e-04        9           2
#> commitment         1.401584e+01 1.812767e-04        9           2
#> goal               1.401584e+01 1.812767e-04        9           2
#> prayer             1.397346e+01 1.854097e-04       11           4
#> move               1.397346e+01 1.854097e-04       11           4
#> small              1.376446e+01 2.072197e-04       18          14
#> way                1.348544e+01 2.404219e-04       37          45
#> is                 1.334199e+01 2.595287e-04      458        1004
#> lead               1.327556e+01 2.688884e-04       17          13
#> generations        1.327556e+01 2.688884e-04       17          13
#> life               1.316876e+01 2.846557e-04       56          81
#> human              1.298574e+01 3.138723e-04       41          53
#> instead            1.287670e+01 3.326987e-04       13           8
#> misery             1.266050e+01 3.734626e-04        6           0
#> hunger             1.266050e+01 3.734626e-04        6           0
#> quest              1.266050e+01 3.734626e-04        6           0
#> night              1.266050e+01 3.734626e-04        6           0
#> deepest            1.266050e+01 3.734626e-04        6           0
#> celebration        1.266050e+01 3.734626e-04        6           0
#> civility           1.266050e+01 3.734626e-04        6           0
#> choices            1.266050e+01 3.734626e-04        6           0
#> ensure             1.266050e+01 3.734626e-04        6           0
#> adversaries        1.266050e+01 3.734626e-04        6           0
#> timeless           1.266050e+01 3.734626e-04        6           0
#> bush               1.266050e+01 3.734626e-04        6           0
#> workers            1.266050e+01 3.734626e-04        6           0
#> streets            1.266050e+01 3.734626e-04        6           0
#> budget             1.266050e+01 3.734626e-04        6           0
#> enemies            1.233125e+01 4.454374e-04       12           6
#> sacrifice          1.231728e+01 4.487849e-04       15          11
#> end                1.221205e+01 4.748196e-04       23          23
#> better             1.196971e+01 5.407227e-04       34          42
#> security           1.194761e+01 5.471746e-04       30          35
#> renewal            1.187330e+01 5.694447e-04        7           1
#> big                1.187330e+01 5.694447e-04        7           1
#> basic              1.187330e+01 5.694447e-04        7           1
#> learn              1.177142e+01 6.014701e-04       10           4
#> ideas              1.161706e+01 6.534950e-04        9           3
#> try                1.161706e+01 6.534950e-04        9           3
#> truly              1.161706e+01 6.534950e-04        9           3
#> welcome            1.161706e+01 6.534950e-04        9           3
#> play               1.161706e+01 6.534950e-04        9           3
#> shape              1.161336e+01 6.547969e-04        8           2
#> heroes             1.161336e+01 6.547969e-04        8           2
#> working            1.161336e+01 6.547969e-04        8           2
#> shared             1.161336e+01 6.547969e-04        8           2
#> anew               1.161336e+01 6.547969e-04        8           2
#> reach              1.156855e+01 6.707663e-04       16          13
#> think              1.156855e+01 6.707663e-04       16          13
#> act                1.133103e+01 7.622259e-04       26          29
#> nations            1.130447e+01 7.732069e-04       75         124
#> once               1.114983e+01 8.403937e-04       21          21
#> stand              1.111035e+01 8.584718e-04       27          31
#> chance             1.095783e+01 9.320904e-04       12           8
#> era                1.095783e+01 9.320904e-04       12           8
#> responsibility     1.074769e+01 1.044094e-03       29          35
#> that               1.070836e+01 1.066523e-03      546        1243
#> ourselves          1.064400e+01 1.104284e-03       35          46
#> "                  1.061619e+01 1.121020e-03       90         158
#> proud              1.033999e+01 1.301788e-03       11           6
#> policies           1.033999e+01 1.301788e-03       11           6
#> programs           1.001269e+01 1.554655e-03        5           0
#> cold               1.001269e+01 1.554655e-03        5           0
#> brothers           1.001269e+01 1.554655e-03        5           0
#> hungry             1.001269e+01 1.554655e-03        5           0
#> knows              1.001269e+01 1.554655e-03        5           0
#> ended              1.001269e+01 1.554655e-03        5           0
#> hatreds            1.001269e+01 1.554655e-03        5           0
#> dawn               1.001269e+01 1.554655e-03        5           0
#> ending             1.001269e+01 1.554655e-03        5           0
#> today's            1.001269e+01 1.554655e-03        5           0
#> breeze             1.001269e+01 1.554655e-03        5           0
#> drugs              1.001269e+01 1.554655e-03        5           0
#> edge               1.001269e+01 1.554655e-03        5           0
#> god's              9.967813e+00 1.593006e-03       10           5
#> building           9.967813e+00 1.593006e-03       10           5
#> few                9.957686e+00 1.601792e-03       17          16
#> light              9.931809e+00 1.624468e-03       15          13
#> say                9.755817e+00 1.787562e-03       20          21
#> fate               9.644512e+00 1.899182e-03        9           4
#> choice             9.594079e+00 1.952058e-03       12           9
#> fight              9.396735e+00 2.173722e-03        8           3
#> allies             9.396735e+00 2.173722e-03        8           3
#> lift               9.396735e+00 2.173722e-03        8           3
#> turning            9.396735e+00 2.173722e-03        8           3
#> watching           9.386652e+00 2.185710e-03        6           1
#> bible              9.386652e+00 2.185710e-03        6           1
#> job                9.386652e+00 2.185710e-03        6           1
#> tens               9.386652e+00 2.185710e-03        6           1
#> feed               9.386652e+00 2.185710e-03        6           1
#> hear               9.274825e+00 2.323249e-03        7           2
#> gift               9.274825e+00 2.323249e-03        7           2
#> loyalty            9.274825e+00 2.323249e-03        7           2
#> soul               9.274825e+00 2.323249e-03        7           2
#> price              9.120181e+00 2.528049e-03       11           8
#>  [ reached getOption("max.print") -- omitted 8982 rows ]

# compare Trump 2017 to other post-war preseidents
pwdfm <- dfm(corpus_subset(data_corpus_inaugural, period == "post-war"))
head(textstat_keyness(pwdfm, target = "2017-Trump"), 10)

#>                  chi2            p n_target n_reference
#> protected    76.79339 0.000000e+00        5           1
#> will         51.63502 6.683543e-13       40         299
#> while        48.33243 3.597567e-12        6           7
#> obama        47.94909 4.374279e-12        3           0
#> we've        47.94909 4.374279e-12        3           0
#> america      31.55748 1.936229e-08       18         112
#> again        27.88512 1.287361e-07        9          33
#> everyone     27.73848 1.388726e-07        4           5
#> your         26.75528 2.309197e-07       11          50
#> transferring 25.59527 4.210694e-07        2           0

# using the likelihood ratio method
head(textstat_keyness(dfm_smooth(pwdfm), measure = "lr", target = "2017-Trump"), 10)

#>                  G2            p n_target n_reference
#> will      24.653617 6.861584e-07       41         317
#> america   14.064791 1.766193e-04       19         130
#> your      10.450999 1.225832e-03       12          68
#> again      9.772128 1.771773e-03       10          51
#> while      9.515287 2.037673e-03        7          25
#> american   8.892732 2.863081e-03       12          76
#> protected  8.829483 2.964023e-03        6          19
#> back       6.862906 8.800333e-03        7          34
#> you        6.728679 9.487483e-03       14         121
#> country    5.833068 1.572767e-02       10          72