calculate keyness statistics

calculate keyness statistics

textstat_keyness(x, target = 1L, measure = c("chi2", "exact", "lr"),
  sort = TRUE)

Arguments

x

a dfm containing the features to be examined for keyness

target

the document index (numeric or character) identifying the document forming the "target" for computing keyness; all other documents' feature frequencies will be combined for use as a reference

measure

(signed) association measure to be used for computing keyness. Currenly available: "chi2" (\(chi^2\) with Yates correction); "exact" (Fisher's exact test); "lr" for the likelihood ratio \(G^2\) statistic with Yates correction.

sort

logical; if TRUE sort features scored in descending order of the measure, otherwise leave in original feature order

Value

a named numeric vector of keyness scores, named with the features. For measure = "chi2" this is the chi-squared value, signed positively if the observed value in the target exceeds its expected value; for measure = "exact" this is the estimate of the odds ratio; for measure = "lr" this is the likelihood ratio $G^2$ statistic.

References

Bondi, Marina, and Mike Scott, eds. 2010. Keyness in Texts. Amsterdam, Philadelphia: John Benjamins, 2010.

Stubbs, Michael. 2010. "Three Concepts of Keywords". In Keyness in Texts, Marina Bondi and Mike Scott, eds. pp21–42. Amsterdam, Philadelphia: John Benjamins.

Scott, M. & Tribble, C. 2006. Textual Patterns: keyword and corpus analysis in language education. Amsterdam: Benjamins, p. 55.

Dunning, Ted. 1993. "Accurate Methods for the Statistics of Surprise and Coincidence", Computational Linguistics, Vol 19, No. 1, pp. 61-74.

Examples

# compare pre- v. post-war terms using grouping
period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
mydfm <- dfm(data_corpus_inaugural, groups = period)
head(mydfm) # make sure 'post-war' is in the first row

#> Document-feature matrix of: 2 documents, 9,232 features (34.6% sparse).
#> (showing first 2 documents and first 6 features)
#>           features
#> docs       chief justice    , mr    . vice
#>   post-war    12      40 2194 25 1804   15
#>   pre-war     24      97 4832  6 3141    3

head(result <- textstat_keyness(mydfm), 10)

#>        we         -         .       our        us         :   america     world 
#>  701.8799  252.2852  227.2380  187.8353  184.9674  175.7657  174.6586  169.3859 
#> americans   freedom 
#>  147.6197  137.7030 

tail(result, 10)

#>         upon           it       public       states constitution           be 
#>    -50.97048    -52.03533    -54.79660    -58.08454    -59.87256    -71.67429 
#>       should        which           of          the 
#>    -81.85968   -158.95020   -177.80869   -297.87187 

# compare Trump 2017 to other post-war preseidents
pwdfm <- dfm(corpus_subset(data_corpus_inaugural, period == "post-war"))
head(textstat_keyness(pwdfm, target = "2017-Trump"), 10)

#>    protected         will        while        obama        we've      america 
#>     76.79339     49.67662     48.33243     47.94909     47.94909     29.11681 
#>        again     everyone         your transferring 
#>     27.88512     27.73848     26.75528     25.59527 

# using the exact method
head(textstat_keyness(dfm_smooth(pwdfm), measure = "exact", target = "2017-Trump"), 10)

#> protected     while     obama     we've  everyone    dreams      back      jobs 
#>  6.160066  5.462699  4.333627  4.333627  4.239974  4.035896  4.016628  3.901020 
#>     again   borders 
#>  3.826084  3.714259