calculate keyness statistics

calculate keyness statistics

textstat_keyness(x, target = 1L, measure = c("chi2", "exact", "lr"),
  sort = TRUE)

Arguments

x

a dfm containing the features to be examined for keyness

target

the document index (numeric, character or logical) identifying the document forming the "target" for computing keyness; all other documents' feature frequencies will be combined for use as a reference

measure

(signed) association measure to be used for computing keyness. Currenly available: "chi2" (\(chi^2\) with Yates correction); "exact" (Fisher's exact test); "lr" for the likelihood ratio \(G\) statistic with Yates correction.

sort

logical; if TRUE sort features scored in descending order of the measure, otherwise leave in original feature order

Value

a data.frame of computed statistics and associated p-values, where the features scored name each row. For measure = "chi2" this is the chi-squared value, signed positively if the observed value in the target exceeds its expected value; for measure = "exact" this is the estimate of the odds ratio; for measure = "lr" this is the likelihood ratio \(G\) statistic.

References

Bondi, Marina, and Mike Scott, eds. 2010. Keyness in Texts. Amsterdam, Philadelphia: John Benjamins, 2010.

Stubbs, Michael. 2010. "Three Concepts of Keywords". In Keyness in Texts, Marina Bondi and Mike Scott, eds. pp21–42. Amsterdam, Philadelphia: John Benjamins.

Scott, M. & Tribble, C. 2006. Textual Patterns: keyword and corpus analysis in language education. Amsterdam: Benjamins, p. 55.

Dunning, Ted. 1993. "Accurate Methods for the Statistics of Surprise and Coincidence", Computational Linguistics, Vol 19, No. 1, pp. 61-74.

Examples

# compare pre- v. post-war terms using grouping
period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
mydfm <- dfm(data_corpus_inaugural, groups = period)
head(mydfm) # make sure 'post-war' is in the first row

#> Document-feature matrix of: 2 documents, 9,232 features (34.6% sparse).
#> (showing first 2 documents and first 6 features)
#>           features
#> docs       chief justice    , mr    . vice
#>   post-war    12      40 2194 25 1804   15
#>   pre-war     24      97 4832  6 3141    3

head(result <- textstat_keyness(mydfm), 10)

#>               chi2 p
#> we        701.8799 0
#> -         252.2852 0
#> .         227.2380 0
#> our       187.8353 0
#> us        184.9674 0
#> :         175.7657 0
#> america   174.6586 0
#> world     169.3859 0
#> americans 147.6197 0
#> freedom   137.7030 0

tail(result, 10)

#>                    chi2            p
#> upon          -50.97048 9.376944e-13
#> it            -52.03533 5.451195e-13
#> public        -54.79660 1.336709e-13
#> states        -58.08454 2.509104e-14
#> constitution  -59.87256 1.010303e-14
#> be            -71.67429 0.000000e+00
#> should        -81.85968 0.000000e+00
#> which        -158.95020 0.000000e+00
#> of           -177.80869 0.000000e+00
#> the          -297.87187 0.000000e+00

# compare pre- v. post-war terms using logical vector
mydfm2 <- dfm(data_corpus_inaugural)
textstat_keyness(mydfm2, docvars(data_corpus_inaugural, "Year") >= 1945)

#>                            chi2            p
#> we                 7.018799e+02 0.000000e+00
#> -                  2.522852e+02 0.000000e+00
#> .                  2.272380e+02 0.000000e+00
#> our                1.878353e+02 0.000000e+00
#> us                 1.849674e+02 0.000000e+00
#> :                  1.757657e+02 0.000000e+00
#> america            1.746586e+02 0.000000e+00
#> world              1.693859e+02 0.000000e+00
#> americans          1.476197e+02 0.000000e+00
#> freedom            1.377030e+02 0.000000e+00
#> new                1.326699e+02 0.000000e+00
#> today              1.264530e+02 0.000000e+00
#> let                1.097794e+02 0.000000e+00
#> together           1.025977e+02 0.000000e+00
#> you                1.006242e+02 0.000000e+00
#> america's          9.051114e+01 0.000000e+00
#> work               8.895165e+01 0.000000e+00
#> help               8.313975e+01 0.000000e+00
#> ,                  6.328131e+01 1.776357e-15
#> know               5.882108e+01 1.731948e-14
#> earth              5.821036e+01 2.353673e-14
#> history            5.607945e+01 6.961098e-14
#> god                5.485544e+01 1.296740e-13
#> live               5.124971e+01 8.133494e-13
#> do                 5.086969e+01 9.870993e-13
#> are                5.052212e+01 1.178280e-12
#> time               5.010091e+01 1.460387e-12
#> century            4.743658e+01 5.681233e-12
#> what               4.740058e+01 5.786593e-12
#> will               4.690854e+01 7.437828e-12
#> journey            4.596549e+01 1.203548e-11
#> dignity            4.577572e+01 1.325962e-11
#> promise            4.429157e+01 2.829315e-11
#> mr                 4.232395e+01 7.733958e-11
#> build              4.131706e+01 1.294346e-10
#> thank              4.064165e+01 1.828669e-10
#> challenges         3.943058e+01 3.399322e-10
#> for                3.936134e+01 3.522012e-10
#> strength           3.807069e+01 6.822725e-10
#> cannot             3.734577e+01 9.893516e-10
#> bless              3.693396e+01 1.221991e-09
#> must               3.649566e+01 1.530062e-09
#> lives              3.606755e+01 1.905946e-09
#> again              3.533471e+01 2.776398e-09
#> generation         3.512850e+01 3.086504e-09
#> across             3.499140e+01 3.311650e-09
#> moment             3.497182e+01 3.345117e-09
#> dream              3.432681e+01 4.659189e-09
#> weapons            3.406121e+01 5.340536e-09
#> young              3.311895e+01 8.668991e-09
#> words              3.275732e+01 1.044124e-08
#> faith              3.219168e+01 1.396880e-08
#> dreams             3.172895e+01 1.772601e-08
#> nation             3.163099e+01 1.864311e-08
#> role               3.137788e+01 2.123865e-08
#> heart              3.131451e+01 2.194340e-08
#> because            3.107581e+01 2.481442e-08
#> resolve            3.107201e+01 2.486304e-08
#> children           3.064151e+01 3.103816e-08
#> friends            3.056931e+01 3.221503e-08
#> your               3.039199e+01 3.529830e-08
#> meaning            2.914201e+01 6.726262e-08
#> peoples            2.893856e+01 7.471069e-08
#> communism          2.869575e+01 8.468931e-08
#> begin              2.869575e+01 8.468931e-08
#> yes                2.869575e+01 8.468931e-08
#> this               2.835429e+01 1.010229e-07
#> unity              2.789357e+01 1.281753e-07
#> story              2.740403e+01 1.650865e-07
#> families           2.656803e+01 2.544198e-07
#> forward            2.656310e+01 2.550694e-07
#> vice               2.603583e+01 3.351402e-07
#> jobs               2.601513e+01 3.387520e-07
#> president          2.601077e+01 3.395174e-07
#> join               2.571630e+01 3.954731e-07
#> world's            2.550260e+01 4.417874e-07
#> strong             2.535083e+01 4.779462e-07
#> day                2.507395e+01 5.517329e-07
#> believe            2.482330e+01 6.283320e-07
#> celebrate          2.477638e+01 6.438161e-07
#> man                2.454282e+01 7.267650e-07
#> back               2.418822e+01 8.736487e-07
#> when               2.402471e+01 9.510743e-07
#> centuries          2.400959e+01 9.585730e-07
#> seek               2.389993e+01 1.014757e-06
#> senator            2.333646e+01 1.359972e-06
#> nuclear            2.333646e+01 1.359972e-06
#> free               2.258075e+01 2.014931e-06
#> peace              2.216887e+01 2.496891e-06
#> women              2.213170e+01 2.545710e-06
#> ask                2.120608e+01 4.124532e-06
#> remember           2.107485e+01 4.416881e-06
#> challenge          2.090517e+01 4.825908e-06
#> change             2.071339e+01 5.334174e-06
#> make               2.046758e+01 6.064988e-06
#> need               1.991480e+01 8.097099e-06
#> courage            1.983500e+01 8.442242e-06
#> old                1.956841e+01 9.706070e-06
#> pledge             1.939449e+01 1.063135e-05
#> come               1.903751e+01 1.281737e-05
#> who                1.900236e+01 1.305570e-05
#> threat             1.895365e+01 1.339328e-05
#> tyranny            1.895365e+01 1.339328e-05
#> values             1.890034e+01 1.377281e-05
#> hope               1.871529e+01 1.517607e-05
#> hard               1.870461e+01 1.526132e-05
#> answer             1.865405e+01 1.567142e-05
#> renew              1.854990e+01 1.655138e-05
#> voices             1.854990e+01 1.655138e-05
#> american           1.839315e+01 1.797024e-05
#> areas              1.798786e+01 2.223178e-05
#> heal               1.798786e+01 2.223178e-05
#> define             1.798786e+01 2.223178e-05
#> compassion         1.798786e+01 2.223178e-05
#> speaker            1.798786e+01 2.223178e-05
#> quiet              1.798786e+01 2.223178e-05
#> don't              1.798786e+01 2.223178e-05
#> here               1.790667e+01 2.320065e-05
#> go                 1.754606e+01 2.804306e-05
#> but                1.735863e+01 3.094898e-05
#> each               1.723627e+01 3.300749e-05
#> founding           1.696872e+01 3.800076e-05
#> third              1.696872e+01 3.800076e-05
#> nation's           1.677276e+01 4.213381e-05
#> poverty            1.677276e+01 4.213381e-05
#> democracy          1.658033e+01 4.663221e-05
#> bring              1.652378e+01 4.804365e-05
#> creed              1.646629e+01 4.952272e-05
#> bold               1.646629e+01 4.952272e-05
#> schools            1.633698e+01 5.301924e-05
#> born               1.633698e+01 5.301924e-05
#> strive             1.633698e+01 5.301924e-05
#> friend             1.625914e+01 5.524243e-05
#> leaders            1.625914e+01 5.524243e-05
#> learned            1.625914e+01 5.524243e-05
#> man's              1.625914e+01 5.524243e-05
#> program            1.625914e+01 5.524243e-05
#> land               1.579734e+01 7.050173e-05
#> planet             1.532040e+01 9.073152e-05
#> sick               1.532040e+01 9.073152e-05
#> productivity       1.532040e+01 9.073152e-05
#> decency            1.532040e+01 9.073152e-05
#> decades            1.532040e+01 9.073152e-05
#> goals              1.532040e+01 9.073152e-05
#> \\                 1.532040e+01 9.073152e-05
#> age                1.504122e+01 1.051886e-04
#> a                  1.493629e+01 1.112029e-04
#> factories          1.440517e+01 1.473968e-04
#> historic           1.440517e+01 1.473968e-04
#> simple             1.439612e+01 1.481074e-04
#> achieve            1.439612e+01 1.481074e-04
#> ceremony           1.401584e+01 1.812767e-04
#> decent             1.401584e+01 1.812767e-04
#> everyone           1.401584e+01 1.812767e-04
#> child              1.401584e+01 1.812767e-04
#> commitment         1.401584e+01 1.812767e-04
#> goal               1.401584e+01 1.812767e-04
#> prayer             1.397346e+01 1.854097e-04
#> move               1.397346e+01 1.854097e-04
#> can                1.385632e+01 1.973322e-04
#> is                 1.312689e+01 2.910882e-04
#> misery             1.266050e+01 3.734626e-04
#> hunger             1.266050e+01 3.734626e-04
#> quest              1.266050e+01 3.734626e-04
#> night              1.266050e+01 3.734626e-04
#> deepest            1.266050e+01 3.734626e-04
#> celebration        1.266050e+01 3.734626e-04
#> civility           1.266050e+01 3.734626e-04
#> choices            1.266050e+01 3.734626e-04
#> ensure             1.266050e+01 3.734626e-04
#> adversaries        1.266050e+01 3.734626e-04
#> timeless           1.266050e+01 3.734626e-04
#> bush               1.266050e+01 3.734626e-04
#> workers            1.266050e+01 3.734626e-04
#> streets            1.266050e+01 3.734626e-04
#> budget             1.266050e+01 3.734626e-04
#> way                1.258826e+01 3.881768e-04
#> life               1.248016e+01 4.112963e-04
#> enemies            1.233125e+01 4.454374e-04
#> small              1.232832e+01 4.461371e-04
#> human              1.216274e+01 4.875365e-04
#> renewal            1.187330e+01 5.694447e-04
#> big                1.187330e+01 5.694447e-04
#> basic              1.187330e+01 5.694447e-04
#> lead               1.182099e+01 5.856680e-04
#> generations        1.182099e+01 5.856680e-04
#> learn              1.177142e+01 6.014701e-04
#> ideas              1.161706e+01 6.534950e-04
#> try                1.161706e+01 6.534950e-04
#> truly              1.161706e+01 6.534950e-04
#> welcome            1.161706e+01 6.534950e-04
#> play               1.161706e+01 6.534950e-04
#> shape              1.161336e+01 6.547969e-04
#> heroes             1.161336e+01 6.547969e-04
#> working            1.161336e+01 6.547969e-04
#> shared             1.161336e+01 6.547969e-04
#> anew               1.161336e+01 6.547969e-04
#> instead            1.117512e+01 8.290139e-04
#> better             1.109329e+01 8.664042e-04
#> end                1.108018e+01 8.725519e-04
#> security           1.100232e+01 9.099777e-04
#> sacrifice          1.081731e+01 1.005555e-03
#> nations            1.077426e+01 1.029212e-03
#> that               1.053397e+01 1.172002e-03
#> proud              1.033999e+01 1.301788e-03
#> policies           1.033999e+01 1.301788e-03
#> act                1.033269e+01 1.306945e-03
#> reach              1.019109e+01 1.411209e-03
#> think              1.019109e+01 1.411209e-03
#> "                  1.015543e+01 1.438768e-03
#> stand              1.014731e+01 1.445119e-03
#> once               1.002062e+01 1.547975e-03
#> programs           1.001269e+01 1.554655e-03
#> cold               1.001269e+01 1.554655e-03
#> brothers           1.001269e+01 1.554655e-03
#> hungry             1.001269e+01 1.554655e-03
#> knows              1.001269e+01 1.554655e-03
#> ended              1.001269e+01 1.554655e-03
#> hatreds            1.001269e+01 1.554655e-03
#> dawn               1.001269e+01 1.554655e-03
#> ending             1.001269e+01 1.554655e-03
#> today's            1.001269e+01 1.554655e-03
#> breeze             1.001269e+01 1.554655e-03
#> drugs              1.001269e+01 1.554655e-03
#> edge               1.001269e+01 1.554655e-03
#> god's              9.967813e+00 1.593006e-03
#> building           9.967813e+00 1.593006e-03
#> responsibility     9.845319e+00 1.702645e-03
#> ourselves          9.843857e+00 1.703999e-03
#> fate               9.644512e+00 1.899182e-03
#> fight              9.396735e+00 2.173722e-03
#> allies             9.396735e+00 2.173722e-03
#> lift               9.396735e+00 2.173722e-03
#> turning            9.396735e+00 2.173722e-03
#> watching           9.386652e+00 2.185710e-03
#> bible              9.386652e+00 2.185710e-03
#> job                9.386652e+00 2.185710e-03
#> tens               9.386652e+00 2.185710e-03
#> feed               9.386652e+00 2.185710e-03
#> chance             9.355701e+00 2.222929e-03
#> era                9.355701e+00 2.222929e-03
#> hear               9.274825e+00 2.323249e-03
#> gift               9.274825e+00 2.323249e-03
#> loyalty            9.274825e+00 2.323249e-03
#> soul               9.274825e+00 2.323249e-03
#> (                  8.887995e+00 2.870519e-03
#> )                  8.887995e+00 2.870519e-03
#> few                8.760054e+00 3.078999e-03
#> say                8.689091e+00 3.201202e-03
#> light              8.636966e+00 3.294096e-03
#> long               8.207655e+00 4.171403e-03
#> choice             8.133566e+00 4.345320e-03
#> greatness          8.006855e+00 4.660060e-03
#> now                7.991973e+00 4.698519e-03
#> meet               7.974277e+00 4.744671e-03
#> price              7.627969e+00 5.746997e-03
#> began              7.611499e+00 5.799724e-03
#> economic           7.605473e+00 5.819141e-03
#> one                7.476244e+00 6.251836e-03
#> millions           7.425536e+00 6.430466e-03
#> weapon             7.386002e+00 6.573345e-03
#> aimed              7.386002e+00 6.573345e-03
#> 4                  7.386002e+00 6.573345e-03
#> symbol             7.386002e+00 6.573345e-03
#> song               7.386002e+00 6.573345e-03
#> stranger           7.386002e+00 6.573345e-03
#> reverend           7.386002e+00 6.573345e-03
#> forebears          7.386002e+00 6.573345e-03
#> beliefs            7.386002e+00 6.573345e-03
#> hardship           7.386002e+00 6.573345e-03
#> uncertain          7.386002e+00 6.573345e-03
#> negotiate          7.386002e+00 6.573345e-03
#> forge              7.386002e+00 6.573345e-03
#> global             7.386002e+00 6.573345e-03
#> match              7.386002e+00 6.573345e-03
#> striving           7.386002e+00 6.573345e-03
#> mall               7.386002e+00 6.573345e-03
#> george             7.386002e+00 6.573345e-03
#> crucial            7.386002e+00 6.573345e-03
#> bigger             7.386002e+00 6.573345e-03
#> engagement         7.386002e+00 6.573345e-03
#> sights             7.386002e+00 6.573345e-03
#> 21st               7.386002e+00 6.573345e-03
#> harness            7.386002e+00 6.573345e-03
#> clinton            7.386002e+00 6.573345e-03
#> middle             7.270389e+00 7.010061e-03
#> else               7.270389e+00 7.010061e-03
#> heritage           7.270389e+00 7.010061e-03
#> turn               7.167030e+00 7.425563e-03
#> came               7.167030e+00 7.425563e-03
#> serve              7.167030e+00 7.425563e-03
#> idea               7.128356e+00 7.587424e-03
#> strengthen         7.056327e+00 7.898579e-03
#> brave              7.024198e+00 8.041547e-03
#> told               7.024198e+00 8.041547e-03
#> proclaim           7.024198e+00 8.041547e-03
#> rise               7.024198e+00 8.041547e-03
#> tomorrow           7.024198e+00 8.041547e-03
#> idealism           7.024198e+00 8.041547e-03
#> opportunity        6.982313e+00 8.231916e-03
#> sworn              6.967831e+00 8.298804e-03
#> pain               6.967831e+00 8.298804e-03
#> seize              6.967831e+00 8.298804e-03
#> capitol            6.967831e+00 8.298804e-03
#> space              6.967831e+00 8.298804e-03
#> fought             6.967831e+00 8.298804e-03
#> potential          6.967831e+00 8.298804e-03
#> going              6.967831e+00 8.298804e-03
#> commitments        6.967831e+00 8.298804e-03
#> hatred             6.967831e+00 8.298804e-03
#> mothers            6.967831e+00 8.298804e-03
#> out                6.956074e+00 8.353515e-03
#> future             6.548609e+00 1.049662e-02
#> society            6.395134e+00 1.144336e-02
#> beginning          6.318486e+00 1.194856e-02
#> globe              6.158104e+00 1.308114e-02
#> growing            6.158104e+00 1.308114e-02
#> dark               6.158104e+00 1.308114e-02
#> where              6.127411e+00 1.331017e-02
#> weak               6.007203e+00 1.424759e-02
#> on                 5.979575e+00 1.447249e-02
#> destiny            5.882746e+00 1.528996e-02
#> down               5.790991e+00 1.610850e-02
#> beyond             5.759626e+00 1.639856e-02
#> liberty            5.737839e+00 1.660321e-02
#> weakness           5.697984e+00 1.698441e-02
#> word               5.697984e+00 1.698441e-02
#> victory            5.697984e+00 1.698441e-02
#> inauguration       5.697984e+00 1.698441e-02
#> poor               5.697984e+00 1.698441e-02
#> achievement        5.697984e+00 1.698441e-02
#> rich               5.574133e+00 1.822769e-02
#> around             5.574133e+00 1.822769e-02
#> vision             5.574133e+00 1.822769e-02
#> sides              5.485059e+00 1.917967e-02
#> lasting            5.485059e+00 1.917967e-02
#> reduce             5.485059e+00 1.917967e-02
#> love               5.342427e+00 2.081247e-02
#> oath               5.309315e+00 2.121169e-02
#> done               5.285741e+00 2.150074e-02
#> gather             5.269341e+00 2.170423e-02
#> borders            5.269341e+00 2.170423e-02
#> forces             5.264090e+00 2.176980e-02
#> created            5.264090e+00 2.176980e-02
#> hopes              5.123795e+00 2.359995e-02
#> crisis             5.043196e+00 2.472284e-02
#> ago                5.039627e+00 2.477382e-02
#> hearts             5.036066e+00 2.482481e-02
#> divided            4.965440e+00 2.585873e-02
#> based              4.965440e+00 2.585873e-02
#> bright             4.899143e+00 2.687002e-02
#> size               4.899143e+00 2.687002e-02
#> agreement          4.899143e+00 2.687002e-02
#> generosity         4.899143e+00 2.687002e-02
#> spoke              4.899143e+00 2.687002e-02
#> charity            4.899143e+00 2.687002e-02
#> storm              4.899143e+00 2.687002e-02
#> ahead              4.899143e+00 2.687002e-02
#> ideals             4.865455e+00 2.739957e-02
#> another            4.840522e+00 2.779848e-02
#> trend              4.801532e+00 2.843444e-02
#> attacks            4.801532e+00 2.843444e-02
#> inalienable        4.801532e+00 2.843444e-02
#> cooperative        4.801532e+00 2.843444e-02
#> survival           4.801532e+00 2.843444e-02
#> victims            4.801532e+00 2.843444e-02
#> technical          4.801532e+00 2.843444e-02
#> housing            4.801532e+00 2.843444e-02
#> technological      4.801532e+00 2.843444e-02
#> bow                4.801532e+00 2.843444e-02
#> amen               4.801532e+00 2.843444e-02
#> unmatched          4.801532e+00 2.843444e-02
#> signs              4.801532e+00 2.843444e-02
#> livelihood         4.801532e+00 2.843444e-02
#> documents          4.801532e+00 2.843444e-02
#> binds              4.801532e+00 2.843444e-02
#> cultural           4.801532e+00 2.843444e-02
#> winning            4.801532e+00 2.843444e-02
#> deserts            4.801532e+00 2.843444e-02
#> skills             4.801532e+00 2.843444e-02
#> strives            4.801532e+00 2.843444e-02
#> blow               4.801532e+00 2.843444e-02
#> flames             4.801532e+00 2.843444e-02
#> shoulders          4.801532e+00 2.843444e-02
#> clergy             4.801532e+00 2.843444e-02
#> explore            4.801532e+00 2.843444e-02
#> conquer            4.801532e+00 2.843444e-02
#> moves              4.801532e+00 2.843444e-02
#> span               4.801532e+00 2.843444e-02
#> mastery            4.801532e+00 2.843444e-02
#> fragile            4.801532e+00 2.843444e-02
#> imagine            4.801532e+00 2.843444e-02
#> unique             4.801532e+00 2.843444e-02
#> 200th              4.801532e+00 2.843444e-02
#> millennium         4.801532e+00 2.843444e-02
#> moon               4.801532e+00 2.843444e-02
#> rhetoric           4.801532e+00 2.843444e-02
#> rebuilding         4.801532e+00 2.843444e-02
#> message            4.801532e+00 2.843444e-02
#> endless            4.801532e+00 2.843444e-02
#> bridges            4.801532e+00 2.843444e-02
#> responsibly        4.801532e+00 2.843444e-02
#> parents            4.801532e+00 2.843444e-02
#> micah              4.801532e+00 2.843444e-02
#> endures            4.801532e+00 2.843444e-02
#> drift              4.801532e+00 2.843444e-02
#> vulnerable         4.801532e+00 2.843444e-02
#> idealistic         4.801532e+00 2.843444e-02
#> spending           4.801532e+00 2.843444e-02
#> inflation          4.801532e+00 2.843444e-02
#> thomas             4.801532e+00 2.843444e-02
#> vietnam            4.801532e+00 2.843444e-02
#> reborn             4.801532e+00 2.843444e-02
#> hall               4.801532e+00 2.843444e-02
#> music              4.801532e+00 2.843444e-02
#> soviet             4.801532e+00 2.843444e-02
#> freedom's          4.801532e+00 2.843444e-02
#> snow               4.801532e+00 2.843444e-02
#> fills              4.801532e+00 2.843444e-02
#> that's             4.801532e+00 2.843444e-02
#> 200                4.801532e+00 2.843444e-02
#> blowing            4.801532e+00 2.843444e-02
#> other's            4.801532e+00 2.843444e-02
#> salute             4.801532e+00 2.843444e-02
#> winter             4.801532e+00 2.843444e-02
#> fascism            4.801532e+00 2.843444e-02
#> technology         4.801532e+00 2.843444e-02
#> shuttered          4.801532e+00 2.843444e-02
#> bridge             4.801532e+00 2.843444e-02
#> guests             4.801532e+00 2.843444e-02
#> raging             4.801532e+00 2.843444e-02
#> debates            4.801532e+00 2.843444e-02
#> businesses         4.801532e+00 2.843444e-02
#> relies             4.801532e+00 2.843444e-02
#> storms             4.801532e+00 2.843444e-02
#> obama              4.801532e+00 2.843444e-02
#> we've              4.801532e+00 2.843444e-02
#> speak              4.736601e+00 2.952735e-02
#> different          4.736601e+00 2.952735e-02
#> transition         4.657508e+00 3.091803e-02
#> philosophy         4.657508e+00 3.091803e-02
#> grows              4.657508e+00 3.091803e-02
#> talk               4.657508e+00 3.091803e-02
#> affirm             4.657508e+00 3.091803e-02
#> deserves           4.657508e+00 3.091803e-02
#> black              4.657508e+00 3.091803e-02
#> environment        4.657508e+00 3.091803e-02
#> magnificent        4.657508e+00 3.091803e-02
#> honorably          4.657508e+00 3.091803e-02
#> transfer           4.657508e+00 3.091803e-02
#> door               4.657508e+00 3.091803e-02
#> food               4.657508e+00 3.091803e-02
#> tell               4.657508e+00 3.091803e-02
#> summon             4.657508e+00 3.091803e-02
#> helping            4.657508e+00 3.091803e-02
#> gladly             4.657508e+00 3.091803e-02
#> mistakes           4.657508e+00 3.091803e-02
#> start              4.657508e+00 3.091803e-02
#> barriers           4.657508e+00 3.091803e-02
#> birthright         4.657508e+00 3.091803e-02
#> isolation          4.657508e+00 3.091803e-02
#> youth              4.657508e+00 3.091803e-02
#> died               4.657508e+00 3.091803e-02
#> image              4.657508e+00 3.091803e-02
#> courses            4.657508e+00 3.091803e-02
#> level              4.657508e+00 3.091803e-02
#> wave               4.657508e+00 3.091803e-02
#> face               4.639065e+00 3.125203e-02
#> makes              4.512273e+00 3.365250e-02
#> forever            4.512273e+00 3.365250e-02
#> heard              4.449601e+00 3.490939e-02
#> politics           4.449601e+00 3.490939e-02
#> arms               4.245223e+00 3.936090e-02
#> problems           4.159790e+00 4.139454e-02
#> before             4.108519e+00 4.266770e-02
#> see                4.030446e+00 4.468614e-02
#> fear               4.024454e+00 4.484513e-02
#> special            4.006785e+00 4.531749e-02
#> eyes               3.981364e+00 4.600630e-02
#> final              3.939803e+00 4.715631e-02
#> choose             3.939803e+00 4.715631e-02
#> raise              3.939803e+00 4.715631e-02
#> jefferson          3.939803e+00 4.715631e-02
#> structure          3.939803e+00 4.715631e-02
#> mission            3.939803e+00 4.715631e-02
#> truths             3.939803e+00 4.715631e-02
#> white              3.939803e+00 4.715631e-02
#> differences        3.861963e+00 4.939250e-02
#> longer             3.861963e+00 4.939250e-02
#> accept             3.861963e+00 4.939250e-02
#> away               3.759319e+00 5.251396e-02
#> home               3.757093e+00 5.258394e-02
#> neighbors          3.688878e+00 5.477644e-02
#> share              3.604562e+00 5.762124e-02
#> economy            3.597135e+00 5.787923e-02
#> knowledge          3.585357e+00 5.829089e-02
#> last               3.576824e+00 5.859109e-02
#> deep               3.496824e+00 6.148665e-02
#> living             3.496824e+00 6.148665e-02
#> fire               3.450469e+00 6.323381e-02
#>  [ reached getOption("max.print") -- omitted 8732 rows ]

# compare Trump 2017 to other post-war preseidents
pwdfm <- dfm(corpus_subset(data_corpus_inaugural, period == "post-war"))
head(textstat_keyness(pwdfm, target = "2017-Trump"), 10)

#>                  chi2            p
#> protected    76.79339 0.000000e+00
#> will         49.67662 1.812883e-12
#> while        48.33243 3.597567e-12
#> obama        47.94909 4.374279e-12
#> we've        47.94909 4.374279e-12
#> america      29.11681 6.814327e-08
#> again        27.88512 1.287361e-07
#> everyone     27.73848 1.388726e-07
#> your         26.75528 2.309197e-07
#> transferring 25.59527 4.210694e-07

# using the likelihood ratio method
head(textstat_keyness(dfm_smooth(pwdfm), measure = "lr", target = "2017-Trump"), 10)

#>                   G            p
#> will      24.653370 6.862464e-07
#> america   14.064544 1.766425e-04
#> your      11.662713 6.376529e-04
#> while     11.166438 8.329039e-04
#> again     11.083498 8.709934e-04
#> protected 10.591417 1.136138e-03
#> american   9.993537 1.570906e-03
#> back       8.209112 4.168054e-03
#> dreams     7.075173 7.815929e-03
#> country    6.800841 9.111495e-03