calculate keyness statistics

calculate keyness statistics

textstat_keyness(x, target = 1L, measure = c("chi2", "exact", "lr"),
  sort = TRUE)

Arguments

x
a dfm containing the features to be examined for keyness

target
the document index (numeric, character or logical) identifying the document forming the "target" for computing keyness; all other documents' feature frequencies will be combined for use as a reference

measure
(signed) association measure to be used for computing keyness. Currenly available: "chi2" (\(chi^2\) with Yates correction); "exact" (Fisher's exact test); "lr" for the likelihood ratio \(G\) statistic with Yates correction.

sort
logical; if TRUE sort features scored in descending order of the measure, otherwise leave in original feature order

Value

a data.frame of computed statistics and associated p-values, where the features scored name each row, and the number of occurrences for both the target and reference groups. For measure = "chi2" this is the chi-squared value, signed positively if the observed value in the target exceeds its expected value; for measure = "exact" this is the estimate of the odds ratio; for measure = "lr" this is the likelihood ratio \(G\) statistic.

References

Bondi, Marina, and Mike Scott, eds. 2010. Keyness in Texts. Amsterdam, Philadelphia: John Benjamins, 2010. Stubbs, Michael. 2010. "Three Concepts of Keywords". In Keyness in Texts, Marina Bondi and Mike Scott, eds. pp21–42. Amsterdam, Philadelphia: John Benjamins. Scott, M. & Tribble, C. 2006. Textual Patterns: keyword and corpus analysis in language education. Amsterdam: Benjamins, p. 55. Dunning, Ted. 1993. "Accurate Methods for the Statistics of Surprise and Coincidence", Computational Linguistics, Vol 19, No. 1, pp. 61-74.

Examples

# compare pre- v. post-war terms using grouping
period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
mydfm <- dfm(data_corpus_inaugural, groups = period)
head(mydfm) # make sure 'post-war' is in the first row

#> Document-feature matrix of: 2 documents, 9,232 features (34.6% sparse).
#> (showing first 2 documents and first 6 features)
#>           features
#> docs       chief justice    , mr    . vice
#>   post-war    12      40 2194 25 1804   15
#>   pre-war     24      97 4832  6 3141    3

head(result <- textstat_keyness(mydfm), 10)

#>               chi2 p n_target n_reference
#> we        701.8799 0      960         779
#> -         252.2852 0      510         532
#> .         227.2380 0     1804        3141
#> our       187.8353 0      874        1307
#> us        184.9674 0      262         216
#> :         175.7657 0      105          29
#> america   174.6586 0      130          54
#> world     169.3859 0      188         126
#> americans 147.6197 0       67           7
#> freedom   137.7030 0      122          64

tail(result, 10)

#>                    chi2            p n_target n_reference
#> upon          -50.97048 9.376944e-13       39         332
#> it            -52.03533 5.451195e-13      257        1132
#> public        -54.79660 1.336709e-13       11         213
#> states        -58.08454 2.509104e-14       28         305
#> constitution  -59.87256 1.010303e-14        6         200
#> be            -71.67429 0.000000e+00      257        1225
#> should        -81.85968 0.000000e+00       15         309
#> which        -158.95020 0.000000e+00       95         911
#> of           -177.80869 0.000000e+00     1437        5666
#> the          -297.87187 0.000000e+00     1988        8094

# compare pre- v. post-war terms using logical vector
mydfm2 <- dfm(data_corpus_inaugural)
textstat_keyness(mydfm2, docvars(data_corpus_inaugural, "Year") >= 1945)

#>                            chi2            p n_target n_reference
#> we                 7.018799e+02 0.000000e+00      960         779
#> -                  2.522852e+02 0.000000e+00      510         532
#> .                  2.272380e+02 0.000000e+00     1804        3141
#> our                1.878353e+02 0.000000e+00      874        1307
#> us                 1.849674e+02 0.000000e+00      262         216
#> :                  1.757657e+02 0.000000e+00      105          29
#> america            1.746586e+02 0.000000e+00      130          54
#> world              1.693859e+02 0.000000e+00      188         126
#> americans          1.476197e+02 0.000000e+00       67           7
#> freedom            1.377030e+02 0.000000e+00      122          64
#> new                1.326699e+02 0.000000e+00      150         102
#> today              1.264530e+02 0.000000e+00       76          21
#> let                1.097794e+02 0.000000e+00      100          54
#> together           1.025977e+02 0.000000e+00       64          19
#> you                1.006242e+02 0.000000e+00      116          80
#> america's          9.051114e+01 0.000000e+00       35           0
#> work               8.895165e+01 0.000000e+00       78          40
#> help               8.313975e+01 0.000000e+00       46          10
#> ,                  6.328131e+01 1.776357e-15     2194        4832
#> know               5.882108e+01 1.731948e-14       63          40
#> earth              5.821036e+01 2.353673e-14       44          18
#> history            5.607945e+01 6.961098e-14       60          38
#> god                5.485544e+01 1.296740e-13       57          35
#> live               5.124971e+01 8.133494e-13       39          16
#> do                 5.086969e+01 9.870993e-13      112         121
#> are                5.052212e+01 1.178280e-12      311         503
#> time               5.010091e+01 1.460387e-12      106         112
#> century            4.743658e+01 5.681233e-12       40          19
#> what               4.740058e+01 5.786593e-12       88          86
#> will               4.690854e+01 7.437828e-12      339         572
#> journey            4.596549e+01 1.203548e-11       20           1
#> dignity            4.577572e+01 1.325962e-11       24           4
#> promise            4.429157e+01 2.829315e-11       34          14
#> mr                 4.232395e+01 7.733958e-11       25           6
#> build              4.131706e+01 1.294346e-10       21           3
#> thank              4.064165e+01 1.828669e-10       18           1
#> challenges         3.943058e+01 3.399322e-10       16           0
#> for                3.936134e+01 3.522012e-10      421         776
#> strength           3.807069e+01 6.822725e-10       45          31
#> cannot             3.734577e+01 9.893516e-10       36          20
#> bless              3.693396e+01 1.221991e-09       18           2
#> must               3.649566e+01 1.530062e-09      151         215
#> lives              3.606755e+01 1.905946e-09       32          16
#> again              3.533471e+01 2.776398e-09       42          29
#> generation         3.512850e+01 3.086504e-09       28          12
#> across             3.499140e+01 3.311650e-09       22           6
#> moment             3.497182e+01 3.345117e-09       27          11
#> dream              3.432681e+01 4.659189e-09       17           2
#> weapons            3.406121e+01 5.340536e-09       14           0
#> young              3.311895e+01 8.668991e-09       19           4
#> words              3.275732e+01 1.044124e-08       26          11
#> faith              3.219168e+01 1.396880e-08       48          40
#> dreams             3.172895e+01 1.772601e-08       16           2
#> nation             3.163099e+01 1.864311e-08      123         171
#> role               3.137788e+01 2.123865e-08       13           0
#> heart              3.131451e+01 2.194340e-08       28          14
#> because            3.107581e+01 2.481442e-08       59          58
#> resolve            3.107201e+01 2.486304e-08       17           3
#> children           3.064151e+01 3.103816e-08       31          18
#> friends            3.056931e+01 3.221503e-08       25          11
#> your               3.039199e+01 3.529830e-08       61          62
#> meaning            2.914201e+01 6.726262e-08       15           2
#> peoples            2.893856e+01 7.471069e-08       26          13
#> communism          2.869575e+01 8.468931e-08       12           0
#> begin              2.869575e+01 8.468931e-08       12           0
#> yes                2.869575e+01 8.468931e-08       12           0
#> this               2.835429e+01 1.010229e-07      295         540
#> unity              2.789357e+01 1.281753e-07       21           8
#> story              2.740403e+01 1.650865e-07       13           1
#> families           2.656803e+01 2.544198e-07       14           2
#> forward            2.656310e+01 2.550694e-07       24          12
#> vice               2.603583e+01 3.351402e-07       15           3
#> jobs               2.601513e+01 3.387520e-07       11           0
#> president          2.601077e+01 3.395174e-07       46          43
#> join               2.571630e+01 3.954731e-07       16           4
#> world's            2.550260e+01 4.417874e-07       18           6
#> strong             2.535083e+01 4.779462e-07       32          23
#> day                2.507395e+01 5.517329e-07       46          44
#> believe            2.482330e+01 6.283320e-07       40          35
#> celebrate          2.477638e+01 6.438161e-07       12           1
#> man                2.454282e+01 7.267650e-07       45          43
#> back               2.418822e+01 8.736487e-07       22          11
#> when               2.402471e+01 9.510743e-07       90         123
#> centuries          2.400959e+01 9.585730e-07       13           2
#> seek               2.389993e+01 1.014757e-06       34          27
#> senator            2.333646e+01 1.359972e-06       10           0
#> nuclear            2.333646e+01 1.359972e-06       10           0
#> free               2.258075e+01 2.014931e-06       79         105
#> peace              2.216887e+01 2.496891e-06      103         152
#> women              2.213170e+01 2.545710e-06       21          11
#> ask                2.120608e+01 4.124532e-06       26          18
#> remember           2.107485e+01 4.416881e-06       17           7
#> challenge          2.090517e+01 4.825908e-06       14           4
#> change             2.071339e+01 5.334174e-06       36          33
#> make               2.046758e+01 6.064988e-06       64          81
#> need               1.991480e+01 8.097099e-06       38          37
#> courage            1.983500e+01 8.442242e-06       23          15
#> old                1.956841e+01 9.706070e-06       39          39
#> pledge             1.939449e+01 1.063135e-05       22          14
#> come               1.903751e+01 1.281737e-05       38          38
#> who                1.900236e+01 1.305570e-05      138         232
#> threat             1.895365e+01 1.339328e-05       11           2
#> tyranny            1.895365e+01 1.339328e-05       11           2
#> values             1.890034e+01 1.377281e-05       16           7
#> hope               1.871529e+01 1.517607e-05       53          64
#> hard               1.870461e+01 1.526132e-05       15           6
#> answer             1.865405e+01 1.567142e-05       12           3
#> renew              1.854990e+01 1.655138e-05       13           4
#> voices             1.854990e+01 1.655138e-05       13           4
#> american           1.839315e+01 1.797024e-05       69          94
#> areas              1.798786e+01 2.223178e-05        8           0
#> heal               1.798786e+01 2.223178e-05        8           0
#> define             1.798786e+01 2.223178e-05        8           0
#> compassion         1.798786e+01 2.223178e-05        8           0
#> speaker            1.798786e+01 2.223178e-05        8           0
#> quiet              1.798786e+01 2.223178e-05        8           0
#> don't              1.798786e+01 2.223178e-05        8           0
#> here               1.790667e+01 2.320065e-05       39          41
#> go                 1.754606e+01 2.804306e-05       26          21
#> but                1.735863e+01 3.094898e-05      225         429
#> each               1.723627e+01 3.300749e-05       55          70
#> founding           1.696872e+01 3.800076e-05        9           1
#> third              1.696872e+01 3.800076e-05        9           1
#> nation's           1.677276e+01 4.213381e-05       15           7
#> poverty            1.677276e+01 4.213381e-05       15           7
#> democracy          1.658033e+01 4.663221e-05       30          28
#> bring              1.652378e+01 4.804365e-05       24          19
#> creed              1.646629e+01 4.952272e-05       10           2
#> bold               1.646629e+01 4.952272e-05       10           2
#> schools            1.633698e+01 5.301924e-05       13           5
#> born               1.633698e+01 5.301924e-05       13           5
#> strive             1.633698e+01 5.301924e-05       13           5
#> friend             1.625914e+01 5.524243e-05       11           3
#> leaders            1.625914e+01 5.524243e-05       11           3
#> learned            1.625914e+01 5.524243e-05       11           3
#> man's              1.625914e+01 5.524243e-05       11           3
#> program            1.625914e+01 5.524243e-05       11           3
#> land               1.579734e+01 7.050173e-05       35          37
#> planet             1.532040e+01 9.073152e-05        7           0
#> sick               1.532040e+01 9.073152e-05        7           0
#> productivity       1.532040e+01 9.073152e-05        7           0
#> decency            1.532040e+01 9.073152e-05        7           0
#> decades            1.532040e+01 9.073152e-05        7           0
#> goals              1.532040e+01 9.073152e-05        7           0
#> \\                 1.532040e+01 9.073152e-05        7           0
#> age                1.504122e+01 1.051886e-04       15           8
#> a                  1.493629e+01 1.112029e-04      690        1556
#> factories          1.440517e+01 1.473968e-04        8           1
#> historic           1.440517e+01 1.473968e-04        8           1
#> simple             1.439612e+01 1.481074e-04       13           6
#> achieve            1.439612e+01 1.481074e-04       13           6
#> ceremony           1.401584e+01 1.812767e-04        9           2
#> decent             1.401584e+01 1.812767e-04        9           2
#> everyone           1.401584e+01 1.812767e-04        9           2
#> child              1.401584e+01 1.812767e-04        9           2
#> commitment         1.401584e+01 1.812767e-04        9           2
#> goal               1.401584e+01 1.812767e-04        9           2
#> prayer             1.397346e+01 1.854097e-04       11           4
#> move               1.397346e+01 1.854097e-04       11           4
#> can                1.385632e+01 1.973322e-04      164         307
#> is                 1.312689e+01 2.910882e-04      458        1004
#> misery             1.266050e+01 3.734626e-04        6           0
#> hunger             1.266050e+01 3.734626e-04        6           0
#> quest              1.266050e+01 3.734626e-04        6           0
#> night              1.266050e+01 3.734626e-04        6           0
#> deepest            1.266050e+01 3.734626e-04        6           0
#> celebration        1.266050e+01 3.734626e-04        6           0
#> civility           1.266050e+01 3.734626e-04        6           0
#> choices            1.266050e+01 3.734626e-04        6           0
#> ensure             1.266050e+01 3.734626e-04        6           0
#> adversaries        1.266050e+01 3.734626e-04        6           0
#> timeless           1.266050e+01 3.734626e-04        6           0
#> bush               1.266050e+01 3.734626e-04        6           0
#> workers            1.266050e+01 3.734626e-04        6           0
#> streets            1.266050e+01 3.734626e-04        6           0
#> budget             1.266050e+01 3.734626e-04        6           0
#> way                1.258826e+01 3.881768e-04       37          45
#> life               1.248016e+01 4.112963e-04       56          81
#> enemies            1.233125e+01 4.454374e-04       12           6
#> small              1.232832e+01 4.461371e-04       18          14
#> human              1.216274e+01 4.875365e-04       41          53
#> renewal            1.187330e+01 5.694447e-04        7           1
#> big                1.187330e+01 5.694447e-04        7           1
#> basic              1.187330e+01 5.694447e-04        7           1
#> lead               1.182099e+01 5.856680e-04       17          13
#> generations        1.182099e+01 5.856680e-04       17          13
#> learn              1.177142e+01 6.014701e-04       10           4
#> ideas              1.161706e+01 6.534950e-04        9           3
#> try                1.161706e+01 6.534950e-04        9           3
#> truly              1.161706e+01 6.534950e-04        9           3
#> welcome            1.161706e+01 6.534950e-04        9           3
#> play               1.161706e+01 6.534950e-04        9           3
#> shape              1.161336e+01 6.547969e-04        8           2
#> heroes             1.161336e+01 6.547969e-04        8           2
#> working            1.161336e+01 6.547969e-04        8           2
#> shared             1.161336e+01 6.547969e-04        8           2
#> anew               1.161336e+01 6.547969e-04        8           2
#> instead            1.117512e+01 8.290139e-04       13           8
#> better             1.109329e+01 8.664042e-04       34          42
#> end                1.108018e+01 8.725519e-04       23          23
#> security           1.100232e+01 9.099777e-04       30          35
#> sacrifice          1.081731e+01 1.005555e-03       15          11
#> nations            1.077426e+01 1.029212e-03       75         124
#> that               1.053397e+01 1.172002e-03      546        1243
#> proud              1.033999e+01 1.301788e-03       11           6
#> policies           1.033999e+01 1.301788e-03       11           6
#> act                1.033269e+01 1.306945e-03       26          29
#> reach              1.019109e+01 1.411209e-03       16          13
#> think              1.019109e+01 1.411209e-03       16          13
#> "                  1.015543e+01 1.438768e-03       90         158
#> stand              1.014731e+01 1.445119e-03       27          31
#> once               1.002062e+01 1.547975e-03       21          21
#> programs           1.001269e+01 1.554655e-03        5           0
#> cold               1.001269e+01 1.554655e-03        5           0
#> brothers           1.001269e+01 1.554655e-03        5           0
#> hungry             1.001269e+01 1.554655e-03        5           0
#> knows              1.001269e+01 1.554655e-03        5           0
#> ended              1.001269e+01 1.554655e-03        5           0
#> hatreds            1.001269e+01 1.554655e-03        5           0
#> dawn               1.001269e+01 1.554655e-03        5           0
#> ending             1.001269e+01 1.554655e-03        5           0
#> today's            1.001269e+01 1.554655e-03        5           0
#> breeze             1.001269e+01 1.554655e-03        5           0
#> drugs              1.001269e+01 1.554655e-03        5           0
#> edge               1.001269e+01 1.554655e-03        5           0
#> god's              9.967813e+00 1.593006e-03       10           5
#> building           9.967813e+00 1.593006e-03       10           5
#> responsibility     9.845319e+00 1.702645e-03       29          35
#> ourselves          9.843857e+00 1.703999e-03       35          46
#> fate               9.644512e+00 1.899182e-03        9           4
#> fight              9.396735e+00 2.173722e-03        8           3
#> allies             9.396735e+00 2.173722e-03        8           3
#> lift               9.396735e+00 2.173722e-03        8           3
#> turning            9.396735e+00 2.173722e-03        8           3
#> watching           9.386652e+00 2.185710e-03        6           1
#> bible              9.386652e+00 2.185710e-03        6           1
#> job                9.386652e+00 2.185710e-03        6           1
#> tens               9.386652e+00 2.185710e-03        6           1
#> feed               9.386652e+00 2.185710e-03        6           1
#> chance             9.355701e+00 2.222929e-03       12           8
#> era                9.355701e+00 2.222929e-03       12           8
#> hear               9.274825e+00 2.323249e-03        7           2
#> gift               9.274825e+00 2.323249e-03        7           2
#> loyalty            9.274825e+00 2.323249e-03        7           2
#> soul               9.274825e+00 2.323249e-03        7           2
#> (                  8.887995e+00 2.870519e-03       11           7
#> )                  8.887995e+00 2.870519e-03       11           7
#> few                8.760054e+00 3.078999e-03       17          16
#> say                8.689091e+00 3.201202e-03       20          21
#> light              8.636966e+00 3.294096e-03       15          13
#>  [ reached getOption("max.print") -- omitted 8982 rows ]

# compare Trump 2017 to other post-war preseidents
pwdfm <- dfm(corpus_subset(data_corpus_inaugural, period == "post-war"))
head(textstat_keyness(pwdfm, target = "2017-Trump"), 10)

#>                  chi2            p n_target n_reference
#> protected    76.79339 0.000000e+00        5           1
#> will         49.67662 1.812883e-12       40         299
#> while        48.33243 3.597567e-12        6           7
#> obama        47.94909 4.374279e-12        3           0
#> we've        47.94909 4.374279e-12        3           0
#> america      29.11681 6.814327e-08       18         112
#> again        27.88512 1.287361e-07        9          33
#> everyone     27.73848 1.388726e-07        4           5
#> your         26.75528 2.309197e-07       11          50
#> transferring 25.59527 4.210694e-07        2           0

# using the likelihood ratio method
head(textstat_keyness(dfm_smooth(pwdfm), measure = "lr", target = "2017-Trump"), 10)

#>                   G            p n_target n_reference
#> will      24.653370 6.862464e-07       41         317
#> america   14.064544 1.766425e-04       19         130
#> your      11.662713 6.376529e-04       12          68
#> while     11.166438 8.329039e-04        7          25
#> again     11.083498 8.709934e-04       10          51
#> protected 10.591417 1.136138e-03        6          19
#> american   9.993537 1.570906e-03       12          76
#> back       8.209112 4.168054e-03        7          34
#> dreams     7.075173 7.815929e-03        6          29
#> country    6.800841 9.111495e-03       10          72