calculate lexical diversity

Calculate the lexical diversity or complexity of text(s).

textstat_lexdiv(x, measure = c("all", "TTR", "C", "R", "CTTR", "U", "S",
  "Maas"), log.base = 10, drop = TRUE, ...)

Arguments

x

an input object, such as a document-feature matrix object

measure

a character vector defining the measure to calculate.

log.base

a numeric value defining the base of the logarithm (for measures using logs)

drop

if TRUE, the result is returned as a numeric vector if only a single measure is requested; otherwise, a data.frame is returned with each column consisting of a requested measure.

...

not used

Value

a data.frame or vector of lexical diversity statistics, each row or vector element corresponding to an input document

Details

textstat_lexdiv calculates a variety of proposed indices for lexical diversity. In the following formulae, \(N\) refers to the total number of tokens, and \(V\) to the number of types:

"TTR":

The ordinary Type-Token Ratio: $$TTR = \frac{V}{N}$$

"C":

Herdan's C (Herdan, 1960, as cited in Tweedie & Baayen, 1998; sometimes referred to as LogTTR): $$C = \frac{\log{V}}{\log{N}}$$

"R":

Guiraud's Root TTR (Guiraud, 1954, as cited in Tweedie & Baayen, 1998): $$R = \frac{V}{\sqrt{N}}$$

"CTTR":

Carroll's Corrected TTR: $$CTTR = \frac{V}{\sqrt{2N}}$$

"U":

Dugast's Uber Index (Dugast, 1978, as cited in Tweedie & Baayen, 1998): $$U = \frac{(\log{N})^2}{\log{N} - \log{V}}$$

"S":

Summer's index: $$S = \frac{\log{\log{V}}}{\log{\log{N}}}$$

"K":

Yule's K (Yule, 1944, as cited in Tweedie & Baayen, 1998) is calculated by: $$K = 10^4 \times \frac{(\sum_{X=1}^{X}{{f_X}X^2}) - N}{N^2}$$ where \(N\) is the number of tokens, \(X\) is a vector with the frequencies of each type, and \(f_X\) is the frequencies for each X.

"Maas":

Maas' indices (\(a\), \(\log{V_0}\) & \(\log{}_{e}{V_0}\)): $$a^2 = \frac{\log{N} - \log{V}}{\log{N}^2}$$ $$\log{V_0} = \frac{\log{V}}{\sqrt{1 - \frac{\log{V}}{\log{N}}^2}}$$ The measure was derived from a formula by Mueller (1969, as cited in Maas, 1972). \(\log{}_{e}{V_0}\) is equivalent to \(\log{V_0}\), only with \(e\) as the base for the logarithms. Also calculated are \(a\), \(\log{V_0}\) (both not the same as before) and \(V'\) as measures of relative vocabulary growth while the text progresses. To calculate these measures, the first half of the text and the full text will be examined (see Maas, 1972, p. 67 ff. for details). Note: for the current method (for a dfm) there is no computation on separate halves of the text.

Note

This implements only the static measures of lexical diversity, not more complex measures based on windows of text such as the Mean Segmental Type-Token Ratio, the Moving-Average Type-Token Ratio (Covington & McFall, 2010), the MLTD or MLTD-MA (Moving-Average Measure of Textual Lexical Diversity) proposed by McCarthy & Jarvis (2010) or Jarvis (no year), or the HD-D version of vocd-D (see McCarthy & Jarvis, 2007). These are available from the package korRpus.

References

Covington, M.A. & McFall, J.D. (2010). Cutting the Gordian Knot: The Moving-Average Type-Token Ratio (MATTR). Journal of Quantitative Linguistics, 17(2), 94--100.

Maas, H.-D., (1972). \"Uber den Zusammenhang zwischen Wortschatzumfang und L\"ange eines Textes. Zeitschrift f\"ur Literaturwissenschaft und Linguistik, 2(8), 73--96.

McCarthy, P.M. & Jarvis, S. (2007). vocd: A theoretical and empirical evaluation. Language Testing, 24(4), 459--488.

McCarthy, P.M. & Jarvis, S. (2010). MTLD, vocd-D, and HD-D: A validation study of sophisticated approaces to lexical diversity assessment. Behaviour Research Methods, 42(2), 381--392.

Michalke, Meik. (2014) koRpus: An R Package for Text Analysis. Version 0.05-5. http://reaktanz.de/?c=hacking&s=koRpus

Tweedie. F.J. & Baayen, R.H. (1998). How Variable May a Constant Be? Measures of Lexical Richness in Perspective. Computers and the Humanities, 32(5), 323--352.

Examples

mydfm <- dfm(corpus_subset(data_corpus_inaugural, Year > 1980), verbose = FALSE)
(results <- textstat_lexdiv(mydfm, c("CTTR", "TTR", "U")))

#>                  CTTR      TTR          U
#> 1981-Reagan  43.29842 1.157613 -186.91243
#> 1985-Reagan  42.27581 1.103578 -280.92307
#> 1989-Bush    44.21663 1.207231 -143.72662
#> 1993-Clinton 53.43691 1.763201  -43.25754
#> 1997-Clinton 46.26199 1.321501  -94.88620
#> 2001-Bush    53.83400 1.789503  -41.99094
#> 2005-Bush    47.49900 1.393118  -78.70659
#> 2009-Obama   43.84239 1.186882 -158.66860
#> 2013-Obama   47.39718 1.387152  -79.82713
#> 2017-Trump   56.17983 1.948857  -35.79428

cor(textstat_lexdiv(mydfm, "all"))

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#>             TTR         C         R      CTTR         U         S Maas lgV0
#> TTR   1.0000000 0.9987134 0.9994190 0.9994190 0.8356166 0.9989957   NA   NA
#> C     0.9987134 1.0000000 0.9998613 0.9998613 0.8594013 0.9999825   NA   NA
#> R     0.9994190 0.9998613 1.0000000 1.0000000 0.8517330 0.9999423   NA   NA
#> CTTR  0.9994190 0.9998613 1.0000000 1.0000000 0.8517330 0.9999423   NA   NA
#> U     0.8356166 0.8594013 0.8517330 0.8517330 1.0000000 0.8567293   NA   NA
#> S     0.9989957 0.9999825 0.9999423 0.9999423 0.8567293 1.0000000   NA   NA
#> Maas         NA        NA        NA        NA        NA        NA    1   NA
#> lgV0         NA        NA        NA        NA        NA        NA   NA    1
#> lgeV0        NA        NA        NA        NA        NA        NA   NA   NA
#>       lgeV0
#> TTR      NA
#> C        NA
#> R        NA
#> CTTR     NA
#> U        NA
#> S        NA
#> Maas     NA
#> lgV0     NA
#> lgeV0     1

# with different settings of drop
textstat_lexdiv(mydfm, "TTR", drop = TRUE)

#>  1981-Reagan  1985-Reagan    1989-Bush 1993-Clinton 1997-Clinton    2001-Bush 
#>     1.157613     1.103578     1.207231     1.763201     1.321501     1.789503 
#>    2005-Bush   2009-Obama   2013-Obama   2017-Trump 
#>     1.393118     1.186882     1.387152     1.948857 

textstat_lexdiv(mydfm, "TTR", drop = FALSE)

#>                   TTR
#> 1981-Reagan  1.157613
#> 1985-Reagan  1.103578
#> 1989-Bush    1.207231
#> 1993-Clinton 1.763201
#> 1997-Clinton 1.321501
#> 2001-Bush    1.789503
#> 2005-Bush    1.393118
#> 2009-Obama   1.186882
#> 2013-Obama   1.387152
#> 2017-Trump   1.948857