Calculate readability

Calculate the readability of text(s) using one of a variety of computed indexes.

textstat_readability(x, measure = c("all", "ARI", "ARI.simple", "Bormuth",
  "Bormuth.GP", "Coleman", "Coleman.C2", "Coleman.Liau", "Coleman.Liau.grade",
  "Coleman.Liau.short", "Dale.Chall", "Dale.Chall.old", "Dale.Chall.PSK",
  "Danielson.Bryan", "Danielson.Bryan.2", "Dickes.Steiwer", "DRP", "ELF",
  "Farr.Jenkins.Paterson", "Flesch", "Flesch.PSK", "Flesch.Kincaid", "FOG",
  "FOG.PSK", "FOG.NRI", "FORCAST", "FORCAST.RGL", "Fucks", "Linsear.Write",
  "LIW", "nWS", "nWS.2", "nWS.3", "nWS.4", "RIX", "Scrabble", "SMOG", "SMOG.C",
  "SMOG.simple",      "SMOG.de", "Spache", "Spache.old", "Strain",
  "Traenkle.Bailer", "Traenkle.Bailer.2", "Wheeler.Smith", "meanSentenceLength",
  "meanWordSyllables"), remove_hyphens = TRUE, min_sentence_length = 1,
  max_sentence_length = 10000, ...)

Arguments

x
a character or corpus object containing the texts

measure
character vector defining the readability measure to calculate. Matches are case-insensitive.

remove_hyphens
if TRUE, treat constituent words in hyphenated as separate terms, for purposes of computing word lengths, e.g. "decision-making" as two terms of lengths 8 and 6 characters respectively, rather than as a single word of 15 characters

min_sentence_length, max_sentence_length
set the minimum and maximum sentence lengths (in tokens, excluding punctuation) to include in the computation of readability. This makes it easy to exclude "sentences" that may not really be sentences, such as section titles, table elements, and other cruft that might be in the texts following conversion.

For finer-grained control, consider filtering sentences prior first, including through pattern-matching, using corpus_trim.

...
not used

Value

textstat_readability returns a data.frame of documents and their readability scores.

Examples

txt <- c("Readability zero one.  Ten, Eleven.", "The cat in a dilapidated tophat.")
textstat_readability(txt, "Flesch.Kincaid")

#>   document Flesch.Kincaid
#> 1    text1      13.705000
#> 2    text2       8.383333

textstat_readability(txt, c("FOG", "FOG.PSK", "FOG.NRI"))

#>   document       FOG  FOG.PSK FOG.NRI
#> 1    text1 17.000000 4.608659 -1.3875
#> 2    text2  9.066667 3.254382 -1.2600

inaugReadability <- textstat_readability(data_corpus_inaugural, "all")
cor(inaugReadability[,-1])

#>                               ARI   ARI.simple     Bormuth  Bormuth.GP
#> ARI                    1.00000000  0.999985813 -0.99188017  0.90232231
#> ARI.simple             0.99998581  1.000000000 -0.99254312  0.90363285
#> Bormuth               -0.99188017 -0.992543125  1.00000000 -0.92665498
#> Bormuth.GP             0.90232231  0.903632850 -0.92665498  1.00000000
#> Coleman               -0.57842561 -0.574425719  0.47841665 -0.34186891
#> Coleman.C2            -0.76781514 -0.764705749  0.68780777 -0.51669363
#> Coleman.Liau          -0.80785899 -0.804736665  0.72721217 -0.55936385
#> Coleman.Liau.grade     0.80785899  0.804736665 -0.72721217  0.55936385
#> Coleman.Liau.short     0.80788731  0.804765212 -0.72724574  0.55939187
#> Dale.Chall            -0.99099423 -0.991693434  0.99996713 -0.92746308
#> Dale.Chall.old         0.99099423  0.991693434 -0.99996713  0.92746308
#> Dale.Chall.PSK         0.99099423  0.991693434 -0.99996713  0.92746308
#> Danielson.Bryan        0.99708146  0.996677212 -0.97962805  0.88486422
#> Danielson.Bryan.2     -0.31030344 -0.305238155  0.18706023 -0.04006627
#> Dickes.Steiwer        -0.99844204 -0.998703929  0.99691688 -0.91329434
#> DRP                    0.99188017  0.992543125 -1.00000000  0.92665498
#> ELF                    0.99594813  0.995835853 -0.98541527  0.89837506
#> Farr.Jenkins.Paterson -0.99154201 -0.992219215  0.99998536 -0.92696268
#> Flesch                -0.96831214 -0.967086924  0.93162624 -0.82257543
#> Flesch.PSK             0.98543988  0.984630595 -0.95852762  0.85630865
#> Flesch.Kincaid         0.99828644  0.998096618 -0.98602857  0.89508980
#>                          Coleman Coleman.C2 Coleman.Liau Coleman.Liau.grade
#> ARI                   -0.5784256 -0.7678151   -0.8078590          0.8078590
#> ARI.simple            -0.5744257 -0.7647057   -0.8047367          0.8047367
#> Bormuth                0.4784167  0.6878078    0.7272122         -0.7272122
#> Bormuth.GP            -0.3418689 -0.5166936   -0.5593638          0.5593638
#> Coleman                1.0000000  0.9578621    0.9027073         -0.9027073
#> Coleman.C2             0.9578621  1.0000000    0.9676044         -0.9676044
#> Coleman.Liau           0.9027073  0.9676044    1.0000000         -1.0000000
#> Coleman.Liau.grade    -0.9027073 -0.9676044   -1.0000000          1.0000000
#> Coleman.Liau.short    -0.9026909 -0.9676038   -1.0000000          1.0000000
#> Dale.Chall             0.4728710  0.6830083    0.7223808         -0.7223808
#> Dale.Chall.old        -0.4728710 -0.6830083   -0.7223808          0.7223808
#> Dale.Chall.PSK        -0.4728710 -0.6830083   -0.7223808          0.7223808
#> Danielson.Bryan       -0.6320177 -0.8068048   -0.8471402          0.8471402
#> Danielson.Bryan.2      0.8922077  0.7931428    0.8066665         -0.8066665
#> Dickes.Steiwer         0.5390065  0.7368282    0.7764052         -0.7764052
#> DRP                   -0.4784167 -0.6878078   -0.7272122          0.7272122
#> ELF                   -0.6127006 -0.7903882   -0.8134246          0.8134246
#> Farr.Jenkins.Paterson  0.4768093  0.6862258    0.7252233         -0.7252233
#> Flesch                 0.7524555  0.8921701    0.9145879         -0.9145879
#> Flesch.PSK            -0.6991742 -0.8564052   -0.8829007          0.8829007
#> Flesch.Kincaid        -0.6124186 -0.7933597   -0.8255002          0.8255002
#>                       Coleman.Liau.short  Dale.Chall Dale.Chall.old
#> ARI                            0.8078873 -0.99099423     0.99099423
#> ARI.simple                     0.8047652 -0.99169343     0.99169343
#> Bormuth                       -0.7272457  0.99996713    -0.99996713
#> Bormuth.GP                     0.5593919 -0.92746308     0.92746308
#> Coleman                       -0.9026909  0.47287100    -0.47287100
#> Coleman.C2                    -0.9676038  0.68300831    -0.68300831
#> Coleman.Liau                  -1.0000000  0.72238083    -0.72238083
#> Coleman.Liau.grade             1.0000000 -0.72238083     0.72238083
#> Coleman.Liau.short             1.0000000 -0.72241463     0.72241463
#> Dale.Chall                    -0.7224146  1.00000000    -1.00000000
#> Dale.Chall.old                 0.7224146 -1.00000000     1.00000000
#> Dale.Chall.PSK                 0.7224146 -1.00000000     1.00000000
#> Danielson.Bryan                0.8471650 -0.97829531     0.97829531
#> Danielson.Bryan.2             -0.8066351  0.18028533    -0.18028533
#> Dickes.Steiwer                -0.7764357  0.99638999    -0.99638999
#> DRP                            0.7272457 -0.99996713     0.99996713
#> ELF                            0.8134514 -0.98451149     0.98451149
#> Farr.Jenkins.Paterson         -0.7252569  0.99998999    -0.99998999
#> Flesch                        -0.9146054  0.92913685    -0.92913685
#> Flesch.PSK                     0.8829221 -0.95657238     0.95657238
#> Flesch.Kincaid                 0.8255270 -0.98488025     0.98488025
#>                       Dale.Chall.PSK Danielson.Bryan Danielson.Bryan.2
#> ARI                       0.99099423      0.99708146       -0.31030344
#> ARI.simple                0.99169343      0.99667721       -0.30523815
#> Bormuth                  -0.99996713     -0.97962805        0.18706023
#> Bormuth.GP                0.92746308      0.88486422       -0.04006627
#> Coleman                  -0.47287100     -0.63201768        0.89220768
#> Coleman.C2               -0.68300831     -0.80680483        0.79314280
#> Coleman.Liau             -0.72238083     -0.84714024        0.80666655
#> Coleman.Liau.grade        0.72238083      0.84714024       -0.80666655
#> Coleman.Liau.short        0.72241463      0.84716497       -0.80663507
#> Dale.Chall               -1.00000000     -0.97829531        0.18028533
#> Dale.Chall.old            1.00000000      0.97829531       -0.18028533
#> Dale.Chall.PSK            1.00000000      0.97829531       -0.18028533
#> Danielson.Bryan           0.97829531      1.00000000       -0.37865977
#> Danielson.Bryan.2        -0.18028533     -0.37865977        1.00000000
#> Dickes.Steiwer           -0.99638999     -0.99178586        0.26101392
#> DRP                       0.99996713      0.97962805       -0.18706023
#> ELF                       0.98451149      0.99535015       -0.32598561
#> Farr.Jenkins.Paterson    -0.99998999     -0.97914589        0.18438177
#> Flesch                   -0.92913685     -0.98231683        0.51737927
#> Flesch.PSK                0.95657238      0.99363576       -0.44828841
#> Flesch.Kincaid            0.98488025      0.99786963       -0.34148774
#>                       Dickes.Steiwer         DRP         ELF
#> ARI                      -0.99844204  0.99188017  0.99594813
#> ARI.simple               -0.99870393  0.99254312  0.99583585
#> Bormuth                   0.99691688 -1.00000000 -0.98541527
#> Bormuth.GP               -0.91329434  0.92665498  0.89837506
#> Coleman                   0.53900648 -0.47841665 -0.61270055
#> Coleman.C2                0.73682822 -0.68780777 -0.79038825
#> Coleman.Liau              0.77640516 -0.72721217 -0.81342458
#> Coleman.Liau.grade       -0.77640516  0.72721217  0.81342458
#> Coleman.Liau.short       -0.77643567  0.72724574  0.81345135
#> Dale.Chall                0.99638999 -0.99996713 -0.98451149
#> Dale.Chall.old           -0.99638999  0.99996713  0.98451149
#> Dale.Chall.PSK           -0.99638999  0.99996713  0.98451149
#> Danielson.Bryan          -0.99178586  0.97962805  0.99535015
#> Danielson.Bryan.2         0.26101392 -0.18706023 -0.32598561
#> Dickes.Steiwer            1.00000000 -0.99691688 -0.99345440
#> DRP                      -0.99691688  1.00000000  0.98541527
#> ELF                      -0.99345440  0.98541527  1.00000000
#> Farr.Jenkins.Paterson     0.99672456 -0.99998536 -0.98524897
#> Flesch                    0.95524130 -0.93162624 -0.97375531
#> Flesch.PSK               -0.97634394  0.95852762  0.98827872
#> Flesch.Kincaid           -0.99512741  0.98602857  0.99727425
#>                       Farr.Jenkins.Paterson     Flesch  Flesch.PSK
#> ARI                             -0.99154201 -0.9683121  0.98543988
#> ARI.simple                      -0.99221922 -0.9670869  0.98463060
#> Bormuth                          0.99998536  0.9316262 -0.95852762
#> Bormuth.GP                      -0.92696268 -0.8225754  0.85630865
#> Coleman                          0.47680930  0.7524555 -0.69917416
#> Coleman.C2                       0.68622584  0.8921701 -0.85640522
#> Coleman.Liau                     0.72522331  0.9145879 -0.88290067
#> Coleman.Liau.grade              -0.72522331 -0.9145879  0.88290067
#> Coleman.Liau.short              -0.72525694 -0.9146054  0.88292206
#> Dale.Chall                       0.99998999  0.9291368 -0.95657238
#> Dale.Chall.old                  -0.99998999 -0.9291368  0.95657238
#> Dale.Chall.PSK                  -0.99998999 -0.9291368  0.95657238
#> Danielson.Bryan                 -0.97914589 -0.9823168  0.99363576
#> Danielson.Bryan.2                0.18438177  0.5173793 -0.44828841
#> Dickes.Steiwer                   0.99672456  0.9552413 -0.97634394
#> DRP                             -0.99998536 -0.9316262  0.95852762
#> ELF                             -0.98524897 -0.9737553  0.98827872
#> Farr.Jenkins.Paterson            1.00000000  0.9307176 -0.95781641
#> Flesch                           0.93071764  1.0000000 -0.99656281
#> Flesch.PSK                      -0.95781641 -0.9965628  1.00000000
#> Flesch.Kincaid                  -0.98561541 -0.9791403  0.99260681
#>                       Flesch.Kincaid         FOG    FOG.PSK    FOG.NRI
#> ARI                       0.99828644  0.99534334  0.9980495  0.4905310
#> ARI.simple                0.99809662  0.99501238  0.9983218  0.4892834
#> Bormuth                  -0.98602857 -0.97973311 -0.9967939 -0.4571761
#> Bormuth.GP                0.89508980  0.88272558  0.9136641  0.3361798
#> Coleman                  -0.61241856 -0.62791369 -0.5401789 -0.4679026
#> Coleman.C2               -0.79335974 -0.80512019 -0.7376615 -0.5350724
#> Coleman.Liau             -0.82550024 -0.83906193 -0.7749720 -0.5426918
#> Coleman.Liau.grade        0.82550024  0.83906193  0.7749720  0.5426918
#> Coleman.Liau.short        0.82552702  0.83908765  0.7750026  0.5426992
#> Dale.Chall               -0.98488025 -0.97841468 -0.9962621 -0.4549252
#> Dale.Chall.old            0.98488025  0.97841468  0.9962621  0.4549252
#> Dale.Chall.PSK            0.98488025  0.97841468  0.9962621  0.4549252
#> Danielson.Bryan           0.99786963  0.99695239  0.9912643  0.5028375
#> Danielson.Bryan.2        -0.34148774 -0.36653699 -0.2590925 -0.3671130
#> Dickes.Steiwer           -0.99512741 -0.99112313 -0.9994543 -0.4649778
#> DRP                       0.98602857  0.97973311  0.9967939  0.4571761
#> ELF                       0.99727425  0.99443432  0.9938624  0.4952706
#> Farr.Jenkins.Paterson    -0.98561541 -0.97924413 -0.9966030 -0.4562044
#> Flesch                   -0.97914032 -0.98344165 -0.9567470 -0.5313693
#> Flesch.PSK                0.99260681  0.99454929  0.9775024  0.5208520
#> Flesch.Kincaid            1.00000000  0.99846226  0.9957626  0.4989686
#>                          FORCAST FORCAST.RGL        Fucks Linsear.Write
#> ARI                    0.5784256   0.5784256  0.997730647    0.98324412
#> ARI.simple             0.5744257   0.5744257  0.998049346    0.98272675
#> Bormuth               -0.4784167  -0.4784167 -0.997515204   -0.96316927
#> Bormuth.GP             0.3418689   0.3418689  0.919893732    0.88236447
#> Coleman               -1.0000000  -1.0000000 -0.530595415   -0.65408730
#> Coleman.C2            -0.9578621  -0.9578621 -0.727965258   -0.81398206
#> Coleman.Liau          -0.9027073  -0.9027073 -0.768184772   -0.84460735
#> Coleman.Liau.grade     0.9027073   0.9027073  0.768184772    0.84460735
#> Coleman.Liau.short     0.9026909   0.9026909  0.768215251    0.84462952
#> Dale.Chall            -0.4728710  -0.4728710 -0.997112805   -0.96173384
#> Dale.Chall.old         0.4728710   0.4728710  0.997112805    0.96173384
#> Dale.Chall.PSK         0.4728710   0.4728710  0.997112805    0.96173384
#> Danielson.Bryan        0.6320177   0.6320177  0.990775706    0.98940268
#> Danielson.Bryan.2     -0.8922077  -0.8922077 -0.249745678   -0.39525523
#> Dickes.Steiwer        -0.5390065  -0.5390065 -0.999407002   -0.97750750
#> DRP                    0.4784167   0.4784167  0.997515204    0.96316927
#> ELF                    0.6127006   0.6127006  0.993623225    0.98844709
#> Farr.Jenkins.Paterson -0.4768093  -0.4768093 -0.997402917   -0.96273644
#> Flesch                -0.7524555  -0.7524555 -0.951964715   -0.98424870
#> Flesch.PSK             0.6991742   0.6991742  0.973922677    0.99144816
#> Flesch.Kincaid         0.6124186   0.6124186  0.993989409    0.98967358
#>                               LIW        nWS       nWS.2       nWS.3
#> ARI                    0.99534363  0.9716012  0.97550799  0.98501356
#> ARI.simple             0.99493616  0.9704382  0.97443531  0.98427247
#> Bormuth               -0.97782198 -0.9362833 -0.94231463 -0.95970981
#> Bormuth.GP             0.88013733  0.8218265  0.82921349  0.85402853
#> Coleman               -0.64151385 -0.7435480 -0.72971888 -0.68262459
#> Coleman.C2            -0.81448329 -0.8869623 -0.87759182 -0.84444185
#> Coleman.Liau          -0.84663781 -0.9108590 -0.90420615 -0.87578843
#> Coleman.Liau.grade     0.84663781  0.9108590  0.90420615  0.87578843
#> Coleman.Liau.short     0.84666264  0.9108773  0.90422529  0.87581059
#> Dale.Chall            -0.97649153 -0.9339636 -0.94010518 -0.95786413
#> Dale.Chall.old         0.97649153  0.9339636  0.94010518  0.95786413
#> Dale.Chall.PSK         0.97649153  0.9339636  0.94010518  0.95786413
#> Danielson.Bryan        0.99817496  0.9848656  0.98751970  0.99235519
#> Danielson.Bryan.2     -0.37953455 -0.5077365 -0.49287612 -0.43696629
#> Dickes.Steiwer        -0.99028645 -0.9594714 -0.96426773 -0.97697873
#> DRP                    0.97782198  0.9362833  0.94231463  0.95970981
#> ELF                    0.99591724  0.9767864  0.97966902  0.98635632
#> Farr.Jenkins.Paterson -0.97739468 -0.9354875 -0.94154405 -0.95902099
#> Flesch                -0.98395156 -0.9980699 -0.99747205 -0.99333954
#> Flesch.PSK             0.99452041  0.9961226  0.99702731  0.99774824
#> Flesch.Kincaid         0.99764432  0.9808885  0.98398342  0.99180645
#>                             nWS.4         RIX      Scrabble        SMOG
#> ARI                    0.99487148  0.99222014 -0.0112182516  0.97393941
#> ARI.simple             0.99451326  0.99191614 -0.0093650938  0.97295999
#> Bormuth               -0.97861519 -0.97704126 -0.0333496851 -0.94303306
#> Bormuth.GP             0.88101179  0.89115856 -0.0158199992  0.82113760
#> Coleman               -0.63167564 -0.62921060  0.3522407942 -0.71333148
#> Coleman.C2            -0.80789970 -0.79924879  0.2472067525 -0.86811865
#> Coleman.Liau          -0.84167664 -0.82960560  0.1918831342 -0.89524825
#> Coleman.Liau.grade     0.84167664  0.82960560 -0.1918831342  0.89524825
#> Coleman.Liau.short     0.84170213  0.82963012 -0.1918617971  0.89526894
#> Dale.Chall            -0.97726151 -0.97599601 -0.0354646177 -0.94089430
#> Dale.Chall.old         0.97726151  0.97599601  0.0354646177  0.94089430
#> Dale.Chall.PSK         0.97726151  0.97599601  0.0354646177  0.94089430
#> Danielson.Bryan        0.99686096  0.99500484 -0.0367962483  0.98460337
#> Danielson.Bryan.2     -0.37128519 -0.35912065  0.3211432536 -0.47640589
#> Dickes.Steiwer        -0.99039744 -0.98825075 -0.0143478880 -0.96367176
#> DRP                    0.97861519  0.97704126  0.0333496851  0.94303306
#> ELF                    0.99411156  0.99677384 -0.0411360767  0.97747350
#> Farr.Jenkins.Paterson -0.97811285 -0.97683788 -0.0335901823 -0.94224804
#> Flesch                -0.98430842 -0.97797435  0.1056725636 -0.99453689
#> Flesch.PSK             0.99497427  0.98969704 -0.0753648675  0.99489006
#> Flesch.Kincaid         0.99823466  0.99457144 -0.0300226725  0.98304183
#>                            SMOG.C SMOG.simple     SMOG.de      Spache
#> ARI                    0.97457741  0.97393941  0.97393941  0.99099423
#> ARI.simple             0.97361454  0.97295999  0.97295999  0.99169343
#> Bormuth               -0.94405248 -0.94303306 -0.94303306 -0.99996713
#> Bormuth.GP             0.82339465  0.82113760  0.82113760  0.92746308
#> Coleman               -0.71131513 -0.71333148 -0.71333148 -0.47287100
#> Coleman.C2            -0.86636103 -0.86811865 -0.86811865 -0.68300831
#> Coleman.Liau          -0.89368168 -0.89524825 -0.89524825 -0.72238083
#> Coleman.Liau.grade     0.89368168  0.89524825  0.89524825  0.72238083
#> Coleman.Liau.short     0.89370243  0.89526894  0.89526894  0.72241463
#> Dale.Chall            -0.94194286 -0.94089430 -0.94089430 -1.00000000
#> Dale.Chall.old         0.94194286  0.94089430  0.94089430  1.00000000
#> Dale.Chall.PSK         0.94194286  0.94089430  0.94089430  1.00000000
#> Danielson.Bryan        0.98509384  0.98460337  0.98460337  0.97829531
#> Danielson.Bryan.2     -0.47358728 -0.47640589 -0.47640589 -0.18028533
#> Dickes.Steiwer        -0.96447120 -0.96367176 -0.96367176 -0.99638999
#> DRP                    0.94405248  0.94303306  0.94303306  0.99996713
#> ELF                    0.97820253  0.97747350  0.97747350  0.98451149
#> Farr.Jenkins.Paterson -0.94328383 -0.94224804 -0.94224804 -0.99998999
#> Flesch                -0.99442699 -0.99453689 -0.99453689 -0.92913685
#> Flesch.PSK             0.99503836  0.99489006  0.99489006  0.95657238
#> Flesch.Kincaid         0.98356657  0.98304183  0.98304183  0.98488025
#>                        Spache.old        Strain Traenkle.Bailer
#> ARI                    0.99099423  0.9976064239    -0.999223850
#> ARI.simple             0.99169343  0.9978635733    -0.999411499
#> Bormuth               -0.99996713 -0.9959605646     0.995931694
#> Bormuth.GP             0.92746308  0.9201367937    -0.910151256
#> Coleman               -0.47287100 -0.5461292792     0.549146837
#> Coleman.C2            -0.68300831 -0.7394040240     0.745171965
#> Coleman.Liau          -0.72238083 -0.7744192348     0.785328440
#> Coleman.Liau.grade     0.72238083  0.7744192348    -0.785328440
#> Coleman.Liau.short     0.72241463  0.7744490784    -0.785358433
#> Dale.Chall            -1.00000000 -0.9954423736     0.995298661
#> Dale.Chall.old         1.00000000  0.9954423736    -0.995298661
#> Dale.Chall.PSK         1.00000000  0.9954423736    -0.995298661
#> Danielson.Bryan        0.97829531  0.9916468644    -0.993491902
#> Danielson.Bryan.2     -0.18028533 -0.2608961180     0.274099704
#> Dickes.Steiwer        -0.99638999 -0.9987560895     0.999594489
#> DRP                    0.99996713  0.9959605646    -0.995931694
#> ELF                    0.98451149  0.9957646921    -0.994372531
#> Farr.Jenkins.Paterson -0.99998999 -0.9958154056     0.995687365
#> Flesch                -0.92913685 -0.9579641510     0.959086871
#> Flesch.PSK             0.95657238  0.9782782859    -0.979131224
#> Flesch.Kincaid         0.98488025  0.9958824210    -0.996329486
#>                       Traenkle.Bailer.2 Wheeler.Smith meanSentenceLength
#> ARI                          -0.7549178    0.99594813         0.99099423
#> ARI.simple                   -0.7514487    0.99583585         0.99169343
#> Bormuth                       0.6663152   -0.98541527        -0.99996713
#> Bormuth.GP                   -0.5147186    0.89837506         0.92746308
#> Coleman                       0.9206771   -0.61270055        -0.47287100
#> Coleman.C2                    0.9574552   -0.79038825        -0.68300831
#> Coleman.Liau                  0.9904793   -0.81342458        -0.72238083
#> Coleman.Liau.grade           -0.9904793    0.81342458         0.72238083
#> Coleman.Liau.short           -0.9904734    0.81345135         0.72241463
#> Dale.Chall                    0.6611792   -0.98451149        -1.00000000
#> Dale.Chall.old               -0.6611792    0.98451149         1.00000000
#> Dale.Chall.PSK               -0.6611792    0.98451149         1.00000000
#> Danielson.Bryan              -0.7999586    0.99535015         0.97829531
#> Danielson.Bryan.2             0.8509255   -0.32598561        -0.18028533
#> Dickes.Steiwer                0.7195290   -0.99345440        -0.99638999
#> DRP                          -0.6663152    0.98541527         0.99996713
#> ELF                          -0.7636349    1.00000000         0.98451149
#> Farr.Jenkins.Paterson         0.6642605   -0.98524897        -0.99998999
#> Flesch                        0.8800874   -0.97375531        -0.92913685
#> Flesch.PSK                   -0.8419885    0.98827872         0.95657238
#> Flesch.Kincaid               -0.7757025    0.99727425         0.98488025
#>                       meanWordSyllables
#> ARI                           0.6762172
#> ARI.simple                    0.6724762
#> Bormuth                      -0.5818750
#> Bormuth.GP                    0.4479544
#> Coleman                      -0.9645525
#> Coleman.C2                   -0.9629188
#> Coleman.Liau                 -0.9543023
#> Coleman.Liau.grade            0.9543023
#> Coleman.Liau.short            0.9542911
#> Dale.Chall                   -0.5763242
#> Dale.Chall.old                0.5763242
#> Dale.Chall.PSK                0.5763242
#> Danielson.Bryan               0.7259320
#> Danielson.Bryan.2            -0.8772134
#> Dickes.Steiwer               -0.6393556
#> DRP                           0.5818750
#> ELF                           0.6978253
#> Farr.Jenkins.Paterson        -0.5798330
#> Flesch                       -0.8376402
#> Flesch.PSK                    0.7895116
#> Flesch.Kincaid                0.7091829
#>  [ reached getOption("max.print") -- omitted 26 rows ]

textstat_readability(data_corpus_inaugural, measure = "Flesch.Kincaid")

#>           document Flesch.Kincaid
#> 1  1789-Washington      28.432155
#> 2  1793-Washington      17.151759
#> 3       1797-Adams      28.692087
#> 4   1801-Jefferson      19.295924
#> 5   1805-Jefferson      21.917088
#> 6     1809-Madison      25.557611
#> 7     1813-Madison      18.034452
#> 8      1817-Monroe      14.443950
#> 9      1821-Monroe      16.773058
#> 10      1825-Adams      19.748273
#> 11    1829-Jackson      21.893148
#> 12    1833-Jackson      19.958757
#> 13   1837-VanBuren      20.263524
#> 14   1841-Harrison      19.355855
#> 15       1845-Polk      16.406306
#> 16     1849-Taylor      23.748805
#> 17     1853-Pierce      16.438241
#> 18   1857-Buchanan      16.110868
#> 19    1861-Lincoln      13.661936
#> 20    1865-Lincoln      11.774929
#> 21      1869-Grant      14.008426
#> 22      1873-Grant      15.082444
#> 23      1877-Hayes      20.297851
#> 24   1881-Garfield      14.042973
#> 25  1885-Cleveland      19.052937
#> 26   1889-Harrison      14.944583
#> 27  1893-Cleveland      18.248838
#> 28   1897-McKinley      15.854298
#> 29   1901-McKinley      12.728640
#> 30  1905-Roosevelt      13.959254
#> 31       1909-Taft      17.099757
#> 32     1913-Wilson      11.707659
#> 33     1917-Wilson      11.780913
#> 34    1921-Harding      13.214469
#> 35   1925-Coolidge      11.626699
#> 36     1929-Hoover      13.259377
#> 37  1933-Roosevelt      11.640683
#> 38  1937-Roosevelt      10.619558
#> 39  1941-Roosevelt       9.687985
#> 40  1945-Roosevelt       9.175680
#> 41     1949-Truman      11.328244
#> 42 1953-Eisenhower      10.328852
#> 43 1957-Eisenhower       8.094908
#> 44    1961-Kennedy      11.822997
#> 45    1965-Johnson       7.550602
#> 46      1969-Nixon       9.235663
#> 47      1973-Nixon      12.288965
#> 48     1977-Carter      11.670742
#> 49     1981-Reagan       9.798608
#> 50     1985-Reagan      10.420294
#> 51       1989-Bush       7.147029
#> 52    1993-Clinton      10.374204
#> 53    1997-Clinton       9.828863
#> 54       2001-Bush       8.918201
#> 55       2005-Bush      11.036277
#> 56      2009-Obama      10.229426
#> 57      2013-Obama      11.734767
#> 58      2017-Trump       9.163084

inaugReadability <- textstat_readability(data_corpus_inaugural, "all")
cor(inaugReadability[,-1])

#>                               ARI   ARI.simple     Bormuth  Bormuth.GP
#> ARI                    1.00000000  0.999985813 -0.99188017  0.90232231
#> ARI.simple             0.99998581  1.000000000 -0.99254312  0.90363285
#> Bormuth               -0.99188017 -0.992543125  1.00000000 -0.92665498
#> Bormuth.GP             0.90232231  0.903632850 -0.92665498  1.00000000
#> Coleman               -0.57842561 -0.574425719  0.47841665 -0.34186891
#> Coleman.C2            -0.76781514 -0.764705749  0.68780777 -0.51669363
#> Coleman.Liau          -0.80785899 -0.804736665  0.72721217 -0.55936385
#> Coleman.Liau.grade     0.80785899  0.804736665 -0.72721217  0.55936385
#> Coleman.Liau.short     0.80788731  0.804765212 -0.72724574  0.55939187
#> Dale.Chall            -0.99099423 -0.991693434  0.99996713 -0.92746308
#> Dale.Chall.old         0.99099423  0.991693434 -0.99996713  0.92746308
#> Dale.Chall.PSK         0.99099423  0.991693434 -0.99996713  0.92746308
#> Danielson.Bryan        0.99708146  0.996677212 -0.97962805  0.88486422
#> Danielson.Bryan.2     -0.31030344 -0.305238155  0.18706023 -0.04006627
#> Dickes.Steiwer        -0.99844204 -0.998703929  0.99691688 -0.91329434
#> DRP                    0.99188017  0.992543125 -1.00000000  0.92665498
#> ELF                    0.99594813  0.995835853 -0.98541527  0.89837506
#> Farr.Jenkins.Paterson -0.99154201 -0.992219215  0.99998536 -0.92696268
#> Flesch                -0.96831214 -0.967086924  0.93162624 -0.82257543
#> Flesch.PSK             0.98543988  0.984630595 -0.95852762  0.85630865
#> Flesch.Kincaid         0.99828644  0.998096618 -0.98602857  0.89508980
#>                          Coleman Coleman.C2 Coleman.Liau Coleman.Liau.grade
#> ARI                   -0.5784256 -0.7678151   -0.8078590          0.8078590
#> ARI.simple            -0.5744257 -0.7647057   -0.8047367          0.8047367
#> Bormuth                0.4784167  0.6878078    0.7272122         -0.7272122
#> Bormuth.GP            -0.3418689 -0.5166936   -0.5593638          0.5593638
#> Coleman                1.0000000  0.9578621    0.9027073         -0.9027073
#> Coleman.C2             0.9578621  1.0000000    0.9676044         -0.9676044
#> Coleman.Liau           0.9027073  0.9676044    1.0000000         -1.0000000
#> Coleman.Liau.grade    -0.9027073 -0.9676044   -1.0000000          1.0000000
#> Coleman.Liau.short    -0.9026909 -0.9676038   -1.0000000          1.0000000
#> Dale.Chall             0.4728710  0.6830083    0.7223808         -0.7223808
#> Dale.Chall.old        -0.4728710 -0.6830083   -0.7223808          0.7223808
#> Dale.Chall.PSK        -0.4728710 -0.6830083   -0.7223808          0.7223808
#> Danielson.Bryan       -0.6320177 -0.8068048   -0.8471402          0.8471402
#> Danielson.Bryan.2      0.8922077  0.7931428    0.8066665         -0.8066665
#> Dickes.Steiwer         0.5390065  0.7368282    0.7764052         -0.7764052
#> DRP                   -0.4784167 -0.6878078   -0.7272122          0.7272122
#> ELF                   -0.6127006 -0.7903882   -0.8134246          0.8134246
#> Farr.Jenkins.Paterson  0.4768093  0.6862258    0.7252233         -0.7252233
#> Flesch                 0.7524555  0.8921701    0.9145879         -0.9145879
#> Flesch.PSK            -0.6991742 -0.8564052   -0.8829007          0.8829007
#> Flesch.Kincaid        -0.6124186 -0.7933597   -0.8255002          0.8255002
#>                       Coleman.Liau.short  Dale.Chall Dale.Chall.old
#> ARI                            0.8078873 -0.99099423     0.99099423
#> ARI.simple                     0.8047652 -0.99169343     0.99169343
#> Bormuth                       -0.7272457  0.99996713    -0.99996713
#> Bormuth.GP                     0.5593919 -0.92746308     0.92746308
#> Coleman                       -0.9026909  0.47287100    -0.47287100
#> Coleman.C2                    -0.9676038  0.68300831    -0.68300831
#> Coleman.Liau                  -1.0000000  0.72238083    -0.72238083
#> Coleman.Liau.grade             1.0000000 -0.72238083     0.72238083
#> Coleman.Liau.short             1.0000000 -0.72241463     0.72241463
#> Dale.Chall                    -0.7224146  1.00000000    -1.00000000
#> Dale.Chall.old                 0.7224146 -1.00000000     1.00000000
#> Dale.Chall.PSK                 0.7224146 -1.00000000     1.00000000
#> Danielson.Bryan                0.8471650 -0.97829531     0.97829531
#> Danielson.Bryan.2             -0.8066351  0.18028533    -0.18028533
#> Dickes.Steiwer                -0.7764357  0.99638999    -0.99638999
#> DRP                            0.7272457 -0.99996713     0.99996713
#> ELF                            0.8134514 -0.98451149     0.98451149
#> Farr.Jenkins.Paterson         -0.7252569  0.99998999    -0.99998999
#> Flesch                        -0.9146054  0.92913685    -0.92913685
#> Flesch.PSK                     0.8829221 -0.95657238     0.95657238
#> Flesch.Kincaid                 0.8255270 -0.98488025     0.98488025
#>                       Dale.Chall.PSK Danielson.Bryan Danielson.Bryan.2
#> ARI                       0.99099423      0.99708146       -0.31030344
#> ARI.simple                0.99169343      0.99667721       -0.30523815
#> Bormuth                  -0.99996713     -0.97962805        0.18706023
#> Bormuth.GP                0.92746308      0.88486422       -0.04006627
#> Coleman                  -0.47287100     -0.63201768        0.89220768
#> Coleman.C2               -0.68300831     -0.80680483        0.79314280
#> Coleman.Liau             -0.72238083     -0.84714024        0.80666655
#> Coleman.Liau.grade        0.72238083      0.84714024       -0.80666655
#> Coleman.Liau.short        0.72241463      0.84716497       -0.80663507
#> Dale.Chall               -1.00000000     -0.97829531        0.18028533
#> Dale.Chall.old            1.00000000      0.97829531       -0.18028533
#> Dale.Chall.PSK            1.00000000      0.97829531       -0.18028533
#> Danielson.Bryan           0.97829531      1.00000000       -0.37865977
#> Danielson.Bryan.2        -0.18028533     -0.37865977        1.00000000
#> Dickes.Steiwer           -0.99638999     -0.99178586        0.26101392
#> DRP                       0.99996713      0.97962805       -0.18706023
#> ELF                       0.98451149      0.99535015       -0.32598561
#> Farr.Jenkins.Paterson    -0.99998999     -0.97914589        0.18438177
#> Flesch                   -0.92913685     -0.98231683        0.51737927
#> Flesch.PSK                0.95657238      0.99363576       -0.44828841
#> Flesch.Kincaid            0.98488025      0.99786963       -0.34148774
#>                       Dickes.Steiwer         DRP         ELF
#> ARI                      -0.99844204  0.99188017  0.99594813
#> ARI.simple               -0.99870393  0.99254312  0.99583585
#> Bormuth                   0.99691688 -1.00000000 -0.98541527
#> Bormuth.GP               -0.91329434  0.92665498  0.89837506
#> Coleman                   0.53900648 -0.47841665 -0.61270055
#> Coleman.C2                0.73682822 -0.68780777 -0.79038825
#> Coleman.Liau              0.77640516 -0.72721217 -0.81342458
#> Coleman.Liau.grade       -0.77640516  0.72721217  0.81342458
#> Coleman.Liau.short       -0.77643567  0.72724574  0.81345135
#> Dale.Chall                0.99638999 -0.99996713 -0.98451149
#> Dale.Chall.old           -0.99638999  0.99996713  0.98451149
#> Dale.Chall.PSK           -0.99638999  0.99996713  0.98451149
#> Danielson.Bryan          -0.99178586  0.97962805  0.99535015
#> Danielson.Bryan.2         0.26101392 -0.18706023 -0.32598561
#> Dickes.Steiwer            1.00000000 -0.99691688 -0.99345440
#> DRP                      -0.99691688  1.00000000  0.98541527
#> ELF                      -0.99345440  0.98541527  1.00000000
#> Farr.Jenkins.Paterson     0.99672456 -0.99998536 -0.98524897
#> Flesch                    0.95524130 -0.93162624 -0.97375531
#> Flesch.PSK               -0.97634394  0.95852762  0.98827872
#> Flesch.Kincaid           -0.99512741  0.98602857  0.99727425
#>                       Farr.Jenkins.Paterson     Flesch  Flesch.PSK
#> ARI                             -0.99154201 -0.9683121  0.98543988
#> ARI.simple                      -0.99221922 -0.9670869  0.98463060
#> Bormuth                          0.99998536  0.9316262 -0.95852762
#> Bormuth.GP                      -0.92696268 -0.8225754  0.85630865
#> Coleman                          0.47680930  0.7524555 -0.69917416
#> Coleman.C2                       0.68622584  0.8921701 -0.85640522
#> Coleman.Liau                     0.72522331  0.9145879 -0.88290067
#> Coleman.Liau.grade              -0.72522331 -0.9145879  0.88290067
#> Coleman.Liau.short              -0.72525694 -0.9146054  0.88292206
#> Dale.Chall                       0.99998999  0.9291368 -0.95657238
#> Dale.Chall.old                  -0.99998999 -0.9291368  0.95657238
#> Dale.Chall.PSK                  -0.99998999 -0.9291368  0.95657238
#> Danielson.Bryan                 -0.97914589 -0.9823168  0.99363576
#> Danielson.Bryan.2                0.18438177  0.5173793 -0.44828841
#> Dickes.Steiwer                   0.99672456  0.9552413 -0.97634394
#> DRP                             -0.99998536 -0.9316262  0.95852762
#> ELF                             -0.98524897 -0.9737553  0.98827872
#> Farr.Jenkins.Paterson            1.00000000  0.9307176 -0.95781641
#> Flesch                           0.93071764  1.0000000 -0.99656281
#> Flesch.PSK                      -0.95781641 -0.9965628  1.00000000
#> Flesch.Kincaid                  -0.98561541 -0.9791403  0.99260681
#>                       Flesch.Kincaid         FOG    FOG.PSK    FOG.NRI
#> ARI                       0.99828644  0.99534334  0.9980495  0.4905310
#> ARI.simple                0.99809662  0.99501238  0.9983218  0.4892834
#> Bormuth                  -0.98602857 -0.97973311 -0.9967939 -0.4571761
#> Bormuth.GP                0.89508980  0.88272558  0.9136641  0.3361798
#> Coleman                  -0.61241856 -0.62791369 -0.5401789 -0.4679026
#> Coleman.C2               -0.79335974 -0.80512019 -0.7376615 -0.5350724
#> Coleman.Liau             -0.82550024 -0.83906193 -0.7749720 -0.5426918
#> Coleman.Liau.grade        0.82550024  0.83906193  0.7749720  0.5426918
#> Coleman.Liau.short        0.82552702  0.83908765  0.7750026  0.5426992
#> Dale.Chall               -0.98488025 -0.97841468 -0.9962621 -0.4549252
#> Dale.Chall.old            0.98488025  0.97841468  0.9962621  0.4549252
#> Dale.Chall.PSK            0.98488025  0.97841468  0.9962621  0.4549252
#> Danielson.Bryan           0.99786963  0.99695239  0.9912643  0.5028375
#> Danielson.Bryan.2        -0.34148774 -0.36653699 -0.2590925 -0.3671130
#> Dickes.Steiwer           -0.99512741 -0.99112313 -0.9994543 -0.4649778
#> DRP                       0.98602857  0.97973311  0.9967939  0.4571761
#> ELF                       0.99727425  0.99443432  0.9938624  0.4952706
#> Farr.Jenkins.Paterson    -0.98561541 -0.97924413 -0.9966030 -0.4562044
#> Flesch                   -0.97914032 -0.98344165 -0.9567470 -0.5313693
#> Flesch.PSK                0.99260681  0.99454929  0.9775024  0.5208520
#> Flesch.Kincaid            1.00000000  0.99846226  0.9957626  0.4989686
#>                          FORCAST FORCAST.RGL        Fucks Linsear.Write
#> ARI                    0.5784256   0.5784256  0.997730647    0.98324412
#> ARI.simple             0.5744257   0.5744257  0.998049346    0.98272675
#> Bormuth               -0.4784167  -0.4784167 -0.997515204   -0.96316927
#> Bormuth.GP             0.3418689   0.3418689  0.919893732    0.88236447
#> Coleman               -1.0000000  -1.0000000 -0.530595415   -0.65408730
#> Coleman.C2            -0.9578621  -0.9578621 -0.727965258   -0.81398206
#> Coleman.Liau          -0.9027073  -0.9027073 -0.768184772   -0.84460735
#> Coleman.Liau.grade     0.9027073   0.9027073  0.768184772    0.84460735
#> Coleman.Liau.short     0.9026909   0.9026909  0.768215251    0.84462952
#> Dale.Chall            -0.4728710  -0.4728710 -0.997112805   -0.96173384
#> Dale.Chall.old         0.4728710   0.4728710  0.997112805    0.96173384
#> Dale.Chall.PSK         0.4728710   0.4728710  0.997112805    0.96173384
#> Danielson.Bryan        0.6320177   0.6320177  0.990775706    0.98940268
#> Danielson.Bryan.2     -0.8922077  -0.8922077 -0.249745678   -0.39525523
#> Dickes.Steiwer        -0.5390065  -0.5390065 -0.999407002   -0.97750750
#> DRP                    0.4784167   0.4784167  0.997515204    0.96316927
#> ELF                    0.6127006   0.6127006  0.993623225    0.98844709
#> Farr.Jenkins.Paterson -0.4768093  -0.4768093 -0.997402917   -0.96273644
#> Flesch                -0.7524555  -0.7524555 -0.951964715   -0.98424870
#> Flesch.PSK             0.6991742   0.6991742  0.973922677    0.99144816
#> Flesch.Kincaid         0.6124186   0.6124186  0.993989409    0.98967358
#>                               LIW        nWS       nWS.2       nWS.3
#> ARI                    0.99534363  0.9716012  0.97550799  0.98501356
#> ARI.simple             0.99493616  0.9704382  0.97443531  0.98427247
#> Bormuth               -0.97782198 -0.9362833 -0.94231463 -0.95970981
#> Bormuth.GP             0.88013733  0.8218265  0.82921349  0.85402853
#> Coleman               -0.64151385 -0.7435480 -0.72971888 -0.68262459
#> Coleman.C2            -0.81448329 -0.8869623 -0.87759182 -0.84444185
#> Coleman.Liau          -0.84663781 -0.9108590 -0.90420615 -0.87578843
#> Coleman.Liau.grade     0.84663781  0.9108590  0.90420615  0.87578843
#> Coleman.Liau.short     0.84666264  0.9108773  0.90422529  0.87581059
#> Dale.Chall            -0.97649153 -0.9339636 -0.94010518 -0.95786413
#> Dale.Chall.old         0.97649153  0.9339636  0.94010518  0.95786413
#> Dale.Chall.PSK         0.97649153  0.9339636  0.94010518  0.95786413
#> Danielson.Bryan        0.99817496  0.9848656  0.98751970  0.99235519
#> Danielson.Bryan.2     -0.37953455 -0.5077365 -0.49287612 -0.43696629
#> Dickes.Steiwer        -0.99028645 -0.9594714 -0.96426773 -0.97697873
#> DRP                    0.97782198  0.9362833  0.94231463  0.95970981
#> ELF                    0.99591724  0.9767864  0.97966902  0.98635632
#> Farr.Jenkins.Paterson -0.97739468 -0.9354875 -0.94154405 -0.95902099
#> Flesch                -0.98395156 -0.9980699 -0.99747205 -0.99333954
#> Flesch.PSK             0.99452041  0.9961226  0.99702731  0.99774824
#> Flesch.Kincaid         0.99764432  0.9808885  0.98398342  0.99180645
#>                             nWS.4         RIX      Scrabble        SMOG
#> ARI                    0.99487148  0.99222014 -0.0112182516  0.97393941
#> ARI.simple             0.99451326  0.99191614 -0.0093650938  0.97295999
#> Bormuth               -0.97861519 -0.97704126 -0.0333496851 -0.94303306
#> Bormuth.GP             0.88101179  0.89115856 -0.0158199992  0.82113760
#> Coleman               -0.63167564 -0.62921060  0.3522407942 -0.71333148
#> Coleman.C2            -0.80789970 -0.79924879  0.2472067525 -0.86811865
#> Coleman.Liau          -0.84167664 -0.82960560  0.1918831342 -0.89524825
#> Coleman.Liau.grade     0.84167664  0.82960560 -0.1918831342  0.89524825
#> Coleman.Liau.short     0.84170213  0.82963012 -0.1918617971  0.89526894
#> Dale.Chall            -0.97726151 -0.97599601 -0.0354646177 -0.94089430
#> Dale.Chall.old         0.97726151  0.97599601  0.0354646177  0.94089430
#> Dale.Chall.PSK         0.97726151  0.97599601  0.0354646177  0.94089430
#> Danielson.Bryan        0.99686096  0.99500484 -0.0367962483  0.98460337
#> Danielson.Bryan.2     -0.37128519 -0.35912065  0.3211432536 -0.47640589
#> Dickes.Steiwer        -0.99039744 -0.98825075 -0.0143478880 -0.96367176
#> DRP                    0.97861519  0.97704126  0.0333496851  0.94303306
#> ELF                    0.99411156  0.99677384 -0.0411360767  0.97747350
#> Farr.Jenkins.Paterson -0.97811285 -0.97683788 -0.0335901823 -0.94224804
#> Flesch                -0.98430842 -0.97797435  0.1056725636 -0.99453689
#> Flesch.PSK             0.99497427  0.98969704 -0.0753648675  0.99489006
#> Flesch.Kincaid         0.99823466  0.99457144 -0.0300226725  0.98304183
#>                            SMOG.C SMOG.simple     SMOG.de      Spache
#> ARI                    0.97457741  0.97393941  0.97393941  0.99099423
#> ARI.simple             0.97361454  0.97295999  0.97295999  0.99169343
#> Bormuth               -0.94405248 -0.94303306 -0.94303306 -0.99996713
#> Bormuth.GP             0.82339465  0.82113760  0.82113760  0.92746308
#> Coleman               -0.71131513 -0.71333148 -0.71333148 -0.47287100
#> Coleman.C2            -0.86636103 -0.86811865 -0.86811865 -0.68300831
#> Coleman.Liau          -0.89368168 -0.89524825 -0.89524825 -0.72238083
#> Coleman.Liau.grade     0.89368168  0.89524825  0.89524825  0.72238083
#> Coleman.Liau.short     0.89370243  0.89526894  0.89526894  0.72241463
#> Dale.Chall            -0.94194286 -0.94089430 -0.94089430 -1.00000000
#> Dale.Chall.old         0.94194286  0.94089430  0.94089430  1.00000000
#> Dale.Chall.PSK         0.94194286  0.94089430  0.94089430  1.00000000
#> Danielson.Bryan        0.98509384  0.98460337  0.98460337  0.97829531
#> Danielson.Bryan.2     -0.47358728 -0.47640589 -0.47640589 -0.18028533
#> Dickes.Steiwer        -0.96447120 -0.96367176 -0.96367176 -0.99638999
#> DRP                    0.94405248  0.94303306  0.94303306  0.99996713
#> ELF                    0.97820253  0.97747350  0.97747350  0.98451149
#> Farr.Jenkins.Paterson -0.94328383 -0.94224804 -0.94224804 -0.99998999
#> Flesch                -0.99442699 -0.99453689 -0.99453689 -0.92913685
#> Flesch.PSK             0.99503836  0.99489006  0.99489006  0.95657238
#> Flesch.Kincaid         0.98356657  0.98304183  0.98304183  0.98488025
#>                        Spache.old        Strain Traenkle.Bailer
#> ARI                    0.99099423  0.9976064239    -0.999223850
#> ARI.simple             0.99169343  0.9978635733    -0.999411499
#> Bormuth               -0.99996713 -0.9959605646     0.995931694
#> Bormuth.GP             0.92746308  0.9201367937    -0.910151256
#> Coleman               -0.47287100 -0.5461292792     0.549146837
#> Coleman.C2            -0.68300831 -0.7394040240     0.745171965
#> Coleman.Liau          -0.72238083 -0.7744192348     0.785328440
#> Coleman.Liau.grade     0.72238083  0.7744192348    -0.785328440
#> Coleman.Liau.short     0.72241463  0.7744490784    -0.785358433
#> Dale.Chall            -1.00000000 -0.9954423736     0.995298661
#> Dale.Chall.old         1.00000000  0.9954423736    -0.995298661
#> Dale.Chall.PSK         1.00000000  0.9954423736    -0.995298661
#> Danielson.Bryan        0.97829531  0.9916468644    -0.993491902
#> Danielson.Bryan.2     -0.18028533 -0.2608961180     0.274099704
#> Dickes.Steiwer        -0.99638999 -0.9987560895     0.999594489
#> DRP                    0.99996713  0.9959605646    -0.995931694
#> ELF                    0.98451149  0.9957646921    -0.994372531
#> Farr.Jenkins.Paterson -0.99998999 -0.9958154056     0.995687365
#> Flesch                -0.92913685 -0.9579641510     0.959086871
#> Flesch.PSK             0.95657238  0.9782782859    -0.979131224
#> Flesch.Kincaid         0.98488025  0.9958824210    -0.996329486
#>                       Traenkle.Bailer.2 Wheeler.Smith meanSentenceLength
#> ARI                          -0.7549178    0.99594813         0.99099423
#> ARI.simple                   -0.7514487    0.99583585         0.99169343
#> Bormuth                       0.6663152   -0.98541527        -0.99996713
#> Bormuth.GP                   -0.5147186    0.89837506         0.92746308
#> Coleman                       0.9206771   -0.61270055        -0.47287100
#> Coleman.C2                    0.9574552   -0.79038825        -0.68300831
#> Coleman.Liau                  0.9904793   -0.81342458        -0.72238083
#> Coleman.Liau.grade           -0.9904793    0.81342458         0.72238083
#> Coleman.Liau.short           -0.9904734    0.81345135         0.72241463
#> Dale.Chall                    0.6611792   -0.98451149        -1.00000000
#> Dale.Chall.old               -0.6611792    0.98451149         1.00000000
#> Dale.Chall.PSK               -0.6611792    0.98451149         1.00000000
#> Danielson.Bryan              -0.7999586    0.99535015         0.97829531
#> Danielson.Bryan.2             0.8509255   -0.32598561        -0.18028533
#> Dickes.Steiwer                0.7195290   -0.99345440        -0.99638999
#> DRP                          -0.6663152    0.98541527         0.99996713
#> ELF                          -0.7636349    1.00000000         0.98451149
#> Farr.Jenkins.Paterson         0.6642605   -0.98524897        -0.99998999
#> Flesch                        0.8800874   -0.97375531        -0.92913685
#> Flesch.PSK                   -0.8419885    0.98827872         0.95657238
#> Flesch.Kincaid               -0.7757025    0.99727425         0.98488025
#>                       meanWordSyllables
#> ARI                           0.6762172
#> ARI.simple                    0.6724762
#> Bormuth                      -0.5818750
#> Bormuth.GP                    0.4479544
#> Coleman                      -0.9645525
#> Coleman.C2                   -0.9629188
#> Coleman.Liau                 -0.9543023
#> Coleman.Liau.grade            0.9543023
#> Coleman.Liau.short            0.9542911
#> Dale.Chall                   -0.5763242
#> Dale.Chall.old                0.5763242
#> Dale.Chall.PSK                0.5763242
#> Danielson.Bryan               0.7259320
#> Danielson.Bryan.2            -0.8772134
#> Dickes.Steiwer               -0.6393556
#> DRP                           0.5818750
#> ELF                           0.6978253
#> Farr.Jenkins.Paterson        -0.5798330
#> Flesch                       -0.8376402
#> Flesch.PSK                    0.7895116
#> Flesch.Kincaid                0.7091829
#>  [ reached getOption("max.print") -- omitted 26 rows ]