calculate readability

Calculate the readability of text(s) using one of a variety of computed indexes.

textstat_readability(x, measure = c("all", "ARI", "ARI.simple", "Bormuth",
  "Bormuth.GP", "Coleman", "Coleman.C2", "Coleman.Liau", "Coleman.Liau.grade",
  "Coleman.Liau.short", "Dale.Chall", "Dale.Chall.old", "Dale.Chall.PSK",
  "Danielson.Bryan", "Danielson.Bryan.2", "Dickes.Steiwer", "DRP", "ELF",
  "Farr.Jenkins.Paterson", "Flesch", "Flesch.PSK", "Flesch.Kincaid", "FOG",
  "FOG.PSK", "FOG.NRI", "FORCAST", "FORCAST.RGL", "Fucks", "Linsear.Write",
  "LIW", "nWS", "nWS.2", "nWS.3", "nWS.4", "RIX", "Scrabble", "SMOG", "SMOG.C",
  "SMOG.simple",      "SMOG.de", "Spache", "Spache.old", "Strain",
  "Traenkle.Bailer", "Traenkle.Bailer.2", "Wheeler.Smith", "meanSentenceLength",
  "meanWordSyllables"), removeHyphens = TRUE, drop = TRUE, ...)

Arguments

x

a character or corpus object containing the texts

measure

character vector defining the readability measure to calculate

removeHyphens

if TRUE, treat constituent words in hyphenated as separate terms, for purposes of computing word lengths, e.g. "decision-making" as two terms of lengths 8 and 6 characters respectively, rather than as a single word of 15 characters

drop

if TRUE, the result is returned as a numeric vector if only a single measure is requested; otherwise, a data.frame is returned with each column consisting of a requested measure.

...

not used

Value

a data.frame object consisting of the documents as rows, and the readability statistics as columns

Examples

txt <- c("Readability zero one.  Ten, Eleven.", "The cat in a dilapidated tophat.")
textstat_readability(txt, "Flesch.Kincaid")

#>     text1     text2 
#> 13.705000  8.383333 

textstat_readability(txt, "Flesch.Kincaid", drop = FALSE)

#>       Flesch.Kincaid
#> text1      13.705000
#> text2       8.383333

textstat_readability(txt, c("FOG", "FOG.PSK", "FOG.NRI"))

#>             FOG  FOG.PSK FOG.NRI
#> text1 17.000000 4.608659   -1.30
#> text2  9.066667 3.254382   -1.11

inaugReadability <- textstat_readability(data_corpus_inaugural, "all")
round(cor(inaugReadability), 2)

#>                         ARI ARI.simple Bormuth Bormuth.GP Coleman Coleman.C2
#> ARI                    1.00       1.00   -0.99       0.90   -0.58      -0.77
#> ARI.simple             1.00       1.00   -0.99       0.90   -0.58      -0.76
#> Bormuth               -0.99      -0.99    1.00      -0.93    0.48       0.68
#> Bormuth.GP             0.90       0.90   -0.93       1.00   -0.34      -0.51
#> Coleman               -0.58      -0.58    0.48      -0.34    1.00       0.96
#> Coleman.C2            -0.77      -0.76    0.68      -0.51    0.96       1.00
#> Coleman.Liau          -0.81      -0.80    0.73      -0.56    0.91       0.97
#> Coleman.Liau.grade     0.81       0.80   -0.73       0.56   -0.91      -0.97
#> Coleman.Liau.short     0.81       0.80   -0.73       0.56   -0.91      -0.97
#> Dale.Chall            -0.99      -0.99    1.00      -0.93    0.48       0.68
#> Dale.Chall.old         0.99       0.99   -1.00       0.93   -0.48      -0.68
#> Dale.Chall.PSK         0.99       0.99   -1.00       0.93   -0.48      -0.68
#> Danielson.Bryan        1.00       1.00   -0.98       0.88   -0.64      -0.80
#> Danielson.Bryan.2     -0.32      -0.32    0.20      -0.04    0.89       0.80
#> Dickes.Steiwer        -0.39      -0.39    0.41      -0.34    0.14       0.25
#> DRP                    0.99       0.99   -1.00       0.93   -0.48      -0.68
#> ELF                    1.00       1.00   -0.99       0.90   -0.62      -0.79
#> Farr.Jenkins.Paterson -0.99      -0.99    1.00      -0.93    0.48       0.68
#> Flesch                -0.97      -0.97    0.93      -0.82    0.76       0.89
#> Flesch.PSK             0.99       0.98   -0.96       0.86   -0.70      -0.86
#> Flesch.Kincaid         1.00       1.00   -0.99       0.90   -0.62      -0.79
#>                       Coleman.Liau Coleman.Liau.grade Coleman.Liau.short
#> ARI                          -0.81               0.81               0.81
#> ARI.simple                   -0.80               0.80               0.80
#> Bormuth                       0.73              -0.73              -0.73
#> Bormuth.GP                   -0.56               0.56               0.56
#> Coleman                       0.91              -0.91              -0.91
#> Coleman.C2                    0.97              -0.97              -0.97
#> Coleman.Liau                  1.00              -1.00              -1.00
#> Coleman.Liau.grade           -1.00               1.00               1.00
#> Coleman.Liau.short           -1.00               1.00               1.00
#> Dale.Chall                    0.72              -0.72              -0.72
#> Dale.Chall.old               -0.72               0.72               0.72
#> Dale.Chall.PSK               -0.72               0.72               0.72
#> Danielson.Bryan              -0.85               0.85               0.85
#> Danielson.Bryan.2             0.81              -0.81              -0.81
#> Dickes.Steiwer                0.23              -0.23              -0.23
#> DRP                          -0.73               0.73               0.73
#> ELF                          -0.81               0.81               0.81
#> Farr.Jenkins.Paterson         0.72              -0.72              -0.72
#> Flesch                        0.91              -0.91              -0.91
#> Flesch.PSK                   -0.88               0.88               0.88
#> Flesch.Kincaid               -0.82               0.82               0.82
#>                       Dale.Chall Dale.Chall.old Dale.Chall.PSK Danielson.Bryan
#> ARI                        -0.99           0.99           0.99            1.00
#> ARI.simple                 -0.99           0.99           0.99            1.00
#> Bormuth                     1.00          -1.00          -1.00           -0.98
#> Bormuth.GP                 -0.93           0.93           0.93            0.88
#> Coleman                     0.48          -0.48          -0.48           -0.64
#> Coleman.C2                  0.68          -0.68          -0.68           -0.80
#> Coleman.Liau                0.72          -0.72          -0.72           -0.85
#> Coleman.Liau.grade         -0.72           0.72           0.72            0.85
#> Coleman.Liau.short         -0.72           0.72           0.72            0.85
#> Dale.Chall                  1.00          -1.00          -1.00           -0.98
#> Dale.Chall.old             -1.00           1.00           1.00            0.98
#> Dale.Chall.PSK             -1.00           1.00           1.00            0.98
#> Danielson.Bryan            -0.98           0.98           0.98            1.00
#> Danielson.Bryan.2           0.19          -0.19          -0.19           -0.39
#> Dickes.Steiwer              0.41          -0.41          -0.41           -0.38
#> DRP                        -1.00           1.00           1.00            0.98
#> ELF                        -0.98           0.98           0.98            1.00
#> Farr.Jenkins.Paterson       1.00          -1.00          -1.00           -0.98
#> Flesch                      0.93          -0.93          -0.93           -0.98
#> Flesch.PSK                 -0.96           0.96           0.96            0.99
#> Flesch.Kincaid             -0.98           0.98           0.98            1.00
#>                       Danielson.Bryan.2 Dickes.Steiwer   DRP   ELF
#> ARI                               -0.32          -0.39  0.99  1.00
#> ARI.simple                        -0.32          -0.39  0.99  1.00
#> Bormuth                            0.20           0.41 -1.00 -0.99
#> Bormuth.GP                        -0.04          -0.34  0.93  0.90
#> Coleman                            0.89           0.14 -0.48 -0.62
#> Coleman.C2                         0.80           0.25 -0.68 -0.79
#> Coleman.Liau                       0.81           0.23 -0.73 -0.81
#> Coleman.Liau.grade                -0.81          -0.23  0.73  0.81
#> Coleman.Liau.short                -0.81          -0.23  0.73  0.81
#> Dale.Chall                         0.19           0.41 -1.00 -0.98
#> Dale.Chall.old                    -0.19          -0.41  1.00  0.98
#> Dale.Chall.PSK                    -0.19          -0.41  1.00  0.98
#> Danielson.Bryan                   -0.39          -0.38  0.98  1.00
#> Danielson.Bryan.2                  1.00           0.00 -0.20 -0.34
#> Dickes.Steiwer                     0.00           1.00 -0.41 -0.39
#> DRP                               -0.20          -0.41  1.00  0.99
#> ELF                               -0.34          -0.39  0.99  1.00
#> Farr.Jenkins.Paterson              0.19           0.41 -1.00 -0.99
#> Flesch                             0.53           0.37 -0.93 -0.97
#> Flesch.PSK                        -0.46          -0.38  0.96  0.99
#> Flesch.Kincaid                    -0.35          -0.40  0.99  1.00
#>                       Farr.Jenkins.Paterson Flesch Flesch.PSK Flesch.Kincaid
#> ARI                                   -0.99  -0.97       0.99           1.00
#> ARI.simple                            -0.99  -0.97       0.98           1.00
#> Bormuth                                1.00   0.93      -0.96          -0.99
#> Bormuth.GP                            -0.93  -0.82       0.86           0.90
#> Coleman                                0.48   0.76      -0.70          -0.62
#> Coleman.C2                             0.68   0.89      -0.86          -0.79
#> Coleman.Liau                           0.72   0.91      -0.88          -0.82
#> Coleman.Liau.grade                    -0.72  -0.91       0.88           0.82
#> Coleman.Liau.short                    -0.72  -0.91       0.88           0.82
#> Dale.Chall                             1.00   0.93      -0.96          -0.98
#> Dale.Chall.old                        -1.00  -0.93       0.96           0.98
#> Dale.Chall.PSK                        -1.00  -0.93       0.96           0.98
#> Danielson.Bryan                       -0.98  -0.98       0.99           1.00
#> Danielson.Bryan.2                      0.19   0.53      -0.46          -0.35
#> Dickes.Steiwer                         0.41   0.37      -0.38          -0.40
#> DRP                                   -1.00  -0.93       0.96           0.99
#> ELF                                   -0.99  -0.97       0.99           1.00
#> Farr.Jenkins.Paterson                  1.00   0.93      -0.96          -0.99
#> Flesch                                 0.93   1.00      -1.00          -0.98
#> Flesch.PSK                            -0.96  -1.00       1.00           0.99
#> Flesch.Kincaid                        -0.99  -0.98       0.99           1.00
#>                         FOG FOG.PSK FOG.NRI FORCAST FORCAST.RGL Fucks
#> ARI                    1.00    1.00    0.49    0.58        0.58  1.00
#> ARI.simple             0.99    1.00    0.49    0.58        0.58  1.00
#> Bormuth               -0.98   -1.00   -0.46   -0.48       -0.48 -1.00
#> Bormuth.GP             0.88    0.91    0.34    0.34        0.34  0.92
#> Coleman               -0.63   -0.54   -0.49   -1.00       -1.00 -0.53
#> Coleman.C2            -0.80   -0.74   -0.55   -0.96       -0.96 -0.73
#> Coleman.Liau          -0.84   -0.77   -0.55   -0.91       -0.91 -0.77
#> Coleman.Liau.grade     0.84    0.77    0.55    0.91        0.91  0.77
#> Coleman.Liau.short     0.84    0.77    0.55    0.91        0.91  0.77
#> Dale.Chall            -0.98   -1.00   -0.45   -0.48       -0.48 -1.00
#> Dale.Chall.old         0.98    1.00    0.45    0.48        0.48  1.00
#> Dale.Chall.PSK         0.98    1.00    0.45    0.48        0.48  1.00
#> Danielson.Bryan        1.00    0.99    0.51    0.64        0.64  0.99
#> Danielson.Bryan.2     -0.38   -0.27   -0.39   -0.89       -0.89 -0.26
#> Dickes.Steiwer        -0.42   -0.41    0.13   -0.14       -0.14 -0.39
#> DRP                    0.98    1.00    0.46    0.48        0.48  1.00
#> ELF                    0.99    0.99    0.50    0.62        0.62  0.99
#> Farr.Jenkins.Paterson -0.98   -1.00   -0.46   -0.48       -0.48 -1.00
#> Flesch                -0.98   -0.96   -0.54   -0.76       -0.76 -0.95
#> Flesch.PSK             0.99    0.98    0.53    0.70        0.70  0.97
#> Flesch.Kincaid         1.00    1.00    0.50    0.62        0.62  0.99
#>                       Linsear.Write   LIW   nWS nWS.2 nWS.3 nWS.4   RIX
#> ARI                            0.87  1.00  0.97  0.98  0.98  0.99  0.99
#> ARI.simple                     0.87  0.99  0.97  0.97  0.98  0.99  0.99
#> Bormuth                       -0.84 -0.98 -0.94 -0.94 -0.96 -0.98 -0.98
#> Bormuth.GP                     0.76  0.88  0.82  0.83  0.85  0.88  0.89
#> Coleman                       -0.68 -0.65 -0.75 -0.73 -0.69 -0.64 -0.63
#> Coleman.C2                    -0.79 -0.81 -0.89 -0.88 -0.84 -0.81 -0.80
#> Coleman.Liau                  -0.83 -0.85 -0.91 -0.90 -0.88 -0.84 -0.83
#> Coleman.Liau.grade             0.83  0.85  0.91  0.90  0.88  0.84  0.83
#> Coleman.Liau.short             0.83  0.85  0.91  0.90  0.88  0.84  0.83
#> Dale.Chall                    -0.84 -0.98 -0.93 -0.94 -0.96 -0.98 -0.98
#> Dale.Chall.old                 0.84  0.98  0.93  0.94  0.96  0.98  0.98
#> Dale.Chall.PSK                 0.84  0.98  0.93  0.94  0.96  0.98  0.98
#> Danielson.Bryan                0.89  1.00  0.98  0.99  0.99  1.00  0.99
#> Danielson.Bryan.2             -0.50 -0.39 -0.52 -0.50 -0.45 -0.38 -0.37
#> Dickes.Steiwer                -0.53 -0.40 -0.39 -0.40 -0.42 -0.42 -0.40
#> DRP                            0.84  0.98  0.94  0.94  0.96  0.98  0.98
#> ELF                            0.88  1.00  0.98  0.98  0.99  0.99  1.00
#> Farr.Jenkins.Paterson         -0.84 -0.98 -0.93 -0.94 -0.96 -0.98 -0.98
#> Flesch                        -0.92 -0.98 -1.00 -1.00 -0.99 -0.98 -0.98
#> Flesch.PSK                     0.91  0.99  1.00  1.00  1.00  0.99  0.99
#> Flesch.Kincaid                 0.89  1.00  0.98  0.98  0.99  1.00  0.99
#>                       Scrabble  SMOG SMOG.C SMOG.simple SMOG.de Spache
#> ARI                      -0.03  0.97   0.97        0.97    0.97   0.99
#> ARI.simple               -0.03  0.97   0.97        0.97    0.97   0.99
#> Bormuth                  -0.01 -0.94  -0.94       -0.94   -0.94  -1.00
#> Bormuth.GP               -0.03  0.82   0.82        0.82    0.82   0.93
#> Coleman                   0.36 -0.72  -0.72       -0.72   -0.72  -0.48
#> Coleman.C2                0.26 -0.87  -0.87       -0.87   -0.87  -0.68
#> Coleman.Liau              0.21 -0.90  -0.89       -0.90   -0.90  -0.72
#> Coleman.Liau.grade       -0.21  0.90   0.89        0.90    0.90   0.72
#> Coleman.Liau.short       -0.21  0.90   0.89        0.90    0.90   0.72
#> Dale.Chall               -0.02 -0.94  -0.94       -0.94   -0.94  -1.00
#> Dale.Chall.old            0.02  0.94   0.94        0.94    0.94   1.00
#> Dale.Chall.PSK            0.02  0.94   0.94        0.94    0.94   1.00
#> Danielson.Bryan          -0.06  0.98   0.98        0.98    0.98   0.98
#> Danielson.Bryan.2         0.32 -0.49  -0.48       -0.49   -0.49  -0.19
#> Dickes.Steiwer           -0.46 -0.42  -0.42       -0.42   -0.42  -0.41
#> DRP                       0.01  0.94   0.94        0.94    0.94   1.00
#> ELF                      -0.06  0.98   0.98        0.98    0.98   0.98
#> Farr.Jenkins.Paterson    -0.01 -0.94  -0.94       -0.94   -0.94  -1.00
#> Flesch                    0.12 -0.99  -0.99       -0.99   -0.99  -0.93
#> Flesch.PSK               -0.09  0.99   0.99        0.99    0.99   0.96
#> Flesch.Kincaid           -0.05  0.98   0.98        0.98    0.98   0.98
#>                       Spache.old Strain Traenkle.Bailer Traenkle.Bailer.2
#> ARI                         0.99   1.00           -1.00             -0.75
#> ARI.simple                  0.99   1.00           -1.00             -0.75
#> Bormuth                    -1.00  -1.00            1.00              0.66
#> Bormuth.GP                  0.93   0.92           -0.91             -0.51
#> Coleman                    -0.48  -0.55            0.55              0.92
#> Coleman.C2                 -0.68  -0.74            0.74              0.96
#> Coleman.Liau               -0.72  -0.77            0.78              0.99
#> Coleman.Liau.grade          0.72   0.77           -0.78             -0.99
#> Coleman.Liau.short          0.72   0.77           -0.78             -0.99
#> Dale.Chall                 -1.00  -1.00            1.00              0.66
#> Dale.Chall.old              1.00   1.00           -1.00             -0.66
#> Dale.Chall.PSK              1.00   1.00           -1.00             -0.66
#> Danielson.Bryan             0.98   0.99           -0.99             -0.80
#> Danielson.Bryan.2          -0.19  -0.27            0.28              0.86
#> Dickes.Steiwer             -0.41  -0.40            0.39              0.18
#> DRP                         1.00   1.00           -1.00             -0.66
#> ELF                         0.98   1.00           -0.99             -0.76
#> Farr.Jenkins.Paterson      -1.00  -1.00            1.00              0.66
#> Flesch                     -0.93  -0.96            0.96              0.88
#> Flesch.PSK                  0.96   0.98           -0.98             -0.84
#> Flesch.Kincaid              0.98   1.00           -1.00             -0.77
#>                       Wheeler.Smith meanSentenceLength meanWordSyllables
#> ARI                            1.00               0.99              0.68
#> ARI.simple                     1.00               0.99              0.68
#> Bormuth                       -0.99              -1.00             -0.59
#> Bormuth.GP                     0.90               0.93              0.45
#> Coleman                       -0.62              -0.48             -0.97
#> Coleman.C2                    -0.79              -0.68             -0.96
#> Coleman.Liau                  -0.81              -0.72             -0.96
#> Coleman.Liau.grade             0.81               0.72              0.96
#> Coleman.Liau.short             0.81               0.72              0.96
#> Dale.Chall                    -0.98              -1.00             -0.58
#> Dale.Chall.old                 0.98               1.00              0.58
#> Dale.Chall.PSK                 0.98               1.00              0.58
#> Danielson.Bryan                1.00               0.98              0.73
#> Danielson.Bryan.2             -0.34              -0.19             -0.88
#> Dickes.Steiwer                -0.39              -0.41             -0.22
#> DRP                            0.99               1.00              0.59
#> ELF                            1.00               0.98              0.70
#> Farr.Jenkins.Paterson         -0.99              -1.00             -0.58
#> Flesch                        -0.97              -0.93             -0.84
#> Flesch.PSK                     0.99               0.96              0.79
#> Flesch.Kincaid                 1.00               0.98              0.71
#>  [ reached getOption("max.print") -- omitted 26 rows ]

textstat_readability(data_corpus_inaugural, measure = "Flesch.Kincaid")

#>                       ARI ARI.simple    Bormuth Bormuth.GP  Coleman Coleman.C2
#> 1789-Washington 32.955509  106.69538 -11.650638 4052786964 42.95252   37.62792
#> 1793-Washington 17.773889   76.41667  -6.105827  585045935 41.81667   38.61296
#> 1797-Adams      32.829402  106.47725 -11.746553 4153690247 40.41730   35.32876
#> 1801-Jefferson  21.973359   84.81767  -7.726099 1184124874 45.51185   41.05809
#> 1805-Jefferson  25.495171   91.81200  -8.879208 1796178727 43.38102   38.70928
#> 1809-Madison    29.515543   99.84711 -10.421061 2901530452 42.87370   37.81890
#> 1813-Madison    19.647832   80.12884  -6.678707  765441426 41.86874   38.30764
#> 1817-Monroe     15.083136   71.01505  -4.985700  319250011 41.51625   39.25900
#> 1821-Monroe     18.446022   77.74325  -6.291383  640467659 43.14048   39.68360
#> 1825-Adams      21.808166   84.37763  -7.212572  963884019 39.27737   35.70021
#> 1829-Jackson    24.487926   89.75298  -8.307031 1471198826 39.88777   35.77340
#> 1833-Jackson    21.984922   84.76803  -7.431765 1054040352 39.91449   36.16385
#> 1837-VanBuren   22.354030   85.46784  -7.415802 1047626796 37.42643   33.93762
#> 1841-Harrison   21.278245   83.40493  -7.366842 1027614901 42.59608   38.60110
#> 1845-Polk       17.279742   75.36993  -5.675277  470579543 38.75698   36.18217
#> 1849-Taylor     26.734135   94.24444  -9.170219 1978195730 38.28786   34.03900
#> 1853-Pierce     17.768031   76.33302  -5.807702  504007628 41.66823   38.70130
#> 1857-Buchanan   17.253578   75.33740  -5.746745  488271691 40.20632   37.42756
#> 1861-Lincoln    14.003186   68.90856  -4.788818  283022981 45.56484   43.08875
#> 1865-Lincoln    13.181939   67.33957  -4.775341  280285380 53.19337   49.94729
#> 1869-Grant      14.562566   70.03627  -5.042459  329959777 44.12825   41.54072
#>                 Coleman.Liau Coleman.Liau.grade Coleman.Liau.short Dale.Chall
#> 1789-Washington     37.52530          12.781868          12.783201  -73.93000
#> 1793-Washington     43.30799          11.197388          11.198519  -54.28750
#> 1797-Adams          39.25280          12.308527          12.309849  -74.28351
#> 1801-Jefferson      42.77284          11.344020          11.345225  -60.11463
#> 1805-Jefferson      39.93897          12.120513          12.121773  -64.21200
#> 1809-Madison        39.33411          12.286246          12.287545  -69.67286
#> 1813-Madison        41.22654          11.767712          11.768885  -56.32091
#> 1817-Monroe         42.94873          11.295827          11.296886  -50.26298
#> 1821-Monroe         42.31716          11.468880          11.470027  -54.94140
#> 1825-Adams          37.35160          12.829464          12.830686  -58.18973
#> 1829-Jackson        37.81340          12.702927          12.704184  -62.13280
#> 1833-Jackson        39.15646          12.334924          12.336143  -59.00448
#> 1837-VanBuren       37.20257          12.870298          12.871530  -58.91958
#> 1841-Harrison       41.74302          11.626196          11.627395  -58.80700
#> 1845-Polk           40.55599          11.951445          11.952569  -52.70118
#> 1849-Taylor         37.54864          12.775472          12.776755  -65.21773
#> 1853-Pierce         39.79472          12.160037          12.161173  -53.16625
#> 1857-Buchanan       41.52530          11.685851          11.686972  -52.97146
#> 1861-Lincoln        45.81604          10.510172          10.511195  -49.59933
#> 1865-Lincoln        49.57042           9.481456           9.482454  -49.60346
#> 1869-Grant          46.08627          10.436128          10.437171  -50.50975
#>                 Dale.Chall.old Dale.Chall.PSK Danielson.Bryan Danielson.Bryan.2
#> 1789-Washington       22.51248       18.52536       10.484647          85.76939
#> 1793-Washington       21.10050       16.82870        7.448070          84.66330
#> 1797-Adams            22.53789       18.55589       10.349469          86.57501
#> 1801-Jefferson        21.51938       17.33203        8.181849          85.82507
#> 1805-Jefferson        21.81391       17.68595        8.958105          85.26919
#> 1809-Madison          22.20646       18.15764        9.733724          85.87015
#> 1813-Madison          21.24667       17.00434        7.835219          84.43901
#> 1817-Monroe           20.81120       16.48107        6.952016          84.00843
#> 1821-Monroe           21.14750       16.88518        7.571643          84.69805
#> 1825-Adams            21.38101       17.16576        8.394643          83.07319
#> 1829-Jackson          21.66445       17.50635        8.879340          83.95698
#> 1833-Jackson          21.43958       17.23614        8.351491          84.00320
#> 1837-VanBuren         21.43347       17.22880        8.503323          83.14546
#> 1841-Harrison         21.42538       17.21908        8.102936          85.15989
#> 1845-Polk             20.98647       16.69168        7.430119          83.45083
#> 1849-Taylor           21.88621       17.77282        9.314201          84.36409
#> 1853-Pierce           21.01990       16.73185        7.547706          83.19705
#> 1857-Buchanan         21.00590       16.71502        7.389945          83.94017
#> 1861-Lincoln          20.76350       16.42375        6.664196          85.17215
#> 1865-Lincoln          20.76379       16.42411        6.397151          86.84168
#> 1869-Grant            20.82894       16.50239        6.752678          85.47130
#>                 Dickes.Steiwer       DRP       ELF Farr.Jenkins.Paterson
#> 1789-Washington     -1204.6860 1265.0638 22.913043             -93.65864
#> 1793-Washington     -3656.7248  710.5827 12.750000             -64.77832
#> 1797-Adams          -1084.7757 1274.6553 24.378378             -94.21009
#> 1801-Jefferson       -877.2216  872.6099 14.731707             -73.30431
#> 1805-Jefferson       -915.9671  987.9208 17.600000             -79.35801
#> 1809-Madison        -1185.8831 1142.1061 20.714286             -87.39730
#> 1813-Madison         -929.4923  767.8707 13.848485             -67.76885
#> 1817-Monroe          -588.6639  598.5700 10.611570             -58.86191
#> 1821-Monroe          -642.9549  729.1383 12.651163             -65.72380
#> 1825-Adams           -769.5543  821.2572 15.662162             -70.55003
#> 1829-Jackson        -1072.4127  930.7031 17.720000             -76.34278
#> 1833-Jackson         -993.5819  843.1765 15.931034             -71.74064
#> 1837-VanBuren        -747.5182  841.5802 16.663158             -71.64659
#> 1841-Harrison        -670.1251  836.6842 14.980952             -71.41691
#> 1845-Polk            -603.6194  667.5277 12.620915             -62.48274
#> 1849-Taylor         -1136.7566 1017.0219 20.090909             -80.90058
#> 1853-Pierce          -651.8634  680.7702 12.163462             -63.13078
#> 1857-Buchanan        -664.7949  674.6745 12.426966             -62.86236
#> 1861-Lincoln         -557.3967  578.8818  9.370370             -57.83550
#> 1865-Lincoln        -1018.8642  577.5341  7.807692             -57.74701
#> 1869-Grant           -827.3953  604.2459 10.175000             -59.19254
#>                    Flesch Flesch.PSK Flesch.Kincaid       FOG   FOG.PSK
#> 1789-Washington  2.093153  10.252742      28.423909 32.909319 18.713936
#> 1793-Washington 32.205417   7.972480      17.151759 22.092593 11.194674
#> 1797-Adams       1.082980  10.318964      28.692087 32.847170 18.786085
#> 1801-Jefferson  31.874466   8.186297      19.295924 23.028338 12.866143
#> 1805-Jefferson  23.658614   8.765996      21.917088 25.587590 14.509155
#> 1809-Madison    11.654058   9.595328      25.557611 29.589821 16.844384
#> 1813-Madison    31.125809   8.098945      18.034452 21.780357 11.620805
#> 1817-Monroe     41.230501   7.351708      14.443950 18.011229  9.195239
#> 1821-Monroe     36.854126   7.744457      16.738786 20.617229 10.993468
#> 1825-Adams      23.662588   8.563201      19.748273 23.718266 12.559762
#> 1829-Jackson    18.463200   8.975477      21.893148 25.955801 14.085469
#> 1833-Jackson    24.256653   8.558658      19.958757 23.813072 12.784604
#> 1837-VanBuren   21.852460   8.685105      20.263524 24.312214 12.886393
#> 1841-Harrison   28.069387   8.346956      19.355855 23.441281 12.644298
#> 1845-Polk       33.472729   7.850958      16.403854 20.153127 10.325207
#> 1849-Taylor     13.122225   9.366501      23.748805 27.755704 15.291266
#> 1853-Pierce     34.452010   7.813935      16.434709 20.153203 10.440256
#> 1857-Buchanan   36.270979   7.709554      16.110868 19.624870 10.262107
#> 1861-Lincoln    45.217073   7.114976      13.648965 17.311489  8.858958
#> 1865-Lincoln    58.663618   6.391926      11.774929 14.893032  8.265039
#> 1869-Grant      44.989986   7.157815      14.008426 17.357745  9.095519
#>                   FOG.NRI   FORCAST FORCAST.RGL     Fucks Linsear.Write
#> 1789-Washington  922.1172 10.534591   10.018050 307.47826    8.00000000
#> 1793-Washington   46.0875 10.666667   10.163333 160.00000    6.75000000
#> 1797-Adams      1503.3862 10.829384   10.342322 304.91892    6.02702703
#> 1801-Jefferson   680.7951 10.236994    9.690694 199.82927    2.24390244
#> 1805-Jefferson   991.2500 10.484765    9.963241 233.60000    2.20000000
#> 1809-Madison     655.9386 10.543755   10.028131 272.76190    4.61904762
#> 1813-Madison     441.0655 10.660611   10.156672 177.09091    2.69696970
#> 1817-Monroe      924.9374 10.701599   10.201758 133.68595    1.72727273
#> 1821-Monroe     1497.0926 10.512735    9.994008 165.95349    2.80620155
#> 1825-Adams      1199.3797 10.961934   10.488128 196.90541    4.06756757
#> 1829-Jackson     527.9832 10.890957   10.410053 223.76000    4.76000000
#> 1833-Jackson     488.1726 10.887850   10.406636 199.24138    3.51724138
#> 1837-VanBuren   1637.4602 11.177159   10.724875 202.33684    4.51578947
#> 1841-Harrison   3401.6335 10.576037   10.063641 193.01429    4.13333333
#> 1845-Polk       1550.1341 11.022444   10.554688 153.47712    2.68627451
#> 1849-Taylor      564.0843 11.076994   10.614693 246.04545    4.81818182
#> 1853-Pierce     1081.1125 10.683927   10.182320 157.79808    2.50961538
#> 1857-Buchanan    891.2100 10.853917   10.369308 153.88764    1.74157303
#> 1861-Lincoln     936.9577 10.230833    9.683916 125.65185    1.77777778
#> 1865-Lincoln     152.8548  9.343795    8.708174 120.96154   -0.46153846
#> 1869-Grant       298.4977 10.397878    9.867666 131.20000    0.87500000
#>                      LIW       nWS     nWS.2     nWS.3     nWS.4       RIX
#> 1789-Washington 88.42284 15.695630 16.316636 16.680578 20.335279 16.304348
#> 1793-Washington 61.15741 11.309407 11.887454 11.679938 13.165519  9.250000
#> 1797-Adams      90.13171 15.964846 16.531852 16.580336 20.288137 17.189189
#> 1801-Jefferson  64.96969 11.172793 11.792965 11.479597 13.733123  9.609756
#> 1805-Jefferson  73.38726 12.539994 13.128956 12.747101 15.436513 12.155556
#> 1809-Madison    82.72561 14.366476 14.970571 14.874422 18.112398 14.952381
#> 1813-Madison    62.29565 11.062005 11.616453 11.136860 12.925391  9.393939
#> 1817-Monroe     53.13938  9.434543  9.964067  9.273761 10.417030  7.041322
#> 1821-Monroe     59.25085 10.377396 10.947202 10.486798 12.145079  8.519380
#> 1825-Adams      68.14340 12.235631 12.756644 12.285814 14.230963 11.324324
#> 1829-Jackson    73.48879 13.303452 13.853641 13.338664 15.715624 12.800000
#> 1833-Jackson    68.11382 12.130759 12.656241 12.231113 14.285609 11.172414
#> 1837-VanBuren   69.46940 12.686306 13.172985 12.613871 14.629103 11.736842
#> 1841-Harrison   65.96466 11.682934 12.259313 11.985989 14.033079 10.342857
#> 1845-Polk       58.19662 10.623377 11.109897 10.486515 11.855276  8.411765
#> 1849-Taylor     79.28843 14.045748 14.557049 14.198919 16.911013 14.727273
#> 1853-Pierce     58.61408 10.509287 11.064102 10.415260 11.849397  8.509615
#> 1857-Buchanan   56.96620 10.240224 10.746825 10.053765 11.489445  8.000000
#> 1861-Lincoln    50.23118  8.584970  9.185073  8.857187  9.945472  6.274074
#> 1865-Lincoln    46.21974  6.812770  7.538727  7.065083  8.286358  5.192308
#> 1869-Grant      51.70559  8.985312  9.563699  8.751855  9.965593  6.625000
#>                 Scrabble      SMOG    SMOG.C SMOG.simple   SMOG.de   Spache
#> 1789-Washington 1.706789 23.309137 22.329086   22.348070 17.348070 16.38730
#> 1793-Washington 1.721875 18.511140 17.775051   17.747881 12.747881 12.94275
#> 1797-Adams      1.624590 23.051898 22.084446   22.101436 17.101436 16.44930
#> 1801-Jefferson  1.678183 17.680133 16.988914   16.951134 11.951134 13.96461
#> 1805-Jefferson  1.663654 18.901063 18.144259   18.121728 13.121728 14.68313
#> 1809-Madison    1.637395 21.237351 20.360094   20.361698 15.361698 15.64076
#> 1813-Madison    1.646475 17.710768 17.017876   16.980506 11.980506 13.29933
#> 1817-Monroe     1.665801 15.614894 15.040537   14.971039  9.971039 12.23700
#> 1821-Monroe     1.686671 16.940438 16.290101   16.241935 11.241935 13.05742
#> 1825-Adams      1.662800 19.122584 18.354099   18.334117 13.334117 13.62705
#> 1829-Jackson    1.662853 20.190999 19.366988   19.358484 14.358484 14.31852
#> 1833-Jackson    1.656629 18.970675 18.210194   18.188471 13.188471 13.76993
#> 1837-VanBuren   1.656920 19.508902 18.720191   18.704508 13.704508 13.75504
#> 1841-Harrison   1.683668 18.644429 17.901236   17.875675 12.875675 13.73530
#> 1845-Polk       1.650196 17.068960 16.411449   16.365158 11.365158 12.66457
#> 1849-Taylor     1.668391 21.029407 20.162651   20.162327 15.162327 14.85950
#> 1853-Pierce     1.665082 16.964547 16.312861   16.265050 11.265050 12.74613
#> 1857-Buchanan   1.666204 16.506134 15.880274   15.825536 10.825536 12.71197
#> 1861-Lincoln    1.650617 15.112258 14.567813   14.489125  9.489125 12.12062
#> 1865-Lincoln    1.680357 12.635684 12.251292   12.114654  7.114654 12.12135
#> 1869-Grant      1.670097 14.940816 14.406736   14.324752  9.324752 12.28028
#>                 Spache.old    Strain Traenkle.Bailer Traenkle.Bailer.2
#> 1789-Washington   18.21165 31.239130       -958.4147         -289.0435
#> 1793-Washington   14.19775 16.800000       -592.7418         -265.6644
#> 1797-Adams        18.28389 31.605405       -959.2671         -286.1880
#> 1801-Jefferson    15.38851 19.770732       -694.3015         -266.3682
#> 1805-Jefferson    16.22580 22.926667       -775.9126         -276.4466
#> 1809-Madison      17.34171 27.485714       -879.7402         -286.6070
#> 1813-Madison      14.61327 18.018182       -637.3946         -278.9883
#> 1817-Monroe       13.37535 13.589256       -524.4979         -271.4463
#> 1821-Monroe       14.33137 16.581395       -608.4767         -272.3736
#> 1825-Adams        14.99516 20.006757       -684.0920         -295.1908
#> 1829-Jackson      15.80092 22.812000       -750.6621         -291.9814
#> 1833-Jackson      15.16166 20.348276       -690.6934         -286.6761
#> 1837-VanBuren     15.14431 20.649474       -695.0190         -292.8928
#> 1841-Harrison     15.12130 19.701429       -678.8259         -275.4081
#> 1845-Polk         13.87359 15.774510       -575.4832         -281.8499
#> 1849-Taylor       16.43132 25.213636       -804.1209         -290.1119
#> 1853-Pierce       13.96862 15.923077       -586.5154         -285.9280
#> 1857-Buchanan     13.92882 15.610112       -577.3427         -278.1131
#> 1861-Lincoln      13.23973 12.833333       -501.4761         -258.1184
#> 1865-Lincoln      13.24058 11.550000       -487.0985         -240.7621
#> 1869-Grant        13.42577 13.350000       -517.4437         -257.1858
#>                 Wheeler.Smith meanSentenceLength meanWordSyllables
#> 1789-Washington     229.13043           62.21739          1.673655
#> 1793-Washington     127.50000           33.75000          1.659259
#> 1797-Adams          243.78378           62.72973          1.679449
#> 1801-Jefferson      147.31707           42.19512          1.561850
#> 1805-Jefferson      176.00000           48.13333          1.587719
#> 1809-Madison        207.14286           56.04762          1.634664
#> 1813-Madison        138.48485           36.69697          1.636664
#> 1817-Monroe         106.11570           27.91736          1.622558
#> 1821-Monroe         126.51163           34.69767          1.592940
#> 1825-Adams          156.62162           39.40541          1.692387
#> 1829-Jackson        177.20000           45.12000          1.685284
#> 1833-Jackson        159.31034           40.58621          1.671198
#> 1837-VanBuren       166.63158           40.46316          1.701093
#> 1841-Harrison       149.80952           40.30000          1.629564
#> 1845-Polk           126.20915           31.45098          1.671862
#> 1849-Taylor         200.90909           49.59091          1.694775
#> 1853-Pierce         121.63462           32.12500          1.652200
#> 1857-Buchanan       124.26966           31.84270          1.634086
#> 1861-Lincoln         93.70370           26.95556          1.586974
#> 1865-Lincoln         78.07692           26.96154          1.427960
#> 1869-Grant          101.75000           28.27500          1.573828
#>  [ reached getOption("max.print") -- omitted 36 rows ]

inaugReadability <- textstat_readability(data_corpus_inaugural, "all")
round(cor(inaugReadability), 2)

#>                         ARI ARI.simple Bormuth Bormuth.GP Coleman Coleman.C2
#> ARI                    1.00       1.00   -0.99       0.90   -0.58      -0.77
#> ARI.simple             1.00       1.00   -0.99       0.90   -0.58      -0.76
#> Bormuth               -0.99      -0.99    1.00      -0.93    0.48       0.68
#> Bormuth.GP             0.90       0.90   -0.93       1.00   -0.34      -0.51
#> Coleman               -0.58      -0.58    0.48      -0.34    1.00       0.96
#> Coleman.C2            -0.77      -0.76    0.68      -0.51    0.96       1.00
#> Coleman.Liau          -0.81      -0.80    0.73      -0.56    0.91       0.97
#> Coleman.Liau.grade     0.81       0.80   -0.73       0.56   -0.91      -0.97
#> Coleman.Liau.short     0.81       0.80   -0.73       0.56   -0.91      -0.97
#> Dale.Chall            -0.99      -0.99    1.00      -0.93    0.48       0.68
#> Dale.Chall.old         0.99       0.99   -1.00       0.93   -0.48      -0.68
#> Dale.Chall.PSK         0.99       0.99   -1.00       0.93   -0.48      -0.68
#> Danielson.Bryan        1.00       1.00   -0.98       0.88   -0.64      -0.80
#> Danielson.Bryan.2     -0.32      -0.32    0.20      -0.04    0.89       0.80
#> Dickes.Steiwer        -0.39      -0.39    0.41      -0.34    0.14       0.25
#> DRP                    0.99       0.99   -1.00       0.93   -0.48      -0.68
#> ELF                    1.00       1.00   -0.99       0.90   -0.62      -0.79
#> Farr.Jenkins.Paterson -0.99      -0.99    1.00      -0.93    0.48       0.68
#> Flesch                -0.97      -0.97    0.93      -0.82    0.76       0.89
#> Flesch.PSK             0.99       0.98   -0.96       0.86   -0.70      -0.86
#> Flesch.Kincaid         1.00       1.00   -0.99       0.90   -0.62      -0.79
#>                       Coleman.Liau Coleman.Liau.grade Coleman.Liau.short
#> ARI                          -0.81               0.81               0.81
#> ARI.simple                   -0.80               0.80               0.80
#> Bormuth                       0.73              -0.73              -0.73
#> Bormuth.GP                   -0.56               0.56               0.56
#> Coleman                       0.91              -0.91              -0.91
#> Coleman.C2                    0.97              -0.97              -0.97
#> Coleman.Liau                  1.00              -1.00              -1.00
#> Coleman.Liau.grade           -1.00               1.00               1.00
#> Coleman.Liau.short           -1.00               1.00               1.00
#> Dale.Chall                    0.72              -0.72              -0.72
#> Dale.Chall.old               -0.72               0.72               0.72
#> Dale.Chall.PSK               -0.72               0.72               0.72
#> Danielson.Bryan              -0.85               0.85               0.85
#> Danielson.Bryan.2             0.81              -0.81              -0.81
#> Dickes.Steiwer                0.23              -0.23              -0.23
#> DRP                          -0.73               0.73               0.73
#> ELF                          -0.81               0.81               0.81
#> Farr.Jenkins.Paterson         0.72              -0.72              -0.72
#> Flesch                        0.91              -0.91              -0.91
#> Flesch.PSK                   -0.88               0.88               0.88
#> Flesch.Kincaid               -0.82               0.82               0.82
#>                       Dale.Chall Dale.Chall.old Dale.Chall.PSK Danielson.Bryan
#> ARI                        -0.99           0.99           0.99            1.00
#> ARI.simple                 -0.99           0.99           0.99            1.00
#> Bormuth                     1.00          -1.00          -1.00           -0.98
#> Bormuth.GP                 -0.93           0.93           0.93            0.88
#> Coleman                     0.48          -0.48          -0.48           -0.64
#> Coleman.C2                  0.68          -0.68          -0.68           -0.80
#> Coleman.Liau                0.72          -0.72          -0.72           -0.85
#> Coleman.Liau.grade         -0.72           0.72           0.72            0.85
#> Coleman.Liau.short         -0.72           0.72           0.72            0.85
#> Dale.Chall                  1.00          -1.00          -1.00           -0.98
#> Dale.Chall.old             -1.00           1.00           1.00            0.98
#> Dale.Chall.PSK             -1.00           1.00           1.00            0.98
#> Danielson.Bryan            -0.98           0.98           0.98            1.00
#> Danielson.Bryan.2           0.19          -0.19          -0.19           -0.39
#> Dickes.Steiwer              0.41          -0.41          -0.41           -0.38
#> DRP                        -1.00           1.00           1.00            0.98
#> ELF                        -0.98           0.98           0.98            1.00
#> Farr.Jenkins.Paterson       1.00          -1.00          -1.00           -0.98
#> Flesch                      0.93          -0.93          -0.93           -0.98
#> Flesch.PSK                 -0.96           0.96           0.96            0.99
#> Flesch.Kincaid             -0.98           0.98           0.98            1.00
#>                       Danielson.Bryan.2 Dickes.Steiwer   DRP   ELF
#> ARI                               -0.32          -0.39  0.99  1.00
#> ARI.simple                        -0.32          -0.39  0.99  1.00
#> Bormuth                            0.20           0.41 -1.00 -0.99
#> Bormuth.GP                        -0.04          -0.34  0.93  0.90
#> Coleman                            0.89           0.14 -0.48 -0.62
#> Coleman.C2                         0.80           0.25 -0.68 -0.79
#> Coleman.Liau                       0.81           0.23 -0.73 -0.81
#> Coleman.Liau.grade                -0.81          -0.23  0.73  0.81
#> Coleman.Liau.short                -0.81          -0.23  0.73  0.81
#> Dale.Chall                         0.19           0.41 -1.00 -0.98
#> Dale.Chall.old                    -0.19          -0.41  1.00  0.98
#> Dale.Chall.PSK                    -0.19          -0.41  1.00  0.98
#> Danielson.Bryan                   -0.39          -0.38  0.98  1.00
#> Danielson.Bryan.2                  1.00           0.00 -0.20 -0.34
#> Dickes.Steiwer                     0.00           1.00 -0.41 -0.39
#> DRP                               -0.20          -0.41  1.00  0.99
#> ELF                               -0.34          -0.39  0.99  1.00
#> Farr.Jenkins.Paterson              0.19           0.41 -1.00 -0.99
#> Flesch                             0.53           0.37 -0.93 -0.97
#> Flesch.PSK                        -0.46          -0.38  0.96  0.99
#> Flesch.Kincaid                    -0.35          -0.40  0.99  1.00
#>                       Farr.Jenkins.Paterson Flesch Flesch.PSK Flesch.Kincaid
#> ARI                                   -0.99  -0.97       0.99           1.00
#> ARI.simple                            -0.99  -0.97       0.98           1.00
#> Bormuth                                1.00   0.93      -0.96          -0.99
#> Bormuth.GP                            -0.93  -0.82       0.86           0.90
#> Coleman                                0.48   0.76      -0.70          -0.62
#> Coleman.C2                             0.68   0.89      -0.86          -0.79
#> Coleman.Liau                           0.72   0.91      -0.88          -0.82
#> Coleman.Liau.grade                    -0.72  -0.91       0.88           0.82
#> Coleman.Liau.short                    -0.72  -0.91       0.88           0.82
#> Dale.Chall                             1.00   0.93      -0.96          -0.98
#> Dale.Chall.old                        -1.00  -0.93       0.96           0.98
#> Dale.Chall.PSK                        -1.00  -0.93       0.96           0.98
#> Danielson.Bryan                       -0.98  -0.98       0.99           1.00
#> Danielson.Bryan.2                      0.19   0.53      -0.46          -0.35
#> Dickes.Steiwer                         0.41   0.37      -0.38          -0.40
#> DRP                                   -1.00  -0.93       0.96           0.99
#> ELF                                   -0.99  -0.97       0.99           1.00
#> Farr.Jenkins.Paterson                  1.00   0.93      -0.96          -0.99
#> Flesch                                 0.93   1.00      -1.00          -0.98
#> Flesch.PSK                            -0.96  -1.00       1.00           0.99
#> Flesch.Kincaid                        -0.99  -0.98       0.99           1.00
#>                         FOG FOG.PSK FOG.NRI FORCAST FORCAST.RGL Fucks
#> ARI                    1.00    1.00    0.49    0.58        0.58  1.00
#> ARI.simple             0.99    1.00    0.49    0.58        0.58  1.00
#> Bormuth               -0.98   -1.00   -0.46   -0.48       -0.48 -1.00
#> Bormuth.GP             0.88    0.91    0.34    0.34        0.34  0.92
#> Coleman               -0.63   -0.54   -0.49   -1.00       -1.00 -0.53
#> Coleman.C2            -0.80   -0.74   -0.55   -0.96       -0.96 -0.73
#> Coleman.Liau          -0.84   -0.77   -0.55   -0.91       -0.91 -0.77
#> Coleman.Liau.grade     0.84    0.77    0.55    0.91        0.91  0.77
#> Coleman.Liau.short     0.84    0.77    0.55    0.91        0.91  0.77
#> Dale.Chall            -0.98   -1.00   -0.45   -0.48       -0.48 -1.00
#> Dale.Chall.old         0.98    1.00    0.45    0.48        0.48  1.00
#> Dale.Chall.PSK         0.98    1.00    0.45    0.48        0.48  1.00
#> Danielson.Bryan        1.00    0.99    0.51    0.64        0.64  0.99
#> Danielson.Bryan.2     -0.38   -0.27   -0.39   -0.89       -0.89 -0.26
#> Dickes.Steiwer        -0.42   -0.41    0.13   -0.14       -0.14 -0.39
#> DRP                    0.98    1.00    0.46    0.48        0.48  1.00
#> ELF                    0.99    0.99    0.50    0.62        0.62  0.99
#> Farr.Jenkins.Paterson -0.98   -1.00   -0.46   -0.48       -0.48 -1.00
#> Flesch                -0.98   -0.96   -0.54   -0.76       -0.76 -0.95
#> Flesch.PSK             0.99    0.98    0.53    0.70        0.70  0.97
#> Flesch.Kincaid         1.00    1.00    0.50    0.62        0.62  0.99
#>                       Linsear.Write   LIW   nWS nWS.2 nWS.3 nWS.4   RIX
#> ARI                            0.87  1.00  0.97  0.98  0.98  0.99  0.99
#> ARI.simple                     0.87  0.99  0.97  0.97  0.98  0.99  0.99
#> Bormuth                       -0.84 -0.98 -0.94 -0.94 -0.96 -0.98 -0.98
#> Bormuth.GP                     0.76  0.88  0.82  0.83  0.85  0.88  0.89
#> Coleman                       -0.68 -0.65 -0.75 -0.73 -0.69 -0.64 -0.63
#> Coleman.C2                    -0.79 -0.81 -0.89 -0.88 -0.84 -0.81 -0.80
#> Coleman.Liau                  -0.83 -0.85 -0.91 -0.90 -0.88 -0.84 -0.83
#> Coleman.Liau.grade             0.83  0.85  0.91  0.90  0.88  0.84  0.83
#> Coleman.Liau.short             0.83  0.85  0.91  0.90  0.88  0.84  0.83
#> Dale.Chall                    -0.84 -0.98 -0.93 -0.94 -0.96 -0.98 -0.98
#> Dale.Chall.old                 0.84  0.98  0.93  0.94  0.96  0.98  0.98
#> Dale.Chall.PSK                 0.84  0.98  0.93  0.94  0.96  0.98  0.98
#> Danielson.Bryan                0.89  1.00  0.98  0.99  0.99  1.00  0.99
#> Danielson.Bryan.2             -0.50 -0.39 -0.52 -0.50 -0.45 -0.38 -0.37
#> Dickes.Steiwer                -0.53 -0.40 -0.39 -0.40 -0.42 -0.42 -0.40
#> DRP                            0.84  0.98  0.94  0.94  0.96  0.98  0.98
#> ELF                            0.88  1.00  0.98  0.98  0.99  0.99  1.00
#> Farr.Jenkins.Paterson         -0.84 -0.98 -0.93 -0.94 -0.96 -0.98 -0.98
#> Flesch                        -0.92 -0.98 -1.00 -1.00 -0.99 -0.98 -0.98
#> Flesch.PSK                     0.91  0.99  1.00  1.00  1.00  0.99  0.99
#> Flesch.Kincaid                 0.89  1.00  0.98  0.98  0.99  1.00  0.99
#>                       Scrabble  SMOG SMOG.C SMOG.simple SMOG.de Spache
#> ARI                      -0.03  0.97   0.97        0.97    0.97   0.99
#> ARI.simple               -0.03  0.97   0.97        0.97    0.97   0.99
#> Bormuth                  -0.01 -0.94  -0.94       -0.94   -0.94  -1.00
#> Bormuth.GP               -0.03  0.82   0.82        0.82    0.82   0.93
#> Coleman                   0.36 -0.72  -0.72       -0.72   -0.72  -0.48
#> Coleman.C2                0.26 -0.87  -0.87       -0.87   -0.87  -0.68
#> Coleman.Liau              0.21 -0.90  -0.89       -0.90   -0.90  -0.72
#> Coleman.Liau.grade       -0.21  0.90   0.89        0.90    0.90   0.72
#> Coleman.Liau.short       -0.21  0.90   0.89        0.90    0.90   0.72
#> Dale.Chall               -0.02 -0.94  -0.94       -0.94   -0.94  -1.00
#> Dale.Chall.old            0.02  0.94   0.94        0.94    0.94   1.00
#> Dale.Chall.PSK            0.02  0.94   0.94        0.94    0.94   1.00
#> Danielson.Bryan          -0.06  0.98   0.98        0.98    0.98   0.98
#> Danielson.Bryan.2         0.32 -0.49  -0.48       -0.49   -0.49  -0.19
#> Dickes.Steiwer           -0.46 -0.42  -0.42       -0.42   -0.42  -0.41
#> DRP                       0.01  0.94   0.94        0.94    0.94   1.00
#> ELF                      -0.06  0.98   0.98        0.98    0.98   0.98
#> Farr.Jenkins.Paterson    -0.01 -0.94  -0.94       -0.94   -0.94  -1.00
#> Flesch                    0.12 -0.99  -0.99       -0.99   -0.99  -0.93
#> Flesch.PSK               -0.09  0.99   0.99        0.99    0.99   0.96
#> Flesch.Kincaid           -0.05  0.98   0.98        0.98    0.98   0.98
#>                       Spache.old Strain Traenkle.Bailer Traenkle.Bailer.2
#> ARI                         0.99   1.00           -1.00             -0.75
#> ARI.simple                  0.99   1.00           -1.00             -0.75
#> Bormuth                    -1.00  -1.00            1.00              0.66
#> Bormuth.GP                  0.93   0.92           -0.91             -0.51
#> Coleman                    -0.48  -0.55            0.55              0.92
#> Coleman.C2                 -0.68  -0.74            0.74              0.96
#> Coleman.Liau               -0.72  -0.77            0.78              0.99
#> Coleman.Liau.grade          0.72   0.77           -0.78             -0.99
#> Coleman.Liau.short          0.72   0.77           -0.78             -0.99
#> Dale.Chall                 -1.00  -1.00            1.00              0.66
#> Dale.Chall.old              1.00   1.00           -1.00             -0.66
#> Dale.Chall.PSK              1.00   1.00           -1.00             -0.66
#> Danielson.Bryan             0.98   0.99           -0.99             -0.80
#> Danielson.Bryan.2          -0.19  -0.27            0.28              0.86
#> Dickes.Steiwer             -0.41  -0.40            0.39              0.18
#> DRP                         1.00   1.00           -1.00             -0.66
#> ELF                         0.98   1.00           -0.99             -0.76
#> Farr.Jenkins.Paterson      -1.00  -1.00            1.00              0.66
#> Flesch                     -0.93  -0.96            0.96              0.88
#> Flesch.PSK                  0.96   0.98           -0.98             -0.84
#> Flesch.Kincaid              0.98   1.00           -1.00             -0.77
#>                       Wheeler.Smith meanSentenceLength meanWordSyllables
#> ARI                            1.00               0.99              0.68
#> ARI.simple                     1.00               0.99              0.68
#> Bormuth                       -0.99              -1.00             -0.59
#> Bormuth.GP                     0.90               0.93              0.45
#> Coleman                       -0.62              -0.48             -0.97
#> Coleman.C2                    -0.79              -0.68             -0.96
#> Coleman.Liau                  -0.81              -0.72             -0.96
#> Coleman.Liau.grade             0.81               0.72              0.96
#> Coleman.Liau.short             0.81               0.72              0.96
#> Dale.Chall                    -0.98              -1.00             -0.58
#> Dale.Chall.old                 0.98               1.00              0.58
#> Dale.Chall.PSK                 0.98               1.00              0.58
#> Danielson.Bryan                1.00               0.98              0.73
#> Danielson.Bryan.2             -0.34              -0.19             -0.88
#> Dickes.Steiwer                -0.39              -0.41             -0.22
#> DRP                            0.99               1.00              0.59
#> ELF                            1.00               0.98              0.70
#> Farr.Jenkins.Paterson         -0.99              -1.00             -0.58
#> Flesch                        -0.97              -0.93             -0.84
#> Flesch.PSK                     0.99               0.96              0.79
#> Flesch.Kincaid                 1.00               0.98              0.71
#>  [ reached getOption("max.print") -- omitted 26 rows ]