calculate readability

Calculate the readability of text(s) using one of a variety of computed indexes.

textstat_readability(x, measure = c("all", "ARI", "ARI.simple", "Bormuth",
  "Bormuth.GP", "Coleman", "Coleman.C2", "Coleman.Liau", "Coleman.Liau.grade",
  "Coleman.Liau.short", "Dale.Chall", "Dale.Chall.old", "Dale.Chall.PSK",
  "Danielson.Bryan", "Danielson.Bryan.2", "Dickes.Steiwer", "DRP", "ELF",
  "Farr.Jenkins.Paterson", "Flesch", "Flesch.PSK", "Flesch.Kincaid", "FOG",
  "FOG.PSK", "FOG.NRI", "FORCAST", "FORCAST.RGL", "Fucks", "Linsear.Write",
  "LIW", "nWS", "nWS.2", "nWS.3", "nWS.4", "RIX", "Scrabble", "SMOG", "SMOG.C",
  "SMOG.simple",      "SMOG.de", "Spache", "Spache.old", "Strain",
  "Traenkle.Bailer", "Traenkle.Bailer.2", "Wheeler.Smith", "meanSentenceLength",
  "meanWordSyllables"), removeHyphens = TRUE, drop = TRUE, ...)

Arguments

x

a character or corpus object containing the texts

measure

character vector defining the readability measure to calculate

removeHyphens

if TRUE, treat constituent words in hyphenated as separate terms, for purposes of computing word lengths, e.g. "decision-making" as two terms of lengths 8 and 6 characters respectively, rather than as a single word of 15 characters

drop

if TRUE, the result is returned as a numeric vector if only a single measure is requested; otherwise, a data.frame is returned with each column consisting of a requested measure.

...

not used

Value

a data.frame object consisting of the documents as rows, and the readability statistics as columns

Examples

txt <- c("Readability zero one.  Ten, Eleven.", "The cat in a dilapidated tophat.")
textstat_readability(txt, "Flesch.Kincaid")

#>     text1     text2 
#> 13.705000  8.383333 

textstat_readability(txt, "Flesch.Kincaid", drop = FALSE)

#>       Flesch.Kincaid
#> text1      13.705000
#> text2       8.383333

textstat_readability(txt, c("FOG", "FOG.PSK", "FOG.NRI"))

#>             FOG  FOG.PSK FOG.NRI
#> text1 17.000000 4.608659   -1.30
#> text2  9.066667 3.254382   -1.11

inaugReadability <- textstat_readability(data_corpus_inaugural, "all")
round(cor(inaugReadability), 2)

#>                         ARI ARI.simple Bormuth Bormuth.GP Coleman Coleman.C2
#> ARI                    1.00       1.00   -0.99       0.90      NA         NA
#> ARI.simple             1.00       1.00   -0.99       0.90      NA         NA
#> Bormuth               -0.99      -0.99    1.00      -0.93      NA         NA
#> Bormuth.GP             0.90       0.90   -0.93       1.00      NA         NA
#> Coleman                  NA         NA      NA         NA       1         NA
#> Coleman.C2               NA         NA      NA         NA      NA          1
#> Coleman.Liau          -0.81      -0.80    0.73      -0.56      NA         NA
#> Coleman.Liau.grade     0.81       0.80   -0.73       0.56      NA         NA
#> Coleman.Liau.short     0.81       0.80   -0.73       0.56      NA         NA
#> Dale.Chall            -0.99      -0.99    1.00      -0.93      NA         NA
#> Dale.Chall.old         0.99       0.99   -1.00       0.93      NA         NA
#> Dale.Chall.PSK         0.99       0.99   -1.00       0.93      NA         NA
#> Danielson.Bryan        1.00       1.00   -0.98       0.88      NA         NA
#> Danielson.Bryan.2     -0.32      -0.32    0.20      -0.04      NA         NA
#> Dickes.Steiwer        -0.39      -0.39    0.41      -0.34      NA         NA
#> DRP                    0.99       0.99   -1.00       0.93      NA         NA
#> ELF                      NA         NA      NA         NA      NA         NA
#> Farr.Jenkins.Paterson    NA         NA      NA         NA      NA         NA
#> Flesch                   NA         NA      NA         NA      NA         NA
#> Flesch.PSK               NA         NA      NA         NA      NA         NA
#> Flesch.Kincaid           NA         NA      NA         NA      NA         NA
#>                       Coleman.Liau Coleman.Liau.grade Coleman.Liau.short
#> ARI                          -0.81               0.81               0.81
#> ARI.simple                   -0.80               0.80               0.80
#> Bormuth                       0.73              -0.73              -0.73
#> Bormuth.GP                   -0.56               0.56               0.56
#> Coleman                         NA                 NA                 NA
#> Coleman.C2                      NA                 NA                 NA
#> Coleman.Liau                  1.00              -1.00              -1.00
#> Coleman.Liau.grade           -1.00               1.00               1.00
#> Coleman.Liau.short           -1.00               1.00               1.00
#> Dale.Chall                    0.72              -0.72              -0.72
#> Dale.Chall.old               -0.72               0.72               0.72
#> Dale.Chall.PSK               -0.72               0.72               0.72
#> Danielson.Bryan              -0.85               0.85               0.85
#> Danielson.Bryan.2             0.81              -0.81              -0.81
#> Dickes.Steiwer                0.23              -0.23              -0.23
#> DRP                          -0.73               0.73               0.73
#> ELF                             NA                 NA                 NA
#> Farr.Jenkins.Paterson           NA                 NA                 NA
#> Flesch                          NA                 NA                 NA
#> Flesch.PSK                      NA                 NA                 NA
#> Flesch.Kincaid                  NA                 NA                 NA
#>                       Dale.Chall Dale.Chall.old Dale.Chall.PSK Danielson.Bryan
#> ARI                        -0.99           0.99           0.99            1.00
#> ARI.simple                 -0.99           0.99           0.99            1.00
#> Bormuth                     1.00          -1.00          -1.00           -0.98
#> Bormuth.GP                 -0.93           0.93           0.93            0.88
#> Coleman                       NA             NA             NA              NA
#> Coleman.C2                    NA             NA             NA              NA
#> Coleman.Liau                0.72          -0.72          -0.72           -0.85
#> Coleman.Liau.grade         -0.72           0.72           0.72            0.85
#> Coleman.Liau.short         -0.72           0.72           0.72            0.85
#> Dale.Chall                  1.00          -1.00          -1.00           -0.98
#> Dale.Chall.old             -1.00           1.00           1.00            0.98
#> Dale.Chall.PSK             -1.00           1.00           1.00            0.98
#> Danielson.Bryan            -0.98           0.98           0.98            1.00
#> Danielson.Bryan.2           0.19          -0.19          -0.19           -0.39
#> Dickes.Steiwer              0.41          -0.41          -0.41           -0.38
#> DRP                        -1.00           1.00           1.00            0.98
#> ELF                           NA             NA             NA              NA
#> Farr.Jenkins.Paterson         NA             NA             NA              NA
#> Flesch                        NA             NA             NA              NA
#> Flesch.PSK                    NA             NA             NA              NA
#> Flesch.Kincaid                NA             NA             NA              NA
#>                       Danielson.Bryan.2 Dickes.Steiwer   DRP ELF
#> ARI                               -0.32          -0.39  0.99  NA
#> ARI.simple                        -0.32          -0.39  0.99  NA
#> Bormuth                            0.20           0.41 -1.00  NA
#> Bormuth.GP                        -0.04          -0.34  0.93  NA
#> Coleman                              NA             NA    NA  NA
#> Coleman.C2                           NA             NA    NA  NA
#> Coleman.Liau                       0.81           0.23 -0.73  NA
#> Coleman.Liau.grade                -0.81          -0.23  0.73  NA
#> Coleman.Liau.short                -0.81          -0.23  0.73  NA
#> Dale.Chall                         0.19           0.41 -1.00  NA
#> Dale.Chall.old                    -0.19          -0.41  1.00  NA
#> Dale.Chall.PSK                    -0.19          -0.41  1.00  NA
#> Danielson.Bryan                   -0.39          -0.38  0.98  NA
#> Danielson.Bryan.2                  1.00           0.00 -0.20  NA
#> Dickes.Steiwer                     0.00           1.00 -0.41  NA
#> DRP                               -0.20          -0.41  1.00  NA
#> ELF                                  NA             NA    NA   1
#> Farr.Jenkins.Paterson                NA             NA    NA  NA
#> Flesch                               NA             NA    NA  NA
#> Flesch.PSK                           NA             NA    NA  NA
#> Flesch.Kincaid                       NA             NA    NA  NA
#>                       Farr.Jenkins.Paterson Flesch Flesch.PSK Flesch.Kincaid
#> ARI                                      NA     NA         NA             NA
#> ARI.simple                               NA     NA         NA             NA
#> Bormuth                                  NA     NA         NA             NA
#> Bormuth.GP                               NA     NA         NA             NA
#> Coleman                                  NA     NA         NA             NA
#> Coleman.C2                               NA     NA         NA             NA
#> Coleman.Liau                             NA     NA         NA             NA
#> Coleman.Liau.grade                       NA     NA         NA             NA
#> Coleman.Liau.short                       NA     NA         NA             NA
#> Dale.Chall                               NA     NA         NA             NA
#> Dale.Chall.old                           NA     NA         NA             NA
#> Dale.Chall.PSK                           NA     NA         NA             NA
#> Danielson.Bryan                          NA     NA         NA             NA
#> Danielson.Bryan.2                        NA     NA         NA             NA
#> Dickes.Steiwer                           NA     NA         NA             NA
#> DRP                                      NA     NA         NA             NA
#> ELF                                      NA     NA         NA             NA
#> Farr.Jenkins.Paterson                     1     NA         NA             NA
#> Flesch                                   NA      1         NA             NA
#> Flesch.PSK                               NA     NA          1             NA
#> Flesch.Kincaid                           NA     NA         NA              1
#>                       FOG FOG.PSK FOG.NRI FORCAST FORCAST.RGL Fucks
#> ARI                    NA      NA      NA      NA          NA  1.00
#> ARI.simple             NA      NA      NA      NA          NA  1.00
#> Bormuth                NA      NA      NA      NA          NA -1.00
#> Bormuth.GP             NA      NA      NA      NA          NA  0.92
#> Coleman                NA      NA      NA      NA          NA    NA
#> Coleman.C2             NA      NA      NA      NA          NA    NA
#> Coleman.Liau           NA      NA      NA      NA          NA -0.77
#> Coleman.Liau.grade     NA      NA      NA      NA          NA  0.77
#> Coleman.Liau.short     NA      NA      NA      NA          NA  0.77
#> Dale.Chall             NA      NA      NA      NA          NA -1.00
#> Dale.Chall.old         NA      NA      NA      NA          NA  1.00
#> Dale.Chall.PSK         NA      NA      NA      NA          NA  1.00
#> Danielson.Bryan        NA      NA      NA      NA          NA  0.99
#> Danielson.Bryan.2      NA      NA      NA      NA          NA -0.26
#> Dickes.Steiwer         NA      NA      NA      NA          NA -0.39
#> DRP                    NA      NA      NA      NA          NA  1.00
#> ELF                    NA      NA      NA      NA          NA    NA
#> Farr.Jenkins.Paterson  NA      NA      NA      NA          NA    NA
#> Flesch                 NA      NA      NA      NA          NA    NA
#> Flesch.PSK             NA      NA      NA      NA          NA    NA
#> Flesch.Kincaid         NA      NA      NA      NA          NA    NA
#>                       Linsear.Write   LIW nWS nWS.2 nWS.3 nWS.4   RIX Scrabble
#> ARI                              NA  1.00  NA    NA    NA    NA  0.99    -0.03
#> ARI.simple                       NA  0.99  NA    NA    NA    NA  0.99    -0.03
#> Bormuth                          NA -0.98  NA    NA    NA    NA -0.98    -0.01
#> Bormuth.GP                       NA  0.88  NA    NA    NA    NA  0.89    -0.03
#> Coleman                          NA    NA  NA    NA    NA    NA    NA       NA
#> Coleman.C2                       NA    NA  NA    NA    NA    NA    NA       NA
#> Coleman.Liau                     NA -0.85  NA    NA    NA    NA -0.83     0.21
#> Coleman.Liau.grade               NA  0.85  NA    NA    NA    NA  0.83    -0.21
#> Coleman.Liau.short               NA  0.85  NA    NA    NA    NA  0.83    -0.21
#> Dale.Chall                       NA -0.98  NA    NA    NA    NA -0.98    -0.02
#> Dale.Chall.old                   NA  0.98  NA    NA    NA    NA  0.98     0.02
#> Dale.Chall.PSK                   NA  0.98  NA    NA    NA    NA  0.98     0.02
#> Danielson.Bryan                  NA  1.00  NA    NA    NA    NA  0.99    -0.06
#> Danielson.Bryan.2                NA -0.39  NA    NA    NA    NA -0.37     0.32
#> Dickes.Steiwer                   NA -0.40  NA    NA    NA    NA -0.40    -0.46
#> DRP                              NA  0.98  NA    NA    NA    NA  0.98     0.01
#> ELF                              NA    NA  NA    NA    NA    NA    NA       NA
#> Farr.Jenkins.Paterson            NA    NA  NA    NA    NA    NA    NA       NA
#> Flesch                           NA    NA  NA    NA    NA    NA    NA       NA
#> Flesch.PSK                       NA    NA  NA    NA    NA    NA    NA       NA
#> Flesch.Kincaid                   NA    NA  NA    NA    NA    NA    NA       NA
#>                       SMOG SMOG.C SMOG.simple SMOG.de Spache Spache.old Strain
#> ARI                     NA     NA          NA      NA   0.99       0.99     NA
#> ARI.simple              NA     NA          NA      NA   0.99       0.99     NA
#> Bormuth                 NA     NA          NA      NA  -1.00      -1.00     NA
#> Bormuth.GP              NA     NA          NA      NA   0.93       0.93     NA
#> Coleman                 NA     NA          NA      NA     NA         NA     NA
#> Coleman.C2              NA     NA          NA      NA     NA         NA     NA
#> Coleman.Liau            NA     NA          NA      NA  -0.72      -0.72     NA
#> Coleman.Liau.grade      NA     NA          NA      NA   0.72       0.72     NA
#> Coleman.Liau.short      NA     NA          NA      NA   0.72       0.72     NA
#> Dale.Chall              NA     NA          NA      NA  -1.00      -1.00     NA
#> Dale.Chall.old          NA     NA          NA      NA   1.00       1.00     NA
#> Dale.Chall.PSK          NA     NA          NA      NA   1.00       1.00     NA
#> Danielson.Bryan         NA     NA          NA      NA   0.98       0.98     NA
#> Danielson.Bryan.2       NA     NA          NA      NA  -0.19      -0.19     NA
#> Dickes.Steiwer          NA     NA          NA      NA  -0.41      -0.41     NA
#> DRP                     NA     NA          NA      NA   1.00       1.00     NA
#> ELF                     NA     NA          NA      NA     NA         NA     NA
#> Farr.Jenkins.Paterson   NA     NA          NA      NA     NA         NA     NA
#> Flesch                  NA     NA          NA      NA     NA         NA     NA
#> Flesch.PSK              NA     NA          NA      NA     NA         NA     NA
#> Flesch.Kincaid          NA     NA          NA      NA     NA         NA     NA
#>                       Traenkle.Bailer Traenkle.Bailer.2 Wheeler.Smith
#> ARI                             -1.00             -0.75            NA
#> ARI.simple                      -1.00             -0.75            NA
#> Bormuth                          1.00              0.66            NA
#> Bormuth.GP                      -0.91             -0.51            NA
#> Coleman                            NA                NA            NA
#> Coleman.C2                         NA                NA            NA
#> Coleman.Liau                     0.78              0.99            NA
#> Coleman.Liau.grade              -0.78             -0.99            NA
#> Coleman.Liau.short              -0.78             -0.99            NA
#> Dale.Chall                       1.00              0.66            NA
#> Dale.Chall.old                  -1.00             -0.66            NA
#> Dale.Chall.PSK                  -1.00             -0.66            NA
#> Danielson.Bryan                 -0.99             -0.80            NA
#> Danielson.Bryan.2                0.28              0.86            NA
#> Dickes.Steiwer                   0.39              0.18            NA
#> DRP                             -1.00             -0.66            NA
#> ELF                                NA                NA            NA
#> Farr.Jenkins.Paterson              NA                NA            NA
#> Flesch                             NA                NA            NA
#> Flesch.PSK                         NA                NA            NA
#> Flesch.Kincaid                     NA                NA            NA
#>                       meanSentenceLength meanWordSyllables
#> ARI                                 0.99                NA
#> ARI.simple                          0.99                NA
#> Bormuth                            -1.00                NA
#> Bormuth.GP                          0.93                NA
#> Coleman                               NA                NA
#> Coleman.C2                            NA                NA
#> Coleman.Liau                       -0.72                NA
#> Coleman.Liau.grade                  0.72                NA
#> Coleman.Liau.short                  0.72                NA
#> Dale.Chall                         -1.00                NA
#> Dale.Chall.old                      1.00                NA
#> Dale.Chall.PSK                      1.00                NA
#> Danielson.Bryan                     0.98                NA
#> Danielson.Bryan.2                  -0.19                NA
#> Dickes.Steiwer                     -0.41                NA
#> DRP                                 1.00                NA
#> ELF                                   NA                NA
#> Farr.Jenkins.Paterson                 NA                NA
#> Flesch                                NA                NA
#> Flesch.PSK                            NA                NA
#> Flesch.Kincaid                        NA                NA
#>  [ reached getOption("max.print") -- omitted 26 rows ]

textstat_readability(data_corpus_inaugural, measure = "Flesch.Kincaid")

#> 1789-Washington 1793-Washington      1797-Adams  1801-Jefferson  1805-Jefferson 
#>              NA       17.151759       28.692087       19.295924       21.917088 
#>    1809-Madison    1813-Madison     1817-Monroe     1821-Monroe      1825-Adams 
#>       25.557611       18.034452       14.443950              NA       19.748273 
#>    1829-Jackson    1833-Jackson   1837-VanBuren   1841-Harrison       1845-Polk 
#>       21.893148       19.958757       20.263524       19.355855              NA 
#>     1849-Taylor     1853-Pierce   1857-Buchanan    1861-Lincoln    1865-Lincoln 
#>       23.748805              NA       16.110868              NA       11.774929 
#>      1869-Grant      1873-Grant      1877-Hayes   1881-Garfield  1885-Cleveland 
#>       14.008426              NA              NA              NA       19.052937 
#>   1889-Harrison  1893-Cleveland   1897-McKinley   1901-McKinley  1905-Roosevelt 
#>              NA       18.248838              NA              NA       13.959254 
#>       1909-Taft     1913-Wilson     1917-Wilson    1921-Harding   1925-Coolidge 
#>              NA       11.707659       11.780913       13.214469              NA 
#>     1929-Hoover  1933-Roosevelt  1937-Roosevelt  1941-Roosevelt  1945-Roosevelt 
#>       13.259377       11.640683              NA              NA              NA 
#>     1949-Truman 1953-Eisenhower 1957-Eisenhower    1961-Kennedy    1965-Johnson 
#>              NA              NA        8.094908              NA              NA 
#>      1969-Nixon      1973-Nixon     1977-Carter     1981-Reagan     1985-Reagan 
#>              NA              NA              NA              NA              NA 
#>       1989-Bush    1993-Clinton    1997-Clinton       2001-Bush       2005-Bush 
#>              NA              NA              NA              NA       10.954613 
#>      2009-Obama      2013-Obama 
#>       10.229426              NA 

inaugReadability <- textstat_readability(data_corpus_inaugural, "all")
round(cor(inaugReadability), 2)

#>                         ARI ARI.simple Bormuth Bormuth.GP Coleman Coleman.C2
#> ARI                    1.00       1.00   -0.99       0.90      NA         NA
#> ARI.simple             1.00       1.00   -0.99       0.90      NA         NA
#> Bormuth               -0.99      -0.99    1.00      -0.93      NA         NA
#> Bormuth.GP             0.90       0.90   -0.93       1.00      NA         NA
#> Coleman                  NA         NA      NA         NA       1         NA
#> Coleman.C2               NA         NA      NA         NA      NA          1
#> Coleman.Liau          -0.81      -0.80    0.73      -0.56      NA         NA
#> Coleman.Liau.grade     0.81       0.80   -0.73       0.56      NA         NA
#> Coleman.Liau.short     0.81       0.80   -0.73       0.56      NA         NA
#> Dale.Chall            -0.99      -0.99    1.00      -0.93      NA         NA
#> Dale.Chall.old         0.99       0.99   -1.00       0.93      NA         NA
#> Dale.Chall.PSK         0.99       0.99   -1.00       0.93      NA         NA
#> Danielson.Bryan        1.00       1.00   -0.98       0.88      NA         NA
#> Danielson.Bryan.2     -0.32      -0.32    0.20      -0.04      NA         NA
#> Dickes.Steiwer        -0.39      -0.39    0.41      -0.34      NA         NA
#> DRP                    0.99       0.99   -1.00       0.93      NA         NA
#> ELF                      NA         NA      NA         NA      NA         NA
#> Farr.Jenkins.Paterson    NA         NA      NA         NA      NA         NA
#> Flesch                   NA         NA      NA         NA      NA         NA
#> Flesch.PSK               NA         NA      NA         NA      NA         NA
#> Flesch.Kincaid           NA         NA      NA         NA      NA         NA
#>                       Coleman.Liau Coleman.Liau.grade Coleman.Liau.short
#> ARI                          -0.81               0.81               0.81
#> ARI.simple                   -0.80               0.80               0.80
#> Bormuth                       0.73              -0.73              -0.73
#> Bormuth.GP                   -0.56               0.56               0.56
#> Coleman                         NA                 NA                 NA
#> Coleman.C2                      NA                 NA                 NA
#> Coleman.Liau                  1.00              -1.00              -1.00
#> Coleman.Liau.grade           -1.00               1.00               1.00
#> Coleman.Liau.short           -1.00               1.00               1.00
#> Dale.Chall                    0.72              -0.72              -0.72
#> Dale.Chall.old               -0.72               0.72               0.72
#> Dale.Chall.PSK               -0.72               0.72               0.72
#> Danielson.Bryan              -0.85               0.85               0.85
#> Danielson.Bryan.2             0.81              -0.81              -0.81
#> Dickes.Steiwer                0.23              -0.23              -0.23
#> DRP                          -0.73               0.73               0.73
#> ELF                             NA                 NA                 NA
#> Farr.Jenkins.Paterson           NA                 NA                 NA
#> Flesch                          NA                 NA                 NA
#> Flesch.PSK                      NA                 NA                 NA
#> Flesch.Kincaid                  NA                 NA                 NA
#>                       Dale.Chall Dale.Chall.old Dale.Chall.PSK Danielson.Bryan
#> ARI                        -0.99           0.99           0.99            1.00
#> ARI.simple                 -0.99           0.99           0.99            1.00
#> Bormuth                     1.00          -1.00          -1.00           -0.98
#> Bormuth.GP                 -0.93           0.93           0.93            0.88
#> Coleman                       NA             NA             NA              NA
#> Coleman.C2                    NA             NA             NA              NA
#> Coleman.Liau                0.72          -0.72          -0.72           -0.85
#> Coleman.Liau.grade         -0.72           0.72           0.72            0.85
#> Coleman.Liau.short         -0.72           0.72           0.72            0.85
#> Dale.Chall                  1.00          -1.00          -1.00           -0.98
#> Dale.Chall.old             -1.00           1.00           1.00            0.98
#> Dale.Chall.PSK             -1.00           1.00           1.00            0.98
#> Danielson.Bryan            -0.98           0.98           0.98            1.00
#> Danielson.Bryan.2           0.19          -0.19          -0.19           -0.39
#> Dickes.Steiwer              0.41          -0.41          -0.41           -0.38
#> DRP                        -1.00           1.00           1.00            0.98
#> ELF                           NA             NA             NA              NA
#> Farr.Jenkins.Paterson         NA             NA             NA              NA
#> Flesch                        NA             NA             NA              NA
#> Flesch.PSK                    NA             NA             NA              NA
#> Flesch.Kincaid                NA             NA             NA              NA
#>                       Danielson.Bryan.2 Dickes.Steiwer   DRP ELF
#> ARI                               -0.32          -0.39  0.99  NA
#> ARI.simple                        -0.32          -0.39  0.99  NA
#> Bormuth                            0.20           0.41 -1.00  NA
#> Bormuth.GP                        -0.04          -0.34  0.93  NA
#> Coleman                              NA             NA    NA  NA
#> Coleman.C2                           NA             NA    NA  NA
#> Coleman.Liau                       0.81           0.23 -0.73  NA
#> Coleman.Liau.grade                -0.81          -0.23  0.73  NA
#> Coleman.Liau.short                -0.81          -0.23  0.73  NA
#> Dale.Chall                         0.19           0.41 -1.00  NA
#> Dale.Chall.old                    -0.19          -0.41  1.00  NA
#> Dale.Chall.PSK                    -0.19          -0.41  1.00  NA
#> Danielson.Bryan                   -0.39          -0.38  0.98  NA
#> Danielson.Bryan.2                  1.00           0.00 -0.20  NA
#> Dickes.Steiwer                     0.00           1.00 -0.41  NA
#> DRP                               -0.20          -0.41  1.00  NA
#> ELF                                  NA             NA    NA   1
#> Farr.Jenkins.Paterson                NA             NA    NA  NA
#> Flesch                               NA             NA    NA  NA
#> Flesch.PSK                           NA             NA    NA  NA
#> Flesch.Kincaid                       NA             NA    NA  NA
#>                       Farr.Jenkins.Paterson Flesch Flesch.PSK Flesch.Kincaid
#> ARI                                      NA     NA         NA             NA
#> ARI.simple                               NA     NA         NA             NA
#> Bormuth                                  NA     NA         NA             NA
#> Bormuth.GP                               NA     NA         NA             NA
#> Coleman                                  NA     NA         NA             NA
#> Coleman.C2                               NA     NA         NA             NA
#> Coleman.Liau                             NA     NA         NA             NA
#> Coleman.Liau.grade                       NA     NA         NA             NA
#> Coleman.Liau.short                       NA     NA         NA             NA
#> Dale.Chall                               NA     NA         NA             NA
#> Dale.Chall.old                           NA     NA         NA             NA
#> Dale.Chall.PSK                           NA     NA         NA             NA
#> Danielson.Bryan                          NA     NA         NA             NA
#> Danielson.Bryan.2                        NA     NA         NA             NA
#> Dickes.Steiwer                           NA     NA         NA             NA
#> DRP                                      NA     NA         NA             NA
#> ELF                                      NA     NA         NA             NA
#> Farr.Jenkins.Paterson                     1     NA         NA             NA
#> Flesch                                   NA      1         NA             NA
#> Flesch.PSK                               NA     NA          1             NA
#> Flesch.Kincaid                           NA     NA         NA              1
#>                       FOG FOG.PSK FOG.NRI FORCAST FORCAST.RGL Fucks
#> ARI                    NA      NA      NA      NA          NA  1.00
#> ARI.simple             NA      NA      NA      NA          NA  1.00
#> Bormuth                NA      NA      NA      NA          NA -1.00
#> Bormuth.GP             NA      NA      NA      NA          NA  0.92
#> Coleman                NA      NA      NA      NA          NA    NA
#> Coleman.C2             NA      NA      NA      NA          NA    NA
#> Coleman.Liau           NA      NA      NA      NA          NA -0.77
#> Coleman.Liau.grade     NA      NA      NA      NA          NA  0.77
#> Coleman.Liau.short     NA      NA      NA      NA          NA  0.77
#> Dale.Chall             NA      NA      NA      NA          NA -1.00
#> Dale.Chall.old         NA      NA      NA      NA          NA  1.00
#> Dale.Chall.PSK         NA      NA      NA      NA          NA  1.00
#> Danielson.Bryan        NA      NA      NA      NA          NA  0.99
#> Danielson.Bryan.2      NA      NA      NA      NA          NA -0.26
#> Dickes.Steiwer         NA      NA      NA      NA          NA -0.39
#> DRP                    NA      NA      NA      NA          NA  1.00
#> ELF                    NA      NA      NA      NA          NA    NA
#> Farr.Jenkins.Paterson  NA      NA      NA      NA          NA    NA
#> Flesch                 NA      NA      NA      NA          NA    NA
#> Flesch.PSK             NA      NA      NA      NA          NA    NA
#> Flesch.Kincaid         NA      NA      NA      NA          NA    NA
#>                       Linsear.Write   LIW nWS nWS.2 nWS.3 nWS.4   RIX Scrabble
#> ARI                              NA  1.00  NA    NA    NA    NA  0.99    -0.03
#> ARI.simple                       NA  0.99  NA    NA    NA    NA  0.99    -0.03
#> Bormuth                          NA -0.98  NA    NA    NA    NA -0.98    -0.01
#> Bormuth.GP                       NA  0.88  NA    NA    NA    NA  0.89    -0.03
#> Coleman                          NA    NA  NA    NA    NA    NA    NA       NA
#> Coleman.C2                       NA    NA  NA    NA    NA    NA    NA       NA
#> Coleman.Liau                     NA -0.85  NA    NA    NA    NA -0.83     0.21
#> Coleman.Liau.grade               NA  0.85  NA    NA    NA    NA  0.83    -0.21
#> Coleman.Liau.short               NA  0.85  NA    NA    NA    NA  0.83    -0.21
#> Dale.Chall                       NA -0.98  NA    NA    NA    NA -0.98    -0.02
#> Dale.Chall.old                   NA  0.98  NA    NA    NA    NA  0.98     0.02
#> Dale.Chall.PSK                   NA  0.98  NA    NA    NA    NA  0.98     0.02
#> Danielson.Bryan                  NA  1.00  NA    NA    NA    NA  0.99    -0.06
#> Danielson.Bryan.2                NA -0.39  NA    NA    NA    NA -0.37     0.32
#> Dickes.Steiwer                   NA -0.40  NA    NA    NA    NA -0.40    -0.46
#> DRP                              NA  0.98  NA    NA    NA    NA  0.98     0.01
#> ELF                              NA    NA  NA    NA    NA    NA    NA       NA
#> Farr.Jenkins.Paterson            NA    NA  NA    NA    NA    NA    NA       NA
#> Flesch                           NA    NA  NA    NA    NA    NA    NA       NA
#> Flesch.PSK                       NA    NA  NA    NA    NA    NA    NA       NA
#> Flesch.Kincaid                   NA    NA  NA    NA    NA    NA    NA       NA
#>                       SMOG SMOG.C SMOG.simple SMOG.de Spache Spache.old Strain
#> ARI                     NA     NA          NA      NA   0.99       0.99     NA
#> ARI.simple              NA     NA          NA      NA   0.99       0.99     NA
#> Bormuth                 NA     NA          NA      NA  -1.00      -1.00     NA
#> Bormuth.GP              NA     NA          NA      NA   0.93       0.93     NA
#> Coleman                 NA     NA          NA      NA     NA         NA     NA
#> Coleman.C2              NA     NA          NA      NA     NA         NA     NA
#> Coleman.Liau            NA     NA          NA      NA  -0.72      -0.72     NA
#> Coleman.Liau.grade      NA     NA          NA      NA   0.72       0.72     NA
#> Coleman.Liau.short      NA     NA          NA      NA   0.72       0.72     NA
#> Dale.Chall              NA     NA          NA      NA  -1.00      -1.00     NA
#> Dale.Chall.old          NA     NA          NA      NA   1.00       1.00     NA
#> Dale.Chall.PSK          NA     NA          NA      NA   1.00       1.00     NA
#> Danielson.Bryan         NA     NA          NA      NA   0.98       0.98     NA
#> Danielson.Bryan.2       NA     NA          NA      NA  -0.19      -0.19     NA
#> Dickes.Steiwer          NA     NA          NA      NA  -0.41      -0.41     NA
#> DRP                     NA     NA          NA      NA   1.00       1.00     NA
#> ELF                     NA     NA          NA      NA     NA         NA     NA
#> Farr.Jenkins.Paterson   NA     NA          NA      NA     NA         NA     NA
#> Flesch                  NA     NA          NA      NA     NA         NA     NA
#> Flesch.PSK              NA     NA          NA      NA     NA         NA     NA
#> Flesch.Kincaid          NA     NA          NA      NA     NA         NA     NA
#>                       Traenkle.Bailer Traenkle.Bailer.2 Wheeler.Smith
#> ARI                             -1.00             -0.75            NA
#> ARI.simple                      -1.00             -0.75            NA
#> Bormuth                          1.00              0.66            NA
#> Bormuth.GP                      -0.91             -0.51            NA
#> Coleman                            NA                NA            NA
#> Coleman.C2                         NA                NA            NA
#> Coleman.Liau                     0.78              0.99            NA
#> Coleman.Liau.grade              -0.78             -0.99            NA
#> Coleman.Liau.short              -0.78             -0.99            NA
#> Dale.Chall                       1.00              0.66            NA
#> Dale.Chall.old                  -1.00             -0.66            NA
#> Dale.Chall.PSK                  -1.00             -0.66            NA
#> Danielson.Bryan                 -0.99             -0.80            NA
#> Danielson.Bryan.2                0.28              0.86            NA
#> Dickes.Steiwer                   0.39              0.18            NA
#> DRP                             -1.00             -0.66            NA
#> ELF                                NA                NA            NA
#> Farr.Jenkins.Paterson              NA                NA            NA
#> Flesch                             NA                NA            NA
#> Flesch.PSK                         NA                NA            NA
#> Flesch.Kincaid                     NA                NA            NA
#>                       meanSentenceLength meanWordSyllables
#> ARI                                 0.99                NA
#> ARI.simple                          0.99                NA
#> Bormuth                            -1.00                NA
#> Bormuth.GP                          0.93                NA
#> Coleman                               NA                NA
#> Coleman.C2                            NA                NA
#> Coleman.Liau                       -0.72                NA
#> Coleman.Liau.grade                  0.72                NA
#> Coleman.Liau.short                  0.72                NA
#> Dale.Chall                         -1.00                NA
#> Dale.Chall.old                      1.00                NA
#> Dale.Chall.PSK                      1.00                NA
#> Danielson.Bryan                     0.98                NA
#> Danielson.Bryan.2                  -0.19                NA
#> Dickes.Steiwer                     -0.41                NA
#> DRP                                 1.00                NA
#> ELF                                   NA                NA
#> Farr.Jenkins.Paterson                 NA                NA
#> Flesch                                NA                NA
#> Flesch.PSK                            NA                NA
#> Flesch.Kincaid                        NA                NA
#>  [ reached getOption("max.print") -- omitted 26 rows ]