calculate readability

Calculate the readability of text(s) using one of a variety of computed indexes.

textstat_readability(x, measure = c("all", "ARI", "ARI.simple", "Bormuth",
  "Bormuth.GP", "Coleman", "Coleman.C2", "Coleman.Liau", "Coleman.Liau.grade",
  "Coleman.Liau.short", "Dale.Chall", "Dale.Chall.old", "Dale.Chall.PSK",
  "Danielson.Bryan", "Danielson.Bryan.2", "Dickes.Steiwer", "DRP", "ELF",
  "Farr.Jenkins.Paterson", "Flesch", "Flesch.PSK", "Flesch.Kincaid", "FOG",
  "FOG.PSK", "FOG.NRI", "FORCAST", "FORCAST.RGL", "Fucks", "Linsear.Write",
  "LIW", "nWS", "nWS.2", "nWS.3", "nWS.4", "RIX", "Scrabble", "SMOG", "SMOG.C",
  "SMOG.simple",      "SMOG.de", "Spache", "Spache.old", "Strain",
  "Traenkle.Bailer", "Traenkle.Bailer.2", "Wheeler.Smith", "meanSentenceLength",
  "meanWordSyllables"), removeHyphens = TRUE, drop = TRUE, ...)

Arguments

x

a character or corpus object containing the texts

measure

character vector defining the readability measure to calculate

removeHyphens

if TRUE, treat constituent words in hyphenated as separate terms, for purposes of computing word lengths, e.g. "decision-making" as two terms of lengths 8 and 6 characters respectively, rather than as a single word of 15 characters

drop

if TRUE, the result is returned as a numeric vector if only a single measure is requested; otherwise, a data.frame is returned with each column consisting of a requested measure.

...

not used

Value

a data.frame object consisting of the documents as rows, and the readability statistics as columns

Examples

txt <- c("Readability zero one.  Ten, Eleven.", "The cat in a dilapidated tophat.")
textstat_readability(txt, "Flesch.Kincaid")

#>     text1     text2 
#> 13.705000  8.383333 

textstat_readability(txt, "Flesch.Kincaid", drop = FALSE)

#>       Flesch.Kincaid
#> text1      13.705000
#> text2       8.383333

textstat_readability(txt, c("FOG", "FOG.PSK", "FOG.NRI"))

#>             FOG  FOG.PSK FOG.NRI
#> text1 17.000000 4.608659   -1.30
#> text2  9.066667 3.254382   -1.11

inaugReadability <- textstat_readability(data_corpus_inaugural, "all")
round(cor(inaugReadability), 2)

#>                         ARI ARI.simple Bormuth Bormuth.GP Coleman Coleman.C2
#> ARI                    1.00       1.00   -0.99       0.90   -0.58      -0.77
#> ARI.simple             1.00       1.00   -0.99       0.90   -0.58      -0.76
#> Bormuth               -0.99      -0.99    1.00      -0.93    0.48       0.68
#> Bormuth.GP             0.90       0.90   -0.93       1.00   -0.34      -0.51
#> Coleman               -0.58      -0.58    0.48      -0.34    1.00       0.96
#> Coleman.C2            -0.77      -0.76    0.68      -0.51    0.96       1.00
#> Coleman.Liau          -0.81      -0.80    0.73      -0.56    0.91       0.97
#> Coleman.Liau.grade     0.81       0.80   -0.73       0.56   -0.91      -0.97
#> Coleman.Liau.short     0.81       0.80   -0.73       0.56   -0.91      -0.97
#> Dale.Chall            -0.99      -0.99    1.00      -0.93    0.48       0.68
#> Dale.Chall.old         0.99       0.99   -1.00       0.93   -0.48      -0.68
#> Dale.Chall.PSK         0.99       0.99   -1.00       0.93   -0.48      -0.68
#> Danielson.Bryan        1.00       1.00   -0.98       0.88   -0.64      -0.80
#> Danielson.Bryan.2     -0.32      -0.32    0.20      -0.04    0.89       0.80
#> Dickes.Steiwer        -0.39      -0.39    0.41      -0.34    0.14       0.25
#> DRP                    0.99       0.99   -1.00       0.93   -0.48      -0.68
#> ELF                    1.00       1.00   -0.99       0.90   -0.62      -0.79
#> Farr.Jenkins.Paterson -0.99      -0.99    1.00      -0.93    0.48       0.68
#> Flesch                -0.97      -0.97    0.93      -0.82    0.76       0.89
#> Flesch.PSK             0.99       0.98   -0.96       0.86   -0.70      -0.86
#> Flesch.Kincaid         1.00       1.00   -0.99       0.90   -0.62      -0.79
#>                       Coleman.Liau Coleman.Liau.grade Coleman.Liau.short
#> ARI                          -0.81               0.81               0.81
#> ARI.simple                   -0.80               0.80               0.80
#> Bormuth                       0.73              -0.73              -0.73
#> Bormuth.GP                   -0.56               0.56               0.56
#> Coleman                       0.91              -0.91              -0.91
#> Coleman.C2                    0.97              -0.97              -0.97
#> Coleman.Liau                  1.00              -1.00              -1.00
#> Coleman.Liau.grade           -1.00               1.00               1.00
#> Coleman.Liau.short           -1.00               1.00               1.00
#> Dale.Chall                    0.72              -0.72              -0.72
#> Dale.Chall.old               -0.72               0.72               0.72
#> Dale.Chall.PSK               -0.72               0.72               0.72
#> Danielson.Bryan              -0.85               0.85               0.85
#> Danielson.Bryan.2             0.81              -0.81              -0.81
#> Dickes.Steiwer                0.23              -0.23              -0.23
#> DRP                          -0.73               0.73               0.73
#> ELF                          -0.81               0.81               0.81
#> Farr.Jenkins.Paterson         0.72              -0.72              -0.72
#> Flesch                        0.91              -0.91              -0.91
#> Flesch.PSK                   -0.88               0.88               0.88
#> Flesch.Kincaid               -0.82               0.82               0.82
#>                       Dale.Chall Dale.Chall.old Dale.Chall.PSK Danielson.Bryan
#> ARI                        -0.99           0.99           0.99            1.00
#> ARI.simple                 -0.99           0.99           0.99            1.00
#> Bormuth                     1.00          -1.00          -1.00           -0.98
#> Bormuth.GP                 -0.93           0.93           0.93            0.88
#> Coleman                     0.48          -0.48          -0.48           -0.64
#> Coleman.C2                  0.68          -0.68          -0.68           -0.80
#> Coleman.Liau                0.72          -0.72          -0.72           -0.85
#> Coleman.Liau.grade         -0.72           0.72           0.72            0.85
#> Coleman.Liau.short         -0.72           0.72           0.72            0.85
#> Dale.Chall                  1.00          -1.00          -1.00           -0.98
#> Dale.Chall.old             -1.00           1.00           1.00            0.98
#> Dale.Chall.PSK             -1.00           1.00           1.00            0.98
#> Danielson.Bryan            -0.98           0.98           0.98            1.00
#> Danielson.Bryan.2           0.19          -0.19          -0.19           -0.39
#> Dickes.Steiwer              0.41          -0.41          -0.41           -0.38
#> DRP                        -1.00           1.00           1.00            0.98
#> ELF                        -0.98           0.98           0.98            1.00
#> Farr.Jenkins.Paterson       1.00          -1.00          -1.00           -0.98
#> Flesch                      0.93          -0.93          -0.93           -0.98
#> Flesch.PSK                 -0.96           0.96           0.96            0.99
#> Flesch.Kincaid             -0.98           0.98           0.98            1.00
#>                       Danielson.Bryan.2 Dickes.Steiwer   DRP   ELF
#> ARI                               -0.32          -0.39  0.99  1.00
#> ARI.simple                        -0.32          -0.39  0.99  1.00
#> Bormuth                            0.20           0.41 -1.00 -0.99
#> Bormuth.GP                        -0.04          -0.34  0.93  0.90
#> Coleman                            0.89           0.14 -0.48 -0.62
#> Coleman.C2                         0.80           0.25 -0.68 -0.79
#> Coleman.Liau                       0.81           0.23 -0.73 -0.81
#> Coleman.Liau.grade                -0.81          -0.23  0.73  0.81
#> Coleman.Liau.short                -0.81          -0.23  0.73  0.81
#> Dale.Chall                         0.19           0.41 -1.00 -0.98
#> Dale.Chall.old                    -0.19          -0.41  1.00  0.98
#> Dale.Chall.PSK                    -0.19          -0.41  1.00  0.98
#> Danielson.Bryan                   -0.39          -0.38  0.98  1.00
#> Danielson.Bryan.2                  1.00           0.00 -0.20 -0.34
#> Dickes.Steiwer                     0.00           1.00 -0.41 -0.39
#> DRP                               -0.20          -0.41  1.00  0.99
#> ELF                               -0.34          -0.39  0.99  1.00
#> Farr.Jenkins.Paterson              0.19           0.41 -1.00 -0.99
#> Flesch                             0.53           0.37 -0.93 -0.97
#> Flesch.PSK                        -0.46          -0.38  0.96  0.99
#> Flesch.Kincaid                    -0.35          -0.40  0.99  1.00
#>                       Farr.Jenkins.Paterson Flesch Flesch.PSK Flesch.Kincaid
#> ARI                                   -0.99  -0.97       0.99           1.00
#> ARI.simple                            -0.99  -0.97       0.98           1.00
#> Bormuth                                1.00   0.93      -0.96          -0.99
#> Bormuth.GP                            -0.93  -0.82       0.86           0.90
#> Coleman                                0.48   0.76      -0.70          -0.62
#> Coleman.C2                             0.68   0.89      -0.86          -0.79
#> Coleman.Liau                           0.72   0.91      -0.88          -0.82
#> Coleman.Liau.grade                    -0.72  -0.91       0.88           0.82
#> Coleman.Liau.short                    -0.72  -0.91       0.88           0.82
#> Dale.Chall                             1.00   0.93      -0.96          -0.98
#> Dale.Chall.old                        -1.00  -0.93       0.96           0.98
#> Dale.Chall.PSK                        -1.00  -0.93       0.96           0.98
#> Danielson.Bryan                       -0.98  -0.98       0.99           1.00
#> Danielson.Bryan.2                      0.19   0.53      -0.46          -0.35
#> Dickes.Steiwer                         0.41   0.37      -0.38          -0.40
#> DRP                                   -1.00  -0.93       0.96           0.99
#> ELF                                   -0.99  -0.97       0.99           1.00
#> Farr.Jenkins.Paterson                  1.00   0.93      -0.96          -0.99
#> Flesch                                 0.93   1.00      -1.00          -0.98
#> Flesch.PSK                            -0.96  -1.00       1.00           0.99
#> Flesch.Kincaid                        -0.99  -0.98       0.99           1.00
#>                         FOG FOG.PSK FOG.NRI FORCAST FORCAST.RGL Fucks
#> ARI                    1.00    1.00    0.49    0.58        0.58  1.00
#> ARI.simple             0.99    1.00    0.49    0.58        0.58  1.00
#> Bormuth               -0.98   -1.00   -0.46   -0.48       -0.48 -1.00
#> Bormuth.GP             0.88    0.91    0.34    0.34        0.34  0.92
#> Coleman               -0.63   -0.54   -0.49   -1.00       -1.00 -0.53
#> Coleman.C2            -0.80   -0.74   -0.55   -0.96       -0.96 -0.73
#> Coleman.Liau          -0.84   -0.77   -0.55   -0.91       -0.91 -0.77
#> Coleman.Liau.grade     0.84    0.77    0.55    0.91        0.91  0.77
#> Coleman.Liau.short     0.84    0.77    0.55    0.91        0.91  0.77
#> Dale.Chall            -0.98   -1.00   -0.45   -0.48       -0.48 -1.00
#> Dale.Chall.old         0.98    1.00    0.45    0.48        0.48  1.00
#> Dale.Chall.PSK         0.98    1.00    0.45    0.48        0.48  1.00
#> Danielson.Bryan        1.00    0.99    0.51    0.64        0.64  0.99
#> Danielson.Bryan.2     -0.38   -0.27   -0.39   -0.89       -0.89 -0.26
#> Dickes.Steiwer        -0.42   -0.41    0.13   -0.14       -0.14 -0.39
#> DRP                    0.98    1.00    0.46    0.48        0.48  1.00
#> ELF                    0.99    0.99    0.50    0.62        0.62  0.99
#> Farr.Jenkins.Paterson -0.98   -1.00   -0.46   -0.48       -0.48 -1.00
#> Flesch                -0.98   -0.96   -0.54   -0.76       -0.76 -0.95
#> Flesch.PSK             0.99    0.98    0.53    0.70        0.70  0.97
#> Flesch.Kincaid         1.00    1.00    0.50    0.62        0.62  0.99
#>                       Linsear.Write   LIW   nWS nWS.2 nWS.3 nWS.4   RIX
#> ARI                            0.87  1.00  0.97  0.98  0.98  0.99  0.99
#> ARI.simple                     0.87  0.99  0.97  0.97  0.98  0.99  0.99
#> Bormuth                       -0.84 -0.98 -0.94 -0.94 -0.96 -0.98 -0.98
#> Bormuth.GP                     0.76  0.88  0.82  0.83  0.85  0.88  0.89
#> Coleman                       -0.68 -0.65 -0.75 -0.73 -0.69 -0.64 -0.63
#> Coleman.C2                    -0.79 -0.81 -0.89 -0.88 -0.84 -0.81 -0.80
#> Coleman.Liau                  -0.83 -0.85 -0.91 -0.90 -0.88 -0.84 -0.83
#> Coleman.Liau.grade             0.83  0.85  0.91  0.90  0.88  0.84  0.83
#> Coleman.Liau.short             0.83  0.85  0.91  0.90  0.88  0.84  0.83
#> Dale.Chall                    -0.84 -0.98 -0.93 -0.94 -0.96 -0.98 -0.98
#> Dale.Chall.old                 0.84  0.98  0.93  0.94  0.96  0.98  0.98
#> Dale.Chall.PSK                 0.84  0.98  0.93  0.94  0.96  0.98  0.98
#> Danielson.Bryan                0.89  1.00  0.98  0.99  0.99  1.00  0.99
#> Danielson.Bryan.2             -0.50 -0.39 -0.52 -0.50 -0.45 -0.38 -0.37
#> Dickes.Steiwer                -0.53 -0.40 -0.39 -0.40 -0.42 -0.42 -0.40
#> DRP                            0.84  0.98  0.94  0.94  0.96  0.98  0.98
#> ELF                            0.88  1.00  0.98  0.98  0.99  0.99  1.00
#> Farr.Jenkins.Paterson         -0.84 -0.98 -0.93 -0.94 -0.96 -0.98 -0.98
#> Flesch                        -0.92 -0.98 -1.00 -1.00 -0.99 -0.98 -0.98
#> Flesch.PSK                     0.91  0.99  1.00  1.00  1.00  0.99  0.99
#> Flesch.Kincaid                 0.89  1.00  0.98  0.98  0.99  1.00  0.99
#>                       Scrabble  SMOG SMOG.C SMOG.simple SMOG.de Spache
#> ARI                      -0.03  0.97   0.97        0.97    0.97   0.99
#> ARI.simple               -0.03  0.97   0.97        0.97    0.97   0.99
#> Bormuth                  -0.01 -0.94  -0.94       -0.94   -0.94  -1.00
#> Bormuth.GP               -0.03  0.82   0.82        0.82    0.82   0.93
#> Coleman                   0.36 -0.72  -0.72       -0.72   -0.72  -0.48
#> Coleman.C2                0.26 -0.87  -0.87       -0.87   -0.87  -0.68
#> Coleman.Liau              0.21 -0.90  -0.89       -0.90   -0.90  -0.72
#> Coleman.Liau.grade       -0.21  0.90   0.89        0.90    0.90   0.72
#> Coleman.Liau.short       -0.21  0.90   0.89        0.90    0.90   0.72
#> Dale.Chall               -0.02 -0.94  -0.94       -0.94   -0.94  -1.00
#> Dale.Chall.old            0.02  0.94   0.94        0.94    0.94   1.00
#> Dale.Chall.PSK            0.02  0.94   0.94        0.94    0.94   1.00
#> Danielson.Bryan          -0.06  0.98   0.98        0.98    0.98   0.98
#> Danielson.Bryan.2         0.32 -0.49  -0.48       -0.49   -0.49  -0.19
#> Dickes.Steiwer           -0.46 -0.42  -0.42       -0.42   -0.42  -0.41
#> DRP                       0.01  0.94   0.94        0.94    0.94   1.00
#> ELF                      -0.06  0.98   0.98        0.98    0.98   0.98
#> Farr.Jenkins.Paterson    -0.01 -0.94  -0.94       -0.94   -0.94  -1.00
#> Flesch                    0.12 -0.99  -0.99       -0.99   -0.99  -0.93
#> Flesch.PSK               -0.09  0.99   0.99        0.99    0.99   0.96
#> Flesch.Kincaid           -0.05  0.98   0.98        0.98    0.98   0.98
#>                       Spache.old Strain Traenkle.Bailer Traenkle.Bailer.2
#> ARI                         0.99   1.00           -1.00             -0.75
#> ARI.simple                  0.99   1.00           -1.00             -0.75
#> Bormuth                    -1.00  -1.00            1.00              0.66
#> Bormuth.GP                  0.93   0.92           -0.91             -0.51
#> Coleman                    -0.48  -0.55            0.55              0.92
#> Coleman.C2                 -0.68  -0.74            0.74              0.96
#> Coleman.Liau               -0.72  -0.77            0.78              0.99
#> Coleman.Liau.grade          0.72   0.77           -0.78             -0.99
#> Coleman.Liau.short          0.72   0.77           -0.78             -0.99
#> Dale.Chall                 -1.00  -1.00            1.00              0.66
#> Dale.Chall.old              1.00   1.00           -1.00             -0.66
#> Dale.Chall.PSK              1.00   1.00           -1.00             -0.66
#> Danielson.Bryan             0.98   0.99           -0.99             -0.80
#> Danielson.Bryan.2          -0.19  -0.27            0.28              0.86
#> Dickes.Steiwer             -0.41  -0.40            0.39              0.18
#> DRP                         1.00   1.00           -1.00             -0.66
#> ELF                         0.98   1.00           -0.99             -0.76
#> Farr.Jenkins.Paterson      -1.00  -1.00            1.00              0.66
#> Flesch                     -0.93  -0.96            0.96              0.88
#> Flesch.PSK                  0.96   0.98           -0.98             -0.84
#> Flesch.Kincaid              0.98   1.00           -1.00             -0.77
#>                       Wheeler.Smith meanSentenceLength meanWordSyllables
#> ARI                            1.00               0.99              0.68
#> ARI.simple                     1.00               0.99              0.68
#> Bormuth                       -0.99              -1.00             -0.59
#> Bormuth.GP                     0.90               0.93              0.45
#> Coleman                       -0.62              -0.48             -0.97
#> Coleman.C2                    -0.79              -0.68             -0.96
#> Coleman.Liau                  -0.81              -0.72             -0.96
#> Coleman.Liau.grade             0.81               0.72              0.96
#> Coleman.Liau.short             0.81               0.72              0.96
#> Dale.Chall                    -0.98              -1.00             -0.58
#> Dale.Chall.old                 0.98               1.00              0.58
#> Dale.Chall.PSK                 0.98               1.00              0.58
#> Danielson.Bryan                1.00               0.98              0.73
#> Danielson.Bryan.2             -0.34              -0.19             -0.88
#> Dickes.Steiwer                -0.39              -0.41             -0.22
#> DRP                            0.99               1.00              0.59
#> ELF                            1.00               0.98              0.70
#> Farr.Jenkins.Paterson         -0.99              -1.00             -0.58
#> Flesch                        -0.97              -0.93             -0.84
#> Flesch.PSK                     0.99               0.96              0.79
#> Flesch.Kincaid                 1.00               0.98              0.71
#>  [ reached getOption("max.print") -- omitted 26 rows ]

textstat_readability(data_corpus_inaugural, measure = "Flesch.Kincaid")

#> 1789-Washington 1793-Washington      1797-Adams  1801-Jefferson  1805-Jefferson 
#>       28.423909       17.151759       28.692087       19.295924       21.917088 
#>    1809-Madison    1813-Madison     1817-Monroe     1821-Monroe      1825-Adams 
#>       25.557611       18.034452       14.443950       16.738786       19.748273 
#>    1829-Jackson    1833-Jackson   1837-VanBuren   1841-Harrison       1845-Polk 
#>       21.893148       19.958757       20.263524       19.355855       16.403854 
#>     1849-Taylor     1853-Pierce   1857-Buchanan    1861-Lincoln    1865-Lincoln 
#>       23.748805       16.434709       16.110868       13.648965       11.774929 
#>      1869-Grant      1873-Grant      1877-Hayes   1881-Garfield  1885-Cleveland 
#>       14.008426       15.064832       20.288377       14.027181       19.052937 
#>   1889-Harrison  1893-Cleveland   1897-McKinley   1901-McKinley  1905-Roosevelt 
#>       14.933849       18.248838       15.836491       12.675439       13.959254 
#>       1909-Taft     1913-Wilson     1917-Wilson    1921-Harding   1925-Coolidge 
#>       17.093249       11.707659       11.780913       13.214469       11.623790 
#>     1929-Hoover  1933-Roosevelt  1937-Roosevelt  1941-Roosevelt  1945-Roosevelt 
#>       13.259377       11.640683       10.606612        9.652893        9.133462 
#>     1949-Truman 1953-Eisenhower 1957-Eisenhower    1961-Kennedy    1965-Johnson 
#>       11.323070       10.285647        8.094908       11.805733        7.526843 
#>      1969-Nixon      1973-Nixon     1977-Carter     1981-Reagan     1985-Reagan 
#>        9.230108       12.269353       11.641868        9.788920       10.360629 
#>       1989-Bush    1993-Clinton    1997-Clinton       2001-Bush       2005-Bush 
#>        7.136847       10.366829        9.779651        8.910756       10.954613 
#>      2009-Obama      2013-Obama 
#>       10.229426       11.706752 

inaugReadability <- textstat_readability(data_corpus_inaugural, "all")
round(cor(inaugReadability), 2)

#>                         ARI ARI.simple Bormuth Bormuth.GP Coleman Coleman.C2
#> ARI                    1.00       1.00   -0.99       0.90   -0.58      -0.77
#> ARI.simple             1.00       1.00   -0.99       0.90   -0.58      -0.76
#> Bormuth               -0.99      -0.99    1.00      -0.93    0.48       0.68
#> Bormuth.GP             0.90       0.90   -0.93       1.00   -0.34      -0.51
#> Coleman               -0.58      -0.58    0.48      -0.34    1.00       0.96
#> Coleman.C2            -0.77      -0.76    0.68      -0.51    0.96       1.00
#> Coleman.Liau          -0.81      -0.80    0.73      -0.56    0.91       0.97
#> Coleman.Liau.grade     0.81       0.80   -0.73       0.56   -0.91      -0.97
#> Coleman.Liau.short     0.81       0.80   -0.73       0.56   -0.91      -0.97
#> Dale.Chall            -0.99      -0.99    1.00      -0.93    0.48       0.68
#> Dale.Chall.old         0.99       0.99   -1.00       0.93   -0.48      -0.68
#> Dale.Chall.PSK         0.99       0.99   -1.00       0.93   -0.48      -0.68
#> Danielson.Bryan        1.00       1.00   -0.98       0.88   -0.64      -0.80
#> Danielson.Bryan.2     -0.32      -0.32    0.20      -0.04    0.89       0.80
#> Dickes.Steiwer        -0.39      -0.39    0.41      -0.34    0.14       0.25
#> DRP                    0.99       0.99   -1.00       0.93   -0.48      -0.68
#> ELF                    1.00       1.00   -0.99       0.90   -0.62      -0.79
#> Farr.Jenkins.Paterson -0.99      -0.99    1.00      -0.93    0.48       0.68
#> Flesch                -0.97      -0.97    0.93      -0.82    0.76       0.89
#> Flesch.PSK             0.99       0.98   -0.96       0.86   -0.70      -0.86
#> Flesch.Kincaid         1.00       1.00   -0.99       0.90   -0.62      -0.79
#>                       Coleman.Liau Coleman.Liau.grade Coleman.Liau.short
#> ARI                          -0.81               0.81               0.81
#> ARI.simple                   -0.80               0.80               0.80
#> Bormuth                       0.73              -0.73              -0.73
#> Bormuth.GP                   -0.56               0.56               0.56
#> Coleman                       0.91              -0.91              -0.91
#> Coleman.C2                    0.97              -0.97              -0.97
#> Coleman.Liau                  1.00              -1.00              -1.00
#> Coleman.Liau.grade           -1.00               1.00               1.00
#> Coleman.Liau.short           -1.00               1.00               1.00
#> Dale.Chall                    0.72              -0.72              -0.72
#> Dale.Chall.old               -0.72               0.72               0.72
#> Dale.Chall.PSK               -0.72               0.72               0.72
#> Danielson.Bryan              -0.85               0.85               0.85
#> Danielson.Bryan.2             0.81              -0.81              -0.81
#> Dickes.Steiwer                0.23              -0.23              -0.23
#> DRP                          -0.73               0.73               0.73
#> ELF                          -0.81               0.81               0.81
#> Farr.Jenkins.Paterson         0.72              -0.72              -0.72
#> Flesch                        0.91              -0.91              -0.91
#> Flesch.PSK                   -0.88               0.88               0.88
#> Flesch.Kincaid               -0.82               0.82               0.82
#>                       Dale.Chall Dale.Chall.old Dale.Chall.PSK Danielson.Bryan
#> ARI                        -0.99           0.99           0.99            1.00
#> ARI.simple                 -0.99           0.99           0.99            1.00
#> Bormuth                     1.00          -1.00          -1.00           -0.98
#> Bormuth.GP                 -0.93           0.93           0.93            0.88
#> Coleman                     0.48          -0.48          -0.48           -0.64
#> Coleman.C2                  0.68          -0.68          -0.68           -0.80
#> Coleman.Liau                0.72          -0.72          -0.72           -0.85
#> Coleman.Liau.grade         -0.72           0.72           0.72            0.85
#> Coleman.Liau.short         -0.72           0.72           0.72            0.85
#> Dale.Chall                  1.00          -1.00          -1.00           -0.98
#> Dale.Chall.old             -1.00           1.00           1.00            0.98
#> Dale.Chall.PSK             -1.00           1.00           1.00            0.98
#> Danielson.Bryan            -0.98           0.98           0.98            1.00
#> Danielson.Bryan.2           0.19          -0.19          -0.19           -0.39
#> Dickes.Steiwer              0.41          -0.41          -0.41           -0.38
#> DRP                        -1.00           1.00           1.00            0.98
#> ELF                        -0.98           0.98           0.98            1.00
#> Farr.Jenkins.Paterson       1.00          -1.00          -1.00           -0.98
#> Flesch                      0.93          -0.93          -0.93           -0.98
#> Flesch.PSK                 -0.96           0.96           0.96            0.99
#> Flesch.Kincaid             -0.98           0.98           0.98            1.00
#>                       Danielson.Bryan.2 Dickes.Steiwer   DRP   ELF
#> ARI                               -0.32          -0.39  0.99  1.00
#> ARI.simple                        -0.32          -0.39  0.99  1.00
#> Bormuth                            0.20           0.41 -1.00 -0.99
#> Bormuth.GP                        -0.04          -0.34  0.93  0.90
#> Coleman                            0.89           0.14 -0.48 -0.62
#> Coleman.C2                         0.80           0.25 -0.68 -0.79
#> Coleman.Liau                       0.81           0.23 -0.73 -0.81
#> Coleman.Liau.grade                -0.81          -0.23  0.73  0.81
#> Coleman.Liau.short                -0.81          -0.23  0.73  0.81
#> Dale.Chall                         0.19           0.41 -1.00 -0.98
#> Dale.Chall.old                    -0.19          -0.41  1.00  0.98
#> Dale.Chall.PSK                    -0.19          -0.41  1.00  0.98
#> Danielson.Bryan                   -0.39          -0.38  0.98  1.00
#> Danielson.Bryan.2                  1.00           0.00 -0.20 -0.34
#> Dickes.Steiwer                     0.00           1.00 -0.41 -0.39
#> DRP                               -0.20          -0.41  1.00  0.99
#> ELF                               -0.34          -0.39  0.99  1.00
#> Farr.Jenkins.Paterson              0.19           0.41 -1.00 -0.99
#> Flesch                             0.53           0.37 -0.93 -0.97
#> Flesch.PSK                        -0.46          -0.38  0.96  0.99
#> Flesch.Kincaid                    -0.35          -0.40  0.99  1.00
#>                       Farr.Jenkins.Paterson Flesch Flesch.PSK Flesch.Kincaid
#> ARI                                   -0.99  -0.97       0.99           1.00
#> ARI.simple                            -0.99  -0.97       0.98           1.00
#> Bormuth                                1.00   0.93      -0.96          -0.99
#> Bormuth.GP                            -0.93  -0.82       0.86           0.90
#> Coleman                                0.48   0.76      -0.70          -0.62
#> Coleman.C2                             0.68   0.89      -0.86          -0.79
#> Coleman.Liau                           0.72   0.91      -0.88          -0.82
#> Coleman.Liau.grade                    -0.72  -0.91       0.88           0.82
#> Coleman.Liau.short                    -0.72  -0.91       0.88           0.82
#> Dale.Chall                             1.00   0.93      -0.96          -0.98
#> Dale.Chall.old                        -1.00  -0.93       0.96           0.98
#> Dale.Chall.PSK                        -1.00  -0.93       0.96           0.98
#> Danielson.Bryan                       -0.98  -0.98       0.99           1.00
#> Danielson.Bryan.2                      0.19   0.53      -0.46          -0.35
#> Dickes.Steiwer                         0.41   0.37      -0.38          -0.40
#> DRP                                   -1.00  -0.93       0.96           0.99
#> ELF                                   -0.99  -0.97       0.99           1.00
#> Farr.Jenkins.Paterson                  1.00   0.93      -0.96          -0.99
#> Flesch                                 0.93   1.00      -1.00          -0.98
#> Flesch.PSK                            -0.96  -1.00       1.00           0.99
#> Flesch.Kincaid                        -0.99  -0.98       0.99           1.00
#>                         FOG FOG.PSK FOG.NRI FORCAST FORCAST.RGL Fucks
#> ARI                    1.00    1.00    0.49    0.58        0.58  1.00
#> ARI.simple             0.99    1.00    0.49    0.58        0.58  1.00
#> Bormuth               -0.98   -1.00   -0.46   -0.48       -0.48 -1.00
#> Bormuth.GP             0.88    0.91    0.34    0.34        0.34  0.92
#> Coleman               -0.63   -0.54   -0.49   -1.00       -1.00 -0.53
#> Coleman.C2            -0.80   -0.74   -0.55   -0.96       -0.96 -0.73
#> Coleman.Liau          -0.84   -0.77   -0.55   -0.91       -0.91 -0.77
#> Coleman.Liau.grade     0.84    0.77    0.55    0.91        0.91  0.77
#> Coleman.Liau.short     0.84    0.77    0.55    0.91        0.91  0.77
#> Dale.Chall            -0.98   -1.00   -0.45   -0.48       -0.48 -1.00
#> Dale.Chall.old         0.98    1.00    0.45    0.48        0.48  1.00
#> Dale.Chall.PSK         0.98    1.00    0.45    0.48        0.48  1.00
#> Danielson.Bryan        1.00    0.99    0.51    0.64        0.64  0.99
#> Danielson.Bryan.2     -0.38   -0.27   -0.39   -0.89       -0.89 -0.26
#> Dickes.Steiwer        -0.42   -0.41    0.13   -0.14       -0.14 -0.39
#> DRP                    0.98    1.00    0.46    0.48        0.48  1.00
#> ELF                    0.99    0.99    0.50    0.62        0.62  0.99
#> Farr.Jenkins.Paterson -0.98   -1.00   -0.46   -0.48       -0.48 -1.00
#> Flesch                -0.98   -0.96   -0.54   -0.76       -0.76 -0.95
#> Flesch.PSK             0.99    0.98    0.53    0.70        0.70  0.97
#> Flesch.Kincaid         1.00    1.00    0.50    0.62        0.62  0.99
#>                       Linsear.Write   LIW   nWS nWS.2 nWS.3 nWS.4   RIX
#> ARI                            0.87  1.00  0.97  0.98  0.98  0.99  0.99
#> ARI.simple                     0.87  0.99  0.97  0.97  0.98  0.99  0.99
#> Bormuth                       -0.84 -0.98 -0.94 -0.94 -0.96 -0.98 -0.98
#> Bormuth.GP                     0.76  0.88  0.82  0.83  0.85  0.88  0.89
#> Coleman                       -0.68 -0.65 -0.75 -0.73 -0.69 -0.64 -0.63
#> Coleman.C2                    -0.79 -0.81 -0.89 -0.88 -0.84 -0.81 -0.80
#> Coleman.Liau                  -0.83 -0.85 -0.91 -0.90 -0.88 -0.84 -0.83
#> Coleman.Liau.grade             0.83  0.85  0.91  0.90  0.88  0.84  0.83
#> Coleman.Liau.short             0.83  0.85  0.91  0.90  0.88  0.84  0.83
#> Dale.Chall                    -0.84 -0.98 -0.93 -0.94 -0.96 -0.98 -0.98
#> Dale.Chall.old                 0.84  0.98  0.93  0.94  0.96  0.98  0.98
#> Dale.Chall.PSK                 0.84  0.98  0.93  0.94  0.96  0.98  0.98
#> Danielson.Bryan                0.89  1.00  0.98  0.99  0.99  1.00  0.99
#> Danielson.Bryan.2             -0.50 -0.39 -0.52 -0.50 -0.45 -0.38 -0.37
#> Dickes.Steiwer                -0.53 -0.40 -0.39 -0.40 -0.42 -0.42 -0.40
#> DRP                            0.84  0.98  0.94  0.94  0.96  0.98  0.98
#> ELF                            0.88  1.00  0.98  0.98  0.99  0.99  1.00
#> Farr.Jenkins.Paterson         -0.84 -0.98 -0.93 -0.94 -0.96 -0.98 -0.98
#> Flesch                        -0.92 -0.98 -1.00 -1.00 -0.99 -0.98 -0.98
#> Flesch.PSK                     0.91  0.99  1.00  1.00  1.00  0.99  0.99
#> Flesch.Kincaid                 0.89  1.00  0.98  0.98  0.99  1.00  0.99
#>                       Scrabble  SMOG SMOG.C SMOG.simple SMOG.de Spache
#> ARI                      -0.03  0.97   0.97        0.97    0.97   0.99
#> ARI.simple               -0.03  0.97   0.97        0.97    0.97   0.99
#> Bormuth                  -0.01 -0.94  -0.94       -0.94   -0.94  -1.00
#> Bormuth.GP               -0.03  0.82   0.82        0.82    0.82   0.93
#> Coleman                   0.36 -0.72  -0.72       -0.72   -0.72  -0.48
#> Coleman.C2                0.26 -0.87  -0.87       -0.87   -0.87  -0.68
#> Coleman.Liau              0.21 -0.90  -0.89       -0.90   -0.90  -0.72
#> Coleman.Liau.grade       -0.21  0.90   0.89        0.90    0.90   0.72
#> Coleman.Liau.short       -0.21  0.90   0.89        0.90    0.90   0.72
#> Dale.Chall               -0.02 -0.94  -0.94       -0.94   -0.94  -1.00
#> Dale.Chall.old            0.02  0.94   0.94        0.94    0.94   1.00
#> Dale.Chall.PSK            0.02  0.94   0.94        0.94    0.94   1.00
#> Danielson.Bryan          -0.06  0.98   0.98        0.98    0.98   0.98
#> Danielson.Bryan.2         0.32 -0.49  -0.48       -0.49   -0.49  -0.19
#> Dickes.Steiwer           -0.46 -0.42  -0.42       -0.42   -0.42  -0.41
#> DRP                       0.01  0.94   0.94        0.94    0.94   1.00
#> ELF                      -0.06  0.98   0.98        0.98    0.98   0.98
#> Farr.Jenkins.Paterson    -0.01 -0.94  -0.94       -0.94   -0.94  -1.00
#> Flesch                    0.12 -0.99  -0.99       -0.99   -0.99  -0.93
#> Flesch.PSK               -0.09  0.99   0.99        0.99    0.99   0.96
#> Flesch.Kincaid           -0.05  0.98   0.98        0.98    0.98   0.98
#>                       Spache.old Strain Traenkle.Bailer Traenkle.Bailer.2
#> ARI                         0.99   1.00           -1.00             -0.75
#> ARI.simple                  0.99   1.00           -1.00             -0.75
#> Bormuth                    -1.00  -1.00            1.00              0.66
#> Bormuth.GP                  0.93   0.92           -0.91             -0.51
#> Coleman                    -0.48  -0.55            0.55              0.92
#> Coleman.C2                 -0.68  -0.74            0.74              0.96
#> Coleman.Liau               -0.72  -0.77            0.78              0.99
#> Coleman.Liau.grade          0.72   0.77           -0.78             -0.99
#> Coleman.Liau.short          0.72   0.77           -0.78             -0.99
#> Dale.Chall                 -1.00  -1.00            1.00              0.66
#> Dale.Chall.old              1.00   1.00           -1.00             -0.66
#> Dale.Chall.PSK              1.00   1.00           -1.00             -0.66
#> Danielson.Bryan             0.98   0.99           -0.99             -0.80
#> Danielson.Bryan.2          -0.19  -0.27            0.28              0.86
#> Dickes.Steiwer             -0.41  -0.40            0.39              0.18
#> DRP                         1.00   1.00           -1.00             -0.66
#> ELF                         0.98   1.00           -0.99             -0.76
#> Farr.Jenkins.Paterson      -1.00  -1.00            1.00              0.66
#> Flesch                     -0.93  -0.96            0.96              0.88
#> Flesch.PSK                  0.96   0.98           -0.98             -0.84
#> Flesch.Kincaid              0.98   1.00           -1.00             -0.77
#>                       Wheeler.Smith meanSentenceLength meanWordSyllables
#> ARI                            1.00               0.99              0.68
#> ARI.simple                     1.00               0.99              0.68
#> Bormuth                       -0.99              -1.00             -0.59
#> Bormuth.GP                     0.90               0.93              0.45
#> Coleman                       -0.62              -0.48             -0.97
#> Coleman.C2                    -0.79              -0.68             -0.96
#> Coleman.Liau                  -0.81              -0.72             -0.96
#> Coleman.Liau.grade             0.81               0.72              0.96
#> Coleman.Liau.short             0.81               0.72              0.96
#> Dale.Chall                    -0.98              -1.00             -0.58
#> Dale.Chall.old                 0.98               1.00              0.58
#> Dale.Chall.PSK                 0.98               1.00              0.58
#> Danielson.Bryan                1.00               0.98              0.73
#> Danielson.Bryan.2             -0.34              -0.19             -0.88
#> Dickes.Steiwer                -0.39              -0.41             -0.22
#> DRP                            0.99               1.00              0.59
#> ELF                            1.00               0.98              0.70
#> Farr.Jenkins.Paterson         -0.99              -1.00             -0.58
#> Flesch                        -0.97              -0.93             -0.84
#> Flesch.PSK                     0.99               0.96              0.79
#> Flesch.Kincaid                 1.00               0.98              0.71
#>  [ reached getOption("max.print") -- omitted 26 rows ]