similarity or distance between dfm documents and/or features

Compute a similarity or distance matrix between documents and/or features from a dfm and return a standard dist object.

textstat_dist(x, selection = character(0), n = NULL,
  margin = c("documents", "features"), method = "euclidean", tri = TRUE,
  diag = FALSE, dist = TRUE)

textstat_simil(x, selection = character(0), n = NULL,
  margin = c("documents", "features"), method = "correlation",
  tri = FALSE, diag = FALSE)

Arguments

x

a dfm object

selection

character or character vector of document names or feature labels from the dfm

n

the top n highest-ranking items will be returned. If n is NULL, return all items.

margin

identifies the margin of the dfm on which similarity or difference will be computed: documents for documents or features for word/term features.

method

method the distance measure to be used; see Details

tri

whether the upper triangle of the symmetric \(V \times V\) matrix is recorded

diag

whether the diagonal of the distance matrix should be recorded

dist

whether the distance matrix should be converted into an object of class "dist". The distance matrix created from some methods, such as "kullback", is not symmetric.

Details

textstat_dist options are: "euclidean" (default), "euclidean", "hamming", "Chisquared", "Chisquared2", and "kullback".

textstat_simil options are: "correlation" (default), "cosine", "correlation", "jaccard", "eJaccard", "dice", "eDice", "simple matching", "hamann", and "faith".

Note

If you want to compute similarity on a "normalized" dfm object (controlling for variable document lengths, for methods such as correlation for which different document lengths matter), then wrap the input dfm in weight(x, "relFreq").

See also

textstat_dist, as.list.dist

Examples

# create a dfm from inaugural addresses from Reagan onwards
presDfm <- dfm(corpus_subset(inaugCorpus, Year > 1980),
               remove = stopwords("english"), stem = TRUE)

#> 

## distance

# compute some document distances
(tmp <- textstat_dist(presDfm, margin = "documents"))

#>              1981-Reagan 1985-Reagan 1989-Bush 1993-Clinton 1997-Clinton
#> 1985-Reagan     57.68015                                                
#> 1989-Bush       64.79969    69.94283                                    
#> 1993-Clinton    91.20855    89.92219  92.27134                          
#> 1997-Clinton    83.06022    79.58643  83.13844     72.92462             
#> 2001-Bush       95.72356   101.74478  93.68031     64.32729     72.31874
#> 2005-Bush       93.33274    96.51943  92.98387     70.39886     79.44810
#> 2009-Obama      81.83520    80.75890  80.61017     84.88816     67.29042
#> 2013-Obama     103.88455   104.52272 104.20653     70.46276     71.86793
#>              2001-Bush 2005-Bush 2009-Obama
#> 1985-Reagan                                
#> 1989-Bush                                  
#> 1993-Clinton                               
#> 1997-Clinton                               
#> 2001-Bush                                  
#> 2005-Bush     56.23166                     
#> 2009-Obama    78.24321  85.39321           
#> 2013-Obama    54.69004  66.68583   72.10409

# for specific comparisons
textstat_dist(presDfm, "1985-Reagan", n = 5, margin = "documents")

#> Warning: non-square matrix

#>              1985-Reagan 1981-Reagan 1989-Bush 1993-Clinton
#> 1981-Reagan     57.68015                                   
#> 1989-Bush       69.94283          NA                       
#> 1993-Clinton    89.92219          NA        NA             
#> 1997-Clinton    79.58643          NA        NA           NA

textstat_dist(presDfm, c("2009-Obama" , "2013-Obama"), n = 5, margin = "documents")

#> Warning: non-square matrix

#>             2009-Obama 2013-Obama 1981-Reagan 1985-Reagan
#> 2013-Obama    72.10409                                   
#> 1981-Reagan   81.83520  103.88455                        
#> 1985-Reagan   80.75890  104.52272          NA            
#> 1989-Bush     80.61017  104.20653          NA          NA

textstat_dist(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents")

#>              2009-Obama 2013-Obama 1981-Reagan 1985-Reagan 1989-Bush
#> 2013-Obama     72.10409                                             
#> 1981-Reagan    81.83520  103.88455                                  
#> 1985-Reagan    80.75890  104.52272          NA                      
#> 1989-Bush      80.61017  104.20653          NA          NA          
#> 1993-Clinton   84.88816   70.46276          NA          NA        NA
#> 1997-Clinton   67.29042   71.86793          NA          NA        NA
#> 2001-Bush      78.24321   54.69004          NA          NA        NA
#> 2005-Bush      85.39321   66.68583          NA          NA        NA
#>              1993-Clinton 1997-Clinton 2001-Bush
#> 2013-Obama                                      
#> 1981-Reagan                                     
#> 1985-Reagan                                     
#> 1989-Bush                                       
#> 1993-Clinton                                    
#> 1997-Clinton           NA                       
#> 2001-Bush              NA           NA          
#> 2005-Bush              NA           NA        NA

textstat_dist(presDfm, "2005-Bush", margin = "documents", method = "eJaccard")

#>              2005-Bush 1981-Reagan 1985-Reagan 1989-Bush 1993-Clinton
#> 1981-Reagan  0.1905877                                               
#> 1985-Reagan  0.2007802          NA                                   
#> 1989-Bush    0.1897176          NA          NA                       
#> 1993-Clinton 0.1558785          NA          NA        NA             
#> 1997-Clinton 0.1747363          NA          NA        NA           NA
#> 2001-Bush    0.1112871          NA          NA        NA           NA
#> 2009-Obama   0.1948222          NA          NA        NA           NA
#> 2013-Obama   0.1583972          NA          NA        NA           NA
#>              1997-Clinton 2001-Bush 2009-Obama
#> 1981-Reagan                                   
#> 1985-Reagan                                   
#> 1989-Bush                                     
#> 1993-Clinton                                  
#> 1997-Clinton                                  
#> 2001-Bush              NA                     
#> 2009-Obama             NA        NA           
#> 2013-Obama             NA        NA         NA

## similarities

# compute some document similarities
(tmp <- textstat_simil(presDfm, margin = "documents"))

#>              1981-Reagan 1985-Reagan 1989-Bush 1993-Clinton 1997-Clinton
#> 1981-Reagan                0.9685409 0.9599217    0.9391137    0.9386791
#> 1985-Reagan    0.9685409             0.9537449    0.9452609    0.9468983
#> 1989-Bush      0.9599217   0.9537449              0.9358174    0.9381904
#> 1993-Clinton   0.9391137   0.9452609 0.9358174                 0.9241745
#> 1997-Clinton   0.9386791   0.9468983 0.9381904    0.9241745             
#> 2001-Bush      0.9498797   0.9371453 0.9543776    0.9258829    0.9321614
#> 2005-Bush      0.9346990   0.9297108 0.9349331    0.9142198    0.9077993
#> 2009-Obama     0.9387730   0.9422163 0.9407657    0.9000505    0.9382904
#> 2013-Obama     0.9389438   0.9409353 0.9370933    0.9118394    0.9387808
#>              2001-Bush 2005-Bush 2009-Obama 2013-Obama
#> 1981-Reagan  0.9498797 0.9346990  0.9387730  0.9389438
#> 1985-Reagan  0.9371453 0.9297108  0.9422163  0.9409353
#> 1989-Bush    0.9543776 0.9349331  0.9407657  0.9370933
#> 1993-Clinton 0.9258829 0.9142198  0.9000505  0.9118394
#> 1997-Clinton 0.9321614 0.9077993  0.9382904  0.9387808
#> 2001-Bush              0.9437295  0.9240388  0.9363852
#> 2005-Bush    0.9437295            0.8976144  0.9195065
#> 2009-Obama   0.9240388 0.8976144             0.9458185
#> 2013-Obama   0.9363852 0.9195065  0.9458185           

# output as a list
as.list(tmp)[1:2]

#> $`1981-Reagan`
#>  1985-Reagan    1989-Bush    2001-Bush 1993-Clinton   2013-Obama   2009-Obama 
#>    0.9685409    0.9599217    0.9498797    0.9391137    0.9389438    0.9387730 
#> 1997-Clinton    2005-Bush 
#>    0.9386791    0.9346990 
#> 
#> $`1985-Reagan`
#>  1981-Reagan    1989-Bush 1997-Clinton 1993-Clinton   2009-Obama   2013-Obama 
#>    0.9685409    0.9537449    0.9468983    0.9452609    0.9422163    0.9409353 
#>    2001-Bush    2005-Bush 
#>    0.9371453    0.9297108 
#> 

# for specific comparisons
textstat_simil(presDfm, "1985-Reagan", n = 5, margin = "documents")

#> Warning: non-square matrix

#>              1985-Reagan 1981-Reagan 1989-Bush 1993-Clinton 1997-Clinton
#> 1985-Reagan                0.9685409 0.9537449    0.9452609    0.9468983
#> 1981-Reagan    0.9685409                    NA           NA           NA
#> 1989-Bush      0.9537449          NA                     NA           NA
#> 1993-Clinton   0.9452609          NA        NA                        NA
#> 1997-Clinton   0.9468983          NA        NA           NA             

textstat_simil(presDfm, c("2009-Obama" , "2013-Obama"), n = 5, margin = "documents")

#> Warning: non-square matrix

#>             2009-Obama 2013-Obama 1981-Reagan 1985-Reagan 1989-Bush
#> 2009-Obama              0.9458185   0.9387730   0.9422163 0.9407657
#> 2013-Obama   0.9458185              0.9389438   0.9409353 0.9370933
#> 1981-Reagan  0.9387730  0.9389438                      NA        NA
#> 1985-Reagan  0.9422163  0.9409353          NA                    NA
#> 1989-Bush    0.9407657  0.9370933          NA          NA          

textstat_simil(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents")

#>              2009-Obama 2013-Obama 1981-Reagan 1985-Reagan 1989-Bush
#> 2009-Obama               0.9458185   0.9387730   0.9422163 0.9407657
#> 2013-Obama    0.9458185              0.9389438   0.9409353 0.9370933
#> 1981-Reagan   0.9387730  0.9389438                      NA        NA
#> 1985-Reagan   0.9422163  0.9409353          NA                    NA
#> 1989-Bush     0.9407657  0.9370933          NA          NA          
#> 1993-Clinton  0.9000505  0.9118394          NA          NA        NA
#> 1997-Clinton  0.9382904  0.9387808          NA          NA        NA
#> 2001-Bush     0.9240388  0.9363852          NA          NA        NA
#> 2005-Bush     0.8976144  0.9195065          NA          NA        NA
#>              1993-Clinton 1997-Clinton 2001-Bush 2005-Bush
#> 2009-Obama      0.9000505    0.9382904 0.9240388 0.8976144
#> 2013-Obama      0.9118394    0.9387808 0.9363852 0.9195065
#> 1981-Reagan            NA           NA        NA        NA
#> 1985-Reagan            NA           NA        NA        NA
#> 1989-Bush              NA           NA        NA        NA
#> 1993-Clinton                        NA        NA        NA
#> 1997-Clinton           NA                     NA        NA
#> 2001-Bush              NA           NA                  NA
#> 2005-Bush              NA           NA        NA          

textstat_simil(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents", method = "cosine")

#>              2009-Obama 2013-Obama 1981-Reagan 1985-Reagan 1989-Bush
#> 2009-Obama               0.9473229   0.9399460   0.9436339 0.9416945
#> 2013-Obama    0.9473229              0.9397146   0.9422001 0.9376157
#> 1981-Reagan   0.9399460  0.9397146                      NA        NA
#> 1985-Reagan   0.9436339  0.9422001          NA                    NA
#> 1989-Bush     0.9416945  0.9376157          NA          NA          
#> 1993-Clinton  0.9015783  0.9125611          NA          NA        NA
#> 1997-Clinton  0.9399306  0.9403480          NA          NA        NA
#> 2001-Bush     0.9255036  0.9371368          NA          NA        NA
#> 2005-Bush     0.9004034  0.9217499          NA          NA        NA
#>              1993-Clinton 1997-Clinton 2001-Bush 2005-Bush
#> 2009-Obama      0.9015783    0.9399306 0.9255036 0.9004034
#> 2013-Obama      0.9125611    0.9403480 0.9371368 0.9217499
#> 1981-Reagan            NA           NA        NA        NA
#> 1985-Reagan            NA           NA        NA        NA
#> 1989-Bush              NA           NA        NA        NA
#> 1993-Clinton                        NA        NA        NA
#> 1997-Clinton           NA                     NA        NA
#> 2001-Bush              NA           NA                  NA
#> 2005-Bush              NA           NA        NA          

# compute some term similarities
textstat_simil(presDfm, c("fair", "health", "terror"), method = "cosine",
               margin = "features", 20)

#> Warning: non-square matrix

#>                fair     terror     health      senat   hatfield          ,
#> fair                0.43033148 0.51639778 0.57535596 0.51639778 0.78760215
#> terror   0.43033148            0.44444444 0.24759378 0.66666667 0.58327489
#> health   0.51639778 0.44444444            0.00000000 0.00000000 0.67707179
#> senat    0.57535596 0.24759378 0.00000000                    NA         NA
#> hatfield 0.51639778 0.66666667 0.00000000         NA                    NA
#> ,        0.78760215 0.58327489 0.67707179         NA         NA           
#> mr       0.26363525 0.29172998 0.09724333         NA         NA         NA
#> .        0.76454880 0.59341577 0.70730365         NA         NA         NA
#> chief    0.57735027 0.29814240 0.29814240         NA         NA         NA
#> justic   0.41403934 0.17817416 0.53452248         NA         NA         NA
#> presid   0.66914961 0.43193421 0.43193421         NA         NA         NA
#> vice     0.77459667 0.44444444 0.33333333         NA         NA         NA
#> bush     0.69282032 0.44721360 0.59628479         NA         NA         NA
#> mondal   0.51639778 0.66666667 0.00000000         NA         NA         NA
#> baker    0.51639778 0.66666667 0.00000000         NA         NA         NA
#> speaker  0.31139958 0.20100756 0.00000000         NA         NA         NA
#> o'neil   0.73029674 0.47140452 0.00000000         NA         NA         NA
#> reverend 0.59628479 0.38490018 0.19245009         NA         NA         NA
#> moomaw   0.51639778 0.66666667 0.00000000         NA         NA         NA
#> fellow   0.55138280 0.64993368 0.64993368         NA         NA         NA
#>                  mr          .      chief     justic     presid       vice
#> fair     0.26363525 0.76454880 0.57735027 0.41403934 0.66914961 0.77459667
#> terror   0.29172998 0.59341577 0.29814240 0.17817416 0.43193421 0.44444444
#> health   0.09724333 0.70730365 0.29814240 0.53452248 0.43193421 0.33333333
#> senat            NA         NA         NA         NA         NA         NA
#> hatfield         NA         NA         NA         NA         NA         NA
#> ,                NA         NA         NA         NA         NA         NA
#> mr                          NA         NA         NA         NA         NA
#> .                NA                    NA         NA         NA         NA
#> chief            NA         NA                    NA         NA         NA
#> justic           NA         NA         NA                    NA         NA
#> presid           NA         NA         NA         NA                    NA
#> vice             NA         NA         NA         NA         NA           
#> bush             NA         NA         NA         NA         NA         NA
#> mondal           NA         NA         NA         NA         NA         NA
#> baker            NA         NA         NA         NA         NA         NA
#> speaker          NA         NA         NA         NA         NA         NA
#> o'neil           NA         NA         NA         NA         NA         NA
#> reverend         NA         NA         NA         NA         NA         NA
#> moomaw           NA         NA         NA         NA         NA         NA
#> fellow           NA         NA         NA         NA         NA         NA
#>                bush     mondal      baker    speaker     o'neil   reverend
#> fair     0.69282032 0.51639778 0.51639778 0.31139958 0.73029674 0.59628479
#> terror   0.44721360 0.66666667 0.66666667 0.20100756 0.47140452 0.38490018
#> health   0.59628479 0.00000000 0.00000000 0.00000000 0.00000000 0.19245009
#> senat            NA         NA         NA         NA         NA         NA
#> hatfield         NA         NA         NA         NA         NA         NA
#> ,                NA         NA         NA         NA         NA         NA
#> mr               NA         NA         NA         NA         NA         NA
#> .                NA         NA         NA         NA         NA         NA
#> chief            NA         NA         NA         NA         NA         NA
#> justic           NA         NA         NA         NA         NA         NA
#> presid           NA         NA         NA         NA         NA         NA
#> vice             NA         NA         NA         NA         NA         NA
#> bush                        NA         NA         NA         NA         NA
#> mondal           NA                    NA         NA         NA         NA
#> baker            NA         NA                    NA         NA         NA
#> speaker          NA         NA         NA                    NA         NA
#> o'neil           NA         NA         NA         NA                    NA
#> reverend         NA         NA         NA         NA         NA           
#> moomaw           NA         NA         NA         NA         NA         NA
#> fellow           NA         NA         NA         NA         NA         NA
#>              moomaw     fellow
#> fair     0.51639778 0.55138280
#> terror   0.66666667 0.64993368
#> health   0.00000000 0.64993368
#> senat            NA         NA
#> hatfield         NA         NA
#> ,                NA         NA
#> mr               NA         NA
#> .                NA         NA
#> chief            NA         NA
#> justic           NA         NA
#> presid           NA         NA
#> vice             NA         NA
#> bush             NA         NA
#> mondal           NA         NA
#> baker            NA         NA
#> speaker          NA         NA
#> o'neil           NA         NA
#> reverend         NA         NA
#> moomaw                      NA
#> fellow           NA