Similarity and distance computation between documents or features

These functions compute matrixes of distances and similarities between documents or features from a dfm and return a dist object (or a matrix if specific targets are selected).

textstat_dist(x, selection = NULL, n = NULL, margin = c("documents",
  "features"), method = "euclidean", upper = FALSE, diag = FALSE, p = 2)

textstat_simil(x, selection = NULL, n = NULL, margin = c("documents",
  "features"), method = "correlation", upper = FALSE, diag = FALSE)

Arguments

x

a dfm object

selection

character vector of document names or feature labels from x. A "dist" object is returned if selection is NULL, otherwise, a matrix is returned.

n

the top n highest-ranking items will be returned. If n is NULL, return all items. Useful if the output object will be coerced into a list, for instance if the top n most similar features to a target feature is desired. (See examples.)

margin

identifies the margin of the dfm on which similarity or difference will be computed: documents for documents or features for word/term features.

method

method the similarity or distance measure to be used; see Details

upper

whether the upper triangle of the symmetric \(V \times V\) matrix is recorded

diag

whether the diagonal of the distance matrix should be recorded

p

The power of the Minkowski distance.

Details

textstat_dist options are: "euclidean" (default), "Chisquared", "Chisquared2", "hamming", "kullback". "manhattan", "maximum", "canberra", and "minkowski".

textstat_simil options are: "correlation" (default), "cosine", "jaccard", "eJaccard", "dice", "eDice", "simple matching", "hamann", and "faith".

Note

If you want to compute similarity on a "normalized" dfm object (controlling for variable document lengths, for methods such as correlation for which different document lengths matter), then wrap the input dfm in weight(x, "relFreq").

See also

textstat_dist, as.list.dist, dist

Examples

# create a dfm from inaugural addresses from Reagan onwards
presDfm <- dfm(corpus_subset(data_corpus_inaugural, Year > 1990),
               remove = stopwords("english"), stem = TRUE, remove_punct = TRUE)

# distances for documents 
(d1 <- textstat_dist(presDfm, margin = "documents"))

#>              1993-Clinton 1997-Clinton 2001-Bush 2005-Bush 2009-Obama
#> 1997-Clinton     62.20932                                            
#> 2001-Bush        55.19058     63.76519                               
#> 2005-Bush        64.79197     73.48469  54.31390                     
#> 2009-Obama       53.48832     60.22458  50.70503  62.36185           
#> 2013-Obama       52.92447     61.21274  49.12230  57.97413   48.45617
#> 2017-Trump       61.70899     67.12675  51.67204  60.43178   58.77925
#>              2013-Obama
#> 1997-Clinton           
#> 2001-Bush              
#> 2005-Bush              
#> 2009-Obama             
#> 2013-Obama             
#> 2017-Trump     58.72819

as.matrix(d1)

#>              1993-Clinton 1997-Clinton 2001-Bush 2005-Bush 2009-Obama
#> 1993-Clinton      0.00000     62.20932  55.19058  64.79197   53.48832
#> 1997-Clinton     62.20932      0.00000  63.76519  73.48469   60.22458
#> 2001-Bush        55.19058     63.76519   0.00000  54.31390   50.70503
#> 2005-Bush        64.79197     73.48469  54.31390   0.00000   62.36185
#> 2009-Obama       53.48832     60.22458  50.70503  62.36185    0.00000
#> 2013-Obama       52.92447     61.21274  49.12230  57.97413   48.45617
#> 2017-Trump       61.70899     67.12675  51.67204  60.43178   58.77925
#>              2013-Obama 2017-Trump
#> 1993-Clinton   52.92447   61.70899
#> 1997-Clinton   61.21274   67.12675
#> 2001-Bush      49.12230   51.67204
#> 2005-Bush      57.97413   60.43178
#> 2009-Obama     48.45617   58.77925
#> 2013-Obama      0.00000   58.72819
#> 2017-Trump     58.72819    0.00000

# distances for specific documents
textstat_dist(presDfm, "2017-Trump", margin = "documents")

#>              2017-Trump
#> 2017-Trump      0.00000
#> 1993-Clinton   61.70899
#> 1997-Clinton   67.12675
#> 2001-Bush      51.67204
#> 2005-Bush      60.43178
#> 2009-Obama     58.77925
#> 2013-Obama     58.72819

textstat_dist(presDfm, "2005-Bush", margin = "documents", method = "eJaccard")

#>              2005-Bush
#> 2005-Bush    1.0000000
#> 1993-Clinton 0.3343904
#> 1997-Clinton 0.3753615
#> 2001-Bush    0.4614823
#> 2009-Obama   0.4031614
#> 2013-Obama   0.4555322
#> 2017-Trump   0.4284820

(d2 <- textstat_dist(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents"))

#>              2009-Obama 2013-Obama
#> 2009-Obama      0.00000   48.45617
#> 2013-Obama     48.45617    0.00000
#> 1993-Clinton   53.48832   52.92447
#> 1997-Clinton   60.22458   61.21274
#> 2001-Bush      50.70503   49.12230
#> 2005-Bush      62.36185   57.97413
#> 2017-Trump     58.77925   58.72819

as.list(d1)

#> $`1993-Clinton`
#>    2005-Bush 1997-Clinton   2017-Trump    2001-Bush   2009-Obama   2013-Obama 
#>     64.79197     62.20932     61.70899     55.19058     53.48832     52.92447 
#> 
#> $`1997-Clinton`
#>    2005-Bush   2017-Trump    2001-Bush 1993-Clinton   2013-Obama   2009-Obama 
#>     73.48469     67.12675     63.76519     62.20932     61.21274     60.22458 
#> 
#> $`2001-Bush`
#> 1997-Clinton 1993-Clinton    2005-Bush   2017-Trump   2009-Obama   2013-Obama 
#>     63.76519     55.19058     54.31390     51.67204     50.70503     49.12230 
#> 
#> $`2005-Bush`
#> 1997-Clinton 1993-Clinton   2009-Obama   2017-Trump   2013-Obama    2001-Bush 
#>     73.48469     64.79197     62.36185     60.43178     57.97413     54.31390 
#> 
#> $`2009-Obama`
#>    2005-Bush 1997-Clinton   2017-Trump 1993-Clinton    2001-Bush   2013-Obama 
#>     62.36185     60.22458     58.77925     53.48832     50.70503     48.45617 
#> 
#> $`2013-Obama`
#> 1997-Clinton   2017-Trump    2005-Bush 1993-Clinton    2001-Bush   2009-Obama 
#>     61.21274     58.72819     57.97413     52.92447     49.12230     48.45617 
#> 
#> $`2017-Trump`
#> 1997-Clinton 1993-Clinton    2005-Bush   2009-Obama   2013-Obama    2001-Bush 
#>     67.12675     61.70899     60.43178     58.77925     58.72819     51.67204 
#> 

# similarities for documents
(s1 <- textstat_simil(presDfm, method = "cosine", margin = "documents"))

#>              1993-Clinton 1997-Clinton 2001-Bush 2005-Bush 2009-Obama
#> 1997-Clinton    0.6660288                                            
#> 2001-Bush       0.5333062    0.6434983                               
#> 2005-Bush       0.5092769    0.5535663 0.6501808                     
#> 2009-Obama      0.6288108    0.6954614 0.6529447 0.5766800           
#> 2013-Obama      0.6278986    0.6826705 0.6656213 0.6292757  0.7144178
#> 2017-Trump      0.5033731    0.6136180 0.6381426 0.6021261  0.5867729
#>              2013-Obama
#> 1997-Clinton           
#> 2001-Bush              
#> 2005-Bush              
#> 2009-Obama             
#> 2013-Obama             
#> 2017-Trump    0.5796322

as.matrix(s1)

#>              1993-Clinton 1997-Clinton 2001-Bush 2005-Bush 2009-Obama
#> 1993-Clinton    1.0000000    0.6660288 0.5333062 0.5092769  0.6288108
#> 1997-Clinton    0.6660288    1.0000000 0.6434983 0.5535663  0.6954614
#> 2001-Bush       0.5333062    0.6434983 1.0000000 0.6501808  0.6529447
#> 2005-Bush       0.5092769    0.5535663 0.6501808 1.0000000  0.5766800
#> 2009-Obama      0.6288108    0.6954614 0.6529447 0.5766800  1.0000000
#> 2013-Obama      0.6278986    0.6826705 0.6656213 0.6292757  0.7144178
#> 2017-Trump      0.5033731    0.6136180 0.6381426 0.6021261  0.5867729
#>              2013-Obama 2017-Trump
#> 1993-Clinton  0.6278986  0.5033731
#> 1997-Clinton  0.6826705  0.6136180
#> 2001-Bush     0.6656213  0.6381426
#> 2005-Bush     0.6292757  0.6021261
#> 2009-Obama    0.7144178  0.5867729
#> 2013-Obama    1.0000000  0.5796322
#> 2017-Trump    0.5796322  1.0000000

as.list(s1)

#> $`1993-Clinton`
#> 1997-Clinton   2009-Obama   2013-Obama    2001-Bush    2005-Bush   2017-Trump 
#>    0.6660288    0.6288108    0.6278986    0.5333062    0.5092769    0.5033731 
#> 
#> $`1997-Clinton`
#>   2009-Obama   2013-Obama 1993-Clinton    2001-Bush   2017-Trump    2005-Bush 
#>    0.6954614    0.6826705    0.6660288    0.6434983    0.6136180    0.5535663 
#> 
#> $`2001-Bush`
#>   2013-Obama   2009-Obama    2005-Bush 1997-Clinton   2017-Trump 1993-Clinton 
#>    0.6656213    0.6529447    0.6501808    0.6434983    0.6381426    0.5333062 
#> 
#> $`2005-Bush`
#>    2001-Bush   2013-Obama   2017-Trump   2009-Obama 1997-Clinton 1993-Clinton 
#>    0.6501808    0.6292757    0.6021261    0.5766800    0.5535663    0.5092769 
#> 
#> $`2009-Obama`
#>   2013-Obama 1997-Clinton    2001-Bush 1993-Clinton   2017-Trump    2005-Bush 
#>    0.7144178    0.6954614    0.6529447    0.6288108    0.5867729    0.5766800 
#> 
#> $`2013-Obama`
#>   2009-Obama 1997-Clinton    2001-Bush    2005-Bush 1993-Clinton   2017-Trump 
#>    0.7144178    0.6826705    0.6656213    0.6292757    0.6278986    0.5796322 
#> 
#> $`2017-Trump`
#>    2001-Bush 1997-Clinton    2005-Bush   2009-Obama   2013-Obama 1993-Clinton 
#>    0.6381426    0.6136180    0.6021261    0.5867729    0.5796322    0.5033731 
#> 

# similarities for for specific documents
textstat_simil(presDfm, "2017-Trump", margin = "documents")

#>              2017-Trump
#> 2017-Trump    1.0000000
#> 1993-Clinton  0.4550233
#> 1997-Clinton  0.5758326
#> 2001-Bush     0.6015599
#> 2005-Bush     0.5606500
#> 2009-Obama    0.5389326
#> 2013-Obama    0.5327830

textstat_simil(presDfm, "2017-Trump", method = "cosine", margin = "documents")

#>              2017-Trump
#> 2017-Trump    1.0000000
#> 1993-Clinton  0.5033731
#> 1997-Clinton  0.6136180
#> 2001-Bush     0.6381426
#> 2005-Bush     0.6021261
#> 2009-Obama    0.5867729
#> 2013-Obama    0.5796322

textstat_simil(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents")

#>              2009-Obama 2013-Obama
#> 2009-Obama    1.0000000  0.6559745
#> 2013-Obama    0.6559745  1.0000000
#> 1993-Clinton  0.5725719  0.5752600
#> 1997-Clinton  0.6495037  0.6371584
#> 2001-Bush     0.5945800  0.6137637
#> 2005-Bush     0.5005582  0.5689492
#> 2017-Trump    0.5389326  0.5327830

# compute some term similarities
(s2 <- textstat_simil(presDfm, c("fair", "health", "terror"), method = "cosine",
                      margin = "features", n = 8))

#>              fair    health    terror
#> fair    1.0000000 0.7559289 0.1543033
#> health  0.7559289 1.0000000 0.5443311
#> terror  0.1543033 0.5443311 1.0000000
#> fellow  0.4265617 0.7181848 0.6701663
#> citizen 0.6787417 0.7144508 0.4966330
#> today   0.6265515 0.8288497 0.5986661
#> celebr  0.4472136 0.6761234 0.4830459
#> mysteri 0.2672612 0.2357023 0.2886751

as.list(s2)

#> $fair
#>    health   citizen     today    celebr    fellow   mysteri    terror 
#> 0.7559289 0.6787417 0.6265515 0.4472136 0.4265617 0.2672612 0.1543033 
#> 
#> $health
#>     today      fair    fellow   citizen    celebr    terror   mysteri 
#> 0.8288497 0.7559289 0.7181848 0.7144508 0.6761234 0.5443311 0.2357023 
#> 
#> $terror
#>    fellow     today    health   citizen    celebr   mysteri      fair 
#> 0.6701663 0.5986661 0.5443311 0.4966330 0.4830459 0.2886751 0.1543033 
#>