Distance matrix between documents and/or features

These functions compute distance matrix between documents and/or features from a dfm and return a standard dist object.

Compute a similarity or distance matrix between documents and/or features from a dfm and return a standard dist object.

textstat_dist(x, selection = character(0), n = NULL,
  margin = c("documents", "features"), method = "euclidean", upper = TRUE,
  diag = FALSE)

textstat_simil(x, selection = character(0), n = NULL,
  margin = c("documents", "features"), method = "correlation",
  upper = FALSE, diag = FALSE)

Arguments

x

a dfm object

selection

character or character vector of document names or feature labels from the dfm

n

the top n highest-ranking items will be returned. If n is NULL, return all items.

margin

identifies the margin of the dfm on which similarity or difference will be computed: documents for documents or features for word/term features.

method

method the distance measure to be used; see Details

upper

whether the upper triangle of the symmetric \(V \times V\) matrix is recorded

diag

whether the diagonal of the distance matrix should be recorded

Details

textstat_dist options are: "euclidean" (default), "euclidean", "hamming", "Chisquared", "Chisquared2", and "kullback".

textstat_simil options are: "correlation" (default), "cosine", "correlation", "jaccard", "eJaccard", "dice", "eDice", "simple matching", "hamann", and "faith".

Note

If you want to compute similarity on a "normalized" dfm object (controlling for variable document lengths, for methods such as correlation for which different document lengths matter), then wrap the input dfm in weight(x, "relFreq").

See also

dfm

textstat_dist, as.list.dist

Examples

# create a dfm from inaugural addresses from Reagan onwards
presDfm <- dfm(corpus_subset(inaugCorpus, Year > 1980),
               remove = stopwords("english"), stem = TRUE)

#> 

## distance

# compute some document distances
(tmp <- textstat_dist(presDfm, margin = "documents"))

#>              1981-Reagan 1985-Reagan 1989-Bush 1993-Clinton 1997-Clinton
#> 1981-Reagan                 57.68015  64.79969     91.20855     83.06022
#> 1985-Reagan     57.68015              69.94283     89.92219     79.58643
#> 1989-Bush       64.79969    69.94283               92.27134     83.13844
#> 1993-Clinton    91.20855    89.92219  92.27134                  72.92462
#> 1997-Clinton    83.06022    79.58643  83.13844     72.92462             
#> 2001-Bush       95.72356   101.74478  93.68031     64.32729     72.31874
#> 2005-Bush       93.33274    96.51943  92.98387     70.39886     79.44810
#> 2009-Obama      81.83520    80.75890  80.61017     84.88816     67.29042
#> 2013-Obama     103.88455   104.52272 104.20653     70.46276     71.86793
#>              2001-Bush 2005-Bush 2009-Obama 2013-Obama
#> 1981-Reagan   95.72356  93.33274   81.83520  103.88455
#> 1985-Reagan  101.74478  96.51943   80.75890  104.52272
#> 1989-Bush     93.68031  92.98387   80.61017  104.20653
#> 1993-Clinton  64.32729  70.39886   84.88816   70.46276
#> 1997-Clinton  72.31874  79.44810   67.29042   71.86793
#> 2001-Bush               56.23166   78.24321   54.69004
#> 2005-Bush     56.23166             85.39321   66.68583
#> 2009-Obama    78.24321  85.39321              72.10409
#> 2013-Obama    54.69004  66.68583   72.10409           

# for specific comparisons
textstat_dist(presDfm, "1985-Reagan", n = 5, margin = "documents")

#> Warning: non-square matrix

#>              1985-Reagan 1981-Reagan 1989-Bush 1993-Clinton 1997-Clinton
#> 1985-Reagan                 57.68015  69.94283     89.92219     79.58643
#> 1981-Reagan     57.68015               0.00000      0.00000      0.00000
#> 1989-Bush       69.94283     0.00000                0.00000      0.00000
#> 1993-Clinton    89.92219     0.00000   0.00000                   0.00000
#> 1997-Clinton    79.58643     0.00000   0.00000      0.00000             

textstat_dist(presDfm, c("2009-Obama" , "2013-Obama"), n = 5, margin = "documents")

#> Warning: non-square matrix

#>             2009-Obama 2013-Obama 1981-Reagan 1985-Reagan 1989-Bush
#> 2009-Obama               72.10409    81.83520    80.75890  80.61017
#> 2013-Obama    72.10409              103.88455   104.52272 104.20653
#> 1981-Reagan   81.83520  103.88455                 0.00000   0.00000
#> 1985-Reagan   80.75890  104.52272     0.00000               0.00000
#> 1989-Bush     80.61017  104.20653     0.00000     0.00000          

textstat_dist(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents")

#>              2009-Obama 2013-Obama 1981-Reagan 1985-Reagan 1989-Bush
#> 2009-Obama                72.10409    81.83520    80.75890  80.61017
#> 2013-Obama     72.10409              103.88455   104.52272 104.20653
#> 1981-Reagan    81.83520  103.88455                 0.00000   0.00000
#> 1985-Reagan    80.75890  104.52272     0.00000               0.00000
#> 1989-Bush      80.61017  104.20653     0.00000     0.00000          
#> 1993-Clinton   84.88816   70.46276     0.00000     0.00000   0.00000
#> 1997-Clinton   67.29042   71.86793     0.00000     0.00000   0.00000
#> 2001-Bush      78.24321   54.69004     0.00000     0.00000   0.00000
#> 2005-Bush      85.39321   66.68583     0.00000     0.00000   0.00000
#>              1993-Clinton 1997-Clinton 2001-Bush 2005-Bush
#> 2009-Obama       84.88816     67.29042  78.24321  85.39321
#> 2013-Obama       70.46276     71.86793  54.69004  66.68583
#> 1981-Reagan       0.00000      0.00000   0.00000   0.00000
#> 1985-Reagan       0.00000      0.00000   0.00000   0.00000
#> 1989-Bush         0.00000      0.00000   0.00000   0.00000
#> 1993-Clinton                   0.00000   0.00000   0.00000
#> 1997-Clinton      0.00000                0.00000   0.00000
#> 2001-Bush         0.00000      0.00000             0.00000
#> 2005-Bush         0.00000      0.00000   0.00000          

textstat_dist(presDfm, "2005-Bush", margin = "documents", method = "eJaccard")

#>              2005-Bush 1981-Reagan 1985-Reagan 1989-Bush 1993-Clinton
#> 2005-Bush                0.1905877   0.2007802 0.1897176    0.1558785
#> 1981-Reagan  0.1905877               0.0000000 0.0000000    0.0000000
#> 1985-Reagan  0.2007802   0.0000000             0.0000000    0.0000000
#> 1989-Bush    0.1897176   0.0000000   0.0000000              0.0000000
#> 1993-Clinton 0.1558785   0.0000000   0.0000000 0.0000000             
#> 1997-Clinton 0.1747363   0.0000000   0.0000000 0.0000000    0.0000000
#> 2001-Bush    0.1112871   0.0000000   0.0000000 0.0000000    0.0000000
#> 2009-Obama   0.1948222   0.0000000   0.0000000 0.0000000    0.0000000
#> 2013-Obama   0.1583972   0.0000000   0.0000000 0.0000000    0.0000000
#>              1997-Clinton 2001-Bush 2009-Obama 2013-Obama
#> 2005-Bush       0.1747363 0.1112871  0.1948222  0.1583972
#> 1981-Reagan     0.0000000 0.0000000  0.0000000  0.0000000
#> 1985-Reagan     0.0000000 0.0000000  0.0000000  0.0000000
#> 1989-Bush       0.0000000 0.0000000  0.0000000  0.0000000
#> 1993-Clinton    0.0000000 0.0000000  0.0000000  0.0000000
#> 1997-Clinton              0.0000000  0.0000000  0.0000000
#> 2001-Bush       0.0000000            0.0000000  0.0000000
#> 2009-Obama      0.0000000 0.0000000             0.0000000
#> 2013-Obama      0.0000000 0.0000000  0.0000000           

## similarities

# compute some document similarities
(tmp <- textstat_simil(presDfm, margin = "documents"))

#>              1981-Reagan 1985-Reagan 1989-Bush 1993-Clinton 1997-Clinton
#> 1985-Reagan    0.9685409                                                
#> 1989-Bush      0.9599217   0.9537449                                    
#> 1993-Clinton   0.9391137   0.9452609 0.9358174                          
#> 1997-Clinton   0.9386791   0.9468983 0.9381904    0.9241745             
#> 2001-Bush      0.9498797   0.9371453 0.9543776    0.9258829    0.9321614
#> 2005-Bush      0.9346990   0.9297108 0.9349331    0.9142198    0.9077993
#> 2009-Obama     0.9387730   0.9422163 0.9407657    0.9000505    0.9382904
#> 2013-Obama     0.9389438   0.9409353 0.9370933    0.9118394    0.9387808
#>              2001-Bush 2005-Bush 2009-Obama
#> 1985-Reagan                                
#> 1989-Bush                                  
#> 1993-Clinton                               
#> 1997-Clinton                               
#> 2001-Bush                                  
#> 2005-Bush    0.9437295                     
#> 2009-Obama   0.9240388 0.8976144           
#> 2013-Obama   0.9363852 0.9195065  0.9458185

# output as a list
as.list(tmp)[1:2]

#> $`1981-Reagan`
#>  1985-Reagan    1989-Bush    2001-Bush 1993-Clinton   2013-Obama   2009-Obama 
#>    0.9685409    0.9599217    0.9498797    0.9391137    0.9389438    0.9387730 
#> 1997-Clinton    2005-Bush 
#>    0.9386791    0.9346990 
#> 
#> $`1985-Reagan`
#>  1981-Reagan    1989-Bush 1997-Clinton 1993-Clinton   2009-Obama   2013-Obama 
#>    0.9685409    0.9537449    0.9468983    0.9452609    0.9422163    0.9409353 
#>    2001-Bush    2005-Bush 
#>    0.9371453    0.9297108 
#> 

# for specific comparisons
textstat_simil(presDfm, "1985-Reagan", n = 5, margin = "documents")

#> Warning: non-square matrix

#>              1985-Reagan 1981-Reagan 1989-Bush 1993-Clinton
#> 1981-Reagan    0.9685409                                   
#> 1989-Bush      0.9537449   0.0000000                       
#> 1993-Clinton   0.9452609   0.0000000 0.0000000             
#> 1997-Clinton   0.9468983   0.0000000 0.0000000    0.0000000

textstat_simil(presDfm, c("2009-Obama" , "2013-Obama"), n = 5, margin = "documents")

#> Warning: non-square matrix

#>             2009-Obama 2013-Obama 1981-Reagan 1985-Reagan
#> 2013-Obama   0.9458185                                   
#> 1981-Reagan  0.9387730  0.9389438                        
#> 1985-Reagan  0.9422163  0.9409353   0.0000000            
#> 1989-Bush    0.9407657  0.9370933   0.0000000   0.0000000

textstat_simil(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents")

#>              2009-Obama 2013-Obama 1981-Reagan 1985-Reagan 1989-Bush
#> 2013-Obama    0.9458185                                             
#> 1981-Reagan   0.9387730  0.9389438                                  
#> 1985-Reagan   0.9422163  0.9409353   0.0000000                      
#> 1989-Bush     0.9407657  0.9370933   0.0000000   0.0000000          
#> 1993-Clinton  0.9000505  0.9118394   0.0000000   0.0000000 0.0000000
#> 1997-Clinton  0.9382904  0.9387808   0.0000000   0.0000000 0.0000000
#> 2001-Bush     0.9240388  0.9363852   0.0000000   0.0000000 0.0000000
#> 2005-Bush     0.8976144  0.9195065   0.0000000   0.0000000 0.0000000
#>              1993-Clinton 1997-Clinton 2001-Bush
#> 2013-Obama                                      
#> 1981-Reagan                                     
#> 1985-Reagan                                     
#> 1989-Bush                                       
#> 1993-Clinton                                    
#> 1997-Clinton    0.0000000                       
#> 2001-Bush       0.0000000    0.0000000          
#> 2005-Bush       0.0000000    0.0000000 0.0000000

textstat_simil(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents", method = "cosine")

#>              2009-Obama 2013-Obama 1981-Reagan 1985-Reagan 1989-Bush
#> 2013-Obama    0.9473229                                             
#> 1981-Reagan   0.9399460  0.9397146                                  
#> 1985-Reagan   0.9436339  0.9422001   0.0000000                      
#> 1989-Bush     0.9416945  0.9376157   0.0000000   0.0000000          
#> 1993-Clinton  0.9015783  0.9125611   0.0000000   0.0000000 0.0000000
#> 1997-Clinton  0.9399306  0.9403480   0.0000000   0.0000000 0.0000000
#> 2001-Bush     0.9255036  0.9371368   0.0000000   0.0000000 0.0000000
#> 2005-Bush     0.9004034  0.9217499   0.0000000   0.0000000 0.0000000
#>              1993-Clinton 1997-Clinton 2001-Bush
#> 2013-Obama                                      
#> 1981-Reagan                                     
#> 1985-Reagan                                     
#> 1989-Bush                                       
#> 1993-Clinton                                    
#> 1997-Clinton    0.0000000                       
#> 2001-Bush       0.0000000    0.0000000          
#> 2005-Bush       0.0000000    0.0000000 0.0000000

# compute some term similarities
textstat_simil(presDfm, c("fair", "health", "terror"), method = "cosine",
               margin = "features", 20)

#> Warning: non-square matrix

#>                fair     terror     health      senat   hatfield          ,
#> terror   0.43033148                                                       
#> health   0.51639778 0.44444444                                            
#> senat    0.57535596 0.24759378 0.00000000                                 
#> hatfield 0.51639778 0.66666667 0.00000000 0.00000000                      
#> ,        0.78760215 0.58327489 0.67707179 0.00000000 0.00000000           
#> mr       0.26363525 0.29172998 0.09724333 0.00000000 0.00000000 0.00000000
#> .        0.76454880 0.59341577 0.70730365 0.00000000 0.00000000 0.00000000
#> chief    0.57735027 0.29814240 0.29814240 0.00000000 0.00000000 0.00000000
#> justic   0.41403934 0.17817416 0.53452248 0.00000000 0.00000000 0.00000000
#> presid   0.66914961 0.43193421 0.43193421 0.00000000 0.00000000 0.00000000
#> vice     0.77459667 0.44444444 0.33333333 0.00000000 0.00000000 0.00000000
#> bush     0.69282032 0.44721360 0.59628479 0.00000000 0.00000000 0.00000000
#> mondal   0.51639778 0.66666667 0.00000000 0.00000000 0.00000000 0.00000000
#> baker    0.51639778 0.66666667 0.00000000 0.00000000 0.00000000 0.00000000
#> speaker  0.31139958 0.20100756 0.00000000 0.00000000 0.00000000 0.00000000
#> o'neil   0.73029674 0.47140452 0.00000000 0.00000000 0.00000000 0.00000000
#> reverend 0.59628479 0.38490018 0.19245009 0.00000000 0.00000000 0.00000000
#> moomaw   0.51639778 0.66666667 0.00000000 0.00000000 0.00000000 0.00000000
#> fellow   0.55138280 0.64993368 0.64993368 0.00000000 0.00000000 0.00000000
#>                  mr          .      chief     justic     presid       vice
#> terror                                                                    
#> health                                                                    
#> senat                                                                     
#> hatfield                                                                  
#> ,                                                                         
#> mr                                                                        
#> .        0.00000000                                                       
#> chief    0.00000000 0.00000000                                            
#> justic   0.00000000 0.00000000 0.00000000                                 
#> presid   0.00000000 0.00000000 0.00000000 0.00000000                      
#> vice     0.00000000 0.00000000 0.00000000 0.00000000 0.00000000           
#> bush     0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> mondal   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> baker    0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> speaker  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> o'neil   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> reverend 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> moomaw   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> fellow   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>                bush     mondal      baker    speaker     o'neil   reverend
#> terror                                                                    
#> health                                                                    
#> senat                                                                     
#> hatfield                                                                  
#> ,                                                                         
#> mr                                                                        
#> .                                                                         
#> chief                                                                     
#> justic                                                                    
#> presid                                                                    
#> vice                                                                      
#> bush                                                                      
#> mondal   0.00000000                                                       
#> baker    0.00000000 0.00000000                                            
#> speaker  0.00000000 0.00000000 0.00000000                                 
#> o'neil   0.00000000 0.00000000 0.00000000 0.00000000                      
#> reverend 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000           
#> moomaw   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> fellow   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>              moomaw
#> terror             
#> health             
#> senat              
#> hatfield           
#> ,                  
#> mr                 
#> .                  
#> chief              
#> justic             
#> presid             
#> vice               
#> bush               
#> mondal             
#> baker              
#> speaker            
#> o'neil             
#> reverend           
#> moomaw             
#> fellow   0.00000000