apply a dictionary to a tokens object

Convert tokens into equivalence classes defined by values of a dictionary object.

tokens_lookup(x, dictionary, valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE, capkeys = FALSE, concatenator = " ",
  verbose = FALSE)

Arguments

x

tokens object to which dictionary or thesaurus will be supplied

dictionary

the dictionary-class object that will be applied to x

valuetype

how to interpret keyword expressions: "glob" for "glob"-style wildcard expressions; "regex" for regular expressions; or "fixed" for exact matching. See valuetype for details.

case_insensitive

ignore the case of dictionary values if TRUE uppercase to distinguish them from other features

capkeys

if TRUE, convert dictionary keys to uppercase to distinguish them from other features

concatenator

a charactor that connect words in multi-words entries

verbose

print status messages if TRUE

Examples

toks <- tokens(data_corpus_inaugural)
dict <- dictionary(list(country = "united states",
                   law=c('law*', 'constitution'),
                   freedom=c('*', 'libert*')))
dfm(tokens_lookup(toks, dict, 'glob', verbose = TRUE))

#> Registering 5 entiries in the dictionary...

#> Searching 9882 types of features...

#> Document-feature matrix of: 57 documents, 3 features (13.5% sparse).

dict_fix <- dictionary(list(country = "united states",
                       law = c('law', 'constitution'),
                       freedom = c('freedom', 'liberty')))
dfm(applyDictionary(toks, dict_fix, valuetype='fixed'))

#> Document-feature matrix of: 57 documents, 3 features (21.1% sparse).

dfm(tokens_lookup(toks, dict_fix, valuetype='fixed'))

#> Document-feature matrix of: 57 documents, 3 features (21.1% sparse).